### 2020-12-04 #2885 @whedon
[PRE REVIEW]: Easy3D: a lightweight, easy-to-use, and efficient C++ library for processing and rendering 3D data
**Submitting author:** @LiangliangNan (<a href="http://orcid.org/0000-0002-5629-9975">Liangliang Nan</a>)
**Repository:** <a href="https://github.com/LiangliangNan/Easy3D" target ="_blank">https://github.com/LiangliangNan/Easy3D</a>
**Version:** v2.4.1
