+++ tutorials/Tutorial_503_SoftShadow/soft_shadow.cpp
@@ -39,7 +39,7 @@ TutorialSoftShadow::TutorialSoftShadow(const std::string& title) : Viewer(title)
 	camera()->setViewDirection(vec3(0, 0, 1));
 
 	shadow_ = new SoftShadow(camera());
-    shadow_->set_background_color(background_color_);
+    shadow_->set_virtual_background_color(background_color());
 	shadow_enabled_ = true;
 }
 
