+++ easy3d/fileio/image_io.h
@@ -71,11 +71,13 @@ namespace easy3d {
          *
          * \return Return true on success or false if failed.
          *
-         * \example
+         * Example usage: \par
+         * \code
          *    int width, height, channels;
          *    std::vector<unsigned char> data;
          *    bool success = ImageIO::load(data, file_name, &width, &height, &channels, 0);
          *    // ... process data if true returned ...
+         * \endcode
          */
         static bool load(
                 const std::string& file_name,
