+++ CMakeLists.txt
@@ -6,8 +6,8 @@ get_filename_component(PROJECT_NAME ${CMAKE_CURRENT_SOURCE_DIR} NAME)
 project(${PROJECT_NAME})
 
 set (PROJECT_MAJOR 2)
-set (PROJECT_MINOR 2)
-set (PROJECT_PATCH 0)
+set (PROJECT_MINOR 4)
+set (PROJECT_PATCH 1)
 set (PROJECT_VERSION "${PROJECT_MAJOR}.${PROJECT_MINOR}.${PROJECT_PATCH}")
 
 
