+++ easy3d/algo_ext/mesh_surfacer.h
@@ -36,8 +36,9 @@ namespace easy3d {
 
     class SurfaceMesh;
 
-    // This class simply collects the related algorithms implemented using CGAL
-    // (to eliminate including CGAL header files from client code)
+    /// \brief Detect/Resolve topological issues (e.g., duplicate faces, self intersection) of a surface mesh.
+    /// This class simply collects the related algorithms implemented using CGAL (to eliminate including CGAL header
+    /// files from client code). \see DuplicatedFaces and SelfIntersection.
 
     class MeshSurfacer
     {
@@ -47,24 +48,22 @@ namespace easy3d {
 
         // ------- detect/remove duplicated faces
 
-        // detect and return the duplicated faces
-        // exact == true: do exact predict; otherwise use the distance threshold.
-        // upon return, the second component of each entry contains the set of faces
-        // duplicating the one stored as the first component.
-        // TODO: implement using threshold (now only 'exact' is implemented).
+        /// \brief Detect and return the duplicated faces.
+        /// @param exact True: do exact predict; otherwise use the distance threshold.
+        /// Upon return, the second component of each entry contains the set of faces duplicating the one stored as
+        /// the first component.
         std::vector< std::pair<SurfaceMesh::Face, std::vector<SurfaceMesh::Face> > >
         detect_duplicated_faces(SurfaceMesh* mesh, bool exact = false, double dist_threshold = 1e-6);
 
-        // detect and remove duplicated faces. Returns the number of deleted faces.
-        // exact == true: do exact predict; otherwise use the distance threshold.
-        // TODO: implement using threshold (now only 'exact' is implemented).
+        /// \brief Detect and remove duplicated faces. Returns the number of deleted faces.
+        /// @param exact True: do exact predict; otherwise use the distance threshold.
         unsigned int remove_duplicated_faces(SurfaceMesh* mesh, bool exact = false, double dist_threshold = 1e-6);
 
 
         // ------- detect/remesh self intersections
 
         /**
-         * Detect intersecting face pairs.
+         * \brief Detect intersecting face pairs.
          * @param mesh The input mesh.
          * @param construct If true, also construct the intersecting geometry.
          * @return The intersecting face pairs.
@@ -73,7 +72,7 @@ namespace easy3d {
         detect_self_intersections(SurfaceMesh* mesh);
 
         /**
-         * Detect and remesh the intersecting faces.
+         * \brief Detect and remesh the intersecting faces.
          * @param mesh The input mesh. If self intersection exists, it carries the remeshed model. Otherwise it remains
          *             unchanged.
          * @param stitch Stitch the borders
