+++ easy3d/core/mat.h
@@ -47,7 +47,7 @@ namespace easy3d {
 
 
     /**
-     * Base class for matrix types. Provides generic functionality for N by M matrices.
+     * \brief Base class for matrix types. Provides generic functionality for N by M matrices.
      *      N: The number of rows in this matrix.
      *		M: The number of columns in this matrix.
      *		T: The scalar type for vector elements.
@@ -61,16 +61,16 @@ namespace easy3d {
     public:
         //	----------------- constructor and destructor -------------------
 
-        /**	Default constructor.
-         * Note: The matrix elements are intentionally not initialized. This is efficient
+        /**	\brief Default constructor.
+         * \note: The matrix elements are intentionally not initialized. This is efficient
          *       if the user assigns their values from subsequent computations. Use Mat(T s)
          *       to initialize the elements during construction. */
         Mat();
 
-        /**	Initialized with diagonal as s and others zeros. */
+        /**	\brief Initialized with diagonal as s and others zeros. */
         explicit Mat(T s);
 
-        /**	Copy constructor for rN >= N, rM >= M.
+        /**	\brief Copy constructor for rN >= N, rM >= M.
         *	For smaller incoming matrices (i.e, rN < N, rM < M ) specialization
         *	is required in order to fill the remaining elements with appropriate
         *	values (usually 0 or 1).
@@ -84,128 +84,134 @@ namespace easy3d {
         template <size_t rN, size_t rM>
         explicit Mat(const Mat<rN, rM, T> &rhs);
 
-        /**	Initialize elements from an array of type T. */
+        /**	\brief Initialize elements from an array of type T. */
         explicit Mat(const T *m);
 
-        /**	Static constructor return an N x M identity matrix.
+        /**	\brief Static constructor return an N x M identity matrix.
         *	see also load_identity() */
         static Mat<N, M, T> identity();
 
         //	------------------- size query ---------------------
 
-        /**	Return the number of rows. */
+        /**	\brief Return the number of rows. */
         size_t num_rows() const { return N; }
 
-        /**	Return the number of columns. */
+        /**	\brief Return the number of columns. */
         size_t num_columns() const { return M; }
 
         //	--------------------- access -----------------------
 
-        /**	Return row r as a vector. */
+        /**	\brief Return row r as a vector. */
         Vec<M, T> row(size_t r) const;
 
-        /**	Return col c as a vector. */
+        /**	\brief Return col c as a vector. */
         Vec<N, T> col(size_t c) const;
 
-        /**	const row/column access to elements. */
+        /**	\brief Const row/column access to elements. */
         const T& operator()(size_t r, size_t c) const;
 
-        /**	Non-const row/column access to elements. */
+        /**	\brief Non-const row/column access to elements. */
         T& operator()(size_t r, size_t c);
 
-        /**	const array/low-level access to elements. */
+        /**	\brief const array/low-level access to elements. */
         operator const T*() const;
 
-        /**	Non-const array/low-level access to elements. */
+        /**	\brief Non-const array/low-level access to elements. */
         operator T*();
 
         //	--------------------- modification  ------------------
 
-        /**	Set all elements 0 */
+        /**	\brief Set all elements 0 */
         void load_zero();
 
[…]
