+++ easy3d/core/quat.h
@@ -41,62 +41,62 @@
 #include <easy3d/core/vec.h>
 #include <easy3d/core/mat.h>
 
-/**
- * Euler angles VS Quaternion
- *
- * Euler angles are the easiest way to represent rotation. This representation simply
- * stores the three rotation angles around the X, Y and Z axes. These 3 rotations are
- * then applied successively, usually in this order: first Y, then Z, then X (but not
- * necessarily). Using a different order yields different results.
- * Euler angles are usually used to set a character's orientation since game characters
- * only rotate on the vertical axis. Therefore, it is easier to write, understand and
- * maintain "float direction" than 3 different orientations. Another good use of Euler
- * angles is an FPS camera: you have one angle for heading (Y), and one for up/down (X).
- * However, when things get more complex, Euler angle will be hard to work with, eg.,:
- *  - Interpolating smoothly between 2 orientations is hard. Naively interpolating the
- *    X, Y, and Z angles will be ugly.
- *  - Applying several rotations is complicated and unprecise: you have to compute the
- *    final rotation matrix, and guess the Euler angles from this matrix.
- *  - A well-known problem, the "Gimbal Lock", will sometimes block your rotations, and
- *    other singularities which will flip your model upside-down.
- *  - Different angles make the same rotation (-180 and 180 degrees, for instance)
- *  - It is a mess - as said above, usually the right order is YZX, but if you also use
- *    a library with a different order, you'll be in trouble.
- *  - Some operations are complicated, e.g., rotation of N degrees around a specific axis.
- * Quaternions are a tool to represent rotations, which solves these problems.
- *
- *
- * The Quaternion class represents 3D rotations and orientations.
- *
- * The Quaternion is an appropriate (although not very intuitive) representation for
- * 3D rotations and orientations. Many tools are provided to ease the definition of a
- * Quaternion: see constructors, set_axis_angle(), set_from_rotation_matrix(),
- * set_from_rotated_basis().
- *
- * You can apply the rotation represented by the Quaternion to 3D points using rotate()
- * and inverse_rotate().
- *
- * You can apply the Quaternion q rotation to the OpenGL matrices using
- *      glMultMatrixd(q.matrix());
- * which is equvalent to
- *      glRotate(q.angle()*180.0/M_PI, q.axis().x, q.axis().y, q.axis().z);
- *
- * Internal representation
- * The internal representation of a Quaternion is a set of 4 numbers, [x y z w], which
- * represents rotations the following way:
- *  x = axis.x * sin(angle / 2)
- *  y = axis.y * sin(angle / 2)
- *  z = axis.z * sin(angle / 2)
- *  w = cos(angle / 2)
- * NOTE:
- *   - the angle is in radians and the axis is a unit vector.
- *   - certain implementations place the cosine term in the first position (instead of last).
- *
- * A Quaternion is always normalized, so that its inverse() is actually its conjugate.
- */
 
 namespace easy3d {
 
+    /**
+     * \brief The Quaternion class represents 3D rotations and orientations.
+     *
+     * The Quaternion is an appropriate (although not very intuitive) representation for
+     * 3D rotations and orientations. Many tools are provided to ease the definition of a
+     * Quaternion: see constructors, set_axis_angle(), set_from_rotation_matrix(),
+     * set_from_rotated_basis().
+     *
+     * You can apply the rotation represented by the Quaternion to 3D points using rotate()
+     * and inverse_rotate().
+     *
+     * You can apply the Quaternion q rotation to the OpenGL matrices using
+     *      glMultMatrixd(q.matrix());
+     * which is equvalent to
+     *      glRotate(q.angle()*180.0/M_PI, q.axis().x, q.axis().y, q.axis().z);
[…]
