+++ easy3d/core/model.h
@@ -37,35 +37,50 @@ namespace easy3d {
     class Renderer;
 
     /**
-     * Model is the base class of renderable 3D models, e.g., SurfaceMesh, PointCloud, Graph.
+     * \brief The base class of renderable 3D models, e.g., SurfaceMesh, PointCloud, Graph.
      * A Model can have multiple drawables, e.g., faces, vertices, edges.
      */
     class Model
     {
     public:
+        /// \brief Default constructor.
+        /// The parameter \p name is optional, but it is useful for handling multiple models with the same viewer.
         Model(const std::string& name = "unknown");
         virtual ~Model();
 
-        const std::string& name() const { return name_; }
+        /// \brief Sets/Changes the name of a model.
+        /// Asigning a name to a model is optional, but it is useful for handling multiple models with the same viewer.
         void set_name(const std::string& n) { name_ = n; }
+        /// \brief The name of a model.
+        const std::string& name() const { return name_; }
 
-        /** The bounding box of the model. */
+        /**
+         * \brief The bounding box of the model.
+         * If the bounding box is unknown, it will be computed first. Otherwise it returns the known bounding box.
+         * \see invalidate_bounding_box().
+         */
         const Box3& bounding_box() const;
 
-        void update_bounding_box();
+        /**
+         * \brief Invalidates the bounding box of the model. So when bounding_box() is called, the bounding box will be
+         * re-computed. This function is typically called when the geometry of a model is changed.
+         */
+        void invalidate_bounding_box();
 
-        /** The vertices of the model. */
+        /** \brief The vertices of the model. */
         virtual const std::vector<vec3>& points() const = 0;
 
-        /** Test if the model is empty. */
+        /** \brief Tests if the model is empty. */
         bool empty() const { return points().empty(); };
 
-        /** prints the names of all properties to an output stream (e.g., std::cout). */
+        /** \brief Prints the names of all properties to an output stream (e.g., std::cout). */
         virtual void property_stats(std::ostream &output) const = 0;
 
-        /** Set/Get the renderer of this model. */
+        /** \brief Sets the renderer of this model. */
         void set_renderer(Renderer* r) { renderer_ = r; }
+        /** \brief Gets the renderer of this model. */
         inline Renderer* renderer() { return renderer_; }
+        /** \brief Gets the constant renderer of this model. */
         inline const Renderer* renderer() const { return renderer_; }
 
     protected:
