+++ easy3d/core/segment.h
@@ -35,6 +35,7 @@
 namespace easy3d {
 
 
+    /// \brief A generic segmentation representation, which supports both 2D and 3D line segments.
     template <int DIM, typename FT>
     class GenericSegment {
     public:
@@ -44,24 +45,36 @@ namespace easy3d {
         typedef GenericSegment<DIM, FT> thisclass;
 
     public:
+        /// \brief Constructs a line segments from its two end points \p s and \p t.
         GenericSegment(const Point& s, const Point& t);
         GenericSegment()  {}
 
+        /// \brief Returns the source end point of this line segment.
         const Point& source() const { return s_; }
+        /// \brief Returns the target end point of this line segment.
         const Point& target() const { return t_; }
+        /// \brief Sets/Changes the source end point of this line segment.
         void  set_source(const Point& s) { s_ = s; }
+        /// \brief Sets/Changes the target end point of this line segment.
         void  set_target(const Point& t) { t_ = t; }
 
+        /// \brief Returns the supporting line of this line segment.
         Line  supporting_line() const { return Line::from_two_points(s_, t_); }
 
+        /// \brief Returns a vector originating from source() and pointing to target().
         Vector to_vector() const { return t_ - s_; }
 
-        // the projection of p on the supporting line
+        /// \brief Returns the projection of a point \p p on the supporting line of this segment.
         Point  projection(const Point &p) const;
 
-        // test if the projection of p is inside the two end points
+        /// \brief Tests if the projection of a point \p p is within the two end points of this segment.
         bool   projected_inside(const Point& p) const;
 
+        /// \brief Returns the squared distance of a point \p p to this segment.
+        /// The return the value is the min of the following:
+        ///     - the squared distance between \p p and the supporting_line() of this segment;
+        ///     - the squared distance between \p p and the source() of this segment;
+        ///     - the squared distance between \p p and the target() of this segment;
         FT squared_ditance(const Point &p) const;
 
     private:
