### 2020-12-06 2e73b4f
documentation; add examples to documentation
