+++ tutorials/Tutorial_503_SoftShadow/main.cpp
@@ -22,7 +22,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "soft_shadow.h"
+#include "viewer.h"
 #include <easy3d/core/model.h>
 #include <easy3d/renderer/drawable_triangles.h>
 #include <easy3d/renderer/renderer.h>
