+++ easy3d/renderer/texture_manager.h
@@ -106,6 +106,17 @@ namespace easy3d {
         static std::unordered_map<std::string, bool> attempt_load_texture_; // avoid multiple attempt
     };
 
+
+    /**
+     * \example Tutorial_308_TexturedMesh
+     * The source file containing the main() function:
+     * \include Tutorial_308_TexturedMesh/main.cpp
+     * The header file of the class:
+     * \include Tutorial_308_TexturedMesh/viewer.h
+     * The source file of the class:
+     * \include Tutorial_308_TexturedMesh/viewer.cpp
+     */
+
 } // namespace easy3d
 
 
