+++ easy3d/renderer/average_color_blending.h
@@ -69,6 +69,17 @@ namespace easy3d {
         AverageColorBlending& operator=(const AverageColorBlending&);
     };
 
+
+    /**
+     * \example Tutorial_504_Transparency
+     * The source file containing the main() function:
+     * \include Tutorial_504_Transparency/main.cpp
+     * The header file of the class:
+     * \include Tutorial_504_Transparency/transparency.h
+     * The source file of the class:
+     * \include Tutorial_504_Transparency/transparency.cpp
+     */
+
 }
 
 #endif // EASY3D_RENDERER_AVERAGE_COLOR_BLENDING_H
