+++ easy3d/renderer/eye_dome_lighting.h
@@ -132,6 +132,17 @@ namespace easy3d {
         EyeDomeLighting& operator=(const EyeDomeLighting&);
     };
 
+
+    /**
+     * \example Tutorial_505_EyeDomeLighting
+     * The source file containing the main() function:
+     * \include Tutorial_505_EyeDomeLighting/main.cpp
+     * The header file of the class:
+     * \include Tutorial_505_EyeDomeLighting/eye_dome_lighting.h
+     * The source file of the class:
+     * \include Tutorial_505_EyeDomeLighting/eye_dome_lighting.cpp
+     */
+
 }
 
 #endif	// EASY3D_RENDERER_EYE_DOME_LIGHTING_H
