+++ easy3d/gui/picker_point_cloud.h
@@ -64,6 +64,17 @@ namespace easy3d {
         int pick_vertices_gpu(PointCloud *model, const Polygon2 &plg, bool deselect, ShaderProgram *program);
     };
 
+
+    /**
+     * \example Tutorial_403_PointSelection
+     * The source file containing the main() function:
+     * \include Tutorial_403_PointSelection/main.cpp
+     * The header file of the class:
+     * \include Tutorial_403_PointSelection/point_selection.h
+     * The source file of the class:
+     * \include Tutorial_403_PointSelection/point_selection.cpp
+     */
+
 }
 
 #endif // EASY3D_GUI_PICKER_POINT_CLOUD_H
