+++ easy3d/renderer/average_color_blending.h
@@ -75,9 +75,9 @@ namespace easy3d {
      * The source file containing the main() function:
      * \include Tutorial_504_Transparency/main.cpp
      * The header file of the class:
-     * \include Tutorial_504_Transparency/transparency.h
+     * \include Tutorial_504_Transparency/viewer.h
      * The source file of the class:
-     * \include Tutorial_504_Transparency/transparency.cpp
+     * \include Tutorial_504_Transparency/viewer.cpp
      */
 
 }
