+++ easy3d/renderer/ambient_occlusion.h
@@ -95,9 +95,9 @@ namespace easy3d {
      * The source file containing the main() function:
      * \include Tutorial_501_AmbientOcclusion/main.cpp
      * The header file of the class:
-     * \include Tutorial_501_AmbientOcclusion/ambient_occlusion.h
+     * \include Tutorial_501_AmbientOcclusion/viewer.h
      * The source file of the class:
-     * \include Tutorial_501_AmbientOcclusion/ambient_occlusion.cpp
+     * \include Tutorial_501_AmbientOcclusion/viewer.cpp
      */
 
 
