+++ easy3d/renderer/text_renderer.h
@@ -195,7 +195,15 @@ namespace easy3d {
     };
 
 
-    /// @example Tutorial_309_TextRendering     \include Tutorial_309_TextRendering/main.cpp
+    /**
+     * \example Tutorial_309_TextRendering
+     * The source file containing the main() function:
+     * \include Tutorial_309_TextRendering/main.cpp
+     * The header file of the class:
+     * \include Tutorial_309_TextRendering/viewer.h
+     * The source file of the class:
+     * \include Tutorial_309_TextRendering/viewer.cpp
+     */
 
 }
 
