+++ easy3d/algo/point_cloud_ransac.h
@@ -98,9 +98,9 @@ namespace easy3d {
      * The source file containing the main() function:
      * \include Tutorial_603_Cloud_PlaneExtraction/main.cpp
      * The header file of the class:
-     * \include Tutorial_603_Cloud_PlaneExtraction/plane_extraction.h
+     * \include Tutorial_603_Cloud_PlaneExtraction/viewer.h
      * The source file of the class:
-     * \include Tutorial_603_Cloud_PlaneExtraction/plane_extraction.cpp
+     * \include Tutorial_603_Cloud_PlaneExtraction/viewer.cpp
      */
 
 }
