+++ easy3d/gui/picker_point_cloud.h
@@ -70,9 +70,9 @@ namespace easy3d {
      * The source file containing the main() function:
      * \include Tutorial_403_PointSelection/main.cpp
      * The header file of the class:
-     * \include Tutorial_403_PointSelection/point_selection.h
+     * \include Tutorial_403_PointSelection/viewer.h
      * The source file of the class:
-     * \include Tutorial_403_PointSelection/point_selection.cpp
+     * \include Tutorial_403_PointSelection/viewer.cpp
      */
 
 }
