### 2020-12-06 62581cb
fix Doxygen variable name
