+++ easy3d/core/properties.h
@@ -50,6 +50,8 @@
 
 namespace easy3d {
 
+    /// \brief Base class for a property array.
+    /// \class BasePropertyArray easy3d/core/properties.h
     class BasePropertyArray
     {
     public:
@@ -99,7 +101,8 @@ namespace easy3d {
 
     //== CLASS DEFINITION =========================================================
 
-
+    /// \brief Implementation of a generic property array.
+    /// \class PropertyArray easy3d/core/properties.h
     template <class T>
     class PropertyArray : public BasePropertyArray
     {
@@ -207,7 +210,8 @@ namespace easy3d {
 
     //== CLASS DEFINITION =========================================================
 
-
+    /// \brief Implementation of a generic property.
+    /// \class Property easy3d/core/properties.h
     template <class T>
     class Property
     {
@@ -295,6 +299,8 @@ namespace easy3d {
     //== CLASS DEFINITION =========================================================
 
 
+    /// \brief Implementation of generic property container.
+    /// \class PropertyContainer easy3d/core/properties.h
     class PropertyContainer
     {
     public:
