### 2020-12-07 0c249b0
put all examples doc in tutorials/examples.h
