+++ easy3d/core/graph.h
@@ -923,11 +923,6 @@ namespace easy3d {
 		return (os << 'e' << e.idx());
 	}
 
-
-    /// @example Tutorial_108_Graph                 \include Tutorial_108_Graph/main.cpp
-    /// @example Tutorial_109_Graph_Connectivity    \include Tutorial_109_Graph_Connectivity/main.cpp
-    /// @example Tutorial_110_Graph_Property        \include Tutorial_110_Graph_Property/main.cpp
-
 } // namespace easy3d
 
 #endif // EASY3D_CORE_GRAPH_H
