+++ easy3d/fileio/graph_io.h
@@ -76,9 +76,6 @@ namespace easy3d {
     } // namespace io
 
 
-    /// @example Tutorial_111_Graph_IO  \include Tutorial_111_Graph_IO/main.cpp
-
-
 } // namespace easy3d
 
 #endif // EASY3D_FILEIO_GRAPH_IO_H
