+++ easy3d/renderer/framebuffer_object.h
@@ -410,16 +410,6 @@ namespace easy3d {
         FramebufferObject& operator=(const FramebufferObject&);
     };
 
-    /**
-     * \example Tutorial_506_DepthMaps
-     * The source file containing the main() function:
-     * \include Tutorial_506_DepthMaps/main.cpp
-     * The header file of the class:
-     * \include Tutorial_506_DepthMaps/viewer.h
-     * The source file of the class:
-     * \include Tutorial_506_DepthMaps/viewer.cpp
-     */
-
 
 } // namespace easy3d
 
