+++ easy3d/renderer/drawable_triangles.h
@@ -71,10 +71,6 @@ namespace easy3d {
         float   opacity_;
 	};
 
-
-    /// @example Tutorial_301_Drawables \include Tutorial_301_Drawables/main.cpp
-    /// @example Tutorial_302_Imposters \include Tutorial_302_Imposters/main.cpp
-
 }
 
 
