### 2020-12-08 ee4cf1d
better handle of concave faces in triangulation
