### 2020-12-08 8ac604b
more readable function names;  add 'prev_around_target()' and 'next_around_target()'
