+++ easy3d/core/surface_mesh.cpp
@@ -232,40 +232,40 @@ namespace easy3d {
 		if (!props.empty())
 		{
             output << "vertex properties:\n";
-			for (unsigned int i = 0; i < props.size(); ++i)
-                output << "\t" << props[i] << std::endl;
+			for (const auto& p : props)
+                output << "\t" << p << std::endl;
 		}
 
 		props = halfedge_properties();
 		if (!props.empty())
 		{
             output << "halfedge properties:\n";
-			for (unsigned int i = 0; i < props.size(); ++i)
-                output << "\t" << props[i] << std::endl;
+            for (const auto& p : props)
+                output << "\t" << p << std::endl;
 		}
 
 		props = edge_properties();
 		if (!props.empty())
 		{
             output<< "edge properties:\n";
-			for (unsigned int i = 0; i < props.size(); ++i)
-                output << "\t" << props[i] << std::endl;
+            for (const auto& p : props)
+                output << "\t" << p << std::endl;
 		}
 
 		props = face_properties();
 		if (!props.empty())
 		{
             output << "face properties:\n";
-			for (unsigned int i = 0; i < props.size(); ++i)
-                output << "\t" << props[i] << std::endl;
+            for (const auto& p : props)
+                output << "\t" << p << std::endl;
 		}
 
 		props = model_properties();
 		if (!props.empty())
 		{
             output << "model properties:\n";
-			for (unsigned int i = 0; i < props.size(); ++i)
-                output << "\t" << props[i] << std::endl;
+            for (const auto& p : props)
+                output << "\t" << p << std::endl;
 		}
     }
 
@@ -301,7 +301,7 @@ namespace easy3d {
             {
                 if (target(h) == end)
                     return h;
-                h = cw_rotated_halfedge(h);
+                h = next_around_source(h);
             }
             while (h != hh);
         }
@@ -363,7 +363,7 @@ namespace easy3d {
                     set_out_halfedge(v, h);
                     return;
                 }
-                h = cw_rotated_halfedge(h);
+                h = next_around_source(h);
             }
             while (h != hh);
         }
@@ -835,20 +835,11 @@ namespace easy3d {
         if (!vnormal_)
             vnormal_ = vertex_property<vec3>("v:normal");
 
-        // TODO: may not stable for concave vertices?
         VertexIterator vit, vend=vertices_end();
-        for (vit=vertices_begin(); vit!=vend; ++vit) {
-            if (!is_border(*vit))
-                vnormal_[*vit] = compute_vertex_normal(*vit);
-            else { // Liangliang: if on border, we use the face normal.
-                auto h = out_halfedge(*vit);
-                if (h.is_valid()) { // might be an isolated vertex
-                    if (is_border(h))
-                        h = opposite(h);
-                    vnormal_[*vit] = compute_face_normal(face(h));
-                }
-            }
-        }
+
+        //\todo not stable for concave vertices
+        for (vit=vertices_begin(); vit!=vend; ++vit)
+            vnormal_[*vit] = compute_vertex_normal(*vit);
     }
 
 
@@ -901,7 +892,7 @@ namespace easy3d {
                     }
                 }
 
-                h  = cw_rotated_halfedge(h);
+                h  = next_around_source(h);
             }
             while (h != hend);
 
