+++ easy3d/kdtree/kdtree_search.h
@@ -52,9 +52,9 @@ namespace easy3d {
      * Test 1: 362, 271 points (the Stanford bunny). radius = 0.001
      * \code
      * --------------------------------------------------------------------------------------
-     *       Build		   |		Single		 |		KNN			  |		Radius
+     *       Build         |       Single        |        KNN         |       Radius
      * --------------------|---------------------|--------------------|---------------------
-     * ANN    ETH   FLANN  |  ANN   ETH   FLANN  |  ANN	  ETH   FLANN |  ANN   ETH   FLANN
+     * ANN    ETH   FLANN  |  ANN   ETH   FLANN  |  ANN   ETH   FLANN |  ANN   ETH   FLANN
      * --------------------|---------------------|--------------------|---------------------
      * 0.14   0.05   0.12  |  0.17  0.11   0.71  |  1.33  1.0   1.48  |  1.32  1.01  1.51
      * --------------------------------------------------------------------------------------
@@ -63,9 +63,9 @@ namespace easy3d {
      * Test 2: 4, 116, 466 points (an noisy urban scan). radius = 0.03
      * \code
      * --------------------------------------------------------------------------------------
-     *       Build		   |		Single		 |		KNN			   |		Radius
+     *       Build         |       Single        |        KNN          |       Radius
      * --------------------|---------------------|---------------------|---------------------
-     * ANN    ETH   FLANN  |  ANN   ETH   FLANN  |  ANN	 ETH   FLANN   |  ANN    ETH   FLANN
+     * ANN    ETH   FLANN  |  ANN   ETH   FLANN  |  ANN   ETH   FLANN  |  ANN    ETH   FLANN
      * --------------------|---------------------|---------------------|---------------------
      * 2.2   0.76   1.88   |  2.61  1.84  11.8   |  20.8  13.5  22.0   |  8.75  4.79   15.1
      * --------------------------------------------------------------------------------------
