### 2020-12-08 d80c425
robust vertex normal computation for concave faces
