### 2020-12-08 a1af6c8
define "PREDEFINED= DOXYGEN"
