+++ easy3d/renderer/manipulated_camera_frame.h
@@ -43,80 +43,85 @@
 namespace easy3d {
 
 
-	/*! \brief A manipulated frame with camera specific mouse bindings.
-	  \class ManipulatedCameraFrame easy3d/renderer/manipulated_camera_frame.h
+    /*! \brief A manipulated frame with camera specific mouse bindings.
+      \class ManipulatedCameraFrame easy3d/renderer/manipulated_camera_frame.h
 
-	  \details A ManipulatedCameraFrame is a specialization of a ManipulatedFrame, designed
-	  to be set as the Camera::frame(). Mouse motions are basically interpreted in a
-	  negated way: when the mouse goes to the right, the ManipulatedFrame
-	  translation goes to the right, while the ManipulatedCameraFrame has to go to
-	  the \e left, so that the \e scene seems to move to the right.
+      \details A ManipulatedCameraFrame is a specialization of a ManipulatedFrame, designed
+      to be set as the Camera::frame(). Mouse motions are basically interpreted in a
+      negated way: when the mouse goes to the right, the ManipulatedFrame
+      translation goes to the right, while the ManipulatedCameraFrame has to go to
+      the \e left, so that the \e scene seems to move to the right.
 
-	  A ManipulatedCameraFrame rotates around its pivotPoint(), which corresponds to
-	  the associated Camera::pivotPoint().
+      A ManipulatedCameraFrame rotates around its pivotPoint(), which corresponds to
+      the associated Camera::pivotPoint().
 
-	  A ManipulatedCameraFrame can also "fly" in the scene. It basically moves
-	  forward, and turns according to the mouse motion. See flySpeed(),
-	  sceneUpVector() and the QGLViewer::MOVE_FORWARD and QGLViewer::MOVE_BACKWARD
-	  QGLViewer::MouseAction.
+      A ManipulatedCameraFrame can also "fly" in the scene. It basically moves
+      forward, and turns according to the mouse motion. See flySpeed(),
+      sceneUpVector() and the QGLViewer::MOVE_FORWARD and QGLViewer::MOVE_BACKWARD
+      QGLViewer::MouseAction.
 
-	  See the <a href="../mouse.html">mouse page</a> for a description of the
-	  possible actions that can be performed using the mouse and their bindings.
-	  \nosubgrouping */
-	class ManipulatedCameraFrame : public ManipulatedFrame
-	{
-	public:
-		ManipulatedCameraFrame();
-		/*! Virtual destructor. Empty. */
-		virtual ~ManipulatedCameraFrame() {}
-
-		ManipulatedCameraFrame(const ManipulatedCameraFrame &mcf);
-		ManipulatedCameraFrame &operator=(const ManipulatedCameraFrame &mcf);
-
-		/*! @name Pivot point */
-		//@{
-	public:
-		/*! Returns the point the ManipulatedCameraFrame pivot point, around which the
-		camera rotates.
-
-		It is defined in the world coordinate system. Default value is (0,0,0).
-
-		When the ManipulatedCameraFrame is associated to a Camera,
-		Camera::pivotPoint() also returns this value. This point can interactively be
-		changed using the mouse (see Camera::setPivotPointFromPixel() and
-		QGLViewer::RAP_FROM_PIXEL and QGLViewer::RAP_IS_CENTER in the <a
-		href="../mouse.html">mouse page</a>). */
-		vec3 pivotPoint() const { return pivotPoint_; }
-		/*! Sets the pivotPoint(), defined in the world coordinate system. */
-		void setPivotPoint(const vec3 &point) { pivotPoint_ = point; }
-
-		/*! @name Camera manipulation */
-		//@{
- 	public:
-		/*! Returns whether or not the QGLViewer::ZOOM action zooms on the pivot
-		  point.
-
-		  When set to \c false (default), a zoom action will move the camera along its
[…]
