### 2020-12-09 c0a2dd8
use properties as arguments (instead of property names)
