+++ easy3d/core/surface_mesh.h
@@ -1443,31 +1443,31 @@ namespace easy3d {
         }
 
 
-        /** get the type_info \c T of vertex property named \c. returns an typeid(void)
+        /** get the type_info \c T of vertex property \p name. returns an typeid(void)
          if the property does not exist or if the type does not match. */
         const std::type_info& get_vertex_property_type(const std::string& name) const
         {
             return vprops_.get_type(name);
         }
-        /** get the type_info \c T of halfedge property named \c. returns an typeid(void)
+        /** get the type_info \c T of halfedge property \p name. returns an typeid(void)
          if the property does not exist or if the type does not match. */
         const std::type_info& get_halfedge_property_type(const std::string& name) const
         {
             return hprops_.get_type(name);
         }
-        /** get the type_info \c T of edge property named \c. returns an typeid(void)
+        /** get the type_info \c T of edge property \p name. returns an typeid(void)
          if the property does not exist or if the type does not match. */
         const std::type_info& get_edge_property_type(const std::string& name) const
         {
             return eprops_.get_type(name);
         }
-        /** get the type_info \c T of face property named \c. returns an typeid(void)
+        /** get the type_info \c T of face property \p name. returns an typeid(void)
          if the property does not exist or if the type does not match. */
         const std::type_info& get_face_property_type(const std::string& name) const
         {
             return fprops_.get_type(name);
         }
-        /** get the type_info \c T of model property named \c. returns an typeid(void)
+        /** get the type_info \c T of model property \p name. returns an typeid(void)
          if the property does not exist or if the type does not match. */
         const std::type_info& get_model_property_type(const std::string& name) const
         {
