+++ easy3d/core/point_cloud.h
@@ -377,13 +377,13 @@ namespace easy3d {
             return mprops_.rename(old_name, new_name);
         }
 
-        /** @brief get the type_info \c T of vertex property named \c. returns an typeid(void)
+        /** @brief get the type_info \c T of vertex property \p name. returns an typeid(void)
          if the property does not exist or if the type does not match. */
         const std::type_info& get_vertex_property_type(const std::string& name) const
         {
             return vprops_.get_type(name);
         }
-        /** @brief get the type_info \c T of model property named \c. returns an typeid(void)
+        /** @brief get the type_info \c T of model property \p name. returns an typeid(void)
         if the property does not exist or if the type does not match. */
         const std::type_info& get_model_property_type(const std::string& name) const
         {
