+++ doc/Doxyfile.in
@@ -758,7 +758,7 @@ WARN_LOGFILE           =
 # spaces.
 # Note: If this tag is empty the current directory is searched.
 
-INPUT                  = @PROJECT_SOURCE_DIR@/easy3d  @PROJECT_SOURCE_DIR@/tutorials/examples.h @PROJECT_SOURCE_DIR@/ReadMe.md @PROJECT_SOURCE_DIR@/resources/images
+INPUT                  = @PROJECT_SOURCE_DIR@/easy3d  @PROJECT_SOURCE_DIR@/tutorials/examples.h @PROJECT_SOURCE_DIR@/ReadMe.md
 
 # This tag can be used to specify the character encoding of the source files
 # that doxygen parses. Internally doxygen uses the UTF-8 encoding. Doxygen uses
@@ -846,7 +846,7 @@ EXAMPLE_RECURSIVE      = NO
 # that contain images that are to be included in the documentation (see the
 # \image command).
 
-IMAGE_PATH             =
+IMAGE_PATH             = @PROJECT_SOURCE_DIR@/resources/images
 
 # The INPUT_FILTER tag can be used to specify a program that doxygen should
 # invoke to filter for each input file. Doxygen will invoke the filter program
