### 2020-12-09 6d8f2b2
use uniform color if requested color property of scalar field doesn't exist
