+++ easy3d/renderer/buffers.cpp
@@ -48,7 +48,8 @@ namespace easy3d {
             // min_value and max_value return the expected value range.
             template<typename FT>
             inline void
-            clamp_scalar_field(const std::vector<FT> &property, float &min_value, float &max_value, float dummy_lower_percent,
+            clamp_scalar_field(const std::vector<FT> &property, float &min_value, float &max_value,
+                               float dummy_lower_percent,
                                float dummy_upper_percent) {
                 if (property.empty()) {
                     LOG(WARNING) << "empty property";
@@ -1542,8 +1543,12 @@ namespace easy3d {
                     auto texcoord = model->get_vertex_property<vec2>(name);
                     if (texcoord)
                         details::update(model, drawable, texcoord);
-                    else
-                        LOG(WARNING) << "texcoord property not found: " << drawable->property_name();
+                    else {
+                        LOG(WARNING) << "texcoord property \'" << drawable->property_name()
+                                     << "\' not found (use uniform coloring)";
+                        drawable->set_coloring_method(State::UNIFORM_COLOR);
+                        return;
+                    }
                     break;
                 }
 
@@ -1551,8 +1556,12 @@ namespace easy3d {
                     auto colors = model->get_vertex_property<vec3>(name);
                     if (colors)
                         details::update(model, drawable, colors);
-                    else
-                        LOG(WARNING) << "color property not found: " << drawable->property_name();
+                    else {
+                        LOG(WARNING) << "color property \'" << drawable->property_name()
+                                     << "\' not found (use uniform coloring)";
+                        drawable->set_coloring_method(State::UNIFORM_COLOR);
+                        return;
+                    }
                     break;
                 }
 
@@ -1576,7 +1585,9 @@ namespace easy3d {
                         auto prop = model->get_vertex_property<unsigned char>(name);
                         details::update(model, drawable, prop);
                     } else {
-                        LOG(WARNING) << "scalar field not found: " << drawable->property_name();
+                        LOG(WARNING) << "scalar field \'" << drawable->property_name()
+                                     << "\' not found (use uniform coloring)";
+                        drawable->set_coloring_method(State::UNIFORM_COLOR);
                         return;
                     }
                     break;
@@ -1643,8 +1654,12 @@ namespace easy3d {
                     auto texcoord = model->get_vertex_property<vec2>(name);
                     if (texcoord)
                         details::update(model, drawable, texcoord);
-                    else
-                        LOG(WARNING) << "texcoord property not found: " << drawable->property_name();
+                    else {
+                        LOG(WARNING) << "texcoord property \'" << drawable->property_name()
+                                     << "\' not found (use uniform coloring)";
+                        drawable->set_coloring_method(State::UNIFORM_COLOR);
+                        return;
+                    }
                     break;
                 }
 
@@ -1652,8 +1667,12 @@ namespace easy3d {
                     auto colors = model->get_vertex_property<vec3>(name);
                     if (colors)
                         details::update(model, drawable, colors);
-                    else
-                        LOG(WARNING) << "color property not found: " << drawable->property_name();
+                    else {
+                        LOG(WARNING) << "color property \'" << drawable->property_name()
+                                     << "\' not found (use uniform coloring)";
+                        drawable->set_coloring_method(State::UNIFORM_COLOR);
+                        return;
+                    }
                     break;
                 }
 
@@ -1677,7 +1696,9 @@ namespace easy3d {
                         auto prop = model->get_vertex_property<unsigned char>(name);
                         details::update(model, drawable, prop);
                     } else {
-                        LOG(WARNING) << "scalar field not found: " << drawable->property_name();
+                        LOG(WARNING) << "scalar field \'" << drawable->property_name()
[…]
