### 2020-12-09 5d6245c
improve property management
