### 2020-12-09 f9553bd
removed degenerate faces after remeshing self intersections
