+++ easy3d/algo_ext/self_intersection.h
@@ -131,7 +131,7 @@ namespace easy3d {
 
         // remove degenerate faces by collapsing tiny length edges.
         // returns the number of faces removed
-        int remove_degenerate_faces(SurfaceMesh *mesh, double coincident_threshold = 1e-4) const;
+        int remove_degenerate_faces(SurfaceMesh *mesh, double coincident_threshold = 1e-5) const;
 
         // test if two triangles intersect
         bool do_intersect(const Triangle &A, const Triangle &B);
