### 2020-12-10 adc3eed
topology statistics
