+++ easy3d/algo_ext/mesh_surfacer.cpp
@@ -24,10 +24,16 @@
 
 #include <easy3d/algo_ext/duplicated_faces.h>
 #include <easy3d/algo_ext/self_intersection.h>
-
 #include <easy3d/algo_ext/mesh_surfacer.h>
-
 #include <easy3d/util/logging.h>
+
+
+#include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
+
+#include <CGAL/Surface_mesh.h>
+#include <CGAL/Polygon_mesh_processing/orient_polygon_soup.h>
+#include <CGAL/Polygon_mesh_processing/polygon_soup_to_polygon_mesh.h>
+#include <CGAL/Polygon_mesh_processing/orientation.h>
 
 
 namespace easy3d {
@@ -39,6 +45,118 @@ namespace easy3d {
 
 
     MeshSurfacer::~MeshSurfacer() {
+    }
+
+
+    bool MeshSurfacer::orient(std::vector<vec3> &input_points, std::vector<Polygon> &input_polygons) {
+        typedef CGAL::Exact_predicates_inexact_constructions_kernel          K;
+        typedef CGAL::Surface_mesh<K::Point_3>                               Polyhedron;
+
+        std::vector<K::Point_3> points;
+        std::vector< std::vector<std::size_t> > polygons(input_polygons.size());
+
+        // helper class for identifying coincident vertices
+        class CmpVec {
+        public:
+            CmpVec(float _eps = FLT_MIN) : eps_(_eps) {}
+            bool operator()(const vec3 &v0, const vec3 &v1) const {
+                if (fabs(v0[0] - v1[0]) <= eps_) {
+                    if (fabs(v0[1] - v1[1]) <= eps_)
+                        return (v0[2] < v1[2] - eps_);
+                    else
+                        return (v0[1] < v1[1] - eps_);
+                } else
+                    return (v0[0] < v1[0] - eps_);
+            }
+        private:
+            float eps_;
+        };
+
+        CmpVec comp(FLT_MIN);
+        std::map<vec3, int, CmpVec>            vMap(comp);
+        std::map<vec3, int, CmpVec>::iterator  vMapIt;
+
+        std::map<int, int> index_map;
+
+        for (std::size_t i=0; i<input_points.size(); ++i) {
+            const vec3& p = input_points[i];
+
+            // has the point been referenced before?
+            vMapIt = vMap.find(p);
+            if (vMapIt == vMap.end()) {
+                // No : add vertex and remember idx/vector mapping
+                vMap[p] = points.size();
+                index_map[i] = points.size();
+                points.push_back(K::Point_3(p.x, p.y, p.z));
+            }
+            else {
+                // Yes : get index from map
+                index_map[i] = vMapIt->second;
+            }
+        }
+        for (std::size_t i=0; i<input_polygons.size(); ++i) {
+            for (auto id : input_polygons[i])
+                polygons[i].push_back(index_map[id]);
+        }
+
+        bool status = CGAL::Polygon_mesh_processing::orient_polygon_soup(points, polygons);
+
+        Polyhedron mesh;
+        CGAL::Polygon_mesh_processing::polygon_soup_to_polygon_mesh(points, polygons, mesh);
+
+        if(CGAL::is_closed(mesh))
+            CGAL::Polygon_mesh_processing::orient_to_bound_a_volume(mesh);
+
+        input_points.clear();
+        input_polygons.clear();
+        for (auto v : mesh.vertices()) {
+            const K::Point_3 &p = mesh.point(v);
+            input_points.emplace_back(
+                    vec3(CGAL::to_double(p.x()),
+                         CGAL::to_double(p.y()),
+                         CGAL::to_double(p.z())
+                    )
+            );
+        }
+
[…]
