### 2020-12-11 eb8ceac
update buffer with UNIFORM_COLOR if requested property not found
