+++ easy3d/renderer/buffers.cpp
@@ -1547,6 +1547,7 @@ namespace easy3d {
                         LOG(WARNING) << "texcoord property \'" << drawable->property_name()
                                      << "\' not found (use uniform coloring)";
                         drawable->set_coloring_method(State::UNIFORM_COLOR);
+                        update(model, drawable);
                         return;
                     }
                     break;
@@ -1560,6 +1561,7 @@ namespace easy3d {
                         LOG(WARNING) << "color property \'" << drawable->property_name()
                                      << "\' not found (use uniform coloring)";
                         drawable->set_coloring_method(State::UNIFORM_COLOR);
+                        update(model, drawable);
                         return;
                     }
                     break;
@@ -1588,6 +1590,7 @@ namespace easy3d {
                         LOG(WARNING) << "scalar field \'" << drawable->property_name()
                                      << "\' not found (use uniform coloring)";
                         drawable->set_coloring_method(State::UNIFORM_COLOR);
+                        update(model, drawable);
                         return;
                     }
                     break;
@@ -1658,6 +1661,7 @@ namespace easy3d {
                         LOG(WARNING) << "texcoord property \'" << drawable->property_name()
                                      << "\' not found (use uniform coloring)";
                         drawable->set_coloring_method(State::UNIFORM_COLOR);
+                        update(model, drawable);
                         return;
                     }
                     break;
@@ -1671,6 +1675,7 @@ namespace easy3d {
                         LOG(WARNING) << "color property \'" << drawable->property_name()
                                      << "\' not found (use uniform coloring)";
                         drawable->set_coloring_method(State::UNIFORM_COLOR);
+                        update(model, drawable);
                         return;
                     }
                     break;
@@ -1699,6 +1704,7 @@ namespace easy3d {
                         LOG(WARNING) << "scalar field \'" << drawable->property_name()
                                      << "\' not found (use uniform coloring)";
                         drawable->set_coloring_method(State::UNIFORM_COLOR);
+                        update(model, drawable);
                         return;
                     }
                     break;
@@ -1747,6 +1753,7 @@ namespace easy3d {
                                 LOG(WARNING) << "texcoord property \'" << drawable->property_name()
                                              << "\' not found on edges (use uniform coloring)";
                                 drawable->set_coloring_method(State::UNIFORM_COLOR);
+                                update(model, drawable);
                                 return;
                             }
                             break;
@@ -1759,6 +1766,7 @@ namespace easy3d {
                                 LOG(WARNING) << "texcoord property \'" << drawable->property_name()
                                              << "\' not found on vertices (use uniform coloring)";
                                 drawable->set_coloring_method(State::UNIFORM_COLOR);
+                                update(model, drawable);
                                 return;
                             }
                             break;
@@ -1781,6 +1789,7 @@ namespace easy3d {
                                 LOG(WARNING) << "color property \'" << drawable->property_name()
                                              << "\' not found (use uniform coloring)";
                                 drawable->set_coloring_method(State::UNIFORM_COLOR);
+                                update(model, drawable);
                                 return;
                             }
                             break;
@@ -1793,6 +1802,7 @@ namespace easy3d {
                                 LOG(WARNING) << "color property \'" << drawable->property_name()
                                              << "\' not found (use uniform coloring)";
                                 drawable->set_coloring_method(State::UNIFORM_COLOR);
+                                update(model, drawable);
                                 return;
                             }
                             break;
@@ -1824,6 +1834,7 @@ namespace easy3d {
                                 LOG(WARNING) << "scalar field \'" << drawable->property_name()
                                              << "\' not found on edges (use uniform coloring)";
                                 drawable->set_coloring_method(State::UNIFORM_COLOR);
+                                update(model, drawable);
                                 return;
                             }
                             break;
[…]
