### 2020-12-11 f781486
stitching and reorientation, rename garbage collection functions
