+++ easy3d/algo_ext/surfacer.cpp
@@ -1,0 +1,315 @@
+/**
+ * Copyright (C) 2015 by Liangliang Nan (liangliang.nan@gmail.com)
+ * https://3d.bk.tudelft.nl/liangliang/
+ *
+ * This file is part of Easy3D. If it is useful in your research/work,
+ * I would be grateful if you show your appreciation by citing it:
+ * ------------------------------------------------------------------
+ *      Liangliang Nan.
+ *      Easy3D: a lightweight, easy-to-use, and efficient C++
+ *      library for processing and rendering 3D data. 2018.
+ * ------------------------------------------------------------------
+ * Easy3D is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License Version 3
+ * as published by the Free Software Foundation.
+ *
+ * Easy3D is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+#include <easy3d/algo_ext/surfacer.h>
+
+#include <easy3d/algo_ext/duplicate_faces.h>
+#include <easy3d/algo_ext/self_intersection.h>
+#include <easy3d/util/logging.h>
+
+
+#include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
+
+#include <CGAL/Surface_mesh.h>
+#include <CGAL/Polygon_mesh_processing/orient_polygon_soup.h>
+#include <CGAL/Polygon_mesh_processing/polygon_soup_to_polygon_mesh.h>
+#include <CGAL/Polygon_mesh_processing/orientation.h>
+
+typedef CGAL::Exact_predicates_inexact_constructions_kernel          K;
+typedef CGAL::Surface_mesh<K::Point_3>                               CGALMesh;
+
+
+
+namespace easy3d {
+
+    namespace details {
+
+        void easy3d_to_cgal(SurfaceMesh *input, CGALMesh &output) {
+            if (input->has_garbage())
+                input->collect_garbage();
+
+            output.clear();
+
+            for (auto v : input->vertices()) {
+                const vec3 &p = input->position(v);
+                output.add_vertex(K::Point_3(p.x, p.y, p.z));
+            }
+
+            for (auto f : input->faces()) {
+                std::vector<CGALMesh::Vertex_index> vts;
+                for (auto v : input->vertices(f))
+                    vts.push_back(CGALMesh::Vertex_index(v.idx()));
+                output.add_face(vts);
+            }
+        }
+
+        void cgal_to_easy3d(CGALMesh &input, SurfaceMesh *output) {
+            if (input.has_garbage())
+                input.collect_garbage();
+
+            output->clear();
+            for (auto v : input.vertices()) {
+                const K::Point_3 &p = input.point(v);
+                output->add_vertex(
+                        vec3(CGAL::to_double(p.x()),
+                             CGAL::to_double(p.y()),
+                             CGAL::to_double(p.z())
+                        )
+                );
+            }
+
+            for (auto f : input.faces()) {
+                std::vector<SurfaceMesh::Vertex> vts;
+                auto h = input.halfedge(f);
+                for (auto v : input.vertices_around_face(h))
+                    vts.push_back(SurfaceMesh::Vertex(v.idx()));
+                output->add_face(vts);
+            }
+        }
+
+
+        void easy3d_mesh_to_polygon_soup(SurfaceMesh *mesh,
[…]
