+++ easy3d/core/surface_mesh.h
@@ -324,7 +324,7 @@ namespace easy3d {
             /// Default constructor
             VertexIterator(Vertex v=Vertex(), const SurfaceMesh* m=nullptr) : hnd_(v), mesh_(m)
             {
-                if (mesh_ && mesh_->garbage()) while (mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) ++hnd_.idx_;
+                if (mesh_ && mesh_->has_garbage()) while (mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) ++hnd_.idx_;
             }
 
             /// get the vertex the iterator refers to
@@ -347,7 +347,7 @@ namespace easy3d {
             {
                 ++hnd_.idx_;
                 assert(mesh_);
-                while (mesh_->garbage() && mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) ++hnd_.idx_;
+                while (mesh_->has_garbage() && mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) ++hnd_.idx_;
                 return *this;
             }
 
@@ -356,7 +356,7 @@ namespace easy3d {
             {
                 --hnd_.idx_;
                 assert(mesh_);
-                while (mesh_->garbage() && mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) --hnd_.idx_;
+                while (mesh_->has_garbage() && mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) --hnd_.idx_;
                 return *this;
             }
 
@@ -376,7 +376,7 @@ namespace easy3d {
             /// Default constructor
             HalfedgeIterator(Halfedge h=Halfedge(), const SurfaceMesh* m=nullptr) : hnd_(h), mesh_(m)
             {
-                if (mesh_ && mesh_->garbage()) while (mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) ++hnd_.idx_;
+                if (mesh_ && mesh_->has_garbage()) while (mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) ++hnd_.idx_;
             }
 
             /// get the halfedge the iterator refers to
@@ -399,7 +399,7 @@ namespace easy3d {
             {
                 ++hnd_.idx_;
                 assert(mesh_);
-                while (mesh_->garbage() && mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) ++hnd_.idx_;
+                while (mesh_->has_garbage() && mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) ++hnd_.idx_;
                 return *this;
             }
 
@@ -408,7 +408,7 @@ namespace easy3d {
             {
                 --hnd_.idx_;
                 assert(mesh_);
-                while (mesh_->garbage() && mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) --hnd_.idx_;
+                while (mesh_->has_garbage() && mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) --hnd_.idx_;
                 return *this;
             }
 
@@ -428,7 +428,7 @@ namespace easy3d {
             /// Default constructor
             EdgeIterator(Edge e=Edge(), const SurfaceMesh* m=nullptr) : hnd_(e), mesh_(m)
             {
-                if (mesh_ && mesh_->garbage()) while (mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) ++hnd_.idx_;
+                if (mesh_ && mesh_->has_garbage()) while (mesh_->is_valid(hnd_) && mesh_->is_deleted(hnd_)) ++hnd_.idx_;
             }
 
[…]
