+++ easy3d/algo/tessellator.h
@@ -35,7 +35,8 @@ namespace easy3d {
      * into triangles or simple contours. \par
      * \class Tessellator easy3d/algo/tessellator.h
      * This implementation is also able to keep track of the unique vertices and respective indices, which allows to
-     * take advantage of the element buffer for efficient rendering (i.e., avoid sending duplicated vertices to GPU).
+     * take advantage of the element buffer for efficient rendering (i.e., avoid sending vertices with the same
+     * geometry to the GPU).
      * \par
      * Typical applications:
      *   - Tessellate concave polygons, polygons with holes, or polygons with intersecting edges;

[…]
