+++ easy3d/algo_ext/surfacer.cpp
@@ -32,19 +32,19 @@
 
 #include <CGAL/Surface_mesh.h>
 #include <CGAL/Polygon_mesh_processing/orient_polygon_soup.h>
+#include <CGAL/Polygon_mesh_processing/repair_polygon_soup.h>
 #include <CGAL/Polygon_mesh_processing/polygon_soup_to_polygon_mesh.h>
 #include <CGAL/Polygon_mesh_processing/orientation.h>
 
-typedef CGAL::Exact_predicates_inexact_constructions_kernel          K;
-typedef CGAL::Surface_mesh<K::Point_3>                               CGALMesh;
-
+typedef CGAL::Exact_predicates_inexact_constructions_kernel K;
+typedef CGAL::Surface_mesh<K::Point_3> CGALMesh;
 
 
 namespace easy3d {
 
     namespace details {
 
-        void easy3d_to_cgal(SurfaceMesh *input, CGALMesh &output) {
+        void to_cgal(SurfaceMesh *input, CGALMesh &output) {
             if (input->has_garbage())
                 input->collect_garbage();
 
@@ -63,7 +63,7 @@ namespace easy3d {
             }
         }
 
-        void cgal_to_easy3d(CGALMesh &input, SurfaceMesh *output) {
+        void to_easy3d(CGALMesh &input, SurfaceMesh *output) {
             if (input.has_garbage())
                 input.collect_garbage();
 
@@ -88,9 +88,109 @@ namespace easy3d {
         }
 
 
-        void easy3d_mesh_to_polygon_soup(SurfaceMesh *mesh,
-                                         std::vector<vec3> &points,
-                                         std::vector<Surfacer::Polygon> &polygons) {
+        void to_cgal(const std::vector<vec3> &input_points,
+                     const std::vector<Surfacer::Polygon> &input_polygons,
+                     std::vector<K::Point_3> &points,
+                     std::vector<std::vector<std::size_t> > &polygons) {
+            points.resize(input_points.size());
+            polygons.resize(input_polygons.size());
+
+            for (std::size_t i = 0; i < input_points.size(); ++i) {
+                const vec3 &p = input_points[i];
+                points[i] = K::Point_3(p.x, p.y, p.z);
+            }
+
+            for (std::size_t i = 0; i < input_polygons.size(); ++i) {
+                const auto &input_plg = input_polygons[i];
+                auto &plg = polygons[i];
+                plg.resize(input_plg.size());
+                for (std::size_t j = 0; j < input_plg.size(); ++j)
+                    plg[j] = input_plg[j];
+            }
+        }
+
+        void to_easy3d(const std::vector<K::Point_3> &input_points,
+                       const std::vector<std::vector<std::size_t> > &input_polygons,
+                       std::vector<vec3> &points,
+                       std::vector<Surfacer::Polygon> &polygons) {
+            points.resize(input_points.size());
+            polygons.resize(input_polygons.size());
+
+            for (std::size_t i = 0; i < input_points.size(); ++i) {
+                const auto &p = input_points[i];
+                points[i] = vec3(
+                        CGAL::to_double(p.x()),
+                        CGAL::to_double(p.y()),
+                        CGAL::to_double(p.z())
+                );
+            }
+
+            for (std::size_t i = 0; i < input_polygons.size(); ++i) {
+                const auto &input_plg = input_polygons[i];
+                auto &plg = polygons[i];
+                plg.resize(input_plg.size());
+                for (std::size_t j = 0; j < input_plg.size(); ++j)
+                    plg[j] = input_plg[j];
+            }
+        }
+
+
+        void remove_duplicate_vertices(std::vector<vec3> &points,
[…]
