+++ applications/Mapple/main_window.h
@@ -108,6 +108,7 @@ private slots:
     void surfaceMeshPlanarPartition();
     void surfaceMeshTriangulation();
 
+    void surfaceMeshClean();
     void surfaceMeshStitchConnectedComponents();
     void surfaceMeshStitchBordersWithoutReorientation();
     void surfaceMeshReverseOrientation();
