+++ applications/Mapple/main_window.cpp
@@ -795,6 +795,7 @@ void MainWindow::createActionsForSurfaceMeshMenu() {
     connect(ui->actionSurfaceMeshRemoveIsolatedVertices, SIGNAL(triggered()), this, SLOT(surfaceMeshRemoveIsolatedVertices()));
     connect(ui->actionSurfaceMeshTriangulation, SIGNAL(triggered()), this, SLOT(surfaceMeshTriangulation()));
 
+    connect(ui->actionCleanPolygonMesh, SIGNAL(triggered()), this, SLOT(surfaceMeshClean()));
     connect(ui->actionStitchConnectedComponents, SIGNAL(triggered()), this, SLOT(surfaceMeshStitchConnectedComponents()));
     connect(ui->actionStitchBordersWithoutReorientation, SIGNAL(triggered()), this, SLOT(surfaceMeshStitchBordersWithoutReorientation()));
     connect(ui->actionReverseOrientation, SIGNAL(triggered()), this, SLOT(surfaceMeshReverseOrientation()));
@@ -917,13 +918,33 @@ void MainWindow::surfaceMeshTriangulation() {
 }
 
 
+
+void MainWindow::surfaceMeshClean() {
+    SurfaceMesh* mesh = dynamic_cast<SurfaceMesh*>(viewer()->currentModel());
+    if (!mesh)
+        return;
+
+#if HAS_CGAL
+    Surfacer::clean_polygon_mesh(mesh);
+
+    mesh->renderer()->update();
+    viewer_->update();
+    updateUi();
+#else
+    SurfaceMeshStitching stitch(mesh);
+    stitch.apply();
+    LOG(WARNING) << "install CGAL to allow stitching connected components with incompatible boundaries";
+#endif
+}
+
+
 void MainWindow::surfaceMeshStitchConnectedComponents() {
     SurfaceMesh* mesh = dynamic_cast<SurfaceMesh*>(viewer()->currentModel());
     if (!mesh)
         return;
 
 #if HAS_CGAL
-    // Surfacer::merge_reversible_connected_components_2(mesh); // does the same, but treats the mesh as a polygon soup
+//     Surfacer::merge_reversible_connected_components_2(mesh); // does the same, but treats the mesh as a polygon soup
     Surfacer::merge_reversible_connected_components(mesh);
 
     mesh->renderer()->update();
