### 2020-12-11 3047079
add two methods for processing polygon soups
