### 2020-12-12 11b4846
remember clipping plane orientation
