### 2020-12-12 b684a69
add is_closed()
