+++ easy3d/core/surface_mesh.cpp
@@ -676,6 +676,21 @@ namespace easy3d {
 
     bool
     SurfaceMesh::
+    is_closed() const
+    {
+        EdgeIterator eit=edges_begin(), eend=edges_end();
+        for (; eit!=eend; ++eit)
+            if (is_border(*eit))
+                return false;
+        return true;
+    }
+
+
+    //-----------------------------------------------------------------------------
+
+
+    bool
+    SurfaceMesh::
     is_triangle_mesh() const
     {
         FaceIterator fit=faces_begin(), fend=faces_end();
