### 2020-12-12 4820178
fix clipping plane effect (wrong side visualized)
