### 2020-12-12 25f8944
add clip(), split(), and slice()
