+++ easy3d/algo_ext/surfacer.h
@@ -36,17 +36,25 @@ namespace easy3d {
 
     class SurfaceMesh;
 
-    /// \brief A collection of mesh (and polygon soup) processing functions.
-    /// \class Surfacer  easy3d/algo_ext/surfacer.h
-    /// \details This class collects some related algorithms implemented using CGAL.
-    /// It allows (re)orientation, detecting and resolving topological issues (e.g., duplicate vertices/faces, self
-    /// intersection) of a surface mesh.
-    /// \see DuplicateFaces and SelfIntersection.
-    class Surfacer
-    {
+    /**
+     * \brief A collection of mesh (and polygon soup) processing functions.
+     *
+     * \class Surfacer  easy3d/algo_ext/surfacer.h
+     *
+     * \details This class collects some related algorithms implemented using CGAL.
+     *          It allows (re)orientation, detecting and resolving topological issues (e.g., duplicate vertices/faces,
+     *          self intersection), and clipping/splitting/slicing of a surface mesh.
+     *
+     * \see DuplicateFaces and SelfIntersection.
+     */
+    class Surfacer {
     public:
-        /// A polygon is represented by a list of vertex indices
+        /// A polygon represented by a list of vertex indices
         typedef std::vector<int> Polygon;
+
+        /// A polyline represented by a list of points.
+        /// If a polyline is closed, its first and last points will be identical.
+        typedef std::vector<vec3> Polyline;
 
     public:
 
@@ -62,7 +70,7 @@ namespace easy3d {
          *
          * \see merge_reversible_connected_components_2()
          */
-        static void merge_reversible_connected_components(SurfaceMesh* mesh);
+        static void merge_reversible_connected_components(SurfaceMesh *mesh);
 
         /**
          * \brief Stitches connected components of a surface mesh
@@ -72,12 +80,19 @@ namespace easy3d {
          *
          * \see merge_reversible_connected_components()
          */
-        static bool merge_reversible_connected_components_2(SurfaceMesh* mesh);
+        static bool merge_reversible_connected_components_2(SurfaceMesh *mesh);
 
         /**
          * \brief Reverses the orientation of the entire mesh
          */
-        static void reverse_orientation(SurfaceMesh* mesh);
+        static void reverse_orientation(SurfaceMesh *mesh);
+
+        /**
+         * \brief Makes each connected component of a closed triangle surface mesh inward or outward oriented.
+         * \pre mesh.is_triangle_mesh(), mesh.is_closed()
+         * @param mesh The input mesh.
+         */
+        static void orient(SurfaceMesh *mesh);
 
         /**
          * \brief Tries to consistently orient a polygon soup.
@@ -93,7 +108,7 @@ namespace easy3d {
          * \return \c true if the orientation operation succeeded. \c false if some points were duplicated, thus
          *         producing a self-intersecting polyhedron.
          */
-        static bool orient_polygon_soup(std::vector<vec3>& points, std::vector<Polygon>& polygons);
+        static bool orient_polygon_soup(std::vector<vec3> &points, std::vector<Polygon> &polygons);
 
         /**
          * \brief Repairs a given polygon soup through various repairing operations.
@@ -108,7 +123,7 @@ namespace easy3d {
          * \note The point and polygon containers will be modified by the repairing operations, and thus the indexation
          * of the polygons will also be changed.
          */
-        static void repair_polygon_soup(std::vector<vec3>& points, std::vector<Polygon>& polygons);
+        static void repair_polygon_soup(std::vector<vec3> &points, std::vector<Polygon> &polygons);
 
         /**
          * \brief Repairs a given polygon mesh through various repairing operations.
@@ -125,7 +140,7 @@ namespace easy3d {
[…]
