### 2020-12-12 cecce00
polgon mesh processing functions
