+++ easy3d/algo_ext/surfacer.cpp
@@ -22,15 +22,15 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 #include <easy3d/algo_ext/surfacer.h>
-
 #include <easy3d/algo_ext/duplicate_faces.h>
 #include <easy3d/algo_ext/self_intersection.h>
 #include <easy3d/util/logging.h>
-
+#include <easy3d/core/manifold_builder.h>
 
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 
 #include <CGAL/Surface_mesh.h>
+#include <CGAL/Polygon_mesh_processing/self_intersections.h>
 #include <CGAL/Polygon_mesh_processing/orient_polygon_soup.h>
 #include <CGAL/Polygon_mesh_processing/repair_polygon_soup.h>
 #include <CGAL/Polygon_mesh_processing/polygon_soup_to_polygon_mesh.h>
@@ -211,16 +211,53 @@ namespace easy3d {
                              const std::vector<Surfacer::Polygon> &polygons,
                              SurfaceMesh *mesh) {
             mesh->clear();
+
+            ManifoldBuilder builder(mesh);
+            builder.begin_surface();
+
             for (auto p : points)
-                mesh->add_vertex(p);
+                builder.add_vertex(p);
 
             for (const auto &plg : polygons) {
                 std::vector<SurfaceMesh::Vertex> vts;
                 for (auto v : plg)
                     vts.push_back(SurfaceMesh::Vertex(v));
-                mesh->add_face(vts);
+                builder.add_face(vts);
             }
+            builder.end_surface(false);
         }
+
+        void to_polygon_mesh(std::vector<K::Point_3> &points,
+                             std::vector<std::vector<std::size_t>> &polygons,
+                             CGALMesh& mesh) {
+            mesh.clear();
+            CGAL::Polygon_mesh_processing::polygon_soup_to_polygon_mesh(points, polygons, mesh);
+        }
+    }
+
+
+
+    int Surfacer::stitch_borders(SurfaceMesh *input) {
+        CGALMesh mesh;
+        details::to_cgal(input, mesh);
+        int num = CGAL::Polygon_mesh_processing::stitch_borders(mesh);
+        details::to_easy3d(mesh, input);
+        return num;
+    }
+
+
+    bool Surfacer::orient_and_stitch_polygon_soup(SurfaceMesh *mesh) {
+        if (!mesh)
+            return false;
+
+        std::vector<vec3> points;
+        std::vector<Polygon> polygons;
+        details::to_polygon_soup(mesh, points, polygons);
+
+        bool status = orient_and_stitch_polygon_soup(points, polygons);
+        details::to_polygon_mesh(points, polygons, mesh);
+
+        return status;
     }
 
 
@@ -230,28 +267,21 @@ namespace easy3d {
 
         typedef boost::property_map<CGALMesh, CGAL::dynamic_face_property_t<std::size_t> >::type Fccmap;
         Fccmap fccmap = CGAL::get(CGAL::dynamic_face_property_t<std::size_t>(), mesh);
-        int count_oirg = CGAL::Polygon_mesh_processing::connected_components(mesh, fccmap);
-        if (count_oirg <= 1) {
-            LOG(INFO) << "model has only one connected components, so nothing to merge";
+        int count_prev = CGAL::Polygon_mesh_processing::connected_components(mesh, fccmap);
+        if (count_prev <= 1) {
             return; // nothing to be merged
         }
 
         // iteratively merge the connected components until the no more connected components can be merged
-        int count_prev = count_oirg;
-        int count_now = -1;
         do {
             CGAL::Polygon_mesh_processing::merge_reversible_connected_components(mesh);
[…]
