+++ applications/Mapple/main_window.h
@@ -108,9 +108,9 @@ private slots:
     void surfaceMeshPlanarPartition();
     void surfaceMeshTriangulation();
 
-    void surfaceMeshStitchConnectedComponents();
-    void surfaceMeshStitchBordersWithoutReorientation();
-    void surfaceMeshOrient();
+    void surfaceMeshStitchWithReorientation();
+    void surfaceMeshStitchWithoutReorientation();
+    void surfaceMeshOrientClosedTriangleMesh();
     void surfaceMeshReverseOrientation();
     void surfaceMeshDetectDuplicateFaces();
     void surfaceMeshRemoveDuplicateFaces();
@@ -118,7 +118,7 @@ private slots:
     void surfaceMeshRemeshSelfIntersections();
 
     void surfaceMeshRepairPolygonSoup();
-    void surfaceMeshStitchPolygonSoup();
+    void surfaceMeshOrientAndStitchPolygonSoup();
 
     void surfaceMeshClip();
     void surfaceMeshSplit();

[…]
