### 2020-12-12 ae08215
detect and remove overlapping faces
