+++ easy3d/algo_ext/overlapping_faces.cpp
@@ -22,7 +22,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <easy3d/algo_ext/duplicate_faces.h>
+#include <easy3d/algo_ext/overlapping_faces.h>
 
 #include <set>
 #include <unordered_map>
@@ -34,7 +34,7 @@
 namespace easy3d {
 
 
-    DuplicateFaces::Triangles DuplicateFaces::mesh_to_cgal_triangle_list(SurfaceMesh* mesh) {
+    OverlappingFaces::Triangles OverlappingFaces::mesh_to_cgal_triangle_list(SurfaceMesh* mesh) {
         auto prop = mesh->get_vertex_property<vec3>("v:point");
 
         Triangles triangles;
@@ -61,8 +61,7 @@ namespace easy3d {
     }
 
 
-    bool DuplicateFaces::do_duplicate(const Triangle &A, const Triangle &B, bool exact, double sqr_eps)
-    {
+    OverlappingFaces::OverlapType OverlappingFaces::do_overlap(const Triangle &A, const Triangle &B, double sqr_eps) {
         // Number of combinatorially shared vertices
         int num_comb_shared_vertices = 0;
 
@@ -70,52 +69,88 @@ namespace easy3d {
         int num_geom_shared_vertices = 0;
 
         // Keep track of shared vertex indices
-        std::vector< std::pair<SurfaceMesh::Vertex, SurfaceMesh::Vertex> > shared;
+        std::vector< std::pair<unsigned short, unsigned short> > shared;
+
         for (unsigned short ea = 0; ea < 3; ++ea) {
             for (unsigned short eb = 0; eb < 3; ++eb) {
                 if (A.vertices[ea] == B.vertices[eb]) {
                     ++num_comb_shared_vertices;
                     shared.emplace_back(ea, eb);
                 }
-                else if (exact) {
-                    if (A.triangle.vertex(ea) == B.triangle.vertex(eb)) {
-                        ++num_geom_shared_vertices;
-                        shared.emplace_back(ea, eb);
-                    }
-                }
-                else {  // use distance threshold
-                    if (CGAL::squared_distance(A.triangle.vertex(ea), B.triangle.vertex(eb)) < sqr_eps) {
-                        ++num_geom_shared_vertices;
-                        shared.emplace_back(ea, eb);
-                    }
+                else if (CGAL::squared_distance(A.triangle.vertex(ea), B.triangle.vertex(eb)) < sqr_eps) {
+                    ++num_geom_shared_vertices;
+                    shared.emplace_back(ea, eb);
                 }
             }
         }
         const int total_shared_vertices = num_comb_shared_vertices + num_geom_shared_vertices;
         if (num_comb_shared_vertices == 3) {
             assert(shared.size() == 3);		// Combinatorially duplicate faces
-            return true;
+            return OT_SAME;
         }
         if (total_shared_vertices == 3) {
             assert(shared.size() == 3);		// Geometrically duplicate faces.
-            return true;
+            return OT_SAME;
         }
 
-        return false;
+#if 1   // Liangliang: let's handle coplanar and partial overlap
+        if (total_shared_vertices == 2) {
+            //
+            //  2  o    o 3
+            //     |\  /|
+            //     | \/ |
+            //     | /\ |
+            //     |/  \|
+            //  0  o----o 1
+            //
+            auto normalize = [](Vector_3& n) -> void { n /= std::sqrt(n.squared_length()); };
+            auto na = A.triangle.supporting_plane().orthogonal_vector();
+            auto nb = B.triangle.supporting_plane().orthogonal_vector();
+            normalize(na);
+            normalize(nb);
[…]
