### 2020-12-13 cd203ed
simpler API for detect/remove overlapping faces
