### 2020-12-13 f535e98
log result info
