### 2020-12-13 a41703d
remove degenerate faces
