### 2020-12-13 364a805
enumerate planar components
