+++ easy3d/algo/surface_mesh_planar_partition.cpp
@@ -1,156 +1,0 @@
-/**
- * Copyright (C) 2015 by Liangliang Nan (liangliang.nan@gmail.com)
- * https://3d.bk.tudelft.nl/liangliang/
- *
- * This file is part of Easy3D. If it is useful in your research/work,
- * I would be grateful if you show your appreciation by citing it:
- * ------------------------------------------------------------------
- *      Liangliang Nan.
- *      Easy3D: a lightweight, easy-to-use, and efficient C++
- *      library for processing and rendering 3D data. 2018.
- * ------------------------------------------------------------------
- * Easy3D is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License Version 3
- * as published by the Free Software Foundation.
- *
- * Easy3D is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-
-#include <easy3d/algo/surface_mesh_planar_partition.h>
-
-#include <set>
-#include <stack>
-
-#include <easy3d/core/surface_mesh.h>
-#include <easy3d/algo/surface_mesh_curvature.h>
-#include <easy3d/algo/surface_mesh_geometry.h>
-#include <easy3d/util/logging.h>
-
-
-namespace easy3d {
-
-    SurfaceMeshPlanarPartition::SurfaceMeshPlanarPartition(SurfaceMesh *mesh)
-            : mesh_(mesh) {
-        mesh_->update_vertex_normals();
-        vertex_normal_ = mesh_->get_vertex_property<vec3>("v:normal");
-    }
-
-
-    SurfaceMeshPlanarPartition::~SurfaceMeshPlanarPartition(void) {
-    }
-
-
-    bool SurfaceMeshPlanarPartition::can_grow(SurfaceMesh::Vertex v) const {
-        for (auto f : mesh_->faces(v)) {
-            if (planar_segments_[f] == -1)
-                return true;
-        }
-        return false;
-    }
-
-
-    float SurfaceMeshPlanarPartition::max_squared_dist_to_plane(SurfaceMesh::Face f, const Plane3& plane) const {
-        float max_sd = -FLT_MAX;
-        for (auto v : mesh_->vertices(f)) {
-            float sd = plane.squared_ditance(mesh_->position(v));
-            if (sd > max_sd)
-                max_sd = sd;
-        }
-        return max_sd;
-    }
-
-
-    void SurfaceMeshPlanarPartition::propagate(SurfaceMesh::Vertex v, int id, float max_allowed_squared_deviation) {
-        std::stack<SurfaceMesh::Face> stack;
-        for (auto f : mesh_->faces(v))
-            stack.push(f);
-
-        const Plane3 plane(mesh_->position(v), vertex_normal_[v]);
-
-        while (!stack.empty()) {
-            SurfaceMesh::Face top = stack.top();
-            stack.pop();
-
-            if (planar_segments_[top] == -1 && max_squared_dist_to_plane(top, plane) <= max_allowed_squared_deviation) {
-                planar_segments_[top] = id;
-
-                for (auto h : mesh_->halfedges(top)) {
-                    auto op = mesh_->opposite(h);
-                    if (!mesh_->is_border(op)) {
-                        auto f = mesh_->face(op);
-                        if (planar_segments_[f] == -1) {
-                            stack.push(f);
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-
[…]
