+++ easy3d/algo/surface_mesh_enumerator.h
@@ -78,6 +78,36 @@ namespace easy3d {
         propagate_connected_component(SurfaceMesh *mesh, SurfaceMesh::FaceProperty<int> id, SurfaceMesh::Face seed,
                                       int cur_id);
 
+        /**
+         * \brief Enumerates planar patches.
+         * @param mesh The input mesh.
+         * @param id The vertex property storing the result.
+         * \param angle_threshold Two faces sharing a common edge are considered coplanar if the dihedral angle is
+         *      smaller than \p angle_threshold (in degrees).
+         * @return The number of connected components.
+         */
+        static int enumerate_planar_components(
+                SurfaceMesh *mesh,
+                SurfaceMesh::FaceProperty<int> id,
+                float angle_threshold = 1.0f
+        );
+
+        /**
+         * \brief Propagates a planar patch from the seed face.
+         * @param mesh The input mesh.
+         * @param id The face property storing the result.
+         * @param seed The seed face from which the planar patch will be propagated.
+         * @param cur_id The index of the connected component to be propagated.
+         * \param angle_threshold Two faces sharing a common edge are considered coplanar if the dihedral angle is
+         *      smaller than \p angle_threshold (in degrees).
+         * @attention The face property id must be initialize with "-1".
+         */
+        static void propagate_planar_component(SurfaceMesh *mesh,
+                                               SurfaceMesh::FaceProperty<int> id,
+                                               SurfaceMesh::Face seed, int cur_id,
+                                               float angle_threshold
+        );
+
     };
 
 }   // namespace easy3d
