### 2020-12-15 e67d4df
add join_edges()
