### 2020-12-15 3bffe75
collapse degenerate edges
