### 2020-12-15 01ed6af
enumerate_planar_components()
