+++ easy3d/algo/surface_mesh_enumerator.h
@@ -100,6 +100,7 @@ namespace easy3d {
          * @param cur_id The index of the connected component to be propagated.
          * \param angle_threshold Two faces sharing a common edge are considered coplanar if the dihedral angle is
          *      smaller than \p angle_threshold (in degrees).
+         * \pre The face normal property (with name \c "f:normal") must exist and be up to date.
          * @attention The face property id must be initialize with "-1".
          */
         static void propagate_planar_component(SurfaceMesh *mesh,
