### 2020-12-15 dde28bf
surface mesh polygonization
