+++ easy3d/algo/CMakeLists.txt
@@ -23,6 +23,7 @@ set(${PROJECT_NAME}_HEADERS
         surface_mesh_geodesic.h
         surface_mesh_hole_filling.h
         surface_mesh_parameterization.h
+        surface_mesh_polygonization.h
         surface_mesh_remeshing.h
         surface_mesh_sampler.h
         surface_mesh_simplification.h
@@ -55,6 +56,7 @@ set(${PROJECT_NAME}_SOURCES
         surface_mesh_geodesic.cpp
         surface_mesh_hole_filling.cpp
         surface_mesh_parameterization.cpp
+        surface_mesh_polygonization.cpp
         surface_mesh_remeshing.cpp
         surface_mesh_sampler.cpp
         surface_mesh_simplification.cpp
