+++ easy3d/core/constant.h
@@ -30,15 +30,15 @@
 
 namespace easy3d {
 
-#ifndef M_PI
-#define M_PI			3.14159265358979323846264338327950288419716939937510582   // pi
+#ifndef EASY3D_PI
+#define EASY3D_PI			3.14159265358979323846264338327950288419716939937510582
 #endif
 
-#define half_pi			M_PI * 0.5
-#define quarter_pi		M_PI * 0.25
-#define two_pi			M_PI * 2.0          
-#define rad2deg(a)		((a) * 180.0 / M_PI)
-#define deg2rad(a)		((a) * M_PI / 180.0)
+#define half_pi			EASY3D_PI * 0.5
+#define quarter_pi		EASY3D_PI * 0.25
+#define two_pi			EASY3D_PI * 2.0
+#define rad2deg(a)		((a) * 180.0 / EASY3D_PI)
+#define deg2rad(a)		((a) * EASY3D_PI / 180.0)
 
     // Function returning min/max for corresponding type
     template <typename FT> inline FT min();
