### 2020-12-15 993c8ef
use constant values (instead of #define)
