### 2020-12-15 dbb8a91
edge collapse
