### 2020-12-15 c025adc
resolve non-manifoldness after deleting faces
