### 2020-12-15 1a18f62
add read_poly() and write_poly() to SurfaceMesh class
