+++ easy3d/core/surface_mesh.cpp
@@ -41,6 +41,7 @@
 #include <easy3d/util/logging.h>
 
 #include <cmath>
+#include <fstream>
 
 namespace easy3d {
 
@@ -166,6 +167,105 @@ namespace easy3d {
         }
 
         return *this;
+    }
+
+
+    //-----------------------------------------------------------------------------
+
+
+    bool SurfaceMesh::read_poly(const std::string &file_name)
+    {
+        clear();
+
+        // open file (in binary mode)
+        std::ifstream input(file_name.c_str(), std::fstream::binary);
+        if (input.fail()) {
+            LOG(ERROR) << "could not open file: " << file_name;
+            return false;
+        }
+
+        // how many elements?
+        unsigned int nv, ne, nh, nf;
+        input.read((char*)&nv, sizeof(unsigned int));
+        input.read((char*)&ne, sizeof(unsigned int));
+        input.read((char*)&nf, sizeof(unsigned int));
+        nh = 2*ne;
+
+        // resize containers
+        resize(nv, ne, nf);
+
+        // get properties
+        auto vconn = vertex_property<SurfaceMesh::VertexConnectivity>("v:connectivity");
+        auto hconn = halfedge_property<SurfaceMesh::HalfedgeConnectivity>("h:connectivity");
+        auto fconn = face_property<SurfaceMesh::FaceConnectivity>("f:connectivity");
+        auto point = vertex_property<vec3>("v:point");
+
+        // read properties from file
+        input.read((char*)vconn.data(), nv * sizeof(SurfaceMesh::VertexConnectivity)  );
+        input.read((char*)hconn.data(), nh * sizeof(SurfaceMesh::HalfedgeConnectivity));
+        input.read((char*)fconn.data(), nf * sizeof(SurfaceMesh::FaceConnectivity)    );
+        input.read((char*)point.data(), nv * sizeof(vec3)                             );
+
+        bool has_colors = false;
+        input.read((char*)&has_colors, sizeof(bool));
+        if (has_colors) {
+            SurfaceMesh::VertexProperty<vec3> color = vertex_property<vec3>("v:color");
+            input.read((char*)color.data(), nv * sizeof(vec3));
+        }
+
+        return n_faces() > 0;
+    }
+
+
+    //-----------------------------------------------------------------------------
+
+
+    bool SurfaceMesh::write_poly(const std::string& file_name) const
+    {
+        if (n_faces() == 0) {
+            LOG(ERROR) << "empty mesh";
+            return false;
+        }
+
+        // open file (in binary mode)
+        std::ofstream output(file_name.c_str(), std::fstream::binary);
+        if (output.fail()) {
+            LOG(ERROR) << "could not open file: " << file_name;
+            return false;
+        }
+
+        // how many elements?
+        unsigned int nv, ne, nh, nf;
+        nv = n_vertices();
+        ne = n_edges();
+        nh = n_halfedges();
+        nf = n_faces();
+        nh = 2*ne;
+
+        output.write((char*)&nv, sizeof(unsigned int));
+        output.write((char*)&ne, sizeof(unsigned int));
+        output.write((char*)&nf, sizeof(unsigned int));
+
+        // get properties
+        auto vconn = get_vertex_property<SurfaceMesh::VertexConnectivity>("v:connectivity");
+        auto hconn = get_halfedge_property<SurfaceMesh::HalfedgeConnectivity>("h:connectivity");
+        auto fconn = get_face_property<SurfaceMesh::FaceConnectivity>("f:connectivity");
+        auto point = get_vertex_property<vec3>("v:point");
+
+        // write properties to file
[…]
