### 2020-12-16 9078980
default lighting for different drawables
