+++ easy3d/renderer/setting.h
@@ -41,9 +41,18 @@ namespace easy3d {
 
         // lighting
         extern vec4 light_position;  // light position defined in camera coordinate system
-        extern bool light_distinct_back_color;
-        extern vec4 back_color;
-        extern bool light_two_sides;
+        // drawable lighting
+        extern bool triangles_drawable_two_side_lighting;
+        extern bool points_drawable_two_side_lighting;
+        extern bool lines_drawable_two_side_lighting;
+        // drawable distinct backside color
+        extern bool triangles_drawable_distinct_backside_color;
+        extern bool points_drawable_distinct_backside_color;
+        extern bool lines_drawable_distinct_backside_color;
+        // drawable backside color
+        extern vec4 triangles_drawable_backside_color;
+        extern vec4 points_drawable_backside_color;
+        extern vec4 lines_drawable_backside_color;
 
         // material
         extern vec4 material_ambient;
@@ -64,11 +73,11 @@ namespace easy3d {
         extern float shadow_darkness;
 
         // point cloud
-        extern bool point_cloud_use_color_property;
-        extern bool point_cloud_show_points;
         extern vec4 point_cloud_points_color;
         extern float point_cloud_point_size;
         extern bool point_cloud_impostors;
+        extern bool point_cloud_two_side_lighting;
+        extern vec4 point_cloud_back_side_color;
 
         // surface mesh - surface
         extern bool surface_mesh_phong_shading;
@@ -76,19 +85,16 @@ namespace easy3d {
         extern vec4 surface_mesh_faces_color;
         extern bool surface_mesh_use_color_property;
         extern float surface_mesh_opacity;
-
         // surface mesh - vertices
         extern bool surface_mesh_show_vertices;
         extern vec4 surface_mesh_vertices_color;
         extern bool surface_mesh_vertices_imposters;
         extern float surface_mesh_vertices_point_size;
-
         // surface mesh - edges
         extern bool surface_mesh_show_edges;
         extern vec4 surface_mesh_edges_color;
         extern bool surface_mesh_edges_imposters;
         extern float surface_mesh_edges_line_width;
-
         // surface mesh - borders
         extern bool surface_mesh_show_borders;
         extern vec4 surface_mesh_borders_color;
