+++ 3rd_party/tetgen/tetgen.h
@@ -1,19 +1,49 @@
-///////////////////////////////////////////////////////////////////////////////
-//                                                                           //
-// TetGen                                                                    //
-//                                                                           //
-// A Quality Tetrahedral Mesh Generator and A 3D Delaunay Triangulator       //
-//                                                                           //
-// Version 1.5                                                               //
-// August 18, 2018                                                           //
-//                                                                           //
-// Copyright (C) 2002--2018                                                  //
-//                                                                           //
-// TetGen is freely available through the website: http://www.tetgen.org.    //
-//   It may be copied, modified, and redistributed for non-commercial use.   //
-//   Please consult the file LICENSE for the detailed copyright notices.     //
-//                                                                           //
-///////////////////////////////////////////////////////////////////////////////
+//============================================================================//
+//                                                                            //
+// TetGen                                                                     //
+//                                                                            //
+// A Quality Tetrahedral Mesh Generator and A 3D Delaunay Triangulator        //
+//                                                                            //
+// Version 1.6.0                                                              //
+// August 31, 2020                                                            //
+//                                                                            //
+// Copyright (C) 2002--2020                                                   //
+//                                                                            //
+// Hang Si                                                                    //
+// Research Group: Numerical Mathematics and Scientific Computing             //
+// Weierstrass Institute for Applied Analysis and Stochastics (WIAS)          //
+// Mohrenstr. 39, 10117 Berlin, Germany                                       //
+// si@wias-berlin.de                                                          //
+//                                                                            //
+// TetGen is a tetrahedral mesh generator. It creates 3d triangulations of    //
+//   polyhedral domains. It generates meshes with well-shaped elements whose  //
+//   sizes are adapted to the geometric features or user-provided sizing      //
+//   functions. It has applications in various applications in scientific     //
+//   computing, such as computer graphics (CG), computer-aided design (CAD),  //
+//   geometry processing (parametrizations and computer animation), and       //
+//   physical simulations (finite element analysis).                          //
+//                                                                            //
+// TetGen computes (weighted) Delaunay triangulations for three-dimensional   //
+//   (weighted) point sets, and constrained Delaunay triangulations for       //
+//   three-dimensional polyhedral domains. In the latter case, input edges    //
+//   and triangles can be completely preserved in the output meshes.  TetGen  //
+//   can refine or coarsen an existing mesh to result in good quality and     //
+//   size-adapted mesh according to the geometric features and user-defined   //
+//   mesh sizing functions.                                                   //
+//                                                                            //
+// TetGen implements theoretically proven algorithms for computing the        //
+//   Delaunay and constrained Delaunay tetrahedralizations. TetGen achieves   //
+//   robustness and efficiency by using advanced techniques in computational  //
+//   geometry. A technical paper describes the algorithms and methods         //
+//   implemented in TetGen is available in ACM-TOMS, Hang Si ``TetGen, a      //
+//   Delaunay-Based Quality Tetrahedral Mesh Generator", ACM Transactions on  //
+//   Mathematical Software, February 2015, https://doi.org/10.1145/2629697.   //
+//                                                                            //
+// TetGen is freely available through the website: http://www.tetgen.org.     //
[…]
