### 2020-12-17 68f284c
log the num of tetrahedra
