### 2020-12-17 58a97d5
add code (commented out) for saving the remeshing result into off file directly (with 20 digits)
