### 2020-12-19 c3befe4
unified name for hash functions
