### 2020-12-19 18f65ca
remove a duplicated function call
