+++ easy3d/renderer/state.cpp
@@ -41,7 +41,7 @@ namespace easy3d {
             : visible_(true), coloring_method_(UNIFORM_COLOR), color_(0.8f, 0.8f, 0.8f, 1.0f),
               property_location_(VERTEX), property_name_("uniform color"), lighting_(true),
               lighting_two_sides_(false), distinct_back_color_(true), back_color_(vec4(1, 0, 0, 1)),
-              texture_(nullptr), texture_repeat_(1.0f), texture_fractional_repeat_(0.0f),
+              texture_(nullptr), texture_repeat_(1.0f), texture_fractional_repeat_(0.0f), ssao_enabled_(false),
               clamp_range_(true), clamp_lower_(0.05f), clamp_upper_(0.05f),
               highlight_(false), highlight_range_(-1, -1) {
     }
@@ -52,6 +52,7 @@ namespace easy3d {
                                    distinct_back_color_(s.distinct_back_color()), back_color_(s.back_color()),
                                    texture_(s.texture()), texture_repeat_(s.texture_repeat()),
                                    texture_fractional_repeat_(s.texture_fractional_repeat()),
+                                   ssao_enabled_(s.is_ssao_enabled()),
                                    clamp_range_(s.clamp_range()), clamp_lower_(s.clamp_lower()),
                                    clamp_upper_(s.clamp_upper()), material_(s.material()),
                                    highlight_(s.highlight()), highlight_range_(s.highlight_range()) {
@@ -73,6 +74,7 @@ namespace easy3d {
         texture_ = s.texture();
         texture_repeat_ = s.texture_repeat();
         texture_fractional_repeat_ = s.texture_fractional_repeat();
+        ssao_enabled_ = s.is_ssao_enabled();
         clamp_range_ = s.clamp_range();
         clamp_lower_ = s.clamp_lower();
         clamp_upper_ = s.clamp_upper();
