+++ applications/Mapple/paint_canvas.cpp
@@ -88,11 +88,8 @@ PaintCanvas::PaintCanvas(MainWindow* window)
         , model_idx_(-1)
         , ssao_(nullptr)
         , transparency_(nullptr)
-        , transparency_enabled_(false)
         , shadow_(nullptr)
-        , shadow_enabled_(false)
         , edl_(nullptr)
-        , edl_enabled_(false)
 {
     // like Qt::StrongFocus plus the widget accepts focus by using the mouse wheel.
     setFocusPolicy(Qt::StrongFocus);
@@ -1139,55 +1136,73 @@ void PaintCanvas::postDraw() {
 }
 
 
-AmbientOcclusion *PaintCanvas::ssao() {
-    if (!ssao_)
-        ssao_ = new AmbientOcclusion(camera_);
-    return ssao_;
-}
+void PaintCanvas::enableSsao(bool b) {
+    if (b) {
+        if (!ssao_)
+            ssao_ = new AmbientOcclusion(camera_);
+    }
+    else {
+        delete ssao_;
+        ssao_ = nullptr;
+    }
 
-
-Shadow *PaintCanvas::shadow() {
-    if (!shadow_)
-        shadow_ = new SoftShadow(camera_);
-    return shadow_;
-}
-
-
-Transparency *PaintCanvas::transparency() {
-    if (!transparency_)
-        transparency_ = new DualDepthPeeling(camera_);
-    return transparency_;
-}
-
-
-easy3d::EyeDomeLighting *PaintCanvas::edl() {
-    if (!edl_)
-        edl_ = new EyeDomeLighting(camera_);
-    return edl_;
+    for (const auto m : models_) {
+        for (auto d : m->renderer()->lines_drawables()) {
+            d->enable_ssao(b);
+        }
+        for (auto d : m->renderer()->points_drawables()) {
+            d->enable_ssao(b);
+        }
+        for (auto d : m->renderer()->triangles_drawables()) {
+            d->enable_ssao(b);
+        }
+    }
 }
 
 
 void PaintCanvas::enableShadow(bool b) {
-    shadow_enabled_ = b;
-    // shadow and transparency cannot co-exist
-    if (shadow_enabled_ && transparency_enabled_)
-        transparency_enabled_ = false;
+    if (b) {
+        if (!shadow_)
+            shadow_ = new SoftShadow(camera_);
+
+        if (transparency()) {
+            delete transparency_;
+            transparency_ = nullptr;
+        }
+    }
+    else {
+        delete shadow_;
+        shadow_ = nullptr;
+    }
 }
 
 
 void PaintCanvas::enableTransparency(bool b) {
-    //    transparency_enabled_ = b;
-    //    // ssao and transparency cannot co-exist
-    //    if (transparency_enabled_ && ssao())
-    //        ssao()->set_algorithm(AmbientOcclusion_HBAO::SSAO_NONE);
-    //    // shadow and transparency cannot co-exist
-    //    if (transparency_enabled_ && shadow_enabled_)
-    //        shadow_enabled_ = false;
+    if (b) {
+        if (!transparency_)
+            transparency_ = new DualDepthPeeling(camera_);
+
+        if (shadow()) {
+            delete shadow_;
+            shadow_ = nullptr;
+        }
+    }
+    else {
+        delete transparency_;
+        transparency_ = nullptr;
+    }
 }
 
 
 void PaintCanvas::enableEyeDomeLighting(bool b) {
-    edl_enabled_ = b;
+    if (b) {
+        if (!edl_)
+            edl_ = new EyeDomeLighting(camera_);
+    }
+    else {
+        delete edl_;
+        edl_ = nullptr;
+    }
 }
 
 
[…]

[…]
