### 2020-12-20 f9849cf
add lexicographic comparison
