+++ applications/Mapple/main_window.cpp
@@ -40,6 +40,7 @@
 #include <easy3d/core/surface_mesh.h>
 #include <easy3d/core/graph.h>
 #include <easy3d/core/point_cloud.h>
+#include <easy3d/core/tetra_mesh.h>
 #include <easy3d/core/random.h>
 #include <easy3d/renderer/setting.h>
 #include <easy3d/renderer/camera.h>
@@ -49,6 +50,7 @@
 #include <easy3d/fileio/point_cloud_io.h>
 #include <easy3d/fileio/graph_io.h>
 #include <easy3d/fileio/surface_mesh_io.h>
+#include <easy3d/fileio/tetra_mesh_io.h>
 #include <easy3d/fileio/ply_reader_writer.h>
 #include <easy3d/fileio/point_cloud_io_ptx.h>
 #include <easy3d/fileio/resources.h>
@@ -326,9 +328,10 @@ bool MainWindow::onOpen() {
                 this,
                 "Open file(s)",
                 curDataDirectory_,
-                "Supported formats (*.ply *.obj *.off *.stl *.poly *.trilist *.bin *.las *.laz *.xyz *.bxyz *.vg *.bvg *.ptx)\n"
-                "Mesh formats (*.ply *.obj *.off *.stl *.poly *.trilist)\n"
-                "Point set formats (*.ply *.bin *.ptx *.las *.laz *.xyz *.bxyz *.vg *.bvg *.ptx)\n"
+                "Supported formats (*.ply *.obj *.off *.stl *.poly *.trilist *.bin *.las *.laz *.xyz *.bxyz *.vg *.bvg *.ptx *.tet)\n"
+                "Surface Mesh (*.ply *.obj *.off *.stl *.poly *.trilist)\n"
+                "Point Cloud (*.ply *.bin *.ptx *.las *.laz *.xyz *.bxyz *.vg *.bvg *.ptx)\n"
+                "Tetrahedral Mesh (*.tet)\n"
                 "All formats (*.*)"
             );
 
@@ -369,9 +372,10 @@ bool MainWindow::onSave() {
                 this,
                 "Open file(s)",
                 QString::fromStdString(default_file_name),
-                "Supported formats (*.ply *.obj *.off *.stl *.poly *.bin *.las *.laz *.xyz *.bxyz *.vg *.bvg)\n"
-                "Mesh formats (*.ply *.obj *.off *.stl *.poly)\n"
-                "Point set formats (*.ply *.bin *.ptx *.las *.laz *.xyz *.bxyz *.vg *.bvg)\n"
+                "Supported formats (*.ply *.obj *.off *.stl *.poly *.bin *.las *.laz *.xyz *.bxyz *.vg *.bvg *.tet)\n"
+                "Surface Mesh (*.ply *.obj *.off *.stl *.poly)\n"
+                "Point Cloud (*.ply *.bin *.ptx *.las *.laz *.xyz *.bxyz *.vg *.bvg)\n"
+                "Tetrahedral Mesh (*.tet)\n"
                 "All formats (*.*)"
     );
 
@@ -390,6 +394,10 @@ bool MainWindow::onSave() {
     else if (dynamic_cast<const Graph*>(model)) {
         const Graph* graph = dynamic_cast<const Graph*>(model);
         saved = GraphIO::save(fileName.toStdString(), graph);
+    }
+    else if (dynamic_cast<const TetraMesh*>(model)) {
+        const TetraMesh* mesh = dynamic_cast<const TetraMesh*>(model);
+        saved = TetraMeshIO::save(fileName.toStdString(), mesh);
     }
 
     if (saved) {
@@ -422,6 +430,8 @@ Model* MainWindow::open(const std::string& file_name) {
     }
[…]
