### 2020-12-20 480b221
default empty property_stats()
