+++ easy3d/core/tetra_mesh.cpp
@@ -2,231 +2,110 @@
 
 #include <set>
 #include <map>
+#include <queue>
 #include <fstream>
-#include <cstdio>
 
 
 namespace easy3d {
 
-    TetraMesh::Tet::Tet(const vec3 &v0, const vec3 &v1, const vec3 &v2, const vec3 &v3) {
-        v[0] = v0;
-        v[1] = v1;
-        v[2] = v2;
-        v[3] = v3;
-    }
+
+#if 1
 
 
-    TetraMesh::Tet TetraMesh::getTet(int i) const {
-        const ivec4 &tet = t[i];
-        return Tet(v[tet[0]],
-                   v[tet[1]],
-                   v[tet[2]],
-                   v[tet[3]]);
-    }
+    void TetraMesh::sortIndexes(PositionIndex *indexes, int numberOfIndexes, Tetra const &tetra) {
 
-
-    void TetraMesh::getAdjacentVertices(int vertex, std::vector<int> &vertices) const {
-        if (dirty || incidenceMap.size() == 0) {
-           const_cast<TetraMesh*>(this)->buildIncidenceMap();
-        }
-
-        assert(incidenceMap.size() == v.size());
-        assert(vertex < static_cast<int>(incidenceMap.size()));
-
-        vertices.clear();
-
-        const std::vector<int> &incidentTets = incidenceMap[vertex];
-        for (size_t i = 0; i < incidentTets.size(); ++i) {
-            for (int u = 0; u < 4; ++u) {
-                // Add the current vertex if it hasn't been added already.
-                int vIndex = t[incidentTets[i]][u];
-                bool alreadyAdded = false;
-                for (size_t j = 0; j < vertices.size(); ++j) {
-                    if (vertices[j] == vIndex) {
-                        alreadyAdded = true;
-                        break;
-                    }
-                }
-                if (!alreadyAdded && vIndex != vertex) {
-                    vertices.push_back(vIndex);
-                }
+        for (int i = 0; i < numberOfIndexes; i++) {
+            int minimum = i;
+            for (int j = i + 1; j < numberOfIndexes; j++) {
+                if (tetra.TV(indexes[j]) < tetra.TV(indexes[minimum])) minimum = j;
             }
+
+            PositionIndex tmp = indexes[i];
+            indexes[i] = indexes[minimum];
+            indexes[minimum] = tmp;
         }
+
     }
 
-
-    void TetraMesh::getIncidentTets(int vertex, std::vector<Tet> &tets) const {
-        if (dirty || incidenceMap.size() == 0) {
-            const_cast<TetraMesh*>(this)->buildIncidenceMap();
-        }
-
-        assert(incidenceMap.size() == v.size());
-        assert(vertex < static_cast<int>(incidenceMap.size()));
-
-        tets.clear();
-
-        const std::vector<int> &incidentTets = incidenceMap[vertex];
-        for (size_t i = 0; i < incidentTets.size(); ++i) {
-            tets.push_back(getTet(incidentTets[i]));
-        }
-    }
+    void TetraMesh::build_adjacency() {
+        unsigned int VTstarSet = 0;
+        unsigned int TTandInverseTTSet = 0;
+        std::vector<std::pair<Triangle, TetraIndex> *> faces;
+        faces.reserve(this->n_tetrahedra() * 4);
 
 
-    void TetraMesh::getIncidentTetIndices(int vertex, std::vector<int> &tetIndices) const {
-        if (dirty || incidenceMap.size() == 0) {
-            const_cast<TetraMesh*>(this)->buildIncidenceMap();
-        }
-
-        assert(incidenceMap.size() == v.size());
[…]

[…]
