+++ easy3d/core/tetra_mesh.h
@@ -9,509 +9,11 @@
 
 namespace easy3d {
 
-#if 1
-
-    typedef int VertexIndex;
-    typedef int TetraIndex;
-
-    typedef unsigned short int PositionIndex;
-
-    static TetraIndex reservedTetraIndexes = 2;
-    static TetraIndex maxTetraNumber = UINTPTR_MAX - reservedTetraIndexes;
-
-    static TetraIndex noAdjacentTetraIndex = maxTetraNumber + 1;
-    static TetraIndex uninitializedTetraIndex = maxTetraNumber + 2;
-
-
-
-    /**
-     * \brief Data structure representing a tetrahedral mesh.
-     * \class TetraMesh easy3d/core/tetra_mesh.h
-     */
-    class TetraMesh : public Model {
-    public:
-
-        // Class storing geometrical information of vertices
-        class Vertex {
-        private:
-            float vertexCoordinates[3]; // Coordinates
-            VertexIndex index;
-            TetraIndex vertexVTstar; // VT* relationship
-
-        public:
-
-            // Constructor
-            inline Vertex() {
-                vertexCoordinates[0] = 0;
-                vertexCoordinates[1] = 0;
-                vertexCoordinates[2] = 0;
-                setVTstar(uninitializedTetraIndex);
-            }
-
-            inline Vertex(float x, float y, float z) {
-                vertexCoordinates[0] = x;
-                vertexCoordinates[1] = y;
-                vertexCoordinates[2] = z;
-                setVTstar(uninitializedTetraIndex);
-            }
-
-            // VT* setter
-            inline void setVTstar(TetraIndex VTstar) {
-                vertexVTstar = VTstar;
-            }
-
-            // First float getter
-            inline float x() const {
-                return vertexCoordinates[0];
-            }
-
-            // Second float getter
-            inline float y() const {
-                return vertexCoordinates[1];
-            }
-
-            // Third float getter
-            inline float z() const {
-                return vertexCoordinates[2];
-            }
-
-            // VT* getter
-            inline TetraIndex VTstar() const {
-                return vertexVTstar;
-            }
-
-
-            // Operators -----------
-            inline bool operator==(Vertex vertex) {
-                return x() == vertex.x() && y() == vertex.y() && z() == vertex.z();
-            }
-        };
-
-
-        // Explicit representation of an edge
-        class Edge {
-        private:
-            VertexIndex vertexes[2]; // Indices of the endpoints
-        public:
-
-            // Constructor
-            inline Edge() {}
-
-            inline Edge(VertexIndex vi1, VertexIndex vi2) {
-                vertexes[0] = std::min(vi1, vi2);
-                vertexes[1] = std::max(vi1, vi2);
-            }
-
-            // Smaller-index endpoint getter
-            inline VertexIndex minindex() const { return vertexes[0]; }
-
-            // Bigger-index endpoint getter
-            inline VertexIndex maxindex() const { return vertexes[1]; }
-
-            // Index of the vertex in position vPos
-            inline VertexIndex EV(PositionIndex vPos) {
-                return vertexes[vPos];
-            }
-
-            // Index of verex v inside the edge
-            inline short vertexIndex(VertexIndex v) {
-                if (vertexes[0] == v) return 0;
-                if (vertexes[1] == v) return 1;
-                return -1;
-            }
-
-            // Operators ----------
-            inline bool operator==(Edge edge) {
-                if (minindex() != edge.minindex()) return false;
-                if (maxindex() != edge.maxindex()) return false;
-                return true;
-            }
-
-            inline bool operator!=(Edge edge) {
-                return !((*this) == (edge));
-            }
-
[…]
