+++ easy3d/core/tetra_mesh.cpp
@@ -2,110 +2,231 @@
 
 #include <set>
 #include <map>
-#include <queue>
 #include <fstream>
+#include <cstdio>
 
 
 namespace easy3d {
 
-
-#if 1
-
-
-    void TetraMesh::sortIndexes(PositionIndex *indexes, int numberOfIndexes, Tetra const &tetra) {
-
-        for (int i = 0; i < numberOfIndexes; i++) {
-            int minimum = i;
-            for (int j = i + 1; j < numberOfIndexes; j++) {
-                if (tetra.TV(indexes[j]) < tetra.TV(indexes[minimum])) minimum = j;
-            }
-
-            PositionIndex tmp = indexes[i];
-            indexes[i] = indexes[minimum];
-            indexes[minimum] = tmp;
-        }
-
+    TetraMesh::Tet::Tet(const vec3 &v0, const vec3 &v1, const vec3 &v2, const vec3 &v3) {
+        v[0] = v0;
+        v[1] = v1;
+        v[2] = v2;
+        v[3] = v3;
     }
 
-    void TetraMesh::build_adjacency() {
-        unsigned int VTstarSet = 0;
-        unsigned int TTandInverseTTSet = 0;
-        std::vector<std::pair<Triangle, TetraIndex> *> faces;
-        faces.reserve(this->n_tetrahedra() * 4);
+
+    TetraMesh::Tet TetraMesh::getTet(int i) const {
+        const ivec4 &tet = t[i];
+        return Tet(v[tet[0]],
+                   v[tet[1]],
+                   v[tet[2]],
+                   v[tet[3]]);
+    }
 
 
-        Triangle *tri;
-        Tetra tetra;
+    void TetraMesh::getAdjacentVertices(int vertex, std::vector<int> &vertices) const {
+        if (dirty || incidenceMap.size() == 0) {
+           const_cast<TetraMesh*>(this)->buildIncidenceMap();
+        }
 
+        assert(incidenceMap.size() == v.size());
+        assert(vertex < static_cast<int>(incidenceMap.size()));
 
-        for (TetraIndex index = 0; index < n_tetrahedra(); index = index + 1) {
-            tetra = getTetra(index);
+        vertices.clear();
 
-            for (PositionIndex vp = 0; vp <= 3; vp++) {
-
-                Vertex &vertex = getVertex(tetra.TV(vp));
-
-                if (vertex.VTstar() == uninitializedTetraIndex) {
-                    VTstarSet++;
-                    vertex.setVTstar(index);
+        const std::vector<int> &incidentTets = incidenceMap[vertex];
+        for (size_t i = 0; i < incidentTets.size(); ++i) {
+            for (int u = 0; u < 4; ++u) {
+                // Add the current vertex if it hasn't been added already.
+                int vIndex = t[incidentTets[i]][u];
+                bool alreadyAdded = false;
+                for (size_t j = 0; j < vertices.size(); ++j) {
+                    if (vertices[j] == vIndex) {
+                        alreadyAdded = true;
+                        break;
+                    }
                 }
-
-                tri = tetra.TF(vp);
-                faces.push_back(new std::pair<Triangle, TetraIndex>(*(tri), index));
-                delete tri;
-
+                if (!alreadyAdded && vIndex != vertex) {
+                    vertices.push_back(vIndex);
+                }
             }
         }
-
-
-        sortFaces(&faces);
-
-        Triangle firstFace;
-        Triangle secondFace;
-        TetraIndex firstIndex, secondIndex;
-
-
-        for (unsigned int i = 0; i < faces.size() - 1; i++) {
-            firstFace = faces[i]->first;
-            secondFace = faces[i + 1]->first;
-
-            if ((firstFace) == (secondFace)) {
-                firstIndex = faces[i]->second;
[…]

[…]
