+++ easy3d/core/tetra_mesh.cpp
@@ -1,272 +1,388 @@
+/**
+ * Copyright (C) 2015 by Liangliang Nan (liangliang.nan@gmail.com)
+ * https://3d.bk.tudelft.nl/liangliang/
+ *
+ * This file is part of Easy3D. If it is useful in your research/work,
+ * I would be grateful if you show your appreciation by citing it:
+ * ------------------------------------------------------------------
+ *      Liangliang Nan.
+ *      Easy3D: a lightweight, easy-to-use, and efficient C++
+ *      library for processing and rendering 3D data. 2018.
+ * ------------------------------------------------------------------
+ * Easy3D is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License Version 3
+ * as published by the Free Software Foundation.
+ *
+ * Easy3D is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+
 #include <easy3d/core/tetra_mesh.h>
+#include <easy3d/util/logging.h>
 
-#include <set>
-#include <map>
+#include <cmath>
 #include <fstream>
-#include <cstdio>
-
 
 namespace easy3d {
 
-    TetraMesh::Tet::Tet(const vec3 &v0, const vec3 &v1, const vec3 &v2, const vec3 &v3) {
-        v[0] = v0;
-        v[1] = v1;
-        v[2] = v2;
-        v[3] = v3;
+    TetraMesh::TetraMesh()
+    {
+        // allocate standard properties
+        // same list is used in operator=() and assign()
+        vconn_    = add_vertex_property<VertexConnectivity>("v:connectivity");
+        econn_    = add_edge_property<EdgeConnectivity>("e:connectivity");
+        fconn_    = add_face_property<FaceConnectivity>("f:connectivity");
+        tconn_    = add_tetra_property<TetraConnectivity>("t:connectivity");
+
+        vpoint_   = add_vertex_property<vec3>("v:point");
+        tindices_ = add_tetra_property<ivec4>("t:indices");
+
+        mprops_.push_back();
     }
 
 
-    TetraMesh::Tet TetraMesh::getTet(int i) const {
-        const ivec4 &tet = t[i];
-        return Tet(v[tet[0]],
-                   v[tet[1]],
-                   v[tet[2]],
-                   v[tet[3]]);
+    //-----------------------------------------------------------------------------
+
+
+    TetraMesh::~TetraMesh()
+    {
     }
 
 
-    void TetraMesh::getAdjacentVertices(int vertex, std::vector<int> &vertices) const {
-        if (dirty || incidenceMap.size() == 0) {
-           const_cast<TetraMesh*>(this)->buildIncidenceMap();
+    //-----------------------------------------------------------------------------
+
+
+    TetraMesh& TetraMesh::operator=(const TetraMesh& rhs)
+    {
+        if (this != &rhs)
+        {
+            // deep copy of property containers
+            vprops_ = rhs.vprops_;
+            eprops_ = rhs.eprops_;
+            fprops_ = rhs.fprops_;
+            tprops_ = rhs.tprops_;
+            mprops_ = rhs.mprops_;
+
+            // property handles contain pointers, have to be reassigned
+            vconn_    = add_vertex_property<VertexConnectivity>("v:connectivity");
+            econn_    = add_edge_property<EdgeConnectivity>("e:connectivity");
[…]

[…]
