+++ easy3d/core/graph.h
@@ -122,8 +122,7 @@ namespace easy3d {
 		/// \sa VertexConnectivity
 		struct EdgeConnectivity
 		{
-			Vertex  source_;    /// vertex the edge points to
-			Vertex  target_;    /// vertex the edge originates from
+            std::vector<Vertex>  vertices_;
 		};
 
 
@@ -610,14 +609,7 @@ namespace easy3d {
 		Vertex vertex(Edge e, unsigned int i) const
 		{
 			assert(i<=1);
-			if (i == 0)
-				return econn_[e].source_;
-			else if (i == 1)
-				return econn_[e].target_;
-			else {
-				std::cerr << "index out of range" << std::endl;
-				return Vertex();
-			}
+			return econn_[e].vertices_[i];
 		}
 
 		//@}
