### 2020-12-21 e4f2d32
add ostream support
