### 2020-12-21 7a99ea7
simpler code
