### 2020-12-21 3c90b04
simpler code
