### 2020-12-21 14a431e
polytope mesh... (ongoing)
