+++ easy3d/core/poly_mesh.cpp
@@ -37,8 +37,8 @@ namespace easy3d {
         // same list is used in operator=() and assign()
         vconn_    = add_vertex_property<VertexConnectivity>("v:connectivity");
         econn_    = add_edge_property<EdgeConnectivity>("e:connectivity");
-        fconn_    = add_face_property<FaceConnectivity>("f:connectivity");
-        tconn_    = add_tetra_property<TetraConnectivity>("t:connectivity");
+        hconn_    = add_halfface_property<HalfFaceConnectivity>("h:connectivity");
+        cconn_    = add_tetra_property<CellConnectivity>("c:connectivity");
 
         vpoint_   = add_vertex_property<vec3>("v:point");
         tindices_ = add_tetra_property<ivec4>("t:indices");
@@ -65,21 +65,22 @@ namespace easy3d {
             // deep copy of property containers
             vprops_ = rhs.vprops_;
             eprops_ = rhs.eprops_;
+            hprops_ = rhs.hprops_;
             fprops_ = rhs.fprops_;
-            tprops_ = rhs.tprops_;
+            cprops_ = rhs.cprops_;
             mprops_ = rhs.mprops_;
 
             // property handles contain pointers, have to be reassigned
             vconn_    = add_vertex_property<VertexConnectivity>("v:connectivity");
             econn_    = add_edge_property<EdgeConnectivity>("e:connectivity");
-            fconn_    = add_face_property<FaceConnectivity>("f:connectivity");
-            tconn_    = add_tetra_property<TetraConnectivity>("t:connectivity");
+            hconn_    = add_halfface_property<HalfFaceConnectivity>("h:connectivity");
+            cconn_    = add_tetra_property<CellConnectivity>("c:connectivity");
 
             vpoint_   = vertex_property<vec3>("v:point");
-            tindices_ = tetra_property<ivec4>("t:indices");
+            tindices_ = cell_property<ivec4>("t:indices");
 
             // normals might be there, therefore use get_property
-            fnormal_  = get_face_property<vec3>("f:normal");
+            fnormal_  = get_halfface_property<vec3>("f:normal");
         }
 
         return *this;
@@ -96,31 +97,32 @@ namespace easy3d {
             // clear properties
             vprops_.clear();
             eprops_.clear();
+            hprops_.clear();
             fprops_.clear();
             mprops_.clear();
-            tprops_.clear();
+            cprops_.clear();
 
             // allocate standard properties
             vconn_    = add_vertex_property<VertexConnectivity>("v:connectivity");
             econn_    = add_edge_property<EdgeConnectivity>("e:connectivity");
-            fconn_    = add_face_property<FaceConnectivity>("f:connectivity");
-            tconn_    = add_tetra_property<TetraConnectivity>("t:connectivity");
+            hconn_    = add_halfface_property<HalfFaceConnectivity>("h:connectivity");
+            cconn_    = add_tetra_property<CellConnectivity>("c:connectivity");
             
             vpoint_   = add_vertex_property<vec3>("v:point");
             tindices_ = add_tetra_property<ivec4>("t:indices");
             
             // normals might be there, therefore use get_property
-            fnormal_  = get_face_property<vec3>("f:normal");
+            fnormal_  = get_halfface_property<vec3>("f:normal");
 
             // copy properties from other mesh
             vconn_.array()     = rhs.vconn_.array();
-            tconn_.array()     = rhs.tconn_.array();
[…]
