+++ tutorials/examples.h
@@ -17,6 +17,8 @@
 /// @example Tutorial_109_Graph_Connectivity        \include Tutorial_109_Graph_Connectivity/main.cpp
 /// @example Tutorial_110_Graph_Property            \include Tutorial_110_Graph_Property/main.cpp
 /// @example Tutorial_111_Graph_IO                  \include Tutorial_111_Graph_IO/main.cpp
+/// @example Tutorial_112_PolyMesh                  \include Tutorial_112_PolyMesh/main.cpp
+/// @example Tutorial_113_PolyMesh_Connectivity     \include Tutorial_113_PolyMesh_Connectivity/main.cpp
 
 /**
  * \example Tutorial_201_Viewer_imgui
