+++ easy3d/core/poly_mesh.h
@@ -22,8 +22,8 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef EASY3D_CORE_POLY_MESH_H
-#define EASY3D_CORE_POLY_MESH_H
+#ifndef EASY3D_CORE_POLYHEDRAL_MESH_H
+#define EASY3D_CORE_POLYHEDRAL_MESH_H
 
 #include <easy3d/core/model.h>
 #include <easy3d/core/types.h>
@@ -32,7 +32,7 @@
 namespace easy3d {
 
     /**
-     * \brief Data structure representing a polytope mesh.
+     * \brief Data structure representing a polyhedral mesh.
      * \class PolyMesh easy3d/core/poly_mesh.h
      */
 
@@ -123,7 +123,7 @@ namespace easy3d {
             std::ostream& operator<<(std::ostream& os) const { return os << 'f' << idx(); }
         };
 
-        /// this type represents a polytope cell (internally it is basically an index)
+        /// this type represents a polyhedral cell (internally it is basically an index)
         /// \sa Vertex, Edge, HalfFace
         struct Cell : public BaseHandle
         {
@@ -151,7 +151,7 @@ namespace easy3d {
         /// \sa VertexConnectivity, HalfFaceConnectivity, CellConnectivity
         struct EdgeConnectivity
         {
-            std::vector<Vertex>     vertices_;
+            std::vector<Vertex>  vertices_;
             std::set<HalfFace>   halffaces_;
             std::set<Cell>       cells_;
         };
@@ -679,7 +679,7 @@ namespace easy3d {
         bool write_tet(const std::string& filename) const;
         //@}
 
-    public: //----------------------------------------------- add new vertex / face
+    public: //----------------------------------------------- add new vertex / face / cell
 
         /// \name Add new elements by hand
         //@{
@@ -1138,6 +1138,12 @@ namespace easy3d {
         {
             return CellContainer(cells_begin(), cells_end());
         }
+        //@}
+
+    public: //--------------------------------------------- adjacency access
+
+        /// \name Adjacency access
+        //@{
 
         /// returns circulator for vertices around vertex \c v
         const std::set<Vertex>& vertices(Vertex v) const
@@ -1150,6 +1156,11 @@ namespace easy3d {
         {
             assert(i<=1);
             return HalfFace((f.idx() << 1) + i);
+        }
+
+        HalfFace opposite(HalfFace f) const
+        {
+            return hconn_[f].opposite_;
         }
 
         /// returns the \c i'th vertex of edge \c e. \c i has to be 0 or 1.
@@ -1190,9 +1201,9 @@ namespace easy3d {
         }
 
         /// returns circulator for vertices around vertex \c v
-        const std::set<Edge>& edges(Cell t) const
+        const std::set<Edge>& edges(Cell c) const
         {
-            return cconn_[t].edges_;
+            return cconn_[c].edges_;
         }
         
         /// returns circulator for faces around vertex \c v
@@ -1230,16 +1241,35 @@ namespace easy3d {
         {
             return hconn_[f].cell_;
         }
+        //@}
 
-        HalfFace opposite(HalfFace f) const
+    public: //--------------------------------------------- higher-level operations
+
+        /// \name Higher-level Topological Operations
+        //@{
+
+        /// returns whether the mesh a tetrahedral mesh.
+        bool is_tetraheral_mesh() const;
+
+        /// returns whether \c f is a boundary face, i.e., it is incident to only one cell.
+        bool is_border(Face f) const
         {
-            return hconn_[f].opposite_;
[…]

[…]
