+++ tutorials/Tutorial_112_PolyMesh/main.cpp
@@ -53,59 +53,59 @@ int main(int argc, char **argv) {
     vertices.push_back(v0);
     vertices.push_back(v1);
     vertices.push_back(v4);
-    PolyMesh::HalfFace f0 = mesh.add_face(vertices);
+    PolyMesh::HalfFace h0 = mesh.add_face(vertices);
 
     vertices.clear();
     vertices.push_back(v1);
     vertices.push_back(v2);
     vertices.push_back(v4);
-    PolyMesh::HalfFace f1 = mesh.add_face(vertices);
+    PolyMesh::HalfFace h1 = mesh.add_face(vertices);
 
     vertices.clear();
     vertices.push_back(v0);
     vertices.push_back(v1);
     vertices.push_back(v2);
-    PolyMesh::HalfFace f2 = mesh.add_face(vertices);
+    PolyMesh::HalfFace h2 = mesh.add_face(vertices);
 
     vertices.clear();
     vertices.push_back(v0);
     vertices.push_back(v4);
     vertices.push_back(v2);
-    PolyMesh::HalfFace f3 = mesh.add_face(vertices);
+    PolyMesh::HalfFace h3 = mesh.add_face(vertices);
 
     vertices.clear();
     vertices.push_back(v0);
     vertices.push_back(v4);
     vertices.push_back(v3);
-    PolyMesh::HalfFace f4 = mesh.add_face(vertices);
+    PolyMesh::HalfFace h4 = mesh.add_face(vertices);
 
     vertices.clear();
     vertices.push_back(v2);
     vertices.push_back(v3);
     vertices.push_back(v4);
-    PolyMesh::HalfFace f5 = mesh.add_face(vertices);
+    PolyMesh::HalfFace h5 = mesh.add_face(vertices);
 
     vertices.clear();
     vertices.push_back(v0);
     vertices.push_back(v2);
     vertices.push_back(v3);
-    PolyMesh::HalfFace f6 = mesh.add_face(vertices);
+    PolyMesh::HalfFace h6 = mesh.add_face(vertices);
 
     std::vector<PolyMesh::HalfFace> halffaces;
 
     // Add first tetrahedron
-    halffaces.push_back(mesh.opposite(f0));
-    halffaces.push_back(mesh.opposite(f1));
-    halffaces.push_back(f2);
-    halffaces.push_back(mesh.opposite(f3));
+    halffaces.push_back(mesh.opposite(h0));
+    halffaces.push_back(mesh.opposite(h1));
+    halffaces.push_back(h2);
+    halffaces.push_back(mesh.opposite(h3));
     mesh.add_cell(halffaces);
 
     // Add second tetrahedron
     halffaces.clear();
-    halffaces.push_back(mesh.opposite(f4));
-    halffaces.push_back(mesh.opposite(f5));
-    halffaces.push_back(f3);
-    halffaces.push_back(f6);
+    halffaces.push_back(mesh.opposite(h4));
+    halffaces.push_back(mesh.opposite(h5));
+    halffaces.push_back(h3);
+    halffaces.push_back(h6);
     mesh.add_cell(halffaces);
 
     std::cout << "#cell:   " << mesh.n_cells() << std::endl;
@@ -115,10 +115,10 @@ int main(int argc, char **argv) {
 
     // Print positions of vertices to std out
     for (auto v : mesh.vertices())
-        std::cout << "Position of " << v << ": " << mesh.position(v) << std::endl;
+        std::cout << "coordinates of " << v << ": " << mesh.position(v) << std::endl;
 
     for (auto c : mesh.cells()) {
-        std::cout << "Vertices of " << c << ": ";
+        std::cout << "vertex indices of " << c << ": ";
         for (auto v : mesh.vertices(c))
             std::cout << v.idx() << " ";
         std::cout << std::endl;
