+++ easy3d/fileio/poly_mesh_io.cpp
@@ -35,9 +35,8 @@
 namespace easy3d {
 
 
-	PolyMesh* PolyMeshIO::load(const std::string& file_name)
-	{
-		std::setlocale(LC_NUMERIC, "C");
+    PolyMesh *PolyMeshIO::load(const std::string &file_name) {
+        std::setlocale(LC_NUMERIC, "C");
 
         std::ifstream input(file_name.c_str());
         if (input.fail()) {
@@ -45,44 +44,48 @@ namespace easy3d {
             return nullptr;
         }
 
-        const std::string& ext = file_system::extension(file_name, true);
-        if (ext != "tet") {
+        const std::string &ext = file_system::extension(file_name, true);
+        if (ext != "plm") {
             LOG(ERROR) << "unsupported file format: " << ext;
             return nullptr;
         }
 
-        PolyMesh* mesh = new PolyMesh;
+        PolyMesh *mesh = new PolyMesh;
         mesh->set_name(file_name);
 
-		StopWatch w;
+        StopWatch w;
 
         std::string dummy;
-        int num_vertices, num_tets;
-        input >> dummy >> num_vertices >> num_tets;
+        int num_vertices, num_cells;
+        input >> dummy >> num_vertices >> dummy >> num_cells;
 
         vec3 p;
-        for (std::size_t i = 0; i < num_vertices; ++i) {
+        for (std::size_t v = 0; v < num_vertices; ++v) {
             input >> p;
             mesh->add_vertex(p);
         }
 
-        ivec4 ids;
-        for (std::size_t i = 0; i < num_tets; ++i) {
-            input >> ids;
-            mesh->add_tetra(
-                    PolyMesh::Vertex(ids[0]),
-                    PolyMesh::Vertex(ids[1]),
-                    PolyMesh::Vertex(ids[2]),
-                    PolyMesh::Vertex(ids[3])
-                    );
+        int num_halffaces, num_valence, idx;
+        for (std::size_t c = 0; c < num_cells; ++c) {
+            input >> num_halffaces;
+            std::vector<PolyMesh::HalfFace> halffaces(num_halffaces);
+            for (std::size_t hf = 0; hf < num_halffaces; ++hf) {
+                input >> num_valence;
+                std::vector<PolyMesh::Vertex> vts(num_valence);
+                for (std::size_t v = 0; v < num_valence; ++v) {
+                    input >> idx;
+                    vts[v] = PolyMesh::Vertex(idx);
+                }
+                halffaces[hf] = mesh->add_face(vts);
+            }
+            mesh->add_cell(halffaces);
         }
 
-		if (num_vertices == 0 || num_tets == 0) {
+        if (num_vertices == 0 || num_cells == 0) {
             LOG(WARNING) << "no valid data in file: " << file_name;
-			delete mesh;
-			return nullptr;
-		}
-		else {
+            delete mesh;
+            return nullptr;
+        } else {
             LOG(INFO) << "polyhedral mesh loaded ("
                       << "#vertex: " << mesh->n_vertices() << ", "
                       << "#edge: " << mesh->n_edges() << ", "
@@ -92,15 +95,14 @@ namespace easy3d {
 
             return mesh;
         }
-	}
+    }
 
 
-	bool PolyMeshIO::save(const std::string& file_name, const PolyMesh* mesh)
-	{
+    bool PolyMeshIO::save(const std::string &file_name, const PolyMesh *mesh) {
         if (!mesh || mesh->n_vertices() == 0 || mesh->n_cells() == 0) {
-			LOG(ERROR) << "surface mesh is null";
-			return false;
-		}
+            LOG(ERROR) << "surface mesh is null";
+            return false;
+        }
 
[…]
