+++ easy3d/renderer/average_color_blending.h
@@ -54,7 +54,7 @@ namespace easy3d {
         /// \brief Renders the scene (a set of surfaces) with transparency effect
         void draw(const std::vector<TrianglesDrawable*>& surfaces) override;
 
-#ifndef DOXYGEN
+
     protected:
 
         // Allocate and configure FBOs, shader programs.
@@ -65,7 +65,6 @@ namespace easy3d {
 
         // Release all FBOs, shader programs.
         void clear();
-#endif
 
     private:
         Camera*	camera_;
