+++ easy3d/core/point_cloud.h
@@ -97,6 +97,11 @@ namespace easy3d {
                 return idx_ < _rhs.idx_;
             }
 
+            /// helper structure to be able to use std::unordered_map
+            struct Hash {
+                std::size_t operator()(const BaseHandle& h) const { return h.idx(); }
+            };
+
         private:
             friend class VertexIterator;
             friend class PointCloud;
