+++ easy3d/algo/surface_mesh_tetrahedralization.cpp
@@ -1,0 +1,193 @@
+/**
+ * Copyright (C) 2015 by Liangliang Nan (liangliang.nan@gmail.com)
+ * https://3d.bk.tudelft.nl/liangliang/
+ *
+ * This file is part of Easy3D. If it is useful in your research/work,
+ * I would be grateful if you show your appreciation by citing it:
+ * ------------------------------------------------------------------
+ *      Liangliang Nan.
+ *      Easy3D: a lightweight, easy-to-use, and efficient C++
+ *      library for processing and rendering 3D data. 2018.
+ * ------------------------------------------------------------------
+ * Easy3D is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License Version 3
+ * as published by the Free Software Foundation.
+ *
+ * Easy3D is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+
+#include <easy3d/algo/surface_mesh_tetrahedralization.h>
+#include <easy3d/core/surface_mesh.h>
+#include <easy3d/core/poly_mesh.h>
+#include <easy3d/util/logging.h>
+
+#include <3rd_party/tetgen/tetgen.h>
+
+
+namespace easy3d {
+
+
+    SurfaceMeshTetrehedralization::SurfaceMeshTetrehedralization()
+            : add_steiner_points_on_exterior_boundary_(true), add_steiner_points_on_interior_boundary_(true),
+              tag_regions_(false), max_tet_shape_(2.0), max_tet_volume_(-1.0) {
+    }
+
+
+    SurfaceMeshTetrehedralization::~SurfaceMeshTetrehedralization() {
+    }
+
+
+    PolyMesh *SurfaceMeshTetrehedralization::tetrahedralize(SurfaceMesh *mesh) {
+        if (!mesh) {
+            LOG(WARNING) << "input mesh is NULL";
+            return nullptr;
+        }
+
+        for (auto e : mesh->edges()) {
+            if (mesh->is_border(e)) {
+                LOG(WARNING) << "input mesh is not closed, cannot perform tetrahedralization";
+                return nullptr;
+            }
+        }
+
+        tetgenbehavior *tetgen_args = new tetgenbehavior;
+        // Create tetgen argument string from options.
+        if (cmdline_ == "") {
+            std::ostringstream s;
+            // Q: quiet
+            // p: input data is surfacic
+            // n: output tet neighbors
+            // q: desired quality
+            // V: verbose
+            s << "Qpnq" << max_tet_shape_;
+            if (max_tet_volume_ > 0.0) {
+                s << "a" << max_tet_volume_;
+            }
+
+            // AA: generate region tags for each shell.
+            if (tag_regions_) {
+                s << "AA";
+            }
+
+            // YY: prohibit steiner points on boundaries (first Y for exterior boundary, second Y for the other ones).
+            if (add_steiner_points_on_exterior_boundary_ && !add_steiner_points_on_interior_boundary_) {
+                LOG(WARNING)
+                        << "invalid combination of flags (does not preserve exterior boundary and preserves interior ones)"
+                        << " - preserving exterior boundary as well ...";
+                add_steiner_points_on_exterior_boundary_ = false;
+            }
+
[…]
