### 2020-12-22 a0ad1db
simpler code
