### 2020-12-22 fae7c21
don't render invisible models
