### 2020-12-22 500fb13
check if effect exists
