### 2020-12-22 816788d
clipping plane
