+++ easy3d/renderer/state.cpp
@@ -42,7 +42,7 @@ namespace easy3d {
               property_location_(VERTEX), property_name_("uniform color"), lighting_(true),
               lighting_two_sides_(false), distinct_back_color_(true), back_color_(vec4(1, 0, 0, 1)),
               texture_(nullptr), texture_repeat_(1.0f), texture_fractional_repeat_(0.0f), ssao_enabled_(false),
-              clamp_range_(true), clamp_lower_(0.05f), clamp_upper_(0.05f),
+              clamp_range_(true), clamp_lower_(0.05f), clamp_upper_(0.05f), plane_clipping_discard_(false),
               highlight_(false), highlight_range_(-1, -1) {
     }
 
@@ -54,7 +54,7 @@ namespace easy3d {
                                    texture_fractional_repeat_(s.texture_fractional_repeat()),
                                    ssao_enabled_(s.is_ssao_enabled()),
                                    clamp_range_(s.clamp_range()), clamp_lower_(s.clamp_lower()),
-                                   clamp_upper_(s.clamp_upper()), material_(s.material()),
+                                   clamp_upper_(s.clamp_upper()), material_(s.material()), plane_clipping_discard_(false),
                                    highlight_(s.highlight()), highlight_range_(s.highlight_range()) {
     }
 
@@ -69,6 +69,7 @@ namespace easy3d {
         lighting_two_sides_ = s.lighting_two_sides();
         distinct_back_color_ = s.distinct_back_color();
         back_color_ = s.back_color();
+        plane_clipping_discard_ = s.plane_clipping_discard();
         highlight_ = s.highlight();
         highlight_range_ = s.highlight_range();
         texture_ = s.texture();
