+++ easy3d/renderer/buffers.cpp
@@ -2423,7 +2423,7 @@ namespace easy3d {
         }
 
 
-        void update(PolyMesh* model, TrianglesDrawable* drawable) {
+        void update(PolyMesh* model, TrianglesDrawable* drawable, bool border) {
             assert(model);
             assert(drawable);
 
@@ -2432,15 +2432,53 @@ namespace easy3d {
                 return;
             }
 
-            std::vector<unsigned int> d_indices;
-            for (auto f : model->faces()) {
-                for (auto v : model->vertices(f)) {
-                    d_indices.push_back(v.idx());
+            if (!model->is_tetraheral_mesh()) {
+                std::vector<unsigned int> d_indices;
+                for (auto f : model->faces()) {
+                    if (model->is_border(f) == border) {
+                        for (auto v : model->vertices(f))
+                            d_indices.push_back(v.idx());
+                    }
                 }
-            }
 
-            drawable->update_vertex_buffer(model->points());
-            drawable->update_element_buffer(d_indices);
+                drawable->update_vertex_buffer(model->points());
+                drawable->update_element_buffer(d_indices);
+            }
+            else {
+                /**
+                 * We use the Tessellator to eliminate duplicate vertices. This allows us to take advantage of element
+                 * buffer to minimize the number of vertices sent to the GPU.
+                 */
+                Tessellator tessellator;
+
+                for (auto f : model->faces()) {
+                    if (model->is_border(f) == border) {
+                        tessellator.begin_polygon(model->compute_face_normal(model->halfface(f, 0)));
+                        // tessellator.set_winding_rule(Tessellator::WINDING_NONZERO);  // or POSITIVE
+                        tessellator.begin_contour();
+                        for (auto v : model->vertices(f)) {
+                            Tessellator::Vertex vertex(model->position(v), v.idx());
+                            tessellator.add_vertex(vertex);
+                        }
+                        tessellator.end_contour();
+                        tessellator.end_polygon();
+                    }
+                }
+
+                const std::vector<Tessellator::Vertex *> &vts = tessellator.vertices();
+                std::vector<vec3> d_points;
+                d_points.reserve(vts.size());
+
+                for (auto v :vts)
+                    d_points.emplace_back(v->data());
+
+                const auto &d_indices = tessellator.elements();
+
+                drawable->update_vertex_buffer(d_points);
+                drawable->update_element_buffer(d_indices);
+
+                DLOG(INFO) << "num of vertices in model/sent to GPU: " << model->n_vertices() << "/" << d_points.size();
+            }
         }
 
 
@@ -2497,7 +2535,7 @@ namespace easy3d {
                 PolyMesh *mesh = dynamic_cast<PolyMesh *>(model);
                 switch (drawable->type()) {
                     case Drawable::DT_TRIANGLES:
-                        update(mesh, dynamic_cast<TrianglesDrawable *>(drawable));
+                        update(mesh, dynamic_cast<TrianglesDrawable *>(drawable), drawable->name() == "faces:border");
                         break;
                     case Drawable::DT_LINES:
                         update(mesh, dynamic_cast<LinesDrawable *>(drawable));
