+++ easy3d/renderer/renderer.cpp
@@ -109,13 +109,21 @@ namespace easy3d {
         } else if (dynamic_cast<PolyMesh *>(model)) {
             PolyMesh *mesh = dynamic_cast<PolyMesh *>(model);
 
-            // faces
-            auto faces = mesh->renderer()->add_triangles_drawable("faces");
-            faces->set_uniform_coloring(setting::surface_mesh_faces_color);
-            faces->set_distinct_back_color(true);
-            faces->set_lighting_two_sides(true);
-            faces->set_visible(true);
-            faces->set_plane_clipping_discard(true);
+            // we have two faces drawables for polyhedral meshes
+            // interior faces
+            auto interior_faces = mesh->renderer()->add_triangles_drawable("faces:interior");
+            interior_faces->set_uniform_coloring(setting::triangles_drawable_backside_color);
+            interior_faces->set_distinct_back_color(false);
+            interior_faces->set_lighting_two_sides(true);
+            interior_faces->set_visible(true);
+            interior_faces->set_plane_clipping_discard(true);
+            // border faces
+            auto border_faces = mesh->renderer()->add_triangles_drawable("faces:border");
+            border_faces->set_uniform_coloring(setting::surface_mesh_faces_color);
+            border_faces->set_distinct_back_color(false);
+            border_faces->set_lighting_two_sides(true);
+            border_faces->set_visible(true);
+            border_faces->set_plane_clipping_discard(true);
 
             // edges
             auto edges = mesh->renderer()->add_lines_drawable("edges");
