+++ easy3d/core/poly_mesh.cpp
@@ -31,6 +31,110 @@
 
 namespace easy3d {
 
+    namespace details {
+
+        template<typename T>
+        inline void read(std::istream &input, std::set<T>& data) {
+            unsigned int size(0);
+            input.read((char*)&size, sizeof(unsigned int));
+            if (size == 0)
+                std::cout << "bug";
+            std::vector<T> array(size);
+            input.read((char*)array.data(), size * sizeof(T));
+            data = std::set<T>(array.begin(), array.end());
+        }
+
+        template<typename T>
+        inline void write(std::ostream &output, const std::set<T>& data) {
+            unsigned int size = data.size();
+            if (size == 0)
+                std::cout << "bug";
+            output.write((char*)&size, sizeof(unsigned int));
+            std::vector<T> array(data.begin(), data.end());
+            output.write((char*)array.data(), size * sizeof(T));
+        }
+
+        template<typename T>
+        inline void read(std::istream &input, std::vector<T>& data) {
+            unsigned int size(0);
+            input.read((char*)&size, sizeof(unsigned int));
+            data.resize(size);
+            if (size == 0)
+                std::cout << "bug";
+            input.read((char*)data.data(), size * sizeof(T));
+        }
+
+        template<typename T>
+        inline void write(std::ostream &output, const std::vector<T>& data) {
+            unsigned int size = data.size();
+            if (size == 0)
+                std::cout << "bug";
+            output.write((char*)&size, sizeof(unsigned int));
+            output.write((char*)data.data(), size * sizeof(T));
+        }
+    }
+
+
+    void PolyMesh::VertexConnectivity::read(std::istream &input) {
+        details::read(input, vertices_);
+        details::read(input, edges_);
+        details::read(input, halffaces_);
+        details::read(input, cells_);
+    }
+    
+
+    void PolyMesh::VertexConnectivity::write(std::ostream &output) const {
+        details::write(output, vertices_);
+        details::write(output, edges_);
+        details::write(output, halffaces_);
+        details::write(output, cells_);
+    }
+
+
+    void PolyMesh::EdgeConnectivity::read(std::istream &input) {
+        details::read(input, vertices_);
+        details::read(input, halffaces_);
+        details::read(input, cells_);
+    }
+
+
+    void PolyMesh::EdgeConnectivity::write(std::ostream &output) const {
+        details::write(output, vertices_);
+        details::write(output, halffaces_);
+        details::write(output, cells_);
+    }
+
+
+    void PolyMesh::HalfFaceConnectivity::read(std::istream &input) {
+        details::read(input, vertices_);
+        details::read(input, edges_);
+        input.read((char*)(&cell_), sizeof(Cell));
+        input.read((char*)(&opposite_), sizeof(HalfFace));
+    }
+
+
+    void PolyMesh::HalfFaceConnectivity::write(std::ostream &output) const {
+        details::write(output, vertices_);
+        details::write(output, edges_);
+        output.write((char*)(&cell_), sizeof(Cell));
+        output.write((char*)(&opposite_), sizeof(HalfFace));
+    }
+
+
+    void PolyMesh::CellConnectivity::read(std::istream &input) {
+        details::read(input, vertices_);
+        details::read(input, edges_);
+        details::read(input, halffaces_);
+    }
+
+
+    void PolyMesh::CellConnectivity::write(std::ostream &output) const {
+        details::write(output, vertices_);
[…]
