### 2020-12-23 64ff2a9
io of adjacency information
