+++ ReadMe.md
@@ -28,7 +28,7 @@
   programmable-shader-style rendering, i.e., no fixed function calls). Client code does not need to touch the low-level 
   APIs of OpenGL. 
   
-* Step-by-step tutorials to get acquainted with the data structures, rendering techniques, and algorithms 
+* Step-by-step tutorials demonstrating various uses of the API, to get acquainted with the data structures, rendering techniques, and algorithms 
   for 3D modeling and geometry processing. 
     
 * A viewer that can be used directly to visualize 3D scenes in various formats, which can also be easily extended.
