+++ easy3d/algo/surface_mesh_smoothing.cpp
@@ -32,8 +32,10 @@
 
 namespace easy3d {
 
+    // \cond
     using SparseMatrix = Eigen::SparseMatrix<double>;
     using Triplet = Eigen::Triplet<double>;
+    // \endcond
 
     //-----------------------------------------------------------------------------
 
