### 2020-12-23 a3d0b50
optimize find_half_face()
