### 2020-12-23 9911ea0
optimize add_face()
