+++ easy3d/algo/surface_mesh_tetrahedralization.h
@@ -42,22 +42,46 @@ namespace easy3d {
         SurfaceMeshTetrehedralization();
         ~SurfaceMeshTetrehedralization();
 
-        void set_add_steiner_points_on_exterior_boundary(bool x) {
-            add_steiner_points_on_exterior_boundary_ = x;
+        /**
+         * Sets if Steiner points are allowed on the boundary edges and faces of the input surface. Default is true.
+         * \details Disable this will preserve the input boundary edges and faces. In this case, Steiner points (if
+         *      there exists any) will appear only in the interior space of the input surface.
+         */
+        void set_allow_steiner_points_on_boundary(bool x) {
+            allow_steiner_points_on_boundary_ = x;
         }
 
-        void set_add_steiner_points_on_interior_boundary(bool x) {
-            add_steiner_points_on_interior_boundary_ = x;
-        }
-
+        /**
+         * Sets the maximum allowable radius-edge ratio. Default value is 2.0.
+         * The value controls how new points can be added to improve the mesh quality.
+         * More complicated constraints can be set by using set_command_line(). See the "-q" switch in tetgen manual.
+         * http://wias-berlin.de/software/tetgen/1.5/doc/manual/manual005.html#cmd-q
+         * \sa set_max_dihedral_angle()
+         */
         void set_max_tet_shape(double x) { max_tet_shape_ = x; }
 
+        /**
+         * Sets the maximum allowable dihedral angle. Default value is 0.
+         * The value controls how new points can be added to improve the mesh quality.
+         * More complicated constraints can be set by using set_command_line(). See the "-q" switch in tetgen manual.
+         * http://wias-berlin.de/software/tetgen/1.5/doc/manual/manual005.html#cmd-q
+         * \sa set_max_tet_shape()
+         */
+        void set_max_dihedral_angle(double x) { max_dihedral_angle_ = x; }
+
+        /**
+         * Sets the maximum volume constraint on all tetrahedra. Default value is -1 (no max volume constraint).
+         * \details A positive value indicates that no tetrahedra is generated whose volume is larger than this value.
+         * \note A negagive value indicates no such a constraint.
+         * More complicated constraints can be set by using set_command_line(). See the "-a" switch in tetgen manual.
+         * http://wias-berlin.de/software/tetgen/1.5/doc/manual/manual005.html#cmd-a
+         */
         void set_max_tet_volume(double x) { max_tet_volume_ = x; }
 
         /**
-         * Enables/Disables marking the cells.
-         * If enables, outer cells will be marked "-1" and inner cells will be marked "1".
-         * \note Effective only when
+         * If enabled, assigns an additional attribute (an integer number) to each tetrahedron that identifies to what
+         * facet-bounded region it belongs. In the output mesh, all tetrahedra in the same region will get a
+         * corresponding non-zero attribute.
          */
         void set_tag_regions(bool x) { tag_regions_ = x; }
 
@@ -73,10 +97,10 @@ namespace easy3d {
 
     private:
 
-        bool add_steiner_points_on_exterior_boundary_;
-        bool add_steiner_points_on_interior_boundary_;
+        bool allow_steiner_points_on_boundary_;
         bool tag_regions_;
         double max_tet_shape_;
+        double max_dihedral_angle_;
         double max_tet_volume_;
         std::string cmdline_;
     };
