### 2020-12-24 6ba6615
clipping plane size relative to scene radius
