### 2020-12-24 f3502b9
don't discard vertices when visualizing cross sections
