+++ applications/Mapple/main_window.cpp
@@ -183,7 +183,7 @@ MainWindow::MainWindow(QWidget *parent)
     readSettings();
     updateWindowTitle();
 
-#ifdef NDEBUG
+#if 0
     QMessageBox::warning(this, "Mapple is not ready yet!",
                          "Mapple is still under development. This version is not feature complete nor is fully tested. Using it is at your own risk.",
                          QMessageBox::Ok);
