### 2020-12-25 fa89407
allow to used color property defined on edges
