+++ sandbox/Experiment/main.cpp
@@ -4,9 +4,12 @@
 #include <easy3d/renderer/texture.h>
 #include <easy3d/core/graph.h>
 #include <easy3d/fileio/graph_io.h>
+#include <easy3d/fileio/surface_mesh_io.h>
 #include <easy3d/core/curve.h>
 #include <easy3d/util/string.h>
+#include <easy3d/util/stop_watch.h>
 #include <easy3d/core/poly_mesh.h>
+#include <easy3d/core/surface_mesh.h>
 
 #include <cstring>
 #include <cstdio>
@@ -18,20 +21,141 @@
 #define STB_TRUETYPE_IMPLEMENTATION
 #include <3rd_party/stb/stb_truetype.h>
 
+#include <3rd_party/meshoptimizer/src/meshoptimizer.h>
 
 using namespace easy3d;
+
+
+void simplify(const SurfaceMesh& mesh)
+{
+    static const size_t lod_count = 5;
+
+    std::vector<vec3> mesh_vertices = mesh.points();
+    std::vector<unsigned int> mesh_indices(mesh.n_faces() * 3);
+    int idx = 0;
+    for (auto f : mesh.faces()) {
+        for (auto v : mesh.vertices(f)) {
+            mesh_indices[idx] = v.idx();
+            ++idx;
+        }
+    }
+
+    StopWatch w;
+
+    // generate 4 LOD levels (1-4), with each subsequent LOD using 70% triangles
+    // note that each LOD uses the same (shared) vertex buffer
+    std::vector<unsigned int> lods[lod_count];
+    lods[0] = mesh_indices;
+
+    for (size_t i = 1; i < lod_count; ++i)
+    {
+        std::vector<unsigned int>& lod = lods[i];
+
+        float threshold = powf(0.7f, float(i));
+        size_t target_index_count = size_t(mesh_indices.size() * threshold) / 3 * 3;
+        float target_error = 1e-2f;
+
+        // we can simplify all the way from base level or from the last result
+        // simplifying from the base level sometimes produces better results, but simplifying from last level is faster
+        const std::vector<unsigned int>& source = lods[i - 1];
+
+        if (source.size() < target_index_count)
+            target_index_count = source.size();
+
+        lod.resize(source.size());
+        lod.resize(meshopt_simplify(&lod[0], &source[0], source.size(), &mesh_vertices[0].x, mesh_vertices.size(), sizeof(vec3), target_index_count, target_error));
+    }
+
+    double middle = w.elapsed_seconds(4);
+    w.restart();
+
+    // optimize each individual LOD for vertex cache & overdraw
+    for (size_t i = 0; i < lod_count; ++i)
+    {
+        std::vector<unsigned int>& lod = lods[i];
+
+        meshopt_optimizeVertexCache(&lod[0], &lod[0], lod.size(), mesh_vertices.size());
+        meshopt_optimizeOverdraw(&lod[0], &lod[0], lod.size(), &mesh_vertices[0].x, mesh_vertices.size(), sizeof(vec3), 1.0f);
+
+        printf("\tLOD %d: %d triangles\n", int(i), int(lod.size() / 3));
+    }
+
+    // concatenate all LODs into one IB
+    // note: the order of concatenation is important - since we optimize the entire IB for vertex fetch,
+    // putting coarse LODs first makes sure that the vertex range referenced by them is as small as possible
+    // some GPUs process the entire range referenced by the index buffer region so doing this optimizes the vertex transform
+    // cost for coarse LODs
+    // this order also produces much better vertex fetch cache coherency for coarse LODs (since they're essentially optimized first)
[…]
