### 2020-12-25 22d4c36
#define M_PI if not defined (an MS issue)
