+++ easy3d/core/constant.h
@@ -36,25 +36,10 @@ namespace easy3d {
 #define M_PI			3.14159265358979323846264338327950288
 #endif
 
-#define half_pi			M_PI * 0.5
-#define quarter_pi		M_PI * 0.25
-#define two_pi			M_PI * 2.0
-#define rad2deg(a)		((a) * 180.0 / M_PI)
-#define deg2rad(a)		((a) * M_PI / 180.0)
-
-    /// \brief The PI.
-    const double pi = 3.14159265358979323846264338327950288;
-//    /// \brief The half PI (PI/2)
-//    const double half_pi = pi * 0.5;
-//    /// \brief The quarter PI (PI/4)
-//    const double quarter_pi = pi * 0.25;
-//    /// \brief The two PI (2 * PI).
-//    const double two_pi	 = pi * 2.0;
-
-//    /// \brief Conversion from radians to degrees
-//    double inline rad2deg(double rad) { return rad * 180.0 / pi; }
-//    /// \brief Conversion from degrees to radians
-//    double inline deg2rad(double deg) { return deg * pi / 180.0; }
+    /// \brief Conversion from radians to degrees
+    double inline rad2deg(double rad) { return rad * 180.0 / M_PI; }
+    /// \brief Conversion from degrees to radians
+    double inline deg2rad(double deg) { return deg * M_PI / 180.0; }
 
     /// Function returning minimum representable value for a given type
     template <typename FT> inline FT min();
