### 2020-12-25 76a28c6
#include <easy3d/core/constant.h>
