+++ 3rd_party/tinyobjloader/tiny_obj_loader.h
@@ -1,3098 +1,0 @@
-/*
-The MIT License (MIT)
-
-Copyright (c) 2012-2018 Syoyo Fujita and many contributors.
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
-*/
-
-//
-// version 2.0.0 : Add new object oriented API. 1.x API is still provided.
-//                 * Support line primitive.
-//                 * Support points primitive.
-//                 * Support multiple search path for .mtl(v1 API).
-//                 * Support vertex weight `vw`(as an tinyobj extension)
-//                 * Support escaped whitespece in mtllib
-// version 1.4.0 : Modifed ParseTextureNameAndOption API
-// version 1.3.1 : Make ParseTextureNameAndOption API public
-// version 1.3.0 : Separate warning and error message(breaking API of LoadObj)
-// version 1.2.3 : Added color space extension('-colorspace') to tex opts.
-// version 1.2.2 : Parse multiple group names.
-// version 1.2.1 : Added initial support for line('l') primitive(PR #178)
-// version 1.2.0 : Hardened implementation(#175)
-// version 1.1.1 : Support smoothing groups(#162)
-// version 1.1.0 : Support parsing vertex color(#144)
-// version 1.0.8 : Fix parsing `g` tag just after `usemtl`(#138)
-// version 1.0.7 : Support multiple tex options(#126)
-// version 1.0.6 : Add TINYOBJLOADER_USE_DOUBLE option(#124)
-// version 1.0.5 : Ignore `Tr` when `d` exists in MTL(#43)
-// version 1.0.4 : Support multiple filenames for 'mtllib'(#112)
-// version 1.0.3 : Support parsing texture options(#85)
-// version 1.0.2 : Improve parsing speed by about a factor of 2 for large
-// files(#105)
-// version 1.0.1 : Fixes a shape is lost if obj ends with a 'usemtl'(#104)
-// version 1.0.0 : Change data structure. Change license from BSD to MIT.
-//
-
-//
-// Use this in *one* .cc
-//   #define TINYOBJLOADER_IMPLEMENTATION
-//   #include "tiny_obj_loader.h"
-//
-
-#ifndef TINY_OBJ_LOADER_H_
-#define TINY_OBJ_LOADER_H_
-
-#include <map>
-#include <string>
-#include <vector>
-
-namespace tinyobj {
-
-// TODO(syoyo): Better C++11 detection for older compiler
[…]
