### 2020-12-26 0b958ba
disable double-sided lighting by default
