+++ tutorials/Tutorial_601_Cloud_NormalEstimation/main.cpp
@@ -55,6 +55,7 @@ int main(int argc, char **argv) {
     auto drawable = model->renderer()->get_points_drawable("vertices");
     drawable->set_uniform_coloring(vec4(0.6f, 0.6f, 1.0f, 1.0f));
     drawable->set_point_size(5.0f);
+    drawable->set_lighting_two_sides(false);
 
     // Run the viewer
     return viewer.run();
