### 2020-12-26 c3d3cb8
add angle_threshold parameter
