### 2020-12-26 7c96cc3
save/restore viewer state; save/retore camera path
