+++ applications/Mapple/main_window.cpp
@@ -606,6 +606,78 @@ void MainWindow::setBackgroundColor() {
 }
 
 
+void MainWindow::saveCameraStateToFile() {
+    QString suggested_name = curDataDirectory_;
+    if (viewer()->currentModel()) {
+        const std::string name = file_system::replace_extension(viewer()->currentModel()->name(), "view");
+        suggested_name = QString::fromStdString(name);
+    }
+    const QString fileName = QFileDialog::getSaveFileName(
+            this,
+            "Save viewer state to file",
+            suggested_name,
+            "Viewer state (*.view)\n"
+            "All formats (*.*)"
+    );
+
+    if (!fileName.isEmpty()) {
+        viewer_->saveStateToFile(fileName.toStdString());
+        // assume the user will soon restore the state from this file.
+        curDataDirectory_ = fileName.left(fileName.lastIndexOf("/"));
+    }
+}
+
+
+void MainWindow::restoreCameraStateFromFile() {
+    const QString fileName = QFileDialog::getOpenFileName(
+            this,
+            "Restore viewer state from file",
+            curDataDirectory_,
+            "Viewer state (*.view)\n"
+            "All formats (*.*)"
+    );
+
+    if (!fileName.isEmpty())
+        viewer_->restoreStateFromFile(fileName.toStdString());
+}
+
+
+void MainWindow::exportCamaraPathToFile() {
+    QString suggested_name = curDataDirectory_;
+    if (viewer()->currentModel()) {
+        const std::string name = file_system::replace_extension(viewer()->currentModel()->name(), "path");
+        suggested_name = QString::fromStdString(name);
+    }
+    const QString fileName = QFileDialog::getSaveFileName(
+            this,
+            "Export camera path to file",
+            suggested_name,
+            "Camera state (*.path)\n"
+            "All formats (*.*)"
+    );
+
+    if (!fileName.isEmpty()) {
+        viewer_->exportCamaraPathToFile(fileName.toStdString());
+        // assume the user will soon restore the state from this file.
+        curDataDirectory_ = fileName.left(fileName.lastIndexOf("/"));
+    }
+}
+
+
+void MainWindow::importCameraPathFromFile() {
+    const QString fileName = QFileDialog::getOpenFileName(
+            this,
+            "Import camera path from file",
+            curDataDirectory_,
+            "Camera path (*.path)\n"
+            "All formats (*.*)"
+    );
+
+    if (!fileName.isEmpty())
+        viewer_->importCameraPathFromFile(fileName.toStdString());
+}
+
+
 bool MainWindow::okToContinue()
 {
     if (isWindowModified()) {
@@ -767,13 +839,13 @@ void MainWindow::createActionsForViewMenu() {
     connect(ui->actionCopyCamera, SIGNAL(triggered()), viewer_, SLOT(copyCamera()));
     connect(ui->actionPasteCamera, SIGNAL(triggered()), viewer_, SLOT(pasteCamera()));
 
-    connect(ui->actionSaveCameraStateToFile, SIGNAL(triggered()), viewer_, SLOT(saveCameraStateToFile()));
-    connect(ui->actionRestoreCameraStateFromFile, SIGNAL(triggered()), viewer_, SLOT(restoreCameraStateFromFile()));
+    connect(ui->actionSaveCameraStateToFile, SIGNAL(triggered()), this, SLOT(saveCameraStateToFile()));
+    connect(ui->actionRestoreCameraStateFromFile, SIGNAL(triggered()), this, SLOT(restoreCameraStateFromFile()));
[…]
