+++ applications/Mapple/paint_canvas.cpp
@@ -64,6 +64,7 @@
 #include <QApplication>
 #include <QClipboard>
 #include <QTime>
+#include <QMessageBox>
 
 
 using namespace easy3d;
@@ -1255,13 +1256,175 @@ void PaintCanvas::pasteCamera() {
 }
 
 
-void PaintCanvas::saveCameraStateToFile() {
-    LOG(WARNING) << "planned to be implemented in a future release";
+void PaintCanvas::saveStateToFile(const std::string& file_name) const {
+    if (file_name.empty())
+        return;
+
+    // Write the state to file
+    std::ofstream output(file_name.c_str());
+    if (output.fail()) {
+        QMessageBox::warning(window(), tr("Save state to file error"), tr("Unable to create file %1").arg(QString::fromStdString(file_name)));
+        return;
+    }
+
+    //-----------------------------------------------------
+
+    // first line is just a comment
+    output << "<Mapple state file>" << std::endl << std::endl;
+
+    //-----------------------------------------------------
+
+    // write foreground and background colors
+    output << "<color>" << std::endl;
+    output << "\t backGroundColor: " << backGroundColor() << std::endl;
+    output << "</color>" << std::endl << std::endl;
+
+    //-----------------------------------------------------
+
+    output << "<display>" << std::endl;
+    output << "\t cameraPathIsDrawn: " << show_camera_path_ << std::endl;
+    output << "</display>" << std::endl << std::endl;
+
+    //-----------------------------------------------------
+
+    output << "<windowState>" << std::endl;
+    output << "\t state: " << window()->windowState() << std::endl;;
+    if (window()->windowState() == Qt::WindowNoState) {
+        output << "\t size: " << window()->width() << " " << window()->height() << std::endl;
+        output << "\t position: " << window()->pos().x() << " " << window()->pos().y() << std::endl;
+    }
+    output << "</windowState>" << std::endl << std::endl;
+
+    //-----------------------------------------------------
+
+    output << "<camera>" << std::endl;
+//    // Restore original QCamera zClippingCoefficient before saving.
+//    if (cameraIsEdited())
+//        camera()->setZClippingCoefficient(previousCameraZClippingCoefficient_);
+
+    switch (camera()->type()) {
+        case Camera::PERSPECTIVE:	output << "\t type: " << "PERSPECTIVE" << std::endl;	break;
+        case Camera::ORTHOGRAPHIC:	output << "\t type: " << "ORTHOGRAPHIC" << std::endl;	break;
+    }
+    output << "\t zClippingCoefficient: " << camera()->zClippingCoefficient() << std::endl;
+    output << "\t zNearCoefficient: " << camera()->zNearCoefficient() << std::endl;
+    output << "\t sceneRadius: " << camera()->sceneRadius() << std::endl;
+    output << "\t fieldOfView: " << camera()->fieldOfView() << std::endl;
+    output << "\t sceneCenter: " << camera()->sceneCenter() << std::endl;
+
+    // ManipulatedCameraFrame
+    output << "\t position: " << camera()->frame()->position() << std::endl;
+    output << "\t orientation: " << camera()->frame()->orientation() << std::endl;
+    output << "\t wheelSens: " << camera()->frame()->wheelSensitivity() << std::endl;
+    output << "\t rotSens: " << camera()->frame()->rotationSensitivity() << std::endl;
+    output << "\t zoomSens: " << camera()->frame()->zoomSensitivity() << std::endl;
[…]
