+++ applications/Mapple/paint_canvas.h
@@ -122,14 +122,21 @@ public slots:
     void copyCamera();
     void pasteCamera();
 
-    void saveCameraStateToFile();
-    void restoreCameraStateFromFile();
     void addKeyFrame();
     void playCameraPath();
     void showCamaraPath();
-    void importCameraPathFromFile();
-    void exportCamaraPathToFile();
     void deleteCameraPath();
+
+public:
+    /*! Save the viewer state (camera state, widget geometry, display flags... etc.) to a file.
+    Use restoreStateFromFile() to restore this state, or you can restore it in your init() method).
+    */
+    void saveStateToFile(const std::string& file_name) const;
+    /*! Restores the viewer state from previously saved file. */
+    void restoreStateFromFile(const std::string& file_name);
+
+    void importCameraPathFromFile(const std::string& file_name);
+    void exportCamaraPathToFile(const std::string& file_name) const;
 
 protected:
 
