### 2020-12-26 6645fe2
add KeyFrameInterpolator::adjust_scene_radius(Camera* cam)
