+++ easy3d/renderer/key_frame_interpolator.cpp
@@ -38,7 +38,7 @@
 #include <easy3d/renderer/frame.h>
 #include <easy3d/renderer/drawable_lines.h>
 #include <easy3d/renderer/primitives.h>
-#include <easy3d/renderer/camera.h>   // for scene raduis (to draw cameras in a proper size)
+#include <easy3d/renderer/camera.h>   // for scene radius (to draw cameras in a proper size)
 
 #include <easy3d/util/timer.h>
 
@@ -58,20 +58,18 @@ namespace easy3d {
         , interpolationTime_(0.0)
         , interpolationSpeed_(1.0)
         , interpolationStarted_(false)
-        , closedPath_(false)
         , loopInterpolation_(false)
         , pathIsValid_(false)
         , valuesAreValid_(true)
         , currentFrameValid_(false)
-        // #CONNECTION# Values cut pasted initFromDOMElement()
         , path_drawable_(nullptr)
         , cameras_drawable_(nullptr)
     {
-        if (keyFrame_.size() < 2)
+        if (keyFrames_.size() < 2)
             return;
 
         for (int i=0; i<4; ++i)
-            currentFrame_[i] = keyFrame_.begin();
+            currentFrame_[i] = keyFrames_.begin();
     }
 
     /*! Virtual destructor. Clears the keyFrame path. */
@@ -98,26 +96,26 @@ namespace easy3d {
 
         interpolationTime_ += interpolationSpeed() * interpolationPeriod() / 1000.0;
 
-        if (interpolationTime() > keyFrame_.back()->time())
+        if (interpolationTime() > keyFrames_.back()->time())
         {
             if (loopInterpolation())
-                setInterpolationTime(keyFrame_.front()->time() + interpolationTime_ - keyFrame_.back()->time());
+                setInterpolationTime(keyFrames_.front()->time() + interpolationTime_ - keyFrames_.back()->time());
             else
             {
                 // Make sure last KeyFrame is reached and displayed
-                interpolateAtTime(keyFrame_.back()->time());
+                interpolateAtTime(keyFrames_.back()->time());
                 stopInterpolation();
             }
             timer_.stop();
         }
-        else if (interpolationTime() < keyFrame_.front()->time())
+        else if (interpolationTime() < keyFrames_.front()->time())
         {
             if (loopInterpolation())
-                setInterpolationTime(keyFrame_.back()->time() - keyFrame_.front()->time() + interpolationTime_);
+                setInterpolationTime(keyFrames_.back()->time() - keyFrames_.front()->time() + interpolationTime_);
             else
             {
                 // Make sure first KeyFrame is reached and displayed
-                interpolateAtTime(keyFrame_.front()->time());
+                interpolateAtTime(keyFrames_.front()->time());
                 stopInterpolation();
             }
             timer_.stop();
@@ -150,12 +148,12 @@ namespace easy3d {
         if (period >= 0)
             setInterpolationPeriod(period);
 
-        if (!keyFrame_.empty())
+        if (!keyFrames_.empty())
         {
-            if ((interpolationSpeed() > 0.0) && (interpolationTime() >= keyFrame_.back()->time()))
-                setInterpolationTime(keyFrame_.front()->time());
-            if ((interpolationSpeed() < 0.0) && (interpolationTime() <= keyFrame_.front()->time()))
-                setInterpolationTime(keyFrame_.back()->time());
+            if ((interpolationSpeed() > 0.0) && (interpolationTime() >= keyFrames_.back()->time()))
+                setInterpolationTime(keyFrames_.front()->time());
[…]
