### 2020-12-26 26c9208
add KeyFrameInterpolator::adjust_scene_radius(Camera* cam)
