+++ easy3d/renderer/key_frame_interpolator.cpp
@@ -46,43 +46,28 @@
 namespace easy3d {
 
 
-    /*! Creates a KeyFrameInterpolator, with \p frame as associated frame().
-
-      The frame() can be set or changed using setFrame().
-
-      interpolationTime(), interpolationSpeed() and interpolationPeriod() are set to their default
-      values. */
-    KeyFrameInterpolator::KeyFrameInterpolator(Frame* frame)
-        : frame_(frame)
-        , period_(40)   // 25 frames per second
-        , interpolationTime_(0.0)
-        , interpolationSpeed_(1.0)
-        , interpolationStarted_(false)
-        , loopInterpolation_(false)
-        , pathIsValid_(false)
-        , valuesAreValid_(true)
-        , currentFrameValid_(false)
-        , path_drawable_(nullptr)
-        , cameras_drawable_(nullptr)
-    {
+    KeyFrameInterpolator::KeyFrameInterpolator(Frame *frame)
+            : frame_(frame), period_(40)   // 25 frames per second
+            , interpolationTime_(0.0), interpolationSpeed_(1.0), interpolationStarted_(false),
+              loopInterpolation_(false), pathIsValid_(false), valuesAreValid_(true), currentFrameValid_(false),
+              path_drawable_(nullptr), cameras_drawable_(nullptr) {
         if (keyFrames_.size() < 2)
             return;
 
-        for (int i=0; i<4; ++i)
+        for (int i = 0; i < 4; ++i)
             currentFrame_[i] = keyFrames_.begin();
     }
 
-    /*! Virtual destructor. Clears the keyFrame path. */
-    KeyFrameInterpolator::~KeyFrameInterpolator()
-    {
+
+    KeyFrameInterpolator::~KeyFrameInterpolator() {
         deletePath();
     }
 
-    /*! Sets the frame() associated to the KeyFrameInterpolator. */
-    void KeyFrameInterpolator::setFrame(Frame* const frame)
-    {
+
+    void KeyFrameInterpolator::setFrame(Frame *const frame) {
         frame_ = frame;
     }
+
 
     /*! Updates frame() state according to current interpolationTime(). Then adds
       interpolationPeriod()*interpolationSpeed() to interpolationTime().
@@ -90,30 +75,24 @@ namespace easy3d {
       This internal method is called by a timer when interpolationIsStarted(). It can be used for
       debugging purpose. stopInterpolation() is called when interpolationTime() reaches firstTime() or
       lastTime(), unless loopInterpolation() is \c true. */
-    void KeyFrameInterpolator::update()
-    {
+    void KeyFrameInterpolator::update() {
         interpolateAtTime(interpolationTime());
 
         interpolationTime_ += interpolationSpeed() * interpolationPeriod() / 1000.0;
 
-        if (interpolationTime() > keyFrames_.back()->time())
-        {
+        if (interpolationTime() > keyFrames_.back()->time()) {
             if (loopInterpolation())
                 setInterpolationTime(keyFrames_.front()->time() + interpolationTime_ - keyFrames_.back()->time());
-            else
-            {
+            else {
                 // Make sure last KeyFrame is reached and displayed
                 interpolateAtTime(keyFrames_.back()->time());
                 stopInterpolation();
             }
             timer_.stop();
-        }
-        else if (interpolationTime() < keyFrames_.front()->time())
-        {
+        } else if (interpolationTime() < keyFrames_.front()->time()) {
             if (loopInterpolation())
                 setInterpolationTime(keyFrames_.back()->time() - keyFrames_.front()->time() + interpolationTime_);
-            else
-            {
+            else {
                 // Make sure first KeyFrame is reached and displayed
[…]
