+++ easy3d/renderer/key_frame_interpolator.h
@@ -137,39 +137,28 @@ namespace easy3d {
     class KeyFrameInterpolator
     {
     public:
+        /*! Creates a KeyFrameInterpolator, with \p frame as associated frame().
+         * \details The frame() can be set or changed using setFrame().
+         * interpolationTime(), interpolationSpeed() and interpolationPeriod() are set to their default values. */
         KeyFrameInterpolator(Frame* fr=nullptr);
+
+        /*! Virtual destructor. Clears the keyFrame path. */
         virtual ~KeyFrameInterpolator();
-
-//    Q_SIGNALS:
-//        /*! This signal is emitted whenever the frame() state is interpolated.
-
-//        The emission of this signal triggers the synchronous emission of the frame()
-//        Frame::interpolated() signal, which may also be useful.
-
-//        This signal should especially be connected to your QCanvas::update() slot, so that the display
-//        is updated after every update of the KeyFrameInterpolator frame():
-//        \code
-//        connect(myKeyFrameInterpolator, SIGNAL(interpolated()), SLOT(update()));
-//        \endcode
-//        Use the QCanvas::QGLViewerPool() to connect the signal to all the viewers.
-
-//        Note that the QCanvas::camera() QCamera::keyFrameInterpolator() created using QCanvas::pathKey()
-//        have their interpolated() signals automatically connected to the QCanvas::update() slot. */
-//        void interpolated();
-
-//        /*! This signal is emitted when the interpolation reaches the first (when interpolationSpeed()
-//          is negative) or the last keyFrame.
-
-//        When loopInterpolation() is \c true, interpolationTime() is reset and the interpolation
-//        continues. It otherwise stops. */
-//        void endReached();
 
         /*! @name Path creation */
         //@{
     public:
+        /*! Appends a new keyFrame to the path.
+         * Same as addKeyFrame(const Frame& frame, double), except that the keyFrameTime() is automatically set
+         * to previous keyFrameTime() plus one second (or 0.0 if there is no previous keyFrame). */
         void addKeyFrame(const Frame& frame);
+
+        /*! Appends a new keyFrame to the path, with its associated \p time (in seconds).
+         * The path will use the current \p frame state.
+         * \attention The keyFrameTime() have to be monotonously increasing over keyFrames. */
         void addKeyFrame(const Frame& frame, double time);
 
+        /*! Removes all keyFrames from the path. The numberOfKeyFrames() is set to 0. */
         void deletePath();
         //@}
 
@@ -185,6 +174,7 @@ namespace easy3d {
         Frame* frame() const { return frame_; }
 
     public:
+        /*! Sets the frame() associated to the KeyFrameInterpolator. */
         void setFrame(Frame* const frame);
         //@}
 
@@ -207,8 +197,19 @@ namespace easy3d {
          */
         double keyFrameTime(int index) const;
 
+        /*! Returns the duration of the KeyFrameInterpolator path, expressed in seconds.
+         * Simply corresponds to lastTime() - firstTime(). Returns 0.0 if the path has less than 2 keyFrames.
+         * \sa keyFrameTime(). */
         double duration() const;
+
+        /*! Returns the time corresponding to the first keyFrame, expressed in seconds.
+         * Returns 0.0 if the path is empty.
+         * \sa lastTime(), duration() and keyFrameTime(). */
         double firstTime() const;
+
+        /*! Returns the time corresponding to the last keyFrame, expressed in seconds.
+         * Returns 0.0 if the path is empty.
+         * \sa firstTime(), duration() and keyFrameTime(). */
         double lastTime() const;
         //@}
 
@@ -270,17 +271,67 @@ namespace easy3d {
         stopInterpolation() or toggleInterpolation() to modify this state. */
[…]
