+++ easy3d/core/graph.h
@@ -40,6 +40,9 @@ namespace easy3d {
 	/**
 	 * @brief A Graph data structure with easy property management.
 	 * \class Graph easy3d/core/graph.h
+     *
+     * This implementation is inspired by Surface_mesh
+     * https://opensource.cit-ec.de/projects/surface_mesh
 	 */
 
 	class Graph : public virtual Model
@@ -104,7 +107,7 @@ namespace easy3d {
 		};
 
 		/// this type represents an edge (internally it is basically an index)
-		/// \sa Vertex, Halfedge, Face
+		/// \sa Vertex
 		struct Edge : public BaseHandle
 		{
 			/// default constructor (with invalid index)
@@ -115,7 +118,7 @@ namespace easy3d {
 	public: //-------------------------------------------------- connectivity types
 
 		/// This type stores the vertex connectivity
-		/// \sa HalfedgeConnectivity, FaceConnectivity
+		/// \sa EdgeConnectivity
 		struct VertexConnectivity
 		{
 			/// all edges connected with the vertex
@@ -127,14 +130,15 @@ namespace easy3d {
 		/// \sa VertexConnectivity
 		struct EdgeConnectivity
 		{
-            std::vector<Vertex>  vertices_;
+            Vertex source_;
+            Vertex target_;
 		};
 
 
 	public: //------------------------------------------------------ property types
 
 		/// Vertex property of type T
-		/// \sa HalfedgeProperty, EdgeProperty, FaceProperty
+		/// \sa EdgeProperty
 		template <class T> class VertexProperty : public Property<T>
 		{
 		public:
@@ -158,7 +162,7 @@ namespace easy3d {
 
 
 		/// Edge property of type T
-		/// \sa VertexProperty, HalfedgeProperty, FaceProperty
+		/// \sa VertexProperty
 		template <class T> class EdgeProperty : public Property<T>
 		{
 		public:
@@ -190,13 +194,13 @@ namespace easy3d {
 			explicit ModelProperty() {}
 			explicit ModelProperty(Property<T> p) : Property<T>(p) {}
 
-			/// access the data stored for the mesh
+			/// access the data stored for the graph
 			typename Property<T>::reference operator[](size_t idx)
 			{
 				return Property<T>::operator[](idx);
 			}
 
-			/// access the data stored for the mesh
+			/// access the data stored for the graph
 			typename Property<T>::const_reference operator[](size_t idx) const
 			{
 				return Property<T>::operator[](idx);
@@ -511,7 +515,7 @@ namespace easy3d {
 		/// default constructor
 		Graph();
 
-		// destructor (is virtual, since we inherit from Geometry_representation)
+		/// destructor
 		virtual ~Graph();
 
 		/// copy constructor: copies \c rhs to \c *this. performs a deep copy of all properties.
@@ -542,25 +546,25 @@ namespace easy3d {
 		/// \name Memory Management
 		//@{
 
-		/// returns number of (deleted and valid) vertices in the mesh
+		/// returns number of (deleted and valid) vertices in the graph
 		unsigned int vertices_size() const { return (unsigned int)vprops_.size(); }
-		/// returns number of (deleted and valid)edges in the mesh
+		/// returns number of (deleted and valid)edges in the graph
 		unsigned int edges_size() const { return (unsigned int)eprops_.size(); }
 
-		/// returns number of vertices in the mesh
+		/// returns number of vertices in the graph
 		unsigned int n_vertices() const { return vertices_size() - deleted_vertices_; }
-		/// returns number of edges in the mesh
+		/// returns number of edges in the graph
 		unsigned int n_edges() const { return edges_size() - deleted_edges_; }
 
[…]
