+++ CMakeLists.txt
@@ -7,7 +7,7 @@ project(${PROJECT_NAME})
 
 set (PROJECT_MAJOR 2)
 set (PROJECT_MINOR 4)
-set (PROJECT_PATCH 1)
+set (PROJECT_PATCH 2)
 set (PROJECT_VERSION "${PROJECT_MAJOR}.${PROJECT_MINOR}.${PROJECT_PATCH}")
 
 
