### 2020-12-27 4a3fd2e
add is_border(Vertex) and is_border(Edge)
also fix a bug where some halffaces were not added to the adjacency of an edge (only those when the edge was created were added)
