### 2020-12-27 d0ae3ce
add  SurfaceMesh::join(const SurfaceMesh& other)
