+++ easy3d/core/surface_mesh.cpp
@@ -24,8 +24,8 @@
 
 /** ----------------------------------------------------------
  *
- * the code is adapted from Surface_mesh with modifications and
- * significant enhancement.
+ * the code is adapted from Surface_mesh with significant modifications
+ * and enhancement.
  *		- Surface_mesh (version 1.1)
  * The original code is available at
  * https://opensource.cit-ec.de/projects/surface_mesh
@@ -45,8 +45,7 @@
 
 namespace easy3d {
 
-    SurfaceMesh::
-    SurfaceMesh()
+    SurfaceMesh::SurfaceMesh()
     {
         // allocate standard properties
         // same list is used in operator=() and assign()
@@ -68,8 +67,7 @@ namespace easy3d {
     //-----------------------------------------------------------------------------
 
 
-    SurfaceMesh::
-    ~SurfaceMesh()
+    SurfaceMesh::~SurfaceMesh()
     {
     }
 
@@ -77,9 +75,7 @@ namespace easy3d {
     //-----------------------------------------------------------------------------
 
 
-    SurfaceMesh&
-    SurfaceMesh::
-    operator=(const SurfaceMesh& rhs)
+    SurfaceMesh& SurfaceMesh::operator=(const SurfaceMesh& rhs)
     {
         if (this != &rhs)
         {
@@ -117,9 +113,106 @@ namespace easy3d {
     //-----------------------------------------------------------------------------
 
 
-    SurfaceMesh&
-    SurfaceMesh::
-    assign(const SurfaceMesh& rhs)
+    SurfaceMesh &SurfaceMesh::join(const SurfaceMesh &other) {
+        // increase capacity
+        const unsigned int nv = vertices_size(), nh = halfedges_size(), nf = faces_size();
+        resize(vertices_size() + other.vertices_size(),
+               edges_size() + other.edges_size(),
+               faces_size() + other.faces_size());
+
+        // append properties in the free space created by resize
+        vprops_.transfer(other.vprops_);
+        hprops_.transfer(other.hprops_);
+        fprops_.transfer(other.fprops_);
+        eprops_.transfer(other.eprops_);
+
+        // translate halfedge index in vertex -> halfedge
+        for (unsigned int i = nv; i < nv + other.vertices_size(); i++) {
+            Vertex vi(i);
+            if (vconn_[vi].halfedge_ != Halfedge()) {
+                vconn_[vi].halfedge_ = Halfedge(vconn_[vi].halfedge_.idx() + nh);
+            }
+        }
+        // translate halfedge index in face -> halfedge
+        for (unsigned int i = nf; i < nf + other.faces_size(); i++) {
+            Face fi(i);
+            if (fconn_[fi].halfedge_ != Halfedge()) {
+                fconn_[fi].halfedge_ = Halfedge(fconn_[fi].halfedge_.idx() + nh);
+            }
+        }
+        // translate indices in halfedge -> face, halfedge -> target, halfedge -> prev, and halfedge -> next
+        for (unsigned int i = nh; i < nh + other.halfedges_size(); i++) {
+            Halfedge hi(i);
+            if (hconn_[hi].face_ != Face()) {
+                hconn_[hi].face_ = Face(hconn_[hi].face_.idx() + nf);
+            }
+            if (hconn_[hi].vertex_ != Vertex()) {
+                hconn_[hi].vertex_ = Vertex(hconn_[hi].vertex_.idx() + nv);
+            }
+            if (hconn_[hi].next_ != Halfedge()) {
+                hconn_[hi].next_ = Halfedge(hconn_[hi].next_.idx() + nh);
+            }
+            if (hconn_[hi].prev_ != Halfedge()) {
+                hconn_[hi].prev_ = Halfedge(hconn_[hi].prev_.idx() + nh);
[…]
