### 2020-12-27 47c4c6b
draw camera path
