+++ applications/Mapple/paint_canvas.cpp
@@ -86,7 +86,7 @@ PaintCanvas::PaintCanvas(MainWindow* window)
         , pressed_key_(-1)
         , show_pivot_point_(false)
         , drawable_axes_(nullptr)
-        , show_camera_path_(true)
+        , show_camera_path_(false)
         , model_idx_(-1)
         , ssao_(nullptr)
         , transparency_(nullptr)
@@ -103,6 +103,8 @@ PaintCanvas::PaintCanvas(MainWindow* window)
     camera_->setViewDirection(vec3(-1, 0, 0)); // X pointing out
     camera_->showEntireScene();
     camera_->connect(this, static_cast<void (PaintCanvas::*)(void)>(&PaintCanvas::update));
+
+    kfi_ = new KeyFrameInterpolator(camera_->frame());
 }
 
 
@@ -124,6 +126,7 @@ void PaintCanvas::cleanup() {
     }
 
     delete camera_;
+    delete kfi_;
     delete drawable_axes_;
     delete ssao_;
     delete shadow_;
@@ -1048,8 +1051,8 @@ void PaintCanvas::postDraw() {
     }
 
     // shown only when it is not animating
-    if (show_camera_path_ && !camera()->keyFrameInterpolator()->interpolationIsStarted())
-        camera()->draw_paths();
+    if (show_camera_path_ && !kfi_->interpolationIsStarted())
+        kfi_->drawPath(camera());
     easy3d_debug_log_gl_error;
 
     if (show_pivot_point_) {
@@ -1251,7 +1254,8 @@ void PaintCanvas::pasteCamera() {
     new_frame.setPosition(pos);
     new_frame.setOrientation(orient);
     const float duration = 0.5f;
-    camera()->interpolateTo(new_frame, duration);
+    camera()->interpolateTo(new_frame, duration); // this will override the camera path
+
     update();
 }
 
@@ -1430,19 +1434,18 @@ void PaintCanvas::restoreStateFromFile(const std::string& file_name) {
 
 void PaintCanvas::addKeyFrame(){
     easy3d::Frame *frame = camera()->frame();
-    camera()->keyFrameInterpolator()->addKeyFrame(*frame);
-    camera()->keyFrameInterpolator()->adjust_scene_radius(camera());
-
+    kfi_->addKeyFrame(*frame);
+    kfi_->adjust_scene_radius(camera());
     LOG(INFO) << "key frame added to camera path";
     update();
 }
 
 
 void PaintCanvas::playCameraPath(){
-    if (camera()->keyFrameInterpolator()->interpolationIsStarted())
-        camera()->keyFrameInterpolator()->stopInterpolation();
+    if (kfi_->interpolationIsStarted())
+        kfi_->stopInterpolation();
     else
-        camera()->keyFrameInterpolator()->startInterpolation();
+        kfi_->startInterpolation();
 }
 
 
@@ -1450,15 +1453,14 @@ void PaintCanvas::showCamaraPath(){
     show_camera_path_ = !show_camera_path_;
 
     if (show_camera_path_)
-        camera()->keyFrameInterpolator()->adjust_scene_radius(camera());
+        kfi_->adjust_scene_radius(camera());
 
     update();
 }
 
 
 void PaintCanvas::exportCamaraPathToFile(const std::string& file_name) const {
-    KeyFrameInterpolator* interpolator = camera()->keyFrameInterpolator();
-    if (!interpolator)
+    if (!kfi_)
         return;
 
     std::ofstream output(file_name.c_str());
@@ -1467,12 +1469,12 @@ void PaintCanvas::exportCamaraPathToFile(const std::string& file_name) const {
         return;
     }
 
-    std::size_t num = interpolator->numberOfKeyFrames();
+    std::size_t num = kfi_->numberOfKeyFrames();
[…]
