### 2020-12-27 7294505
update scene bounding box when camera path is hidden
