### 2020-12-27 7bd43c5
same short cuts as Mapple
