### 2020-12-27 c464e28
add overload function add_tetra()
