### 2020-12-27 ce838ad
larger clipping plane
