+++ easy3d/fileio/poly_mesh_io_mesh.cpp
@@ -1,0 +1,272 @@
+/**
+ * Copyright (C) 2015 by Liangliang Nan (liangliang.nan@gmail.com)
+ * https://3d.bk.tudelft.nl/liangliang/
+ *
+ * This file is part of Easy3D. If it is useful in your research/work,
+ * I would be grateful if you show your appreciation by citing it:
+ * ------------------------------------------------------------------
+ *      Liangliang Nan.
+ *      Easy3D: a lightweight, easy-to-use, and efficient C++
+ *      library for processing and rendering 3D data. 2018.
+ * ------------------------------------------------------------------
+ * Easy3D is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License Version 3
+ * as published by the Free Software Foundation.
+ *
+ * Easy3D is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include <easy3d/fileio/poly_mesh_io.h>
+
+#include <iostream>
+
+#include <easy3d/core/poly_mesh.h>
+
+
+namespace easy3d {
+
+    namespace io {
+
+        bool load_mesh(const std::string &file_name, PolyMesh *mesh) {
+            if (!mesh) {
+                LOG(ERROR) << "null mesh pointer";
+                return false;
+            }
+
+            FILE *mesh_file = fopen(file_name.c_str(), "r");
+            if (NULL == mesh_file) {
+                LOG(ERROR) << "could not open file: " << file_name;
+                return false;
+            }
+
+#ifndef LINE_MAX
+#  define LINE_MAX 2048
+#endif
+
+            // eat comments at beginning of file
+            const auto eat_comments = [](FILE *file, char* a_line) -> bool {
+                bool still_comments = true;
+                bool has_line = false;
+                while (still_comments && !feof(file)) {
+                    has_line = fgets(a_line, LINE_MAX, file) != NULL;
+                    still_comments = (a_line[0] == '#' || a_line[0] == '\n');
+                }
+                return has_line;
+            };
+
+            char line[LINE_MAX];
+            eat_comments(mesh_file, line);
+
+            char str[LINE_MAX];
+            sscanf(line, " %s", str);
+            // check that first word is MeshVersionFormatted
+            if (0 != strcmp(str, "MeshVersionFormatted")) {
+                LOG(ERROR) << "first word should be MeshVersionFormatted instead of " << str;
+                fclose(mesh_file);
+                return false;
+            }
+            int version = -1;
+            if (2 != sscanf(line, "%s %d", str, &version))
+                fscanf(mesh_file, " %d", &version);
+            if (version != 1 && version != 2) {
+                LOG(ERROR) << "second word should be 1 or 2 instead of " << version;
+                fclose(mesh_file);
+                return false;
+            }
+
+            std::vector<PolyMesh::Vertex> vertices;
+//            std::vector<PolyMesh::HalfFace> faces; // PolyMesh adds the faces when adding the tetrahedra
+            while (eat_comments(mesh_file, line)) {
+                sscanf(line, " %s", str);
+                int extra;
+                // check that third word is Dimension
+                if (0 == strcmp(str, "Dimension")) {
+                    int three = -1;
+                    if (2 != sscanf(line, "%s %d", str, &three)) {
+                        // 1 appears on next line?
+                        fscanf(mesh_file, " %d", &three);
+                    }
+                    if (three != 3) {
[…]
