### 2020-12-27 54dc483
commented code for drawing face ids
