### 2020-12-28 e1ecac2
allow to highlight a range of faces
