### 2020-12-29 e3765aa
a better action name: actionSetBackgroundColor
