### 2020-12-29 40742ff
log failures for glCreateShader()
