### 2020-12-29 9fd8a4a
don't attempt again if load shader has failed
