+++ applications/Mapple/paint_canvas.cpp
@@ -54,8 +54,9 @@
 #include <easy3d/renderer/text_renderer.h>
 #include <easy3d/renderer/buffers.h>
 #include <easy3d/fileio/resources.h>
-#include <easy3d/util/logging.h>
+#include <easy3d/gui/picker_surface_mesh.h>
 #include <easy3d/util/file_system.h>
+#include <easy3d/util/logging.h>
 
 #include <QKeyEvent>
 #include <QPainter>
@@ -87,6 +88,9 @@ PaintCanvas::PaintCanvas(MainWindow* window)
         , show_pivot_point_(false)
         , drawable_axes_(nullptr)
         , show_camera_path_(false)
+        , show_labels_under_mouse_(true)
+        , picked_face_index_(-1)
+        , show_coordinates_under_mouse_(true)
         , model_idx_(-1)
         , ssao_(nullptr)
         , transparency_(nullptr)
@@ -392,11 +396,26 @@ void PaintCanvas::mouseMoveEvent(QMouseEvent *e) {
     mouse_current_pos_ = e->pos();
     QOpenGLWidget::mouseMoveEvent(e);
 
-    bool found = false;
-    makeCurrent();
-    easy3d::vec3 p = pointUnderPixel(e->pos(), found);
-    doneCurrent();
-    window_->showPointUnderMouse(p, found);
+    if (show_labels_under_mouse_) {
+        auto mesh = dynamic_cast<SurfaceMesh*>(currentModel());
+        if (mesh) {
+            makeCurrent();
+            SurfaceMeshPicker picker(camera());
+            picked_face_index_ = picker.pick_face(mesh, mouse_current_pos_.x(), mouse_current_pos_.y()).idx();
+            doneCurrent();
+        }
+        else
+            picked_face_index_ = -1;
+    }
+
+    if (show_coordinates_under_mouse_) {
+        bool found = false;
+        const easy3d::vec3 p = pointUnderPixel(e->pos(), found);
+        QString coords= "XYZ = [-, -, -]";
+        if (found)
+            coords = QString("XYZ = [%1, %2, %3]").arg(p.x).arg(p.y).arg(p.z);
+        window_->setPointUnderMouse(coords);
+    }
 
     update();
 }
@@ -1005,6 +1024,44 @@ void PaintCanvas::drawCornerAxes() {
 }
 
 
+void PaintCanvas::drawFaceAndVertexLabels(const QColor& face_color, const QColor& vertex_color) {
+    SurfaceMesh* mesh = dynamic_cast<SurfaceMesh*>(currentModel());
+    if (!mesh || picked_face_index_ < 0 || picked_face_index_ >= mesh->n_faces())
+        return;
+
+    // using QPainter
+    QPainter painter;
+    painter.begin(this);
+    painter.setRenderHint(QPainter::HighQualityAntialiasing);
+    painter.setRenderHint(QPainter::TextAntialiasing);
+    painter.beginNativePainting();
+
+    // draw labels for the vertices of the picked face
+    vec3 face_center(0, 0, 0);
+    int valence = 0;
+    for (auto v : mesh->vertices(SurfaceMesh::Face(picked_face_index_))) {
+        const auto& p = mesh->position(v);
+        face_center += p;
+        const vec3 proj = camera()->projectedCoordinatesOf(p);
+        painter.setPen(vertex_color);
+        painter.drawText(proj.x, proj.y, QString::fromStdString("v" + std::to_string(v.idx())));
+        ++valence;
+    }
+
+    // draw label for the picked face
+    const vec3 proj = camera()->projectedCoordinatesOf(face_center/valence);
[…]
