### 2020-12-29 6b30307
by default don't show coordinates and IDs of face (and its vertices) under mouse
