+++ applications/Mapple/paint_canvas.cpp
@@ -88,9 +88,9 @@ PaintCanvas::PaintCanvas(MainWindow* window)
         , show_pivot_point_(false)
         , drawable_axes_(nullptr)
         , show_camera_path_(false)
-        , show_labels_under_mouse_(true)
+        , show_labels_under_mouse_(false)
         , picked_face_index_(-1)
-        , show_coordinates_under_mouse_(true)
+        , show_coordinates_under_mouse_(false)
         , model_idx_(-1)
         , ssao_(nullptr)
         , transparency_(nullptr)
