### 2020-12-29 0497c88
add a few shortcuts for camera path manipulation
