### 2020-12-29 b31c771
log has_nan() for camera->projectedCoordinatesOf(pivotPoint()).
