+++ easy3d/renderer/manipulated_camera_frame.cpp
@@ -76,8 +76,12 @@ namespace easy3d {
 
         const vec3& trans = camera->projectedCoordinatesOf(pivotPoint());
         DLOG_IF(ERROR, has_nan(trans))
-                        << "camera->projectedCoordinatesOf(pivotPoint()): " << trans << ", pivotPoint(): "
-                        << pivotPoint();
+                        << "projectedCoordinatesOf(pivotPoint()): " << trans
+                        << ", pivotPoint(): " << pivotPoint()
+                        << ", camera position: " << camera->position()
+                        << ", camera orientation: " << camera->orientation();
+        if (has_nan(trans))
+            return;
 
 		if (screen) {
             const float pre_x = float(x - dx);
