+++ applications/Mapple/paint_canvas.cpp
@@ -52,7 +52,7 @@
 #include <easy3d/renderer/clipping_plane.h>
 #include <easy3d/renderer/texture_manager.h>
 #include <easy3d/renderer/text_renderer.h>
-#include <easy3d/renderer/buffers.h>
+#include <easy3d/renderer/walk_throuth.h>
 #include <easy3d/fileio/resources.h>
 #include <easy3d/gui/picker_surface_mesh.h>
 #include <easy3d/util/file_system.h>
@@ -87,7 +87,6 @@ PaintCanvas::PaintCanvas(MainWindow* window)
         , pressed_key_(-1)
         , show_pivot_point_(false)
         , drawable_axes_(nullptr)
-        , show_camera_path_(false)
         , show_labels_under_mouse_(false)
         , picked_face_index_(-1)
         , show_coordinates_under_mouse_(false)
@@ -108,7 +107,7 @@ PaintCanvas::PaintCanvas(MainWindow* window)
     camera_->showEntireScene();
     camera_->connect(this, static_cast<void (PaintCanvas::*)(void)>(&PaintCanvas::update));
 
-    kfi_ = new KeyFrameInterpolator(camera_->frame());
+    walk_through_ = new WalkThrough(camera_);
 }
 
 
@@ -130,7 +129,7 @@ void PaintCanvas::cleanup() {
     }
 
     delete camera_;
-    delete kfi_;
+    delete walk_through_;
     delete drawable_axes_;
     delete ssao_;
     delete shadow_;
@@ -277,6 +276,14 @@ void PaintCanvas::mousePressEvent(QMouseEvent *e) {
                     camera()->interpolateToFitScene();
             }
         }
+
+        if (e->modifiers() == Qt::ControlModifier) {
+            bool found = false;
+            const vec3 p = pointUnderPixel(e->pos(), found);
+            if (found && walkThrough()->is_active() && !walkThrough()->interpolator()->interpolationIsStarted()) {
+                walkThrough()->add_position(p);
+            }
+        }
     }
 
     QOpenGLWidget::mousePressEvent(e);
@@ -351,7 +358,7 @@ void PaintCanvas::mouseReleaseEvent(QMouseEvent *e) {
 
 void PaintCanvas::mouseMoveEvent(QMouseEvent *e) {
     int x = e->pos().x(), y = e->pos().y();
-    if (x < 0 || x > width() || y < 0 || y > height()) {
+    if (x < 0 || x > width() || y < 0 || y > height() || walkThrough()->interpolator()->interpolationIsStarted()) {
         e->ignore();
         return;
     }
@@ -935,6 +942,8 @@ void PaintCanvas::paintGL() {
 
     // Add visual hints: axis, camera, grid...
     postDraw();
+
+    Q_EMIT drawFinished();
 }
 
 
@@ -1108,8 +1117,8 @@ void PaintCanvas::postDraw() {
     }
 
     // shown only when it is not animating
-    if (show_camera_path_ && !kfi_->interpolationIsStarted())
-        kfi_->draw_path(camera());
+    if (walk_through_ && walk_through_->show_path() && !walk_through_->interpolator()->interpolationIsStarted())
+        walk_through_->draw();
     easy3d_debug_log_gl_error;
 
     if (show_pivot_point_) {
@@ -1280,6 +1289,9 @@ void PaintCanvas::copyCamera() {
 
 
 void PaintCanvas::pasteCamera() {
+    if (walkThrough()->interpolator()->interpolationIsStarted())
+        return;
+
     // get the camera parameters from clipboard string
     const QString str = qApp->clipboard()->text();
     const QStringList list = str.split(" ", QString::SkipEmptyParts);
[…]
