### 2020-12-30 7836b90
fix the error of call timer from another thread
