### 2020-12-30 1297d71
rename Signal::trigger to Signal::send; clearer signal-slot relationship
