### 2020-12-30 d2b892b
rename Signal::trigger to Signal::send; clearer signal-slot relationship
