### 2020-12-30 71beaa6
add a few view options
