### 2020-12-30 afce972
finer tuning of parameters
