### 2020-12-30 0480fb1
set pivot point ahead after each step
