+++ easy3d/renderer/primitives.h
@@ -192,10 +192,11 @@ namespace easy3d {
 
         /**
          * \brief Prepares data (points) for representing a camera in the 3D world as a set of lines.
-         * \param size The size of the camera. A good value can be, e.g., 5% of the scene raidus.
+         * \param width The width of the camera. A good value can be 5% of the scene radius, or 10% of the
+         *      character height (in walking mode).
          * \param hw_ratio The aspect ratio of the base quad defined as height/width (default 0.6).
          */
-        void prepare_camera(std::vector<vec3>& points, float size, float hw_ratio = 0.6f);
+        void prepare_camera(std::vector<vec3>& points, float width, float hw_ratio = 0.6f);
     }
 
 }
