### 2020-12-30 017d9ee
provide camera size
