### 2020-12-30 0e34fd9
dpi_scaling() is float
