### 2020-12-31 b2e19db
set scene radius before creating a camera path
