+++ easy3d/util/timer.h
@@ -48,53 +48,53 @@ namespace easy3d {
      ** Code examples:
      *### Example 1: calling to non-member functions
      *   // the function you want to call at a constant interval
-     *   \code
-     *   void foo(int value);
-     *
-     *   // the function to be called when timeout.
-     *   void timeout();
-     *
-     *   const int value = 333;
-     *   // function "foo" will be executed every second.
-     *   t.set_interval(1000, &foo, value);
-     *
-     *   // function "timeout()" will be executed after 3 seconds.
-     *   t.set_timeout(3000, &timeout);
-     *   \endcode
+     *      \code
+     *          void foo(int value);
+     *      
+     *          // the function to be called when timeout.
+     *          void timeout();
+     *      
+     *          const int value = 333;
+     *          // function "foo" will be executed every second.
+     *          t.set_interval(1000, &foo, value);
+     *      
+     *          // function "timeout()" will be executed after 3 seconds.
+     *          t.set_timeout(3000, &timeout);
+     *      \endcode
      *
      *### Example 2: calling to member functions
-     *    \code
-     *   // a trivial class
-     *   class Car {
-     *   public:
-     *       Car(const std::string& name, float speed);
-     *       void print_speed();
-     *       void stop();
-     *   };
-     *
-     *   // car 1 reports its speed every 2 seconds and it stops after 10 seconds
-     *   Car car1("BMW", 180);
-     *   t.set_interval(2000, &Car::print_speed, &car1);
-     *   t.set_timeout(10000, &Car::stop, &car1);
-     *
-     *   // car 2 reports its speed every 3 seconds and it stops after 20 seconds
-     *   Car car2("Chevrolet", 120);
-     *  t.set_interval(3000, &Car::print_speed, &car2);
-     *  t.set_timeout(20000, &Car::stop, &car2);
-     *  \endcode
+     *      \code
+     *          // a trivial class
+     *          class Car {
+     *          public:
+     *              Car(const std::string& name, float speed);
+     *              void print_speed();
+     *              void stop();
+     *          };
+     *      
+     *          // car 1 reports its speed every 2 seconds and it stops after 10 seconds
+     *          Car car1("BMW", 180);
+     *          t.set_interval(2000, &car1, &Car::print_speed);
+     *          t.set_timeout(10000, &car1, &Car::stop);
+     *      
+     *          // car 2 reports its speed every 3 seconds and it stops after 20 seconds
+     *          Car car2("Chevrolet", 120);
+     *          t.set_interval(3000, &car2, &Car::print_speed);
+     *          t.set_timeout(20000, &car2, &Car::stop);
+     *      \endcode
      *
      *### Example 3: calling to lambda functions
-     *    \code
-     *  const float value = 5;
-     *  t.set_interval(3000, [&](float, const std::string&) {
-     *      std::cout << "After every 3 sec. value: " << value << ", message: " << msg << std::endl;
-     *  }, value, msg);
-     *
-     *  t.set_timeout(8000, [&]() {
-     *      t.stop();
-     *      std::cout << "After 8 sec, the timer is stopped!" << std::endl;
-     *  });
-     *  \endcode
+     *      \code
+     *          const float value = 5;
+     *          t.set_interval(3000, [&](float, const std::string&) {
+     *              std::cout << "After every 3 sec. value: " << value << ", message: " << msg << std::endl;
+     *          }, value, msg);
+     *      
+     *          t.set_timeout(8000, [&]() {
+     *              t.stop();
+     *              std::cout << "After 8 sec, the timer is stopped!" << std::endl;
+     *          });
+     *      \endcode
      */
 
     class Timer {
@@ -102,93 +102,117 @@ namespace easy3d {
[…]
