### 2020-12-31 be78b95
add commented code for animating using only the keyframes
