### 2020-12-31 c564e44
simplify code
