+++ easy3d/renderer/key_frame_interpolator.cpp
@@ -45,42 +45,16 @@
 #include <easy3d/renderer/camera.h>   // for scene radius (to draw cameras in a proper size)
 
 
-
-//#define DEBUG_INTERPOLATED_FRAMES
-#ifdef DEBUG_INTERPOLATED_FRAMES
-static int num_frames = 0;
-static std::vector<easy3d::Frame> frames;
-
-void save_interpolation() {
-    std::string file = "interpolated_frames.path";
-    std::ofstream output(file.c_str());
-    if (output.fail())
-        std::cout << "could not open file: " << file << std::endl;
-    output << "\tnum_key_frames: " << frames.size() << std::endl;
-    for (std::size_t id = 0; id < frames.size(); ++id) {
-        output << "\tframe: " << id << std::endl;
-        const easy3d::Frame& frame = frames[id];
-        output << "\t\tposition: " << frame.position() << std::endl;
-        output << "\t\torientation: " << frame.orientation() << std::endl;
-    }
-}
-
-#endif
-
-
 namespace easy3d {
 
 
     KeyFrameInterpolator::KeyFrameInterpolator(Frame *frame)
             : frame_(frame), period_(40)   // 25 frames per second
             , interpolationTime_(0.0), interpolationSpeed_(1.0), interpolationStarted_(false),
-              loopInterpolation_(false), pathIsValid_(false), valuesAreValid_(true), currentFrameValid_(false),
+              loopInterpolation_(false), pathIsValid_(false),
               path_drawable_(nullptr), cameras_drawable_(nullptr) {
         if (keyFrames_.size() < 2)
             return;
-
-        for (int i = 0; i < 4; ++i)
-            currentFrame_[i] = keyFrames_.begin();
     }
 
 
@@ -131,10 +105,6 @@ namespace easy3d {
                 stopInterpolation();
             }
             end_reached.send();
-#ifdef DEBUG_INTERPOLATED_FRAMES
-            std::cout << "total frames: " << num_frames << std::endl;
-            save_interpolation();
-#endif
         } else if (interpolationTime() < keyFrames_.front()->time()) {
             if (loopInterpolation())
                 setInterpolationTime(keyFrames_.back()->time() - keyFrames_.front()->time() + interpolationTime_);
@@ -144,38 +114,22 @@ namespace easy3d {
                 stopInterpolation();
             }
             end_reached.send();
-#ifdef DEBUG_INTERPOLATED_FRAMES
-            std::cout << "total frames: " << num_frames << std::endl;
-            save_interpolation();
-#endif
         }
-
-#ifdef DEBUG_INTERPOLATED_FRAMES
-        ++num_frames;
-#endif
         mutex.unlock();
     }
 
 
-    void KeyFrameInterpolator::startInterpolation(int period) {
-        if (period >= 0)
-            setInterpolationPeriod(period);
+    void KeyFrameInterpolator::startInterpolation() {
+        if (keyFrames_.empty())
+            return;
 
-        if (!keyFrames_.empty()) {
-            if ((interpolationSpeed() > 0.0) && (interpolationTime() >= keyFrames_.back()->time()))
-                setInterpolationTime(keyFrames_.front()->time());
-            if ((interpolationSpeed() < 0.0) && (interpolationTime() <= keyFrames_.front()->time()))
-                setInterpolationTime(keyFrames_.back()->time());
-            interpolationStarted_ = true;
-
-#ifdef DEBUG_INTERPOLATED_FRAMES
-            num_frames = 0;
-            frames.clear();
-#endif
-            update();  // for the starting view point
-
-            timer_.set_interval(interpolationPeriod(), this, &KeyFrameInterpolator::update);
-        }
[…]
