+++ easy3d/renderer/key_frame_interpolator.h
@@ -126,8 +126,8 @@ namespace easy3d {
       KeyFrameInterpolator kfi( new Frame() );
       // calls to kfi.addKeyFrame() to define the path.
 
-      const double deltaTime = 0.04; // output a position every deltaTime seconds
-      for (double time=kfi.firstTime(); time<=kfi.lastTime(); time += deltaTime)
+      const float deltaTime = 0.04f; // output a position every deltaTime seconds
+      for (float time=kfi.firstTime(); time<=kfi.lastTime(); time += deltaTime)
       {
         kfi.interpolateAtTime(time);
         cout << "t=" << time << "\tpos=" << kfi.frame()->position() << endl;
@@ -150,14 +150,14 @@ namespace easy3d {
         //@{
     public:
         /*! Appends a new keyFrame to the path.
-         * Same as addKeyFrame(const Frame& frame, double), except that the keyFrameTime() is automatically set
+         * Same as addKeyFrame(const Frame& frame, float), except that the keyFrameTime() is automatically set
          * to previous keyFrameTime() plus one second (or 0.0 if there is no previous keyFrame). */
         void addKeyFrame(const Frame& frame);
 
         /*! Appends a new keyFrame to the path, with its associated \p time (in seconds).
          * The path will use the current \p frame state.
          * \attention The keyFrameTime() have to be monotonously increasing over keyFrames. */
-        void addKeyFrame(const Frame& frame, double time);
+        void addKeyFrame(const Frame& frame, float time);
 
         /*! Removes all keyFrames from the path. The numberOfKeyFrames() is set to 0. */
         void deletePath();
@@ -191,27 +191,27 @@ namespace easy3d {
          * \note If this keyFrame was defined using a pointer to a Frame (see addKeyFrame(const Frame* const)),
          *      the \e current pointed Frame state is returned.
          */
-        Frame keyFrame(int index) const;
+        Frame keyFrame(std::size_t index) const;
 
         /*! Returns the time corresponding to the \p index keyFrame.
          * See also keyFrame(). \p index has to be in the range [0, numberOfKeyFrames()-1].
          */
-        double keyFrameTime(int index) const;
+        float keyFrameTime(std::size_t index) const;
 
         /*! Returns the duration of the KeyFrameInterpolator path, expressed in seconds.
          * Simply corresponds to lastTime() - firstTime(). Returns 0.0 if the path has less than 2 keyFrames.
          * \sa keyFrameTime(). */
-        double duration() const;
+        float duration() const;
 
         /*! Returns the time corresponding to the first keyFrame, expressed in seconds.
          * Returns 0.0 if the path is empty.
          * \sa lastTime(), duration() and keyFrameTime(). */
-        double firstTime() const;
+        float firstTime() const;
 
         /*! Returns the time corresponding to the last keyFrame, expressed in seconds.
          * Returns 0.0 if the path is empty.
          * \sa firstTime(), duration() and keyFrameTime(). */
-        double lastTime() const;
+        float lastTime() const;
         //@}
 
         /*! @name Interpolation parameters */
@@ -221,7 +221,7 @@ namespace easy3d {
 
         This time is regularly updated when interpolationIsStarted(). Can be set directly with
         setInterpolationTime() or interpolateAtTime(). */
-        double interpolationTime() const { return interpolationTime_; }
+        float interpolationTime() const { return interpolationTime_; }
         /*! Returns the current interpolation speed.
 
         Default value is 1.0, which means keyFrameTime() will be matched during the interpolation
@@ -229,7 +229,7 @@ namespace easy3d {
 
         A negative value will result in a reverse interpolation of the keyFrames. See also
         interpolationPeriod(). */
[…]
