### 2020-12-31 83e41ab
use compensated tangent for path interpolation
