### 2020-12-31 19b6f3a
add is_stopped()
