### 2020-12-31 44a614f
simpler API
