+++ easy3d/renderer/key_frame_interpolator.h
@@ -59,7 +59,7 @@ namespace easy3d {
      \note The default duration of any two consecutive keyframes is the same. So for a smoother animation, it is
      suggested to regularly (as equally as possible) sample the viewpoints.
      \todo Allow edit the duration for each keyframe?  (not sure due to many keyframes and can be annoying).
-     
+
       Here is a typical utilization example (see also the <a href="../examples/keyFrames.html">keyFrames
       example</a>):
       \code
@@ -220,11 +220,6 @@ namespace easy3d {
         /*! @name Interpolation parameters */
         //@{
     public:
-        /*! Returns the current interpolation time (in seconds) along the KeyFrameInterpolator path.
-
-        This time is regularly updated when interpolationIsStarted(). Can be set directly with
-        setInterpolationTime() or interpolateAtTime(). */
-        float interpolationTime() const { return interpolationTime_; }
         /*! Returns the current interpolation speed.
 
         Default value is 1.0, which means keyFrameTime() will be matched during the interpolation
@@ -242,30 +237,12 @@ namespace easy3d {
         update, and the frame() state is modified accordingly (see interpolateAtTime()). Default value
         is 40 milliseconds. */
         int interpolationPeriod() const { return period_; }
-        /*! Returns \c true when the interpolation is played in an infinite loop.
-
-        When \c false (default), the interpolation stops when interpolationTime() reaches firstTime()
-        (with negative interpolationSpeed()) or lastTime().
-
-        interpolationTime() is otherwise reset to firstTime() (+ interpolationTime() - lastTime()) (and
-        inversely for negative interpolationSpeed()) and interpolation continues.
-
-        In both cases, the endReached() signal is emitted. */
-        bool loopInterpolation() const { return loopInterpolation_; }
 
     public:
-        /*! Sets the interpolationTime().
-
-        \attention The frame() state is not affected by this method. Use this function to define the
-        starting time of a future interpolation (see startInterpolation()). Use interpolateAtTime() to
-        actually interpolate at a given time. */
-        void setInterpolationTime(float time) { interpolationTime_ = time; }
         /*! Sets the interpolationSpeed(). Negative or null values are allowed. */
         void setInterpolationSpeed(float speed) { interpolationSpeed_ = speed; }
         /*! Sets the interpolationPeriod(). */
         void setInterpolationPeriod(int period) { period_ = period; }
-        /*! Sets the loopInterpolation() value. */
-        void setLoopInterpolation(bool loop=true) { loopInterpolation_ = loop; }
         //@}
 
         /*! @name Interpolation */
@@ -295,29 +272,16 @@ namespace easy3d {
          * \sa startInterpolation(), interpolationIsStarted(), and toggleInterpolation(). */
         void stopInterpolation();
 
-        /*! Stops the interpolation and resets interpolationTime() to the firstTime().
-         * If desired, call interpolateAtTime() after this method to actually move the frame() to firstTime(). */
-        void resetInterpolation();
-
         /*! Calls startInterpolation() or stopInterpolation(), depending on interpolationIsStarted(). */
         void toggleInterpolation() { if (interpolationIsStarted()) stopInterpolation(); else startInterpolation(); }
 
-        /*! Interpolate frame() at time \p time (expressed in seconds). interpolationTime() is set
-         * to \p time and frame() is set accordingly.
-         *
-         * If you simply want to change interpolationTime() but not the frame() state, use setInterpolationTime()
-         * instead.  */
-        virtual void interpolateAtTime(float time);
-        //@}
-
-        /*! @name Computes the interpolation */
-        //@{
-    public:
[…]
