+++ easy3d/renderer/key_frame_interpolator.cpp
@@ -37,7 +37,8 @@
 #include <easy3d/renderer/key_frame_interpolator.h>
 
 #include <fstream>
-#include <mutex>
+#include <chrono>
+#include <thread>
 
 #include <easy3d/renderer/frame.h>
 #include <easy3d/renderer/drawable_lines.h>
@@ -68,8 +69,7 @@ void save_interpolation(const std::vector<easy3d::Frame>& frames) {
 
     KeyFrameInterpolator::KeyFrameInterpolator(Frame *frame)
             : frame_(frame), period_(40)   // 25 frames per second
-            , interpolationTime_(0.0), interpolationSpeed_(1.0), interpolationStarted_(false),
-              loopInterpolation_(false), pathIsValid_(false),
+            , interpolationSpeed_(1.0), interpolationStarted_(false), start_position_(0), pathIsValid_(false),
               path_drawable_(nullptr), cameras_drawable_(nullptr) {
         if (keyFrames_.size() < 2)
             return;
@@ -86,68 +86,32 @@ void save_interpolation(const std::vector<easy3d::Frame>& frames) {
     }
 
 
-    /*! Updates frame() state according to current interpolationTime(). Then adds
-      interpolationPeriod()*interpolationSpeed() to interpolationTime().
-
-      This internal method is called by a timer when interpolationIsStarted(). It can be used for
-      debugging purpose. stopInterpolation() is called when interpolationTime() reaches firstTime() or
-      lastTime(), unless loopInterpolation() is \c true. */
-    void KeyFrameInterpolator::update() {
-#if 0 // animate using only the keyframes (i.e., no interpolation)
-        static int idx = 0;
-        if (idx >= numberOfKeyFrames())
-            return;
-        Frame f = keyFrame(idx++);
-        frame()->setPositionAndOrientation(f.position(), f.orientation());
-        if (idx == numberOfKeyFrames()) {
-            idx = 0; // so I can restart again
-            stopInterpolation();
-            end_reached.send();
-        }
-        return;
-#endif
-
-        static std::mutex mutex;
-        mutex.try_lock();
-
-        interpolateAtTime(interpolationTime());
-
-        interpolationTime_ += interpolationSpeed() * interpolationPeriod() / 1000.0f;
-
-        if (interpolationTime() > keyFrames_.back()->time()) {
-            if (loopInterpolation())
-                setInterpolationTime(keyFrames_.front()->time() + interpolationTime_ - keyFrames_.back()->time());
-            else {
-                // Make sure last KeyFrame is reached and displayed
-                interpolateAtTime(keyFrames_.back()->time());
-                stopInterpolation();
-            }
-            end_reached.send();
-        } else if (interpolationTime() < keyFrames_.front()->time()) {
-            if (loopInterpolation())
-                setInterpolationTime(keyFrames_.back()->time() - keyFrames_.front()->time() + interpolationTime_);
-            else {
-                // Make sure first KeyFrame is reached and displayed
-                interpolateAtTime(keyFrames_.front()->time());
-                stopInterpolation();
-            }
-            end_reached.send();
-        }
-        mutex.unlock();
-    }
-
-
     void KeyFrameInterpolator::startInterpolation() {
         if (keyFrames_.empty())
             return;
 
-        if ((interpolationSpeed() > 0.0f) && (interpolationTime() >= keyFrames_.back()->time()))
-            setInterpolationTime(keyFrames_.front()->time());
-        if ((interpolationSpeed() < 0.0f) && (interpolationTime() <= keyFrames_.front()->time()))
-            setInterpolationTime(keyFrames_.back()->time());
+        if (!pathIsValid_)
+            interpolate();
+
+        // all done in another thread.
[…]
