### 2021-01-01 a9483ff
QVideoEncoder compiles but not linking
