+++ easy3d/QVideoEncoder.cpp
@@ -1,0 +1,420 @@
+#include "QVideoEncoder.h"
+
+//FFmpeg
+extern "C"
+{
+	#include "libavcodec/avcodec.h"
+	#include "libavformat/avformat.h"
+	#include "libavformat/avio.h"
+
+	#include "libavutil/mathematics.h"
+	#include "libavutil/opt.h"
+	#include "libavutil/error.h"
+	#include "libavutil/rational.h"
+	#include "libavutil/frame.h"
+	#include "libavutil/avstring.h"
+	#include "libavutil/imgutils.h"
+
+	#include "libswscale/swscale.h"
+}
+
+//system
+#include <assert.h>
+
+struct FFmpegStuffEnc
+{
+	AVFormatContext *formatContext;
+	AVCodecContext *codecContext;
+	AVStream *videoStream;
+	AVFrame *frame;
+	SwsContext *swsContext;
+
+	FFmpegStuffEnc()
+		: formatContext(0)
+		, codecContext(0)
+		, videoStream(0)
+		, frame(0)
+		, swsContext(0)
+	{}
+};
+
+QVideoEncoder::QVideoEncoder(QString filename, int width, int height, unsigned bitrate, int gop, int fps)
+	: m_filename(filename)
+	, m_width(width)
+	, m_height(height)
+	, m_bitrate(bitrate)
+	, m_gop(gop)
+	, m_fps(fps)
+	, m_isOpen(false)
+	, m_ff(new FFmpegStuffEnc)
+{
+}
+
+QVideoEncoder::~QVideoEncoder()
+{
+	close();
+
+	if (m_ff)
+	{
+		delete m_ff;
+		m_ff = 0;
+	}
+}
+
+bool QVideoEncoder::isSizeValid()
+{
+	return (m_width % 8) == 0 && (m_height % 8) == 0;
+}
+
+bool QVideoEncoder::initFrame()
+{
+	assert(!m_ff->frame);
+
+	m_ff->frame = av_frame_alloc();
+	if (!m_ff->frame)
+	{
+		return false;
+	}
+
+	m_ff->frame->format = m_ff->codecContext->pix_fmt;
+	m_ff->frame->width  = m_ff->codecContext->width;
+	m_ff->frame->height = m_ff->codecContext->height;
+	m_ff->frame->pts = 0;
+
+    /* allocate the buffers for the frame data */
+    int ret = av_frame_get_buffer(m_ff->frame, 32);
+    if (ret < 0)
+	{
+        fprintf(stderr, "Could not allocate frame data.\n");
+		return false;
+    }
+
+	return true;
+}
+
+void QVideoEncoder::freeFrame()
+{
+	if (m_ff->frame)
+	{
+		av_free(m_ff->frame);
+		m_ff->frame = 0;
+	}
+}
+
+bool QVideoEncoder::open(QString* errorString/*=0*/)
+{
+	if (m_isOpen)
+	{
+		//the stream is already opened!
+		return false;
+	}
+
+	if (!isSizeValid())
+	{
+		if (errorString)
+			*errorString = "Invalid video size";
+		return false;
+	}
+
+	// find the output format
+	avformat_alloc_output_context2(&m_ff->formatContext, NULL, NULL, qPrintable(m_filename));
+	if (!m_ff->formatContext)
+	{
+		if (errorString)
+			*errorString = "Could not deduce output format from file extension: using MPEG";
+		
+		avformat_alloc_output_context2(&m_ff->formatContext, NULL, "mpeg", qPrintable(m_filename));
+		if (!m_ff->formatContext)
+		{
+			if (errorString)
+				*errorString = "Codec not found";
+			return false;
+		}
+	}
+
[…]
