+++ applications/Mapple/dialogs/dialog_walk_through.cpp
@@ -53,7 +53,7 @@ DialogWalkThrough::DialogWalkThrough(MainWindow *window)
     connect(clearPathButton, SIGNAL(clicked()), this, SLOT(clearPath()));
 
     connect(previewButton, SIGNAL(toggled(bool)), this, SLOT(preview(bool)));
-    connect(recordButton, SIGNAL(clicked()), this, SLOT(record()));
+    connect(recordButton, SIGNAL(toggled(bool)), this, SLOT(record(bool)));
 
     easy3d::connect(&walkThrough()->path_modified, this, &DialogWalkThrough::newPositionAdded);
 }
@@ -61,7 +61,6 @@ DialogWalkThrough::DialogWalkThrough(MainWindow *window)
 
 DialogWalkThrough::~DialogWalkThrough()
 {
-    easy3d::disconnect(&walkThrough()->path_modified, this);
 }
 
 
@@ -182,6 +181,23 @@ void DialogWalkThrough::clearPath() {
 }
 
 
+void DialogWalkThrough::enableAllButtons(bool b) {
+    labelDefineStartingPoint->setEnabled(b);
+    labelDefinePath->setEnabled(b);
+    labelCharacterHeightFactor->setEnabled(b);
+    labelCharacterDistanceToEye->setEnabled(b);
+    doubleSpinBoxCharacterHeightFactor->setEnabled(b);
+    doubleSpinBoxCharacterDistanceFactor->setEnabled(b);
+    previousPositionButton->setEnabled(b);
+    nextPositionButton->setEnabled(b);
+    removeLastPositionButton->setEnabled(b);
+    horizontalSliderPreview->setEnabled(b);
+    previewButton->setEnabled(b);
+    recordButton->setEnabled(b);
+    clearPathButton->setEnabled(b);
+}
+
+
 void DialogWalkThrough::preview(bool b) {
     if (walkThrough()->num_positions() == 0 && walkThrough()->interpolator()->numberOfKeyFrames() == 0) {
         previewButton->setChecked(false);
@@ -191,18 +207,8 @@ void DialogWalkThrough::preview(bool b) {
     auto interpolationStopped = [this]() -> void { emit animationStopped(); };
 
     if (b) {
-        labelDefineStartingPoint->setEnabled(false);
-        labelDefinePath->setEnabled(false);
-        labelCharacterHeightFactor->setEnabled(false);
-        labelCharacterDistanceToEye->setEnabled(false);
-        doubleSpinBoxCharacterHeightFactor->setEnabled(false);
-        doubleSpinBoxCharacterDistanceFactor->setEnabled(false);
-        previousPositionButton->setEnabled(false);
-        nextPositionButton->setEnabled(false);
-        removeLastPositionButton->setEnabled(false);
-        horizontalSliderPreview->setEnabled(false);
-        recordButton->setEnabled(false);
-        clearPathButton->setEnabled(false);
+        enableAllButtons(false);
+        previewButton->setEnabled(true);
 
         easy3d::connect(&walkThrough()->interpolator()->interpolation_stopped, 0, interpolationStopped);
         QObject::connect(this, &DialogWalkThrough::animationStopped, this, &DialogWalkThrough::resetUIAfterAnimationStopped);
@@ -211,18 +217,7 @@ void DialogWalkThrough::preview(bool b) {
         LOG(INFO) << "animation started...";
     }
     else {
-        labelDefineStartingPoint->setEnabled(true);
-        labelDefinePath->setEnabled(true);
-        labelCharacterHeightFactor->setEnabled(true);
-        labelCharacterDistanceToEye->setEnabled(true);
-        doubleSpinBoxCharacterHeightFactor->setEnabled(true);
[…]
