### 2021-01-01 0030688
fix viewer size when recoring video
