### 2021-01-01 b582b86
Update paint_canvas_snapshot.cpp
