### 2021-01-01 5442908
load camera path if found along side the model
