### 2021-01-01 cf07014
return load/save status
