+++ easy3d/renderer/key_frame_interpolator.h
@@ -301,9 +301,9 @@ namespace easy3d {
         //@{
     public:
         /// saves the camera path to a file
-        void save_keyframes(const std::string& file_name) const;
+        bool save_keyframes(const std::string& file_name) const;
         /// reads camera path from a file
-        void read_keyframes(const std::string& file_name);
+        bool read_keyframes(const std::string& file_name);
         //@}
 
     private:
