+++ applications/Mapple/paint_canvas_snapshot.cpp
@@ -36,12 +36,15 @@
 
 #include <QOpenGLFunctions>
 #include <QMessageBox>
+#include <QApplication>
 
 #include "paint_canvas.h"
 #include "main_window.h"
 
 #include <easy3d/renderer/camera.h>
 #include <easy3d/renderer/transform.h>
+#include <easy3d/renderer/walk_throuth.h>
+#include <easy3d/renderer/key_frame_interpolator.h>
 #include <easy3d/core//model.h>
 #include <easy3d/util/logging.h>
 #include <easy3d/util/file_system.h>
@@ -207,14 +210,9 @@ bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_na
 }
 
 
-#include <QFileInfo>
-#include <QFileDialog>
+#if 0
 
-#include <easy3d/renderer/walk_throuth.h>
-#include <easy3d/renderer/key_frame_interpolator.h>
-
-// true to start the recording and false to stop the recording.
-void PaintCanvas::renderToImages() {
+void PaintCanvas::recordAnimation(const QString &file_name, bool bk_white) {
     auto kfi = walkThrough()->interpolator();
     if (kfi->numberOfKeyFrames() == 0) {
         LOG(WARNING) << "recording aborted (camera path is empty). You may import a camera path from a file or"
@@ -222,107 +220,145 @@ void PaintCanvas::renderToImages() {
         return;
     }
 
-    // must be static because of the lambda function called in the animation thread
-    static QString recordDir;
-    static int snapshotCounter = 0;
-    static QMetaObject::Connection connection;
-
-    auto snapshotFramebuffer = [&]() -> void {
-        const QString count = QString("%1").arg(snapshotCounter++, 4, 10, QChar('0'));
-        const QString fileName = recordDir + QString("/snapshot-") + count + QString(".png");
-        raise(); // to correctly grab the frame buffer, the viewer window must be raised in front of other windows
-        const QImage snapshot = grabFramebuffer();
-        if (!snapshot.save(fileName, "png"))
-            LOG(WARNING) << "unable to save snapshot in " << fileName.toStdString();
-    };
-
-//    auto interpolationStopped = [this]() -> void { emit recordingStopped(); };
-//
-//    if (b) {
-//        std::string model_dir = file_system::parent_directory(currentModel()->name());
-//        recordDir = QFileDialog::getExistingDirectory(
-//                this, tr("Please choose a directory"), QString::fromStdString(model_dir)
-//        );
-//        if (recordDir.isEmpty()) {
-//            window_->stopRecordAnimation();
-//            return false;
-//        }
-//
-//        snapshotCounter = 0;
-//        connection = QObject::connect(this, &PaintCanvas::drawFinished, snapshotFramebuffer);
-//
-//#if 0
-//        // Liangliang: MainWindow::stopRecordAnimation will be called from the animation thread, which causes the
-//        //             error: "QObject::startTimer: Timers cannot be started from another thread". To workaround,
-//        //             I invoke MainWindow::stopRecordAnimation by a viewer signal.
-////        easy3d::connect(kfi, window_, &MainWindow::stopRecordAnimation);
-//#else
-//        easy3d::connect(&kfi->interpolation_stopped, 0, interpolationStopped);
-//        QObject::connect(this, &PaintCanvas::recordingStopped, window_, &MainWindow::stopRecordAnimation);
-//#endif
-//
-//        if (kfi->interpolationIsStarted()) // stop first if is playing now
-//            kfi->stopInterpolation();
-//
-//        if (walk_through_->interpolator()->interpolationIsStarted())
-//            walk_through_->interpolator()->stopInterpolation();
-//        else
[…]
