### 2021-01-01 1247625
fix a bug when progress deleted after the destruction of the mainwinow
