### 2021-01-01 7a5dc62
add recordAnimation() using ffmpeg
