+++ applications/Mapple/paint_canvas.cpp
@@ -1530,61 +1530,6 @@ void PaintCanvas::restoreStateFromFile(const std::string& file_name) {
 }
 
 
-void PaintCanvas::addKeyFrame() {
-    easy3d::Frame *frame = camera()->frame();
-    walk_through_->add_key_frame(*frame);
-    update();
-}
-
-
-void PaintCanvas::showCameraPath(bool b) {
-    walk_through_->set_path_visible(b);
-    if (b) {
-        const int count = walk_through_->interpolator()->numberOfKeyFrames();
-        float radius = camera_->sceneRadius();
-        for (int i=0; i<count; ++i) {
-            radius = std::max( radius,
-                    distance(camera_->sceneCenter(), walk_through_->interpolator()->keyFrame(i).position())
-            );
-        }
-        camera_->setSceneRadius(radius);
-    }
-    else {
-        Box3 box;
-        for (auto m : models_)
-            box.add_box(m->bounding_box());
-        camera_->setSceneBoundingBox(box.min(), box.max());
-    }
-    update();
-}
-
-void PaintCanvas::exportCamaraPathToFile(const std::string& file_name) const {
-    if (walk_through_)
-        walk_through_->interpolator()->save_keyframes(file_name);
-}
-
-
-void PaintCanvas::importCameraPathFromFile(const std::string& file_name) {
-    if (walk_through_) {
-        walk_through_->interpolator()->read_keyframes(file_name);
-        if (walk_through_->is_path_visible()) {
-            // update scene radius to make sure the path is within the view frustum
-            int num = walk_through_->interpolator()->numberOfKeyFrames();
-            float radius = camera_->sceneRadius();
-            for (int i = 0; i < num; ++i) {
-                radius = std::max(
-                        radius,
-                        distance(camera_->sceneCenter(), walk_through_->interpolator()->keyFrame(i).position())
-                );
-            }
-            camera_->setSceneRadius(radius);
-        }
-
-        update();
-    }
-}
-
-
 void PaintCanvas::showFaceVertexLabelsUnderMouse(bool b) {
     show_labels_under_mouse_ = b;
     update();
