+++ applications/Mapple/dialogs/dialog_walk_through.cpp
@@ -25,12 +25,19 @@
 #include "dialog_walk_through.h"
 #include "paint_canvas.h"
 #include "main_window.h"
+
+#include <easy3d/core/model.h>
 #include <easy3d/renderer/walk_throuth.h>
 #include <easy3d/renderer/key_frame_interpolator.h>
+#include <easy3d/renderer/manipulated_camera_frame.h>
+#include <easy3d/renderer/camera.h>
 #include <easy3d/util/logging.h>
+#include <easy3d/util/file_system.h>
+#include <easy3d/util/progress.h>
 
 #include <QFileDialog>
 #include <QMessageBox>
+#include <QButtonGroup>
 
 
 using namespace easy3d;
@@ -44,23 +51,40 @@ DialogWalkThrough::DialogWalkThrough(MainWindow *window)
 	connect(doubleSpinBoxCharacterDistanceFactor, SIGNAL(valueChanged(double)), this, SLOT(setCharacterDistanceFactor(double)));
     connect(doubleSpinBoxInterpolationSpeed, SIGNAL(valueChanged(double)), this, SLOT(setInterpolationSpeed(double)));
 
+    connect(importCameraPathButton, SIGNAL(clicked()), this, SLOT(importCameraPathFromFile()));
+    connect(exportCameraPathButton, SIGNAL(clicked()), this, SLOT(exportCameraPathToFile()));
+    connect(checkBoxShowCameraPath, SIGNAL(toggled(bool)), this, SLOT(showCameraPath(bool)));
+
+    connect(addKeyframeButton, SIGNAL(clicked()), this, SLOT(addKeyFrame()));
+
+    connect(radioButtonWalkingMode, SIGNAL(toggled(bool)), this, SLOT(setWalkingMode(bool)));
+
 	connect(previousPositionButton, SIGNAL(clicked()), this, SLOT(goToPreviousPosition()));
 	connect(nextPositionButton, SIGNAL(clicked()), this, SLOT(goToNextPosition()));
     connect(removeLastPositionButton, SIGNAL(clicked()), this, SLOT(removeLastPosition()));
 
     connect(horizontalSliderPreview, SIGNAL(valueChanged(int)), this, SLOT(goToPosition(int)));
 
-    connect(clearPathButton, SIGNAL(clicked()), this, SLOT(clearPath()));
-
+    connect(clearCameraPathButton, SIGNAL(clicked()), this, SLOT(clearPath()));
     connect(previewButton, SIGNAL(toggled(bool)), this, SLOT(preview(bool)));
     connect(recordButton, SIGNAL(toggled(bool)), this, SLOT(record(bool)));
 
+    connect(browseButton, SIGNAL(clicked()), this, SLOT(browse()));
+
     easy3d::connect(&walkThrough()->path_modified, this, &DialogWalkThrough::newPositionAdded);
+
+    progress_ = new ProgressLogger;
+
+    QButtonGroup* group = new QButtonGroup(this);
+    group->addButton(radioButtonFreeMode);
+    group->addButton(radioButtonWalkingMode);
+    radioButtonWalkingMode->setChecked(true);
 }
 
 
 DialogWalkThrough::~DialogWalkThrough()
 {
+    delete progress_;
 }
 
 
@@ -84,9 +108,26 @@ easy3d::WalkThrough* DialogWalkThrough::walkThrough() {
 }
 
 
+easy3d::KeyFrameInterpolator* DialogWalkThrough::interpolator() {
+    return walkThrough()->interpolator();
+}
+
+
 void DialogWalkThrough::showEvent(QShowEvent* e) {
 	doubleSpinBoxCharacterHeightFactor->setValue(walkThrough()->height_factor());
[…]
