+++ applications/Mapple/paint_canvas_snapshot.cpp
@@ -53,7 +53,7 @@
 
 using namespace easy3d;
 
-bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_name, bool bk_white, bool expand) {
+bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_name, bool bk_white, bool expand, ProgressLogger* progress) {
     int max_samples = 0;
     makeCurrent();
     func_->glGetIntegerv(GL_MAX_SAMPLES, &max_samples);
@@ -107,7 +107,8 @@ bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_na
     if (nbX * sub_w < w) ++nbX;
     if (nbY * sub_h < h) ++nbY;
 
-    ProgressLogger progress(nbX * nbY * 1.2f); // the extra 0.2 is for saving the "big" image
+    if (progress)
+        progress->reset(nbX * nbY * 1.2f, true); // the extra 0.2 is for saving the "big" image
 
     // remember the current projection matrix
     // const mat4& proj_matrix = camera()->projectionMatrix(); // Liangliang: This will definitely NOT work !!!
@@ -130,6 +131,11 @@ bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_na
     int count = 0;
     for (int i = 0; i < nbX; i++) {
         for (int j = 0; j < nbY; j++) {
+            if (progress && progress->is_canceled()) {
+                LOG(WARNING) << "snapshot cancelled";
+                break;
+            }
+
             if (camera()->type() == Camera::PERSPECTIVE) {
                 // change the projection matrix of the camera.
                 const mat4 &proj = transform::frustum(-xMin + i * deltaX, -xMin + (i + 1) * deltaX,
@@ -187,7 +193,8 @@ bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_na
             ++count;
 
 #ifdef SHOW_PROGRESS
-            progress.next(false);
+            if (progress)
+                progress->next(false);
             // this very important (the progress bar may interfere the framebuffer
             makeCurrent();
 #endif
@@ -203,128 +210,16 @@ bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_na
     func_->glClearColor(background_color_[0], background_color_[1], background_color_[2], background_color_[3]);
     doneCurrent();
 
-    bool saved = image.save(file_name);
-    progress.done();
-
-    return saved;
+    return image.save(file_name);
 }
 
 
-#if 0
+#ifdef HAS_FFMPEG
 
-void PaintCanvas::recordAnimation(const QString &file_name, bool bk_white) {
-    auto kfi = walkThrough()->interpolator();
-    if (kfi->numberOfKeyFrames() == 0) {
-        LOG(WARNING) << "recording aborted (camera path is empty). You may import a camera path from a file or"
-                        " creat it by adding key frames";
-        return;
-    }
-
-    // make the image dimensions multiples of 8 (to have the same size as if video saved using ffmpeg)
-    const int original_width = width();
-    const int original_height = height();
-    int w = original_width;
-    int h = original_height;
-    //find the nearest multiples of 8
-    if (original_width % 8)
-        w = (original_width / 8 + 1) * 8;
-    if (original_height % 8)
-        h = (original_height / 8 + 1) * 8;
-    if (w != original_width || h != original_height) {
-        camera_->setScreenWidthAndHeight(w, h);
-        update();
-        QApplication::processEvents();
-    }
-
-    const int fw = w * dpi_scaling();
-    const int fh = h * dpi_scaling();
-    const auto &frames = kfi->interpolate();
-    makeCurrent();
-
-#ifdef USE_QT_FBO
[…]

[…]
