### 2021-01-01 244ffba
add a warning when user deletes the path
