### 2021-01-02 7237546
allow to compile without FFMPEG
