### 2021-01-02 58c34bc
set ffmpeg quite in release mode
