### 2021-01-02 b8442bf
larger default point size for point clouds
