### 2021-01-02 bb75131
add camera path animation
