### 2021-01-02 366a9d0
rename LogClient::output() to LogClient::send()
