### 2021-01-02 c0398e0
add deleteLastKeyFrame()
