### 2021-01-02 3feb06d
log error if occured
