### 2021-01-02 060c491
show pivot also when mouse pressed
