### 2021-01-02 60a7ddd
set pivot point after each move (easier for subsequent manipulation)
