+++ easy3d/renderer/walk_throuth.cpp
@@ -26,11 +26,10 @@
 #include <easy3d/renderer/walk_throuth.h>
 #include <easy3d/core//model.h>
 #include <easy3d/renderer/camera.h>
-#include <easy3d/renderer/drawable_points.h>
-#include <easy3d/renderer/drawable_lines.h>
 #include <easy3d/renderer/key_frame_interpolator.h>
 #include <easy3d/renderer/manipulated_camera_frame.h>
 #include <easy3d/util/logging.h>
+
 
 
 namespace easy3d {
@@ -42,12 +41,7 @@ namespace easy3d {
             , height_factor_(0.2f)
             , third_person_forward_factor_(1.8f)
             , path_visible_(false)
-            , current_position_idx_(-1)
-            , buffer_up_to_date_(false)
-            , character_drawable_(nullptr)
-            , character_head_drawable_(nullptr)
-            , view_direction_base_drawable_(nullptr)
-            , view_direction_arraw_drawable_(nullptr)
+            , current_frame_idx_(-1)
     {
         kfi_ = new KeyFrameInterpolator(camera_->frame());
     }
@@ -68,6 +62,7 @@ namespace easy3d {
             camera_->setSceneRadius(dist);
 
         LOG(INFO) << "a key frame added to camera path";
+        path_modified.send();
     }
 
 
@@ -77,148 +72,92 @@ namespace easy3d {
             box += m->bounding_box();
         scene_box_ = box;
         camera_->setSceneRadius(box.radius() * 1.1f);
-        buffer_up_to_date_ = false;
     }
 
 
-    void WalkThrough::walk_to(const vec3 &pos) {
-        const vec3 &head = character_head(pos);
-        const vec3 view_dir = head - camera_->position();
-        add_position(pos, view_dir);
+    void WalkThrough::walk_to(const vec3 &ground_point) {
+        const vec3 head = character_head(ground_point);
+        vec3 view_dir = head - camera_->position();
+        view_dir.z = 0;
+        view_dir.normalize();
+        const vec3 &cam_pos = head - view_dir * third_person_forward_distance();
+        add_key_frame(cam_pos, view_dir);
 
-        // set the pivot point ahead of he new position.
-        camera_->setPivotPoint(pos + view_dir);
+        // set the pivot point ahead of the character.
+        camera_->setPivotPoint(ground_point + view_dir * third_person_forward_distance());
     }
 
 
     void WalkThrough::set_height_factor(float f) {
         height_factor_ = f;
-        buffer_up_to_date_ = false;
-        move_to(current_position_idx_);
+        move_to(current_frame_idx_);
     }
 
 
     void WalkThrough::set_third_person_forward_factor(float f) {
         third_person_forward_factor_ = f;
-        buffer_up_to_date_ = false;
-        move_to(current_position_idx_);
+        move_to(current_frame_idx_);
     }
 
 
     void WalkThrough::delete_last_position() {
-        if (path_.empty())
+        if (kfi_->numberOfKeyFrames() == 0)
             return;
 
-        if (current_position_idx_ == path_.size() - 1)
-            --current_position_idx_;
-        path_.pop_back();
-        buffer_up_to_date_ = false;
+        kfi_->deleteLastKeyFrame();
+        --current_frame_idx_;
     }
 
 
     void WalkThrough::delete_path() {
         kfi_->deletePath();
-        path_.clear();
-
-        if (character_drawable_) {
-            delete character_drawable_;
-            character_drawable_ = 0;
-        }
-        if (character_head_drawable_) {
-            delete character_head_drawable_;
-            character_head_drawable_ = 0;
-        }
-        if (view_direction_base_drawable_) {
-            delete view_direction_base_drawable_;
-            view_direction_base_drawable_ = 0;
[…]
