### 2021-01-02 caab839
fix a few warnings
