### 2021-01-02 e8fb4fb
improve ProgressLogger
