+++ applications/Mapple/paint_canvas_snapshot.cpp
@@ -107,7 +107,7 @@ bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_na
     if (nbX * sub_w < w) ++nbX;
     if (nbY * sub_h < h) ++nbY;
 
-    ProgressLogger progress(nbX * nbY * 1.1f); // the extra 0.1 is for saving the "big" image (time consuming)
+    ProgressLogger progress(nbX * nbY * 1.1f, false, false); // the extra 0.1 is for saving the "big" image (time consuming)
 
     // remember the current projection matrix
     // const mat4& proj_matrix = camera()->projectionMatrix(); // Liangliang: This will definitely NOT work !!!
@@ -192,7 +192,7 @@ bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_na
             ++count;
 
 #ifdef SHOW_PROGRESS
-            progress.next(false);
+            progress.next();
             // this very important (the progress bar may interfere the framebuffer
             makeCurrent();
 #endif
@@ -272,7 +272,7 @@ void PaintCanvas::recordAnimation(const QString &file_name, int fps, int bit_rat
 #endif
 
     const QString ext_less_name = file_name.left(file_name.lastIndexOf('.'));
-    ProgressLogger progress(frames.size());
+    ProgressLogger progress(frames.size(), false, false);
     for (std::size_t frame_index = 0; frame_index < frames.size(); ++frame_index) {
         if (progress.is_canceled()) {
             success = false;
@@ -322,7 +322,7 @@ void PaintCanvas::recordAnimation(const QString &file_name, int fps, int bit_rat
         currentTime += timeStep;
 
 #ifdef SHOW_PROGRESS
-        progress.next(false);
+        progress.next();
         // this very important (the progress bar may interfere the framebuffer
         makeCurrent();
 #endif
@@ -394,7 +394,7 @@ void PaintCanvas::recordAnimation(const QString &file_name, int, int, bool bk_wh
 
     bool success = true;
     const QString ext_less_name = file_name.left(file_name.lastIndexOf('.'));
-    ProgressLogger progress(frames.size());
+    ProgressLogger progress(frames.size(), false, false);
     for (std::size_t frame_index = 0; frame_index < frames.size(); ++frame_index) {
         if (progress.is_canceled()) {
             success = false;
@@ -439,7 +439,7 @@ void PaintCanvas::recordAnimation(const QString &file_name, int, int, bool bk_wh
         }
 
 #ifdef SHOW_PROGRESS
-        progress.next(false);
+        progress.next();
         // this very important (the progress bar may interfere the framebuffer
         makeCurrent();
 #endif
