### 2021-01-02 3044674
distance weighted timing
