+++ easy3d/renderer/key_frame_interpolator.cpp
@@ -43,7 +43,7 @@
 #include <easy3d/renderer/frame.h>
 #include <easy3d/renderer/drawable_lines.h>
 #include <easy3d/renderer/primitives.h>
-#include <easy3d/renderer/camera.h>   // for scene radius (to draw cameras in a proper size)
+#include <easy3d/renderer/camera.h>   // for drawing the camera path drawables
 
 
 namespace easy3d {
@@ -53,7 +53,7 @@ namespace easy3d {
 //#define DEBUG_INTERPOLATED_FRAMES
 #ifdef DEBUG_INTERPOLATED_FRAMES
 void save_interpolation(const std::vector<easy3d::Frame>& frames) {
-    std::string file = "interpolated_frames.kf";
+    std::string file = "interpolated_frames2.kf";
     std::ofstream output(file.c_str());
     if (output.fail())
         std::cout << "could not open file: " << file << std::endl;
@@ -68,7 +68,7 @@ void save_interpolation(const std::vector<easy3d::Frame>& frames) {
 #endif
 
     KeyFrameInterpolator::KeyFrameInterpolator(Frame *frame)
-            : frame_(frame), period_(40)   // 25 frames per second
+            : frame_(frame), period_(10)   // 25 frames per second
             , interpolationSpeed_(1.0), interpolationStarted_(false), start_position_(0), pathIsValid_(false),
               path_drawable_(nullptr), cameras_drawable_(nullptr) {
         if (keyFrames_.size() < 2)
@@ -122,10 +122,11 @@ void save_interpolation(const std::vector<easy3d::Frame>& frames) {
 
 
     void KeyFrameInterpolator::addKeyFrame(const Frame &frame, float time) {
-        if ((!keyFrames_.empty()) && (keyFrames_.back()->time() > time))
+        if ((!keyFrames_.empty()) && (keyFrames_.back()->time() >= time))
             LOG(ERROR) << "time is not monotone";
-        else
+        else {
             keyFrames_.push_back(new KeyFrame(frame, time));
+        }
 
         pathIsValid_ = false;
         stopInterpolation();
@@ -140,11 +141,27 @@ void save_interpolation(const std::vector<easy3d::Frame>& frames) {
 
 
     void KeyFrameInterpolator::addKeyFrame(const Frame &frame) {
+#if 0
+        // interval of each consecutive keyframes is 1.0 sec.
         float time = 0.0f;
         if (keyFrames_.empty())
             time = 0.0;
         else
             time = keyFrames_.back()->time() + 1.0f;
+#else
+        static float period_reference_distance = 0.0f;
+        // interval between first two keyframes is 1.0 sec.
+        // Others are computed relative to distance between two consecutive keyframes
+        float time = 0.0f;
+        if (keyFrames_.empty())
+            time = 0.0;
+        else if (keyFrames_.size() == 1) {
+            time = 1.0f;
+            period_reference_distance = distance(keyFrames_[0]->position(), frame.position());
+        }
+        else
+            time = keyFrames_.back()->time() + 1.0f * distance(keyFrames_.back()->position(), frame.position()) / period_reference_distance;
+#endif
 
         addKeyFrame(frame, time);
     }
@@ -387,20 +404,21 @@ void save_interpolation(const std::vector<easy3d::Frame>& frames) {
     }
 
     void KeyFrameInterpolator::KeyFrame::computeTangent(const KeyFrame *const prev, const KeyFrame *const next) {
-#if 1   // distance(prev, cur) and distance(cur, next) can have a big difference.
+#if 1
+        // distance(prev, cur) and distance(cur, next) can have a big difference.
         // we compensate this
         float sd_prev = distance2(prev->position(), position());
[…]
