+++ applications/Mapple/dialogs/dialog_walk_through.cpp
@@ -400,6 +400,7 @@ void DialogWalkThrough::importCameraPathFromFile() {
                 );
             }
             viewer_->camera()->setSceneRadius(radius);
+            viewer_->update();
         }
     }
 
