### 2021-01-02 9caa17f
remove interpolate(float time)
