### 2021-01-02 c807931
add KeyFrameInterpolator::set_frame_rate()
