### 2021-01-02 486946f
add KeyFrameInterpolator::set_frame_rate()
