### 2021-01-02 09895f7
add Status control to WalkThrough; improve UI; clean code
