### 2021-01-02 9a477a7
add hints to user
