### 2021-01-03 220594c
smoothing works, code needs cleaned
