+++ easy3d/renderer/key_frame_interpolator.cpp
@@ -209,6 +209,8 @@ void save_interpolation(const std::vector<easy3d::Frame>& frames) {
             interpolated_path_ = interpolate();
             pathIsValid_ = true;
         }
+        if (interpolated_path_.empty()) // interpolation may have failed.
+            return;
 
         if (!path_drawable_) {
             std::vector<vec3> points;
@@ -304,20 +306,20 @@ void save_interpolation(const std::vector<easy3d::Frame>& frames) {
     }
 
 
-    void KeyFrameInterpolator::updateModifiedFrameValues() {
-        quat prevQ = keyFrames_.front()->orientation();
+    void KeyFrameInterpolator::updateModifiedFrameValues(std::vector<KeyFrame*>& keyFrames) {
+        quat prevQ = keyFrames.front()->orientation();
         KeyFrame *kf = nullptr;
-        for (std::size_t i = 0; i < keyFrames_.size(); ++i) {
-            kf = keyFrames_.at(i);
+        for (std::size_t i = 0; i < keyFrames.size(); ++i) {
+            kf = keyFrames.at(i);
             kf->flipOrientationIfNeeded(prevQ);
             prevQ = kf->orientation();
         }
 
-        KeyFrame *prev = keyFrames_.front();
-        kf = keyFrames_.front();
+        KeyFrame *prev = keyFrames.front();
+        kf = keyFrames.front();
         std::size_t index = 1;
         while (kf) {
-            KeyFrame *next = (index < keyFrames_.size()) ? keyFrames_.at(index) : nullptr;
+            KeyFrame *next = (index < keyFrames.size()) ? keyFrames.at(index) : nullptr;
             index++;
             if (next)
                 kf->computeTangent(prev, next);
@@ -408,38 +410,38 @@ void save_interpolation(const std::vector<easy3d::Frame>& frames) {
     }
 
 
-    void KeyFrameInterpolator::getRelatedKeyFramesForTime(float time, std::vector<KeyFrame *>::const_iterator *relatedFrames) const {
+    void KeyFrameInterpolator::getRelatedKeyFramesForTime(float time, const std::vector<KeyFrame*>& keyFrames, std::vector<KeyFrame *>::const_iterator *relatedFrames) const {
         // Assertion: times are sorted in monotone order.
         // Assertion: keyFrames_ is not empty
 
         // TODO: Special case for loops when closed path is implemented !!
         // Recompute everything from scratch
-        relatedFrames[1] = keyFrames_.begin(); // points to the first one
+        relatedFrames[1] = keyFrames.begin(); // points to the first one
 
         while ((*relatedFrames[1])->time() > time) {
-            if (relatedFrames[1] == keyFrames_.begin()) // already the first one
+            if (relatedFrames[1] == keyFrames.begin()) // already the first one
                 break;
             --relatedFrames[1]; // points to the previous one
         }
 
         relatedFrames[2] = relatedFrames[1];
         while ((*relatedFrames[2])->time() < time) {
-            if (*relatedFrames[2] == keyFrames_.back()) // already the last one
+            if (*relatedFrames[2] == keyFrames.back()) // already the last one
                 break;
             ++relatedFrames[2];// points to the next one
         }
 
         relatedFrames[1] = relatedFrames[2];
-        if ((relatedFrames[1] != keyFrames_.begin()) && (time < (*relatedFrames[2])->time()))
+        if ((relatedFrames[1] != keyFrames.begin()) && (time < (*relatedFrames[2])->time()))
             --relatedFrames[1];
 
         relatedFrames[0] = relatedFrames[1];
-        if (relatedFrames[0] != keyFrames_.begin())
+        if (relatedFrames[0] != keyFrames.begin())
             --relatedFrames[0];
 
[…]
