### 2021-01-03 3184cae
add is_nan(FT v)
