### 2021-01-03 d088161
create default "locks" drawable if propery found
