### 2021-01-03 55c3465
log drawable names if error occured
