+++ easy3d/renderer/drawable_triangles.cpp
@@ -64,7 +64,7 @@ namespace easy3d {
 
     void TrianglesDrawable::_draw_triangles(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
 
@@ -122,7 +122,7 @@ namespace easy3d {
 
     void TrianglesDrawable::_draw_triangles_with_texture(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
         if (texcoord_buffer() == 0) {
