+++ easy3d/renderer/drawable_lines.cpp
@@ -81,7 +81,7 @@ namespace easy3d {
 
     void LinesDrawable::_draw_plain_lines(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
 
@@ -141,7 +141,7 @@ namespace easy3d {
 
     void LinesDrawable::_draw_cylinders(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
 
@@ -183,7 +183,7 @@ namespace easy3d {
 
     void LinesDrawable::_draw_cones(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
 
@@ -224,7 +224,7 @@ namespace easy3d {
 
     void LinesDrawable::_draw_plain_lines_with_texture(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
         if (texcoord_buffer() == 0) {
@@ -294,7 +294,7 @@ namespace easy3d {
 
     void LinesDrawable::_draw_cylinders_with_texture(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
         if (texcoord_buffer() == 0) {
@@ -344,7 +344,7 @@ namespace easy3d {
 
     void LinesDrawable::_draw_cones_with_texture(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
         if (texcoord_buffer() == 0) {
