+++ easy3d/renderer/drawable_points.cpp
@@ -89,7 +89,7 @@ namespace easy3d {
 
     void PointsDrawable::_draw_plain_points(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
 
@@ -143,7 +143,7 @@ namespace easy3d {
 
     void PointsDrawable::_draw_spheres_sprite(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
 
@@ -195,7 +195,7 @@ namespace easy3d {
 
     void PointsDrawable::_draw_spheres_geometry(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
 
@@ -242,7 +242,7 @@ namespace easy3d {
 
     void PointsDrawable::_draw_plain_points_with_texture(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
         if (texcoord_buffer() == 0) {
@@ -301,7 +301,7 @@ namespace easy3d {
 
     void PointsDrawable::_draw_spheres_with_texture_sprite(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
         if (texcoord_buffer() == 0) {
@@ -314,7 +314,7 @@ namespace easy3d {
 
     void PointsDrawable::_draw_spheres_with_texture_geometry(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
         if (texcoord_buffer() == 0) {
@@ -366,7 +366,7 @@ namespace easy3d {
 
     void PointsDrawable::_draw_surfels(const Camera *camera, bool with_storage_buffer) const {
         if (vertex_buffer() == 0) {
-            LOG_FIRST_N(ERROR, 1) << "vertex buffer not created (this is the first record)";
+            LOG_FIRST_N(ERROR, 1) << "drawable \'" << name() << "\': vertex buffer not created (this is the first record)";
             return;
         }
         if (normal_buffer() == 0) {
@@ -425,7 +425,7 @@ namespace easy3d {
 
[…]

[…]
