### 2021-01-03 31736e6
fix a bug where the last stopped index was not correctly set
