### 2021-01-03 f277327
improve KeyFrameInterpolator API design
