+++ easy3d/renderer/key_frame_interpolator.cpp
@@ -49,24 +49,6 @@
 namespace easy3d {
 
 
-    // allow to save interpolated frames as keyframes (so I can visualize them)
-//#define DEBUG_INTERPOLATED_FRAMES
-#ifdef DEBUG_INTERPOLATED_FRAMES
-void save_interpolation(const std::vector<easy3d::Frame>& frames) {
-    std::string file = "interpolated_frames2.kf";
-    std::ofstream output(file.c_str());
-    if (output.fail())
-        std::cout << "could not open file: " << file << std::endl;
-    output << "\tnum_key_frames: " << frames.size() << std::endl;
-    for (std::size_t id = 0; id < frames.size(); ++id) {
-        output << "\tframe: " << id << std::endl;
-        const easy3d::Frame& frame = frames[id];
-        output << "\t\tposition: " << frame.position() << std::endl;
-        output << "\t\torientation: " << frame.orientation() << std::endl;
-    }
-}
-#endif
-
     KeyFrameInterpolator::KeyFrameInterpolator(Frame *frame)
             : frame_(frame)
             , fps_(30)
@@ -75,9 +57,8 @@ void save_interpolation(const std::vector<easy3d::Frame>& frames) {
             , last_stopped_index_(0)
             , pathIsValid_(false)
             , path_drawable_(nullptr)
-            , cameras_drawable_(nullptr) {
-        if (keyframes_.size() < 2)
-            return;
+            , cameras_drawable_(nullptr)
+    {
     }
 
 
@@ -86,53 +67,23 @@ void save_interpolation(const std::vector<easy3d::Frame>& frames) {
     }
 
 
-    void KeyFrameInterpolator::setFrame(Frame *const frame) {
-        frame_ = frame;
-    }
-
-
-    void KeyFrameInterpolator::start_interpolation() {
+    void KeyFrameInterpolator::add_keyframe(const Frame &frame) {
+        // interval of each consecutive keyframes is 1.0 sec.
+        float time = 0.0f;
         if (keyframes_.empty())
-            return;
+            time = 0.0;
+        else
+            time = keyframes_.back().time() + 1.0f;
 
-        if (!pathIsValid_)
-            interpolate();
-
-        // all done in another thread.
-        interpolation_started_ = true;
-        timer_.set_timeout(0, [this]() {
-                               for (int id = last_stopped_index_; id < interpolated_path_.size(); ++id) {
-                                   if (timer_.is_stopped()) {
-                                       last_stopped_index_ = id;
-                                       break;
-                                   }
-                                   const auto &f = interpolated_path_[id];
-                                   frame()->setPositionAndOrientation(f.position(), f.orientation());
-                                   // interval in ms. 0.9 for (approximately) compensating the timer overhead
-                                   const int interval = interpolation_period() / interpolation_speed() * 0.9f;
-                                   std::this_thread::sleep_for(std::chrono::milliseconds(interval));
-                                   if (id == interpolated_path_.size() - 1)  // reaches the end frame
-                                       last_stopped_index_ = 0;
-                                   frame_interpolated.send();
-                               }
-                               interpolation_stopped.send();
-                               interpolation_started_ = false;
-                           }
-        );
-    }
-
-
-    void KeyFrameInterpolator::stop_interpolation() {
-        timer_.stop();
-        interpolation_started_ = false;
+        add_keyframe(frame, time);
     }
 
 
     void KeyFrameInterpolator::add_keyframe(const Frame &frame, float time) {
[…]
