+++ easy3d/renderer/key_frame_interpolator.h
@@ -104,7 +104,7 @@ namespace easy3d {
          * \details The frame() can be set or changed using set_frame(). The interpolation_speed() and 
          *      interpolation_period() are set to their default values. 
          */
-        KeyFrameInterpolator(Frame* frame=nullptr);
+        KeyFrameInterpolator(Frame *frame = nullptr);
 
         /** Virtual destructor. Clears the keyframe path. */
         virtual ~KeyFrameInterpolator();
@@ -133,6 +133,28 @@ namespace easy3d {
         void delete_last_keyframe();
 
         /**
+         * \brief Sets the time corresponding to the \p index-th keyframe.
+         * \details The \p index has to be in the range [0, number_of_keyframes()-1].
+         * \note The time \c t have to be monotonously increasing over keyframes.
+         * \sa keyframe_time()
+         */
+        void set_keyframe_time(std::size_t index, float t);
+
+        /**
+         * \brief Sets the position of the \p index-th keyframe.
+         * \details The \p index has to be in the range [0, number_of_keyframes()-1].
+         * \sa keyframe_time()
+         */
+        void set_keyframe_position(std::size_t index, const vec3& pos);
+
+        /**
+         * \brief Sets the orientation of the \p index-th keyframe.
+         * \details The \p index has to be in the range [0, number_of_keyframes()-1].
+         * \sa keyframe_time()
+         */
+        void set_keyframe_orientation(std::size_t index, const quat& q);
+
+        /**
          * \brief Removes all keyframes from the path.
          * \details Upon return, the number_of_keyframes() will return 0.
          */
@@ -153,10 +175,10 @@ namespace easy3d {
 
     public:
         /** Sets the frame() associated to the KeyFrameInterpolator. */
-        void setFrame(Frame* const frame);
+        void set_frame(Frame* const frame) { frame_ = frame; }
         //@}
 
-        /*! @name Path parameters */
+        /*! @name Access keyframe and path parameters */
         //@{
     public:
         /**
@@ -175,9 +197,22 @@ namespace easy3d {
         /**
          * \brief Returns the time corresponding to the \p index-th keyframe.
          * \details The \p index has to be in the range [0, number_of_keyframes()-1].
-         * \sa keyframe().
+         * \sa set_keyframe_time().
          */
         float keyframe_time(std::size_t index) const;
+
+        /**
+         * \brief Returns the position of the \p index-th keyframe.
+         * \details The \p index has to be in the range [0, number_of_keyframes()-1].
+         * \sa set_keyframe_position()
+         */
+        const vec3& keyframe_position(std::size_t index);
+        /**
+         * \brief Returns the orientation of the \p index-th keyframe.
+         * \details The \p index has to be in the range [0, number_of_keyframes()-1].
+         * \sa set_keyframe_orientation()
+         */
+        const quat& keyframe_orientation(std::size_t index);
 
         /**
          * \brief Returns the duration of the KeyFrameInterpolator path, expressed in seconds.
@@ -276,7 +311,7 @@ namespace easy3d {
 
     public:
         /// Computes and returns all the interpolated frames.
-        const std::vector<Frame>& interpolate();
+        const std::vector<Frame>& interpolate(bool smoothing = true);
         //@}
 
         /*! @name Path drawing */
@@ -309,20 +344,22 @@ namespace easy3d {
         KeyFrameInterpolator& operator=(const KeyFrameInterpolator& kfi);
 
 #ifndef DOXYGEN
-        // Internal private KeyFrame representation
-        class KeyFrame
+        // Internal private Keyframe representation
+        class Keyframe
[…]
