### 2021-01-04 3db85da
skip interpolation and smoothing when there is only 1 keyframe
