### 2021-01-04 13fc84c
add Test1() and Test2() for only debug mode
