+++ applications/Mapple/main_window.cpp
@@ -192,10 +192,13 @@ MainWindow::MainWindow(QWidget *parent)
     readSettings();
     updateWindowTitle();
 
-#if 0
-    QMessageBox::warning(this, "Mapple is not ready yet!",
-                         "Mapple is still under development. This version is not feature complete nor is fully tested. Using it is at your own risk.",
-                         QMessageBox::Ok);
+#ifndef NDEBUG  // temporal menus/tools
+		QToolBar* toolBarExperimental = new QToolBar(this);
+		addToolBar(toolBarExperimental);
+		QAction* actionTest1 = toolBarExperimental->addAction("Test1");
+		connect(actionTest1, SIGNAL(triggered()), this, SLOT(test1()));
+        QAction* actionTest2 = toolBarExperimental->addAction("Test2");
+        connect(actionTest2, SIGNAL(triggered()), this, SLOT(test2()));
 #endif
 }
 
@@ -1970,3 +1973,20 @@ void MainWindow::pointCloudDelaunayTriangulation3D() {
     updateUi();
     viewer_->update();
 }
+
+
+#ifndef NDEBUG
+void MainWindow::test1() {
+    float coeff = viewer_->camera()->zNearCoefficient();
+    viewer_->camera()->setZNearCoefficient(coeff + 0.005);
+    std::cout << "camera()->zNearCoefficient(): " << coeff << " -> " << coeff + 0.005 << std::endl;
+    viewer_->update();
+}
+
+void MainWindow::test2() {
+    float coeff = viewer_->camera()->zNearCoefficient();
+    viewer_->camera()->setZNearCoefficient(coeff - 0.001);
+    std::cout << "camera()->zNearCoefficient(): " << coeff << " -> " << coeff - 0.001 << std::endl;
+    viewer_->update();
+}
+#endif
