### 2021-01-04 a70d933
call av_packet_unref() if error occured
