### 2021-01-04 211209b
plane clip behavior control
