+++ easy3d/renderer/average_color_blending.cpp
@@ -110,7 +110,7 @@ namespace easy3d {
                 program->set_uniform("per_vertex_color", d->coloring_method() != State::UNIFORM_COLOR && d->color_buffer());
                 program->set_uniform("default_color", d->color());
                 if (setting::clipping_plane)
-                    setting::clipping_plane->set_program(program, d->plane_clipping_discard());
+                    setting::clipping_plane->set_program(program, d->plane_clip_discard_primitive());
                 d->gl_draw(false);
             }
         }
