+++ easy3d/renderer/soft_shadow.cpp
@@ -93,7 +93,7 @@ namespace easy3d {
         for (auto d : surfaces) {
             if (d->is_visible()) {
                 if (setting::clipping_plane)
-                    setting::clipping_plane->set_program(program, d->plane_clipping_discard());
+                    setting::clipping_plane->set_program(program, d->plane_clip_discard_primitive());
                 d->gl_draw(false);
             }
         }
@@ -156,7 +156,7 @@ namespace easy3d {
                 program->set_uniform("per_vertex_color", d->coloring_method() != State::UNIFORM_COLOR && d->color_buffer());		easy3d_debug_log_gl_error;
                 program->set_uniform("is_background", false);
                 if (setting::clipping_plane)
-                    setting::clipping_plane->set_program(program, d->plane_clipping_discard());
+                    setting::clipping_plane->set_program(program, d->plane_clip_discard_primitive());
                 d->gl_draw(false);
             }
         }
