### 2021-01-04 5646f5c
fix a bug where the OpenGL context was not makeCurrent() before rendering into the framebuffer
