+++ applications/Mapple/paint_canvas_snapshot.cpp
@@ -27,9 +27,7 @@
 #define SHOW_PROGRESS
 
 #ifdef  USE_QT_FBO
-
 #include <QOpenGLFramebufferObject>
-
 #else
 #include <easy3d/renderer/framebuffer_object.h>
 #endif
@@ -48,7 +46,10 @@
 #include <easy3d/core//model.h>
 #include <easy3d/util/logging.h>
 #include <easy3d/util/file_system.h>
+
+#ifdef SHOW_PROGRESS
 #include <easy3d/util/progress.h>
+#endif
 
 
 using namespace easy3d;
@@ -107,11 +108,12 @@ bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_na
     if (nbX * sub_w < w) ++nbX;
     if (nbY * sub_h < h) ++nbY;
 
-    ProgressLogger progress(nbX * nbY * 1.1f, false, false); // the extra 0.1 is for saving the "big" image (time consuming)
-
     // remember the current projection matrix
     // const mat4& proj_matrix = camera()->projectionMatrix(); // Liangliang: This will definitely NOT work !!!
     const mat4 proj_matrix = camera()->projectionMatrix();
+
+    // temporarily don't allow updating rendering when the camera parameters are changing.
+    easy3d::disconnect(&camera_->frame_modified, this);
 
     makeCurrent();
 
@@ -128,13 +130,17 @@ bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_na
 #endif
 
     int count = 0;
+#ifdef SHOW_PROGRESS
+    ProgressLogger progress(nbX * nbY * 1.1f, false, false); // the extra 0.1 is for saving the "big" image (time consuming)
+#endif
     for (int i = 0; i < nbX; i++) {
         for (int j = 0; j < nbY; j++) {
+#ifdef SHOW_PROGRESS
             if (progress.is_canceled()) {
                 LOG(WARNING) << "snapshot cancelled";
                 break;
             }
-
+#endif
             if (camera()->type() == Camera::PERSPECTIVE) {
                 // change the projection matrix of the camera.
                 const mat4 &proj = transform::frustum(-xMin + i * deltaX, -xMin + (i + 1) * deltaX,
@@ -148,6 +154,9 @@ bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_na
             }
 
             //---------------------------------------------------------------------------
+
+            // this very important (the progress bar may interfere the framebuffer
+            makeCurrent();
 
             fbo->bind();
 
@@ -163,6 +172,7 @@ bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_na
             fbo->release();
 
             //---------------------------------------------------------------------------
+
 
 #ifdef USE_QT_FBO
             const QImage subImage = fbo->toImage();
@@ -193,20 +203,23 @@ bool PaintCanvas::saveSnapshot(int w, int h, int samples, const QString &file_na
 
 #ifdef SHOW_PROGRESS
             progress.next();
-            // this very important (the progress bar may interfere the framebuffer
-            makeCurrent();
 #endif
         }
     }
 
+    // this very important (the progress bar may interfere the framebuffer
+    makeCurrent();
+    // clean
     delete fbo;
+    // restore the clear color
+    func_->glClearColor(background_color_[0], background_color_[1], background_color_[2], background_color_[3]);
+    doneCurrent();
 
     // restore the projection matrix
     camera()->set_projection_matrix(proj_matrix);
 
-    // restore the clear color
-    func_->glClearColor(background_color_[0], background_color_[1], background_color_[2], background_color_[3]);
-    doneCurrent();
+    // enable updating the rendering
[…]
