+++ applications/Mapple/dialogs/dialog_walk_through.cpp
@@ -69,7 +69,7 @@ DialogWalkThrough::DialogWalkThrough(MainWindow *window)
 
     connect(clearCameraPathButton, SIGNAL(clicked()), this, SLOT(clearPath()));
     connect(previewButton, SIGNAL(toggled(bool)), this, SLOT(preview(bool)));
-    connect(recordButton, SIGNAL(toggled(bool)), this, SLOT(record(bool)));
+    connect(recordButton, SIGNAL(clicked()), this, SLOT(record()));
 
     connect(browseButton, SIGNAL(clicked()), this, SLOT(browse()));
 
@@ -113,6 +113,9 @@ easy3d::KeyFrameInterpolator* DialogWalkThrough::interpolator() {
 
 
 void DialogWalkThrough::showEvent(QShowEvent* e) {
+    viewer_->camera()->setZClippingCoefficient(5.0f);
+    viewer_->camera()->setZNearCoefficient(0.0001);
+
     if (radioButtonWalkingMode->isChecked())
         walkThrough()->set_status(easy3d::WalkThrough::WALKING_MODE);
     else
@@ -137,6 +140,9 @@ void DialogWalkThrough::showEvent(QShowEvent* e) {
 
 
 void DialogWalkThrough::closeEvent(QCloseEvent* e) {
+    viewer_->camera()->setZClippingCoefficient(std::sqrt(3.0f));
+    viewer_->camera()->setZNearCoefficient(0.005);
+
     walkThrough()->set_status(easy3d::WalkThrough::STOPPED);
     QDialog::closeEvent(e);
 	viewer_->update();
@@ -260,17 +266,21 @@ void DialogWalkThrough::browse() {
 
 
 void DialogWalkThrough::preview(bool b) {
-#if 1
-    // this single line works: very efficient (in another thread without overhead).
+#if 0
+    // this single line works: very efficient (in another thread without overhead), but I want a better UI.
     walkThrough()->preview();
     viewer_->update();
 
-#elif 0 // this also works. But std::this_thread::sleep_for() is not efficient and frame rate is low.
+#elif 0 // this also works, but low framerate (overhead from std::this_thread::sleep_for() and ProgressLogger)
     if (b) {
-        if (interpolator()->number_of_keyframes() == 0 && interpolator()->number_of_keyframes() == 0) {
-            recordButton->setChecked(false);
+        if (!interpolator() || interpolator()->number_of_keyframes() == 0) {
+            LOG_IF(WARNING, interpolator()->number_of_keyframes() == 0)
+                            << "nothing to preview (camera path is empty). You may import a camera path from a file or"
+                               " creat it by adding keyframes";
+            previewButton->setChecked(false);
             return;
         }
+
         static int last_stopped_index = 0;
         const auto &frames = interpolator()->interpolate();
 
@@ -287,7 +297,6 @@ void DialogWalkThrough::preview(bool b) {
             const auto &f = frames[id];
             viewer_->camera()->frame()->setPositionAndOrientation(f.position(), f.orientation());
             const int interval = interpolator()->interpolation_period() / interpolator()->interpolation_speed();
-            std::cout << "interval: " << interval << std::endl;
             std::this_thread::sleep_for(std::chrono::milliseconds(interval));
             if (id == frames.size() - 1)  // reaches the end frame
                 last_stopped_index = 0;
@@ -302,75 +311,95 @@ void DialogWalkThrough::preview(bool b) {
         previewButton->setChecked(false);
     }
     else
-        recordButton->setChecked(false);
+        previewButton->setChecked(false);
 
 #else
 
-    // this also handles the UI, but too complicated. I don't like it.
-//    auto interpolationStopped = [this]() -> void { emit animationStopped(); };
[…]

[…]
