### 2021-01-05 ad801eb
saveState(), restoreState(), save_keyframes(), and read_keyframes() use stream as arguments
