+++ easy3d/renderer/key_frame_interpolator.h
@@ -332,9 +332,9 @@ namespace easy3d {
         //@{
     public:
         /// saves the camera path to a file
-        bool save_keyframes(const std::string& file_name) const;
+        bool save_keyframes(std::ostream& os) const;
         /// reads camera path from a file
-        bool read_keyframes(const std::string& file_name);
+        bool read_keyframes(std::istream& is);
         //@}
 
     private:
