+++ easy3d/renderer/key_frame_interpolator.cpp
@@ -439,12 +439,15 @@ namespace easy3d {
         if (pathIsValid_ || keyframes_.size() == 1)
             return interpolated_path_;
 
-        if (smoothing)
-            adjust_keyframe_times(keyframes_);
+        if (smoothing && interpolated_path_.size() > 2)
+            adjust_keyframe_times(keyframes_, true);
 
-        LOG_IF(INFO, keyframes_.size() > 2) << "interpolating " << keyframes_.size() << " keyframes...";
+        LOG_IF(INFO, keyframes_.size() > 2) << "interpolating " << keyframes_.size() << " keyframes";
         do_interpolate(interpolated_path_, keyframes_);
-        LOG_IF(INFO, keyframes_.size() > 2) << "keyframe interpolation done, " << interpolated_path_.size() << " frames";
+        LOG_IF(INFO, keyframes_.size() > 2)
+                        << "keyframe interpolation done: "
+                        << interpolated_path_.size() << " frames, "
+                        << duration() / interpolation_speed() << " seconds (at x" << interpolation_speed() << " speed)";
 
         if (smoothing && interpolated_path_.size() > 2) { // more iterations do not provide further improvement
             std::vector<Keyframe> as_key_frames;
@@ -453,24 +456,23 @@ namespace easy3d {
             as_key_frames.front().set_time(firstTime());
             as_key_frames.back().set_time(lastTime());
 
-            adjust_keyframe_times(as_key_frames);
+            adjust_keyframe_times(as_key_frames, false);
             do_interpolate(interpolated_path_, as_key_frames);
         }
 
         pathIsValid_ = true;
+        last_stopped_index_ = 0; // not valid any more
         return interpolated_path_;
     }
 
 
 //    int idx = 0;
-
     void KeyFrameInterpolator::do_interpolate(std::vector<Frame>& frames, std::vector<Keyframe>& keyframes) {
         frames.clear();
         if (keyframes.size() < 2)
             return;
 
 //        std::ofstream  output ("tmp-iner-" + std::to_string(idx++) + ".xyz");
-
         update_keyframe_values(keyframes);
 
         const float interval = interpolation_speed() * interpolation_period() / 1000.0f;
@@ -495,33 +497,70 @@ namespace easy3d {
                                        related[2]->tgQ(), related[2]->orientation(), alpha);
 
             frames.emplace_back(Frame(pos, q));
-
 //            output << pos << std::endl;
         }
     }
 
 
-    void KeyFrameInterpolator::adjust_keyframe_times(std::vector<Keyframe>& keyframes) {
+    void KeyFrameInterpolator::adjust_keyframe_times(std::vector<Keyframe>& keyframes, bool slower_turning) {
         if (keyframes.size() < 2)
             return;
 
-        // use step-distance-weighted keyframe timing
+        const float duration = keyframes.back().time() - keyframes.front().time();
 
-        std::vector<float> sub_lengths(keyframes.size(), 0.0f);
-        float total_length = 0.0f;
-        for (std::size_t i=1; i<keyframes.size(); ++i) {
-            total_length += distance(keyframes[i - 1].position(), keyframes[i].position());
-            sub_lengths[i] = total_length;
-        }
-        if (total_length < epsilon_sqr<float>())
-            total_length = 1.0f;
+        if (slower_turning) {   // make turning slower
+            // use step-distance-weighted keyframe timing + turning times.
+            // the time needed for turning from current orientation to the orientation of the next frame.
+            //  - 0 degree: 0 second
+            //  - 90 degrees: 1.0 second
+            // so we use 1 - acos(angle) to approximate this time
+            auto turning_time = [](const Keyframe &frame, const Keyframe &prev) -> float {
+                const float cos_angle = std::abs(quat::dot(frame.orientation(), prev.orientation()));
[…]
