+++ easy3d/renderer/key_frame_interpolator.h
@@ -375,9 +375,10 @@ namespace easy3d {
         void compute_spline(const std::vector<Keyframe>::const_iterator* related, vec3& v1, vec3& v2) const;
         void do_interpolate(std::vector<Frame>& frames, std::vector<Keyframe>& keyframes);
 
-        // adjusts the keyframe times with respect to the accumulated path length
-        // so keyframes.front().time() and keyframes.back().time() are both preserved.
-        void adjust_keyframe_times(std::vector<Keyframe>& keyframes);
+        // stride-length weighted keyframe timing.
+        // both keyframes.front().time() and keyframes.back().time() are preserved.
+        // slower_turning: true to make turning slower
+        void adjust_keyframe_times(std::vector<Keyframe>& keyframes, bool slower_turning);
 #endif
 
     private:
