### 2021-01-05 6560b2a
rename delete_last_position() to delete_last_keyframe()
