+++ applications/Mapple/dialogs/dialog_walk_through.cpp
@@ -61,11 +61,11 @@ DialogWalkThrough::DialogWalkThrough(MainWindow *window)
 
     connect(radioButtonWalkingMode, SIGNAL(toggled(bool)), this, SLOT(setWalkingMode(bool)));
 
-	connect(previousPositionButton, SIGNAL(clicked()), this, SLOT(goToPreviousPosition()));
-	connect(nextPositionButton, SIGNAL(clicked()), this, SLOT(goToNextPosition()));
-    connect(removeLastPositionButton, SIGNAL(clicked()), this, SLOT(removeLastPosition()));
+	connect(previousKeyframeButton, SIGNAL(clicked()), this, SLOT(goToPreviousKeyframe()));
+	connect(nextKeyframeButton, SIGNAL(clicked()), this, SLOT(goToNextKeyframe()));
+    connect(removeLastKeyframeButton, SIGNAL(clicked()), this, SLOT(removeLastKeyframe()));
 
-    connect(horizontalSliderPreview, SIGNAL(valueChanged(int)), this, SLOT(goToPosition(int)));
+    connect(horizontalSliderPreview, SIGNAL(valueChanged(int)), this, SLOT(goToKeyframe(int)));
 
     connect(clearCameraPathButton, SIGNAL(clicked()), this, SLOT(clearPath()));
     connect(previewButton, SIGNAL(toggled(bool)), this, SLOT(preview(bool)));
@@ -88,7 +88,7 @@ DialogWalkThrough::~DialogWalkThrough()
 
 
 void DialogWalkThrough::numKeyramesChanged() {
-    disconnect(horizontalSliderPreview, SIGNAL(valueChanged(int)), this, SLOT(goToPosition(int)));
+    disconnect(horizontalSliderPreview, SIGNAL(valueChanged(int)), this, SLOT(goToKeyframe(int)));
     int num = interpolator()->number_of_keyframes();
     if (num == 1) // range is [0, 0]
         horizontalSliderPreview->setEnabled(false);
@@ -99,7 +99,7 @@ void DialogWalkThrough::numKeyramesChanged() {
     int pos = walkThrough()->current_keyframe_index();
     if (pos >= 0)
         horizontalSliderPreview->setValue(pos);
-    connect(horizontalSliderPreview, SIGNAL(valueChanged(int)), this, SLOT(goToPosition(int)));
+    connect(horizontalSliderPreview, SIGNAL(valueChanged(int)), this, SLOT(goToKeyframe(int)));
 }
 
 
@@ -189,7 +189,7 @@ void DialogWalkThrough::setWalkingMode(bool b) {
 }
 
 
-void DialogWalkThrough::goToPreviousPosition()
+void DialogWalkThrough::goToPreviousKeyframe()
 {
     int pos = walkThrough()->current_keyframe_index();
     if (pos <= 0)  // if not started (or at the 1st keyframe), move to the start view point
@@ -197,11 +197,11 @@ void DialogWalkThrough::goToPreviousPosition()
     else
         walkThrough()->move_to(pos - 1);
     viewer_->update();
-    LOG(INFO) << "moved to position " << walkThrough()->current_keyframe_index();
+    LOG(INFO) << "moved to keyframe " << walkThrough()->current_keyframe_index();
 }
 
 
-void DialogWalkThrough::goToNextPosition()
+void DialogWalkThrough::goToNextKeyframe()
 {
     int pos = walkThrough()->current_keyframe_index();
     if (pos >= interpolator()->number_of_keyframes() - 1)  // if already at the end, move to the last view point
@@ -209,27 +209,27 @@ void DialogWalkThrough::goToNextPosition()
     else
         walkThrough()->move_to(pos + 1);
     viewer_->update();
-    LOG(INFO) << "moved to position " << walkThrough()->current_keyframe_index();
[…]
