### 2021-01-05 cc23daa
add Bezier, BSpline, and CatmullRom
