### 2021-01-06 1447152
allow screen-axis constrained camera manipulation
  Left drag:           Rotate the camera
  Right drag:          Move the camera
  'x' + Left drag:     Rotate the camera around horizontal axis
  'x' + Right drag:    Move the camera along horizontal axis
  'y' + Left drag:     Rotate the camera around vertical axis
  'y' + Right drag:    Move the camera along vertical axis
  'o' + Left drag:     Rotate th camera around ortho-screen axis
