+++ easy3d/renderer/manipulated_camera_frame.cpp
@@ -69,7 +69,7 @@ namespace easy3d {
 	//                 M o u s e    h a n d l i n g                               //
 	////////////////////////////////////////////////////////////////////////////////
 
-	void ManipulatedCameraFrame::action_rotate(int x, int y, int dx, int dy, Camera *const camera, bool screen /* = false */)
+	void ManipulatedCameraFrame::action_rotate(int x, int y, int dx, int dy, Camera *const camera, ScreenAxis axis)
 	{
         if (dx == 0 && dy == 0)
             return;
@@ -83,21 +83,36 @@ namespace easy3d {
         if (has_nan(trans))
             return;
 
-		if (screen) {
-            const float pre_x = float(x - dx);
-            const float pre_y = float(y - dy);
-			const float prev_angle = atan2(pre_y - trans[1], pre_x - trans[0]);
-			const float angle = atan2(y - trans[1], x - trans[0]);
-
-			// The incremental rotation defined in the ManipulatedCameraFrame coordinate system.
-            const quat rot(vec3(0.0, 0.0, 1.0), angle - prev_angle);
-			// Rotates the ManipulatedCameraFrame around its pivotPoint() instead of its origin.
-			rotateAroundPoint(rot, pivotPoint());
-		}
-		else {
+        const int w = camera->screenWidth();
+        const int h = camera->screenHeight();
+        if (axis == NONE) {    // free rotation
             const int pre_x = x - dx;
             const int pre_y = y - dy;
-            const quat& rot = deformedBallQuaternion(x, y, pre_x, pre_y, trans[0], trans[1], camera);
+            const quat& rot = deformedBallQuaternion(x, y, pre_x, pre_y, trans.x, trans.y, w, h);
+            // Rotates the ManipulatedCameraFrame around its pivotPoint() instead of its origin.
+            rotateAroundPoint(rot, pivotPoint());
+        }
+        else {
+            const float pre_x = float(x - dx);
+            const float pre_y = float(y - dy);
+            quat rot;
+            if (axis == ORTHOGONAL) {
+                const float prev_angle = atan2(pre_y - trans[1], pre_x - trans[0]);
+                const float angle = atan2(y - trans[1], x - trans[0]);
+                // The incremental rotation defined in the ManipulatedCameraFrame's coordinate system.
+                rot = quat(vec3(0.0, 0.0, 1.0), angle - prev_angle);
+            }
+            else if (axis == VERTICAL) {
+                const int pre_x = x - dx;
+                const int pre_y = y;    // restricts the movement to be horizontal (so purl vertical rotation)
+                rot = deformedBallQuaternion(x, y, pre_x, pre_y, trans.x, trans.y, w, h);
+            }
+            else if (axis == HORIZONTAL) {
+                const int pre_x = x;    // restricts the movement to be vertical (so purl horizontal rotation)
+                const int pre_y = y - dy;
+                rot = deformedBallQuaternion(x, y, pre_x, pre_y, trans.x, trans.y, w, h);
+            }
+
 			// Rotates the ManipulatedCameraFrame around its pivotPoint() instead of its origin.
 			rotateAroundPoint(rot, pivotPoint());
 		}
@@ -105,18 +120,39 @@ namespace easy3d {
 	}
 
 
-	void ManipulatedCameraFrame::action_translate(int x, int y, int dx, int dy, Camera *const camera, bool screen /* = false */)
+	void ManipulatedCameraFrame::action_translate(int x, int y, int dx, int dy, Camera *const camera, ScreenAxis axis)
 	{
         if (dx == 0 && dy == 0)
             return;
 
-		if (screen) {
[…]
