+++ easy3d/renderer/manipulated_camera_frame.h
@@ -116,8 +116,8 @@ namespace easy3d {
         /*! @name Frame manipulation */
         //@{
     public:
-        void action_rotate(int mouse_x, int mouse_y, int mouse_dx, int mouse_dy, Camera *const camera, bool screen) override;
-        void action_translate(int mouse_x, int mouse_y, int mouse_dx, int mouse_dy, Camera *const camera, bool screen) override;
+        void action_rotate(int mouse_x, int mouse_y, int mouse_dx, int mouse_dy, Camera *const camera, ScreenAxis axis = NONE) override;
+        void action_translate(int mouse_x, int mouse_y, int mouse_dx, int mouse_dy, Camera *const camera, ScreenAxis axis = NONE) override;
         void action_zoom(int dy_wheel, Camera *const camera) override;
 
         virtual void action_turn(float angle_radian, Camera *const camera);        // The rotation around camera Y

[…]
