### 2021-01-06 76ae952
update viewer when animation finishes
