+++ sandbox/Experiment/main.cpp
@@ -10,6 +10,9 @@
 #include <easy3d/util/stop_watch.h>
 #include <easy3d/core/poly_mesh.h>
 #include <easy3d/core/surface_mesh.h>
+#include <easy3d/renderer/key_frame_interpolator.h>
+#include <easy3d/core/curve.h>
+#include <easy3d/util/file_system.h>
 
 #include <cstring>
 #include <cstdio>
@@ -21,142 +24,36 @@
 #define STB_TRUETYPE_IMPLEMENTATION
 #include <3rd_party/stb/stb_truetype.h>
 
-#include <3rd_party/meshoptimizer/src/meshoptimizer.h>
-
 using namespace easy3d;
 
 
-void simplify(const SurfaceMesh& mesh)
+int main()
 {
-    static const size_t lod_count = 5;
+    std::string file = "/Users/lnan/Downloads/LargeMeshAnimation/kf/path_short.kf";
+//    std::string file = "/Users/lnan/Dropbox/Students/2020-PhD-Zexin/data/Lille-3D/test.kf";
 
-    std::vector<vec3> mesh_vertices = mesh.points();
-    std::vector<unsigned int> mesh_indices(mesh.n_faces() * 3);
-    int idx = 0;
-    for (auto f : mesh.faces()) {
-        for (auto v : mesh.vertices(f)) {
-            mesh_indices[idx] = v.idx();
-            ++idx;
-        }
+    std::ifstream is(file.c_str());
+    KeyFrameInterpolator kfi;
+    kfi.read_keyframes(is);
+
+    std::vector<vec3> points(kfi.number_of_keyframes());
+    for (std::size_t i=0; i<kfi.number_of_keyframes(); ++i)
+        points[i] = kfi.keyframe(i).position();
+
+
+    const int resolution = 1000; // Number of line subdivisions to display the spline
+    const int order = 120; // Smoothness of the spline (min 2)
+//    curve::spline::Spline<vec3> spline_curve(order, curve::spline::Spline<vec3>::eUNIFORM);
+    curve::spline::Spline<vec3> spline_curve(order, curve::spline::Spline<vec3>::eOPEN_UNIFORM);
+    spline_curve.set_ctrl_points(points);
+
+
+    std::string output_file = file_system::name_less_extension(file) +  "-Spline-order-" + std::to_string(order) + "-eOPEN_UNIFORM.xyz";
+    std::ofstream output(output_file.c_str());
+    for(int i = 0; i < resolution; ++i) {
+        const vec3 p = spline_curve.eval_f( float(i) / float(resolution-1) );
+        output << p << std::endl;
     }
 
-    StopWatch w;
-
-    // generate 4 LOD levels (1-4), with each subsequent LOD using 70% triangles
-    // note that each LOD uses the same (shared) vertex buffer
-    std::vector<unsigned int> lods[lod_count];
-    lods[0] = mesh_indices;
-
-    for (size_t i = 1; i < lod_count; ++i)
-    {
-        std::vector<unsigned int>& lod = lods[i];
-
-        float threshold = powf(0.7f, float(i));
-        size_t target_index_count = size_t(mesh_indices.size() * threshold) / 3 * 3;
-        float target_error = 1e-2f;
-
-        // we can simplify all the way from base level or from the last result
-        // simplifying from the base level sometimes produces better results, but simplifying from last level is faster
-        const std::vector<unsigned int>& source = lods[i - 1];
-
-        if (source.size() < target_index_count)
-            target_index_count = source.size();
-
-        lod.resize(source.size());
-        lod.resize(meshopt_simplify(&lod[0], &source[0], source.size(), &mesh_vertices[0].x, mesh_vertices.size(), sizeof(vec3), target_index_count, target_error));
-    }
-
-    double middle = w.elapsed_seconds(4);
[…]
