+++ easy3d/core/curve.h
@@ -76,15 +76,15 @@ namespace easy3d {
          * }
          * \endcode
          */
-        template<typename Point>
+        template<typename Point_t>
         inline
-        void quadratic(const Point &A, const Point &B, const Point &C, std::vector<Point> &curve,
+        void quadratic(const Point_t &A, const Point_t &B, const Point_t &C, std::vector<Point_t> &curve,
                        unsigned int bezier_steps = 4, bool include_end = false) {
-            typedef typename Point::FT FT;
+            typedef typename Point_t::FT FT;
             for (unsigned int i = 0; i < bezier_steps; i++) {
                 const FT t = static_cast<FT>(i) / bezier_steps;
-                const Point U = (1.0 - t) * A + t * B;
-                const Point V = (1.0 - t) * B + t * C;
+                const Point_t U = (1.0 - t) * A + t * B;
+                const Point_t V = (1.0 - t) * B + t * C;
                 curve.push_back((1.0 - t) * U + t * V);
             }
 
@@ -128,18 +128,18 @@ namespace easy3d {
          * }
          * \endcode
          */
-        template<typename Point>
+        template<typename Point_t>
         inline
-        void cubic(const Point &A, const Point &B, const Point &C, const Point &D, std::vector<Point> &curve,
+        void cubic(const Point_t &A, const Point_t &B, const Point_t &C, const Point_t &D, std::vector<Point_t> &curve,
                    unsigned int bezier_steps = 4, bool include_end = false) {
-            typedef typename Point::FT FT;
+            typedef typename Point_t::FT FT;
             for (unsigned int i = 0; i < bezier_steps; i++) {
                 const FT t = static_cast<FT>(i) / bezier_steps;
-                const Point U = (1.0 - t) * A + t * B;
-                const Point V = (1.0 - t) * B + t * C;
-                const Point W = (1.0 - t) * C + t * D;
-                const Point M = (1.0 - t) * U + t * V;
-                const Point N = (1.0 - t) * V + t * W;
+                const Point_t U = (1.0 - t) * A + t * B;
+                const Point_t V = (1.0 - t) * B + t * C;
+                const Point_t W = (1.0 - t) * C + t * D;
+                const Point_t M = (1.0 - t) * U + t * V;
+                const Point_t N = (1.0 - t) * V + t * W;
                 curve.push_back((1.0 - t) * M + t * N);
             }
             if (include_end)
@@ -147,27 +147,32 @@ namespace easy3d {
         }
 
 
+        // =============================================================================
+
+
         /**
          * Base class for curve fitting/interpolation
          * \sa Bezier, BSpline, and CatmullRom
          */
-        template<typename Point>
+        template<typename Point_t>
         class Curve {
         public:
-            typedef typename Point::FT FT;
+            typedef typename Point_t::FT FT;
         public:
             Curve() : steps_(10) {}
+
             virtual ~Curve() = default;
 
             void set_steps(int steps) { steps_ = steps; }
 
-            void add_way_point(const Point &point){
+            void add_way_point(const Point_t &point) {
                 way_points_.push_back(point);
                 on_way_point_added();
             }
 
             int node_count() const { return static_cast<int>(nodes_.size()); }
-            const Point& node(int i) const { return nodes_[i]; }
+
+            const Point_t &node(int i) const { return nodes_[i]; }
 
             FT length_from_starting_point(int i) const { return distances_[i]; }
 
@@ -183,25 +188,27 @@ namespace easy3d {
             }
 
         protected:
-            void add_node(const Point &node) {
+            void add_node(const Point_t &node) {
                 nodes_.push_back(node);
-                if(nodes_.size()==1)
+                if (nodes_.size() == 1)
[…]
