### 2021-01-07 a41f2ad
SSAO for textured rendering
