### 2021-01-07 2a80b19
fix a but where the frames were cleared before return; frame rate was not correct in animation
