### 2021-01-07 a89e71e
add string::time_string(double time, int num_digits = 1)
