+++ easy3d/util/string.h
@@ -33,175 +33,160 @@
 
 namespace easy3d {
 
-	/**
-	 * \brief String manipulation utilities.
-	 * \namespace easy3d::string
-	 */
-	namespace string {
-
-		/**
-		 * \brief Splits a string into parts
-		 * \details Splits the string \p in into a list of substrings \p out
-		 * wherever \p separator occurs.
-		 * \param[in] in the input string to split
-		 * \param[in] separator the separator character
-		 * \param[in] out the resulting list of substrings
-		 * \param[in] skip_empty_fields specifies whether empty parts should
-		 * be ignored and not stored in list \p out (this is true by default).
-		 * \see join_strings()
-		 */
-		void split_string(
-			const std::string& in,
-			char separator,
-			std::vector<std::string>& out,
-			bool skip_empty_fields = true
-		);
-
-		/**
-		 * \brief Splits a string into two parts.
-		 * \param[in] in the input string to split
-		 * \param[in] separator the separator character
-		 * \param[in] left the part of the input string on the left
-		 *   of the separator or the empty string if the separator
-		 *   did not appear in the input string
-		 * \param[in] right the right of the input string on the left
-		 *   of the separator or the empty string if the separator
-		 *   did not appear in the input string
-		 * \retval true if the separator was found in the input string
-		 * \retval false otherwise
-		 */
-		bool split_string(
-			const std::string& in,
-			char separator,
-			std::string& left,
-			std::string& right
-		);
-
-		/**
-		 * \brief Join multiple strings
-		 * \details Joins all the strings in list \p in into a single string
-		 * with each element separated by the given \p separator character.
-		 * \param[in] in the list of strings to join
-		 * \param[in] separator the separator character
-		 * \return the joined string
-		 * \see split_string()
-		 */
-		std::string join_strings(
-			const std::vector<std::string>& in,
-			char separator
-		);
-
-		/**
-		 * \brief Join multiple strings
-		 * \details Joins all the strings in list \p in into a single string
-		 * with each element separated by the given \p separator string.
-		 * \param[in] in the list of strings to join
-		 * \param[in] separator the separator string (can be an empty string)
-		 * \return the joined string
-		 * \see split_string()
-		 */
-		std::string join_strings(
-			const std::vector<std::string>& in,
-			const std::string& separator
-		);
-
-		/**
-		 * \brief Replaces the old sub-string by a new sub-string
-		 */
-		void replace_substring(
-			std::string& in,
-			const std::string& old_substr,
-			const std::string& new_substr
-		);
-
-		/**
-		 * \brief Converts a string to lowercase
-		 * \details The conversion is done in place in the string \p s.
-		 * \param[in,out] s The string to convert
-		 * \see to_uppercase()
-		 */
-		std::string to_lowercase(const std::string& s);
-
-		/**
-		 * \brief Converts a string to uppercase
-		 * \details The conversion is done in place in the string \p s.
-		 * \param[in,out] s The string to convert
-		 * \see to_lowercase()
-		 */
-		std::string to_uppercase(const std::string& s);
-
-		/**
-		 * \brief Creates a one char string
-		 * \param[in] c the character to convert to a string
-		 * \return a string that contains character \p c
-		 */
-		inline std::string char_to_string(char c) {
-			char s[2];
-			s[0] = c;
-			s[1] = '\0';
-			return std::string(s);
-		}
-
-		/**
-		 * \brief Adds quotes to a string
-		 * \details Adds character \p quote at the beginning and the end of
-		 * string \p s and returns the resulting string.
-		 * \param[in] s the string to quote
-		 * \param[in] quotes the quoting char (default is '"')
-		 * \return the quoted string
-		 */
[…]
