### 2021-01-07 572287a
allow to pick and highlight models
