+++ easy3d/renderer/renderer.h
@@ -78,11 +78,15 @@ namespace easy3d {
 
         //-------------------- rendering functionalities  -----------------------
 
+        /// Returns whether the model is currently visible.
         bool is_visible() const { return visible_; }
+        /// Shows/Hides the model.
         void set_visible(bool b) { visible_ = b; }
 
+        /// Returns whether the model has been selected.
         bool is_selected() const { return selected_; }
-        void set_selected(bool b) { selected_ = b; }
+        /// Select/Deselect the model. The state of all its drawables will change accordingly.
+        void set_selected(bool b);
 
         /**
          * @brief Invalidates the rendering buffers of the model and thus updates the rendering (delayed in rendering).
