+++ easy3d/core/interpolation.h
@@ -1,0 +1,482 @@
+/*
+ * spline.h
+ *
+ * simple cubic spline interpolation library without external
+ * dependencies
+ *
+ * ---------------------------------------------------------------------
+ * Copyright (C) 2011, 2014 Tino Kluge (ttk448 at gmail.com)
+ *
+ *  This program is free software; you can redistribute it and/or
+ *  modify it under the terms of the GNU General Public License
+ *  as published by the Free Software Foundation; either version 2
+ *  of the License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * ---------------------------------------------------------------------
+ *
+ */
+
+
+#ifndef TK_SPLINE_H
+#define TK_SPLINE_H
+
+#include <cstdio>
+#include <cassert>
+#include <vector>
+#include <algorithm>
+
+
+// https://kluge.in-chemnitz.de/opensource/spline/
+
+namespace interpolation {
+
+    // band matrix solver
+    class band_matrix {
+    private:
+        std::vector<std::vector<double> > m_upper;  // upper band
+        std::vector<std::vector<double> > m_lower;  // lower band
+    public:
+        band_matrix() {};                             // constructor
+        band_matrix(int dim, int n_u, int n_l);       // constructor
+        ~band_matrix() {};                            // destructor
+        void resize(int dim, int n_u, int n_l);      // init with dim,n_u,n_l
+        int dim() const;                             // matrix dimension
+        int num_upper() const {
+            return m_upper.size() - 1;
+        }
+
+        int num_lower() const {
+            return m_lower.size() - 1;
+        }
+
+        // access operator
+        double &operator()(int i, int j);            // write
+        double operator()(int i, int j) const;      // read
+        // we can store an additional diogonal (in m_lower)
+        double &saved_diag(int i);
+
+        double saved_diag(int i) const;
+
+        void lu_decompose();
+
+        std::vector<double> r_solve(const std::vector<double> &b) const;
+
+        std::vector<double> l_solve(const std::vector<double> &b) const;
+
+        std::vector<double> lu_solve(const std::vector<double> &b,
+                                     bool is_lu_decomposed = false);
+
+    };
+
+
+    /**
+     * \brief Cubic spline interpolation.
+     * Example for 3D curve interpolation
+     *      \code
+     *          std::vector<double> T(points.size()), X(points.size()), Y(points.size()), Z(points.size());
+     *          double t = 0.0;
+     *          for (std::size_t i = 0; i < points.size(); ++i) {
+     *              const auto &p = points[i];
+     *              if (i > 0)
+     *                  t += std::max(distance(kfi.keyframe(i-1).position(), p), float(1e-6));
+     *              T[i] = t;
+     *              X[i] = p.x;
+     *              Y[i] = p.y;
+     *              Z[i] = p.z;
+     *          }
+     *
+     *          interpolation::spline x_spline;     x_spline.set_points(T, X);
+     *          interpolation::spline y_spline;     y_spline.set_points(T, Y);
[…]
