### 2021-01-07 d41275c
add SplineCurveFitting<Point>, SplineInterpolation<FT>, and Matrix<FT>
