+++ easy3d/core/spline_interpolation.h
@@ -1,0 +1,503 @@
+/**
+ * Copyright (C) 2015 by Liangliang Nan (liangliang.nan@gmail.com)
+ * https://3d.bk.tudelft.nl/liangliang/
+ *
+ * This file is part of Easy3D. If it is useful in your research/work,
+ * I would be grateful if you show your appreciation by citing it:
+ * ------------------------------------------------------------------
+ *      Liangliang Nan.
+ *      Easy3D: a lightweight, easy-to-use, and efficient C++
+ *      library for processing and rendering 3D data. 2018.
+ * ------------------------------------------------------------------
+ * Easy3D is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License Version 3
+ * as published by the Free Software Foundation.
+ *
+ * Easy3D is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef EASY3D_CORE_CUBIC_SPLINE_INTERPOLATION_H
+#define EASY3D_CORE_CUBIC_SPLINE_INTERPOLATION_H
+
+#include <cstdio>
+#include <cassert>
+#include <vector>
+#include <algorithm>
+
+#include <easy3d/util/logging.h>
+
+
+namespace easy3d {
+
+
+    /**
+     * \brief Cubic spline interpolation.
+     * \details Spline interpolation have many applications, e.g., curve interpolation (for any dimensions).
+     *      The math behind this implementation is described here: https://kluge.in-chemnitz.de/opensource/spline/
+     * Example for 3D curve interpolation
+     *      \code
+     *          std::vector<double> T(points.size()), X(points.size()), Y(points.size()), Z(points.size());
+     *          double t = 0.0;
+     *          for (std::size_t i = 0; i < points.size(); ++i) {
+     *              const auto &p = points[i];
+     *              if (i > 0)
+     *                  t += distance(kfi.keyframe(i-1).position(), p);
+     *              T[i] = t;
+     *              X[i] = p.x;
+     *              Y[i] = p.y;
+     *              Z[i] = p.z;
+     *          }
+     *          SplineInterpolation<double> x_spline;x_spline.set_data(T, X);
+     *          SplineInterpolation<double> y_spline;y_spline.set_data(T, Y);
+     *          SplineInterpolation<double> z_spline; z_spline.set_data(T, Z);
+     *          double total_length = t;
+     *          int steps = 1000;
+     *          double interval = total_length / steps;
+     *          for (int i=0; i<=steps; ++i) {
+     *              double d = i * interval;
+     *              vec3 p(x_spline(d), y_spline(d), z_spline(d));
+     *              std::cout << p << std::endl;
+     *          }
+     *      \endcode
+     * \class SplineInterpolation easy3d/core/spline_interpolation.h
+     */
+    template<typename FT>
+    class SplineInterpolation {
+    public:
+        enum BoundaryType {
+            first_deriv = 1,
+            second_deriv = 2
+        };
+
+    public:
+        /// Constructor.
+        /// Sets default boundary condition to be zero curvature at both ends
+        SplineInterpolation() : left_(second_deriv), right_(second_deriv),
+                        left_value_(0.0), right_value_(0.0),
[…]
