+++ easy3d/core/spline_curve_fitting.h
@@ -1,0 +1,314 @@
+/**
+ * Copyright (C) 2015 by Liangliang Nan (liangliang.nan@gmail.com)
+ * https://3d.bk.tudelft.nl/liangliang/
+ *
+ * This file is part of Easy3D. If it is useful in your research/work,
+ * I would be grateful if you show your appreciation by citing it:
+ * ------------------------------------------------------------------
+ *      Liangliang Nan.
+ *      Easy3D: a lightweight, easy-to-use, and efficient C++
+ *      library for processing and rendering 3D data. 2018.
+ * ------------------------------------------------------------------
+ * Easy3D is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License Version 3
+ * as published by the Free Software Foundation.
+ *
+ * Easy3D is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef EASY3D_CORE_SPLINE_CURVE_FITTING_H
+#define EASY3D_CORE_SPLINE_CURVE_FITTING_H
+
+
+#include <vector>
+#include <cassert>
+
+
+namespace easy3d {
+
+
+    /**
+     * @brief Spline curve fitting for arbitrary dimensions.
+     * \details It allow to set the order of the spline and the end points behavior (passing or not).
+     *      The class can be instantiated with any point type (1D (float), 2D, 3D etc.) as long as the
+     *      necessary operator overloads are implemented.
+     * @note This class use the efficient blossom algorithm to compute a position on the curve.
+     * @tparam Point_t: type of a point operators such as '+' '*' must be correctly overloaded. The default
+     *      constructor must be defined to return the null vector (0, 0 ,0 ...)
+     * @tparam Real_t: floating point representation of the points (float, double etc.)
+     * Example usage:
+     *      \code
+     *          const int resolution = 1000;    // Number of line subdivisions to display the spline
+     *          const int order = 3;            // Smoothness of the spline (min 2)
+     *          SplineCurveFitting<vec3> fitter(order, SplineCurveFitting<vec3>::eOPEN_UNIFORM);
+     *          fitter.set_ctrl_points(points);
+     *          for(int i = 0; i < resolution; ++i) {
+     *               const vec3 p = fitter.eval_f( float(i) / float(resolution-1) );
+     *               std::cout << p << std::endl;
+     *          }
+     *      \endcode
+     * @class SplineCurveFitting easy3d/core/spline_curve_fitting.h
+     */
+    template<typename Point_t>
+    class SplineCurveFitting {
+    public:
+        typedef typename Point_t::FT Real_t;
+        enum Node_e {
+            eUNIFORM,
+            eOPEN_UNIFORM ///< Connected to the first and last control points
+        };
+
+    public:
+
+        /// Type of the nodal vector
+        /// @param k : order of the spline (minimum is two)
+        /// @param node_type : nodal vector type (uniform, open_uniform)
+        /// This will define the behavior of the spline with its control points
+        /// as well as its speed according to its parameter.
+        SplineCurveFitting(int k = 2, Node_e node_type = eOPEN_UNIFORM);
+
+        /// Set the position of the spline control points.
+        void set_ctrl_points(const std::vector<Point_t> &point);
+
+        /// Get the control points of the spline
+        void get_ctrl_points(std::vector<Point_t> &points) const;
+
+        /// The the nodal vector type
+        void set_node_type(Node_e type);
+
[…]
