+++ easy3d/core/CMakeLists.txt
@@ -26,6 +26,8 @@ set(${PROJECT_NAME}_HEADERS
         rect.h
         segment.h
         signal.h
+        spline_curve_fitting.h
+        spline_interpolation.h
         surface_mesh.h
         poly_mesh.h
         polygon.h

[…]
