+++ easy3d/util/timer.h
@@ -40,10 +40,8 @@ namespace easy3d {
      * (e.g., the [QTimer of Qt](http://doc.qt.io/archives/qt-5.5/qtimer.html)).
      * This Timer class provides a single-header implementation.
      *
-     * With Timer, tasks (i.e., calling to functions) can be easily
-     * scheduled at either constant intervals or after a specified
-     * period. Timer supports any types of functions with any number
-     * of arguments.
+     * With Timer, tasks (i.e., calling to functions) can be easily scheduled at either constant intervals or after
+     * a specified period. Timer supports any types of functions with any number of arguments.
      *
      ** Code examples:
      *### Example 1: calling to non-member functions
@@ -105,16 +103,18 @@ namespace easy3d {
          * \brief Executes function \p func after \p delay milliseconds.
          * \param func The pointer to the function.
          * \param delay The time to be delayed, in milliseconds.
-         * \note Arguments must be \c const.
+         * \note Arguments must be \c const. When an argument is a pointer, it has to be of type "Class* const".
          *      \code
+         *          class MyClass {};
+         *          MyClass* const data = new MyClass;   // MyClass* data = new MyClass; will not work
          *          const int taskID = 3;
          *          const std::string msg = "Finished";
-         *          Timer::single_shot(5000, &reset, taskID, msg);
+         *          Timer::single_shot(5000, &reset, taskID, msg, data);
          *      \endcode
-         * so this won't work
+         * So this won't work
          *      \code
-         *          Timer::single_shot(5000, &reset, 3, "Finished");
-         *      \endcode       
+         *          Timer::single_shot(5000, &reset, 3, "Finished", data);
+         *      \endcode
          */
         template < class Function, class... Args >
         static void single_shot(int delay, Function&& func, Args&&... args);
@@ -124,34 +124,38 @@ namespace easy3d {
          * \param delay The time to be delayed, in milliseconds.
          * \param owner The pointer to \c Class instance, e.g., '&a' for 'Class a' or 'this' within Class.
          * \param func The pointer to the member function of \c owner, e.g., '&Class::foo()'.
-         * \note Arguments must be \c const.
+         * \note Arguments must be \c const. When an argument is a pointer, it has to be of type "Class* const".
          *      \code
+         *          class MyClass {};
+         *          MyClass* const data = new MyClass;   // MyClass* data = new MyClass; will not work
          *          const int taskID = 3;
          *          const std::string msg = "Finished";
-         *          Timer::single_shot(5000, viewer, &Viewer::reset, taskID, msg);
+         *          Timer::single_shot(5000, viewer, &Viewer::reset, taskID, msg, data);
          *      \endcode    
          * so this won't work
          *      \code
-         *          Timer::single_shot(5000, viewer, &Viewer::reset, 3, "Finished");
-         *      \endcode       
+         *          Timer::single_shot(5000, viewer, &Viewer::reset, 3, "Finished", data);
+         *      \endcode
          */
         template < class Class, class Function, class... Args >
-        static void single_shot(int delay, Class&& owner, Function&& func, Args&&... args);
+        static void single_shot(int delay, Function&& func, Class&& owner, Args&&... args);
 
         /**
          * \brief Executes function \p func after \p delay milliseconds. 
          * \details This is the same as single_shot() except that it is not static.
          * \param delay The time to be delayed, in milliseconds.
          * \param func The pointer to the function.
-         * \note Arguments must be \c const.
+         * \note Arguments must be \c const. When an argument is a pointer, it has to be of type "Class* const".
          *      \code
+         *          class MyClass {};
[…]
