### 2021-01-07 f525d2d
fix a bug where the manipulated frame of a model was not initialized
