+++ easy3d/core/model.cpp
@@ -31,6 +31,7 @@ namespace easy3d {
             : name_(name)
             , bbox_known_(false)
             , renderer_(nullptr)
+            , frame_(nullptr)
     {
     }
 
