+++ easy3d/util/logging_stl.h
@@ -44,12 +44,9 @@
 // - GLOG_STL_LOGGING_FOR_EXT_SLIST     - <ext/slist>
 //
 
-#ifndef UTIL_GTL_STL_LOGGING_INL_H_
-#define UTIL_GTL_STL_LOGGING_INL_H_
+#ifndef EASY3D_UTIL_LOGGING_STL_H
+#define EASY3D_UTIL_LOGGING_STL_H
 
-#if !@ac_cv_cxx_using_operator@
-# error We do not support stl_logging for this compiler
-#endif
 
 #include <deque>
 #include <list>
@@ -59,51 +56,33 @@
 #include <utility>
 #include <vector>
 
-#ifdef GLOG_STL_LOGGING_FOR_UNORDERED
-# include <unordered_map>
-# include <unordered_set>
-#endif
-
-#ifdef GLOG_STL_LOGGING_FOR_TR1_UNORDERED
-# include <tr1/unordered_map>
-# include <tr1/unordered_set>
-#endif
-
-#ifdef GLOG_STL_LOGGING_FOR_EXT_HASH
-# include <ext/hash_set>
-# include <ext/hash_map>
-#endif
-#ifdef GLOG_STL_LOGGING_FOR_EXT_SLIST
-# include <ext/slist>
-#endif
 
 // Forward declare these two, and define them after all the container streams
 // operators so that we can recurse from pair -> container -> container -> pair
 // properly.
 template<class First, class Second>
-std::ostream& operator<<(std::ostream& out, const std::pair<First, Second>& p);
+std::ostream &operator<<(std::ostream &out, const std::pair<First, Second> &p);
 
-@ac_google_start_namespace@
+namespace google {
 
-template<class Iter>
-void PrintSequence(std::ostream& out, Iter begin, Iter end);
+    template<class Iter>
+    void PrintSequence(std::ostream &out, Iter begin, Iter end);
 
-@ac_google_end_namespace@
+}
 
 #define OUTPUT_TWO_ARG_CONTAINER(Sequence) \
 template<class T1, class T2> \
 inline std::ostream& operator<<(std::ostream& out, \
                                 const Sequence<T1, T2>& seq) { \
-  @ac_google_namespace@::PrintSequence(out, seq.begin(), seq.end()); \
+  google::PrintSequence(out, seq.begin(), seq.end()); \
   return out; \
 }
 
 OUTPUT_TWO_ARG_CONTAINER(std::vector)
+
 OUTPUT_TWO_ARG_CONTAINER(std::deque)
+
 OUTPUT_TWO_ARG_CONTAINER(std::list)
-#ifdef GLOG_STL_LOGGING_FOR_EXT_SLIST
-OUTPUT_TWO_ARG_CONTAINER(__gnu_cxx::slist)
-#endif
 
 #undef OUTPUT_TWO_ARG_CONTAINER
 
@@ -111,11 +90,12 @@ OUTPUT_TWO_ARG_CONTAINER(__gnu_cxx::slist)
 template<class T1, class T2, class T3> \
 inline std::ostream& operator<<(std::ostream& out, \
                                 const Sequence<T1, T2, T3>& seq) { \
-  @ac_google_namespace@::PrintSequence(out, seq.begin(), seq.end()); \
+  google::PrintSequence(out, seq.begin(), seq.end()); \
   return out; \
 }
 
 OUTPUT_THREE_ARG_CONTAINER(std::set)
+
 OUTPUT_THREE_ARG_CONTAINER(std::multiset)
 
 #undef OUTPUT_THREE_ARG_CONTAINER
@@ -124,24 +104,13 @@ OUTPUT_THREE_ARG_CONTAINER(std::multiset)
 template<class T1, class T2, class T3, class T4> \
[…]
