### 2021-01-08 9f13bf0
fix a bug where the maniplator attached to a model is null
