+++ easy3d/util/logging.h
@@ -28,9 +28,7 @@
 // Portions of this code are taken from the GLOG package.  This code is only a
 // small subset of the GLOG functionality. Notable differences from GLOG
 // behavior include lack of support for displaying unprintable characters and
-// lack of stack trace information upon failure of the CHECK macros.  On
-// non-Android systems, log output goes to std::cerr and is not written to a
-// file.
+// lack of stack trace information upon failure of the CHECK macros.
 //
 // CHECK macros are defined to test for conditions within code.  Any CHECK that
 // fails will log the failure and terminate the application.
@@ -86,40 +84,33 @@ namespace google {
     const int ERROR = ::ERROR;
     const int FATAL = ::FATAL;
 
-
-    // Set the file to which the log messages will be written into.
-    void set_log_file(std::ofstream *os);
-
-// Sink class used for integration with mock and test functions. If sinks are
-// added, all log output is also sent to each sink through the send function.
-// In this implementation, WaitTillSent() is called immediately after the send.
-// This implementation is not thread safe.
+    // Sink class used for integration with mock and test functions. If sinks are
+    // added, all log output is also sent to each sink through the send function.
+    // In this implementation, WaitTillSent() is called immediately after the send.
+    // This implementation is not thread safe.
     class LogSink {
     public:
         virtual ~LogSink() {}
 
-        virtual void write_log(int severity, const std::string& msg) = 0;
-        virtual void WaitTillSent() = 0;
+        virtual void send(google::LogSeverity severity,
+                          int line_number,
+                          const std::string& full_path,
+                          const std::string& short_name,
+                          const std::string& time,
+                          const std::string& pid_tid,
+                          const std::string& msg) = 0;
+        virtual void WaitTillSent() { /* client code should implement this */ }
     };
 
-// Global set of log sinks. The actual object is defined in logging.cc.
-    extern std::set<LogSink *> log_sinks_global;
 
-    inline void InitGoogleLogging(char *argv) {
-        (void) argv;
-        // Do nothing; this is ignored.
-    }
+    // Note: the Log sink functions are not thread safe.
+    void AddLogSink(LogSink *sink);
 
-// Note: the Log sink functions are not thread safe.
-    inline void AddLogSink(LogSink *sink) {
-        // TODO(settinger): Add locks for thread safety.
-        log_sinks_global.insert(sink);
-    }
+    void RemoveLogSink(LogSink *sink);
 
-    inline void RemoveLogSink(LogSink *sink) {
-        log_sinks_global.erase(sink);
-    }
-
+    // the function that will be called before abortion.
+    typedef void (*logging_fail_func_t)();
+    void InstallFailureFunction(void (*fail_func)());
 
     // ---------------------------- Logger Class --------------------------------
 
@@ -132,7 +123,6 @@ namespace google {
     class MessageLogger {
     public:
         MessageLogger(const char *file, int line, google::LogSeverity severity);
-
         MessageLogger(const char *file, int line, google::LogSeverity severity, int ctr);
 
         // Output the contents of the stream to the proper channel on destruction.
@@ -141,14 +131,14 @@ namespace google {
         // Return the stream associated with the logger object.
         std::stringstream &stream() { return stream_; }
 
-    private:
         static const std::string severity_labels[4];
-        static const char *severity_color_code[4];
 
+    private:
+        static const char *severity_color_code[4];
         static bool terminal_supports_colors();
 
     private:
-        void LogToSinks(const std::string& msg) const;
+        void LogToSinks() const;
[…]
