+++ applications/Mapple/main_window.h
@@ -196,7 +196,10 @@ private:
     QString strippedName(const QString &fullFileName);
 
     void notify(std::size_t percent, bool update_viewer) override;
-    void write_log(int severity, const std::string &message) override;
+
+    void send(google::LogSeverity severity, int line_number, const std::string &full_path,
+              const std::string &short_name, const std::string &time,
+              const std::string &pid_tid, const std::string &msg) override;
 
 private:
     PaintCanvas*   viewer_;
