### 2021-01-09 233a106
allow to clear log
