+++ applications/Mapple/widgets/widget_log.h
@@ -16,6 +16,7 @@ private Q_SLOTS :
     void showContextMenu(const QPoint &p);
     void copySelected();
     void exportLog();
+    void clearLog();
 
 private:
     QMenu *popupMenu_;
