### 2021-01-09 2d6b311
Update logging.h
