+++ easy3d/util/stack_tracer.cpp
@@ -29,6 +29,8 @@
 #include <cstdlib>
 #include <string>
 #include <sstream>
+#include <iostream>
+
 
 #if defined(WIN32) || defined(_WIN32) || defined(__WIN32__)
 
@@ -36,6 +38,8 @@
 #include <execinfo.h> // for backtrace
 #include <dlfcn.h>    // for dladdr
 #include <cxxabi.h>   // for __cxa_demangle
+#include <signal.h>
+#include<unistd.h>
 #endif
 
 
@@ -46,7 +50,7 @@ namespace easy3d {
     // If it doesn't work on Windows or Linux, maybe I should consider using 'backward':
     //      https://github.com/bombela/backward-cpp
 
-    std::string StackTracer::dump(int skip, int amount) const {
+    std::string StackTracer::back_trace(int skip, int amount) {
 #if 0
         using namespace backward;
         StackTrace st; st.load_here(32);
@@ -70,7 +74,7 @@ namespace easy3d {
 
         std::ostringstream trace_buffer;
 
-        void* callstack[128];
+        void *callstack[128];
         const int nMaxFrames = sizeof(callstack) / sizeof(callstack[0]);
 
         int nFrames = backtrace(callstack, nMaxFrames);
@@ -82,7 +86,7 @@ namespace easy3d {
             nFrames = (skip + amount) < nFrames ? skip + amount : nFrames;
         }
 
-        char** symbols = backtrace_symbols(callstack, nFrames);
+        char **symbols = backtrace_symbols(callstack, nFrames);
 
         for (int i = skip; i < nFrames; i++) {
             // this is the raw data
@@ -91,8 +95,8 @@ namespace easy3d {
             Dl_info info;
             if (dladdr(callstack[i], &info)) {
                 int status(0);
-                char* demangled = abi::__cxa_demangle(info.dli_sname, nullptr, 0, &status);
-                const char* name;
+                char *demangled = abi::__cxa_demangle(info.dli_sname, nullptr, 0, &status);
+                const char *name;
                 if (status == 0)
                     name = demangled ? demangled : "(no demangled name)";
                 else
@@ -104,8 +108,7 @@ namespace easy3d {
                              << "\t" << name << "\n";
                 if (demangled)
                     free(demangled);
-            }
-            else {
+            } else {
                 // the "- skip" makes sure the most recent trace has an index of 0.
                 trace_buffer << "\t" << std::setw(5) << std::left << std::setfill(' ') << i - skip
                              << "\t" << std::setw(20) << std::setfill(' ') << "[unknown]"
@@ -124,4 +127,91 @@ namespace easy3d {
 
         return trace_buffer.str();
     }
+
+
+    StackTracer::StackTracer() {
+        struct sigaction sa;
+        sa.sa_sigaction = (void (*)(int, siginfo_t *, void *))signal_handler;
+        sigemptyset(&sa.sa_mask);
+        sa.sa_flags = SA_RESTART | SA_SIGINFO;
+        if (sigaction(SIGSEGV, &sa, nullptr) < 0) {
+            perror("DeathHandler - sigaction(SIGSEGV)");
+        }
+        if (sigaction(SIGABRT, &sa, nullptr) < 0) {
+            perror("DeathHandler - sigaction(SIGABBRT)");
+        }
+        if (sigaction(SIGFPE, &sa, nullptr) < 0) {
+            perror("DeathHandler - sigaction(SIGFPE)");
+        }
+    }
+
+
+    StackTracer::~StackTracer() {
+        // Disable alternative signal handler stack
+        stack_t altstack;
+        altstack.ss_sp = nullptr;
+        altstack.ss_size = 0;
+        altstack.ss_flags = SS_DISABLE;
[…]
