### 2021-01-09 95a0697
Update stack_tracer.cpp
