### 2021-01-09 3e3a0bb
format logging
