### 2021-01-09 ee3feaa
now failure handling is based on backward (cross platform)
