+++ easy3d/util/stack_tracer.cpp
@@ -23,201 +23,104 @@
  */
 
 #include <easy3d/util/stack_tracer.h>
-#include <easy3d/util/file_system.h>
 
 #include <iomanip>
-#include <cstdlib>
 #include <string>
-#include <sstream>
-#include <iostream>
 
-
-#if defined(WIN32) || defined(_WIN32) || defined(__WIN32__)
-
-#else
-#include <execinfo.h> // for backtrace
-#include <dlfcn.h>    // for dladdr
-#include <cxxabi.h>   // for __cxa_demangle
-#include <signal.h>
-#include <unistd.h>
-#include <sys/wait.h>   // for waitpid
-#endif
-
-#include <easy3d/util/backward.h>
+#include <easy3d/util/file_system.h>
 #include <easy3d/util/logging.h>
+
+#include <3rd_party/backward/backward.hpp>
+
 
 namespace easy3d {
 
 
-    // This function has been test on macOS only.
-    // If it doesn't work on Windows or Linux, maybe I should consider using 'backward':
-    //      https://github.com/bombela/backward-cpp
-
-    std::string StackTracer::back_trace(int skip, int amount) {
-#if 1
-        using namespace backward;
-        StackTrace st;
-        st.load_here(amount + skip);
-
-        TraceResolver tr;
-        tr.load_stacktrace(st);
-        std::stringstream trace_buffer;
-        for (size_t i = skip; i < st.size(); ++i) {
-            const ResolvedTrace trace = tr.resolve(st[i]);
-            // the "- skip" makes sure the most recent trace has an index of 0.
-            trace_buffer << "\t" << std::setw(5) << std::left << std::setfill(' ') << i - skip
-                         << "\t" << std::setw(20) << std::setfill(' ') << file_system::base_name(trace.object_filename)
-                         << "\t" << trace.object_function;
-            if (i != st.size() - 1)
-                trace_buffer << "\n";
-        }
-        return trace_buffer.str();
-#else
-
-
-        std::ostringstream trace_buffer;
-
-        void *callstack[128];
-        const int nMaxFrames = sizeof(callstack) / sizeof(callstack[0]);
-
-        int nFrames = backtrace(callstack, nMaxFrames);
-
-        if (amount == 0) {
-            return trace_buffer.str();
-        } else if (amount > 0) {
-            // limit the number of frames
-            nFrames = (skip + amount) < nFrames ? skip + amount : nFrames;
-        }
-
-        char **symbols = backtrace_symbols(callstack, nFrames);
-
-        for (int i = skip; i < nFrames; i++) {
-            // this is the raw data
-            // printf("%s\n", symbols[i]);
-
-            Dl_info info;
-            if (dladdr(callstack[i], &info)) {
-                int status(0);
-                char *demangled = abi::__cxa_demangle(info.dli_sname, nullptr, 0, &status);
-                const char *name;
-                if (status == 0)
-                    name = demangled ? demangled : "(no demangled name)";
-                else
-                    name = info.dli_sname ? info.dli_sname : "(no dli_sname)";
-
-                // the "- skip" makes sure the most recent trace has an index of 0.
-                trace_buffer << "\t" << std::setw(5) << std::left << std::setfill(' ') << i - skip
-                             << "\t" << std::setw(20) << std::setfill(' ') << file_system::base_name(info.dli_fname)
-                             << "\t" << name << "\n";
-                if (demangled)
-                    free(demangled);
-            } else {
-                // the "- skip" makes sure the most recent trace has an index of 0.
-                trace_buffer << "\t" << std::setw(5) << std::left << std::setfill(' ') << i - skip
-                             << "\t" << std::setw(20) << std::setfill(' ') << "[unknown]"
-                             << "\t" << "[unknown]" << "\n";
-            }
-        }
-
-        if (symbols)
[…]
