+++ 3rd_party/CMakeLists.txt
@@ -16,5 +16,7 @@ add_subdirectory(triangle)
 add_subdirectory(tetgen)
 add_subdirectory(libtess)
 
+add_subdirectory(backward)
+
 
 

[…]
