### 2021-01-09 aae063e
allow easylogging++
