+++ sandbox/Experiment/main.cpp
@@ -1,59 +1,67 @@
-#include <easy3d/fileio/resources.h>
+#if 0
+
+#define ELPP_FEATURE_ALL
+#define ELPP_FEATURE_CRASH_LOG
+//#define ELPP_EXPERIMENTAL_ASYNC
+
+#include <easy3d/util/easylogging++.h>
+
+
+INITIALIZE_EASYLOGGINGPP
+
+void myCrashHandler(int sig) {
+    LOG(ERROR) << "Woops! Crashed!";
+    // FOLLOWING LINE IS OPTIONAL
+    el::Helpers::logCrashReason(sig, true);
+    // FOLLOWING LINE IS ABSOLUTELY NEEDED AT THE END IN ORDER TO ABORT APPLICATION
+    el::Helpers::crashAbort(sig);
+}
+int main(void) {
+    el::Helpers::setCrashHandler(myCrashHandler);
+
+    LOG(INFO) << "My crash handler!";
+
+    std::cout << std::endl << "    ======= Backtrace: =========" << std::endl << el::base::debug::StackTrace();
+
+    int* i;
+    *i = 0; // Crash!
+
+    return 0;
+}
+
+#else
+
+#include <easy3d/util/stack_tracer.h>
 #include <easy3d/util/logging.h>
-#include <easy3d/viewer/viewer.h>
-#include <easy3d/renderer/texture.h>
-#include <easy3d/core/graph.h>
-#include <easy3d/fileio/graph_io.h>
-#include <easy3d/fileio/surface_mesh_io.h>
-#include <easy3d/core/curve.h>
-#include <easy3d/util/string.h>
-#include <easy3d/util/stop_watch.h>
-#include <easy3d/core/poly_mesh.h>
-#include <easy3d/core/surface_mesh.h>
-#include <easy3d/renderer/key_frame_interpolator.h>
-#include <easy3d/core/curve.h>
-#include <easy3d/util/file_system.h>
 
-#include <cstring>
-#include <cstdio>
-#include <vector>
-#include <fstream>
-#include <sstream>
-#include <unordered_map>
-
-#define STB_TRUETYPE_IMPLEMENTATION
-#include <3rd_party/stb/stb_truetype.h>
+#include <iostream>
 
 using namespace easy3d;
 
 
-int main()
-{
-    std::string file = "/Users/lnan/Downloads/LargeMeshAnimation/kf/path_short.kf";
-//    std::string file = "/Users/lnan/Dropbox/Students/2020-PhD-Zexin/data/Lille-3D/test.kf";
 
-    std::ifstream is(file.c_str());
-    KeyFrameInterpolator kfi;
-    kfi.read_keyframes(is);
+int print(int * v) {
+     std::cout << *v << std::endl;
+    LOG_IF(FATAL, v== nullptr) << "v== nullptr";
 
-    std::vector<vec3> points(kfi.number_of_keyframes());
-    for (std::size_t i=0; i<kfi.number_of_keyframes(); ++i)
-        points[i] = kfi.keyframe(i).position();
-
-
-    const int resolution = 1000; // Number of line subdivisions to display the spline
-    const int order = 120; // Smoothness of the spline (min 2)
-//    curve::spline::Spline<vec3> spline_curve(order, curve::spline::Spline<vec3>::eUNIFORM);
-    curve::spline::Spline<vec3> spline_curve(order, curve::spline::Spline<vec3>::eOPEN_UNIFORM);
-    spline_curve.set_ctrl_points(points);
-
-
-    std::string output_file = file_system::name_less_extension(file) +  "-Spline-order-" + std::to_string(order) + "-eOPEN_UNIFORM.xyz";
[…]
