### 2021-01-09 46f2d96
Update test_logging.cpp
