+++ tutorials/Tutorial_203_CompositeView/main.cpp
@@ -34,7 +34,7 @@ using namespace easy3d;
 
 int main(int argc, char** argv) {
     // Initialize logging.
-    logging::initialize();
+    logging::initialize(WARNING);
 
     CompositeView viewer("Tutorial_203_CompositeView");
 
