+++ tutorials/Tutorial_104_SurfaceMesh/main.cpp
@@ -36,7 +36,7 @@ using namespace easy3d;
 
 int main(int argc, char** argv) {
     // Initialize logging.
-    logging::initialize();
+    logging::initialize(WARNING);
 
 	// Easy3D provides two options to construct a surface mesh.
     //  - Option 1: use the add_vertex() and add_[face/triangle/quad]() functions of SurfaceMesh. You can only choose
