+++ tutorials/Tutorial_502_HardShadow/main.cpp
@@ -38,7 +38,7 @@ using namespace easy3d;
 
 int main(int argc, char **argv) {
     // Initialize logging.
-    logging::initialize();
+    logging::initialize(WARNING);
 
     const std::string file = resource::directory() + "/data/room.obj";
 
