+++ tutorials/Tutorial_113_PolyMesh_Connectivity/main.cpp
@@ -125,7 +125,7 @@ PolyMesh *old_mesh_from_previous_example() {
 
 int main(int argc, char **argv) {
     // Initialize logging.
-    logging::initialize();
+    logging::initialize(WARNING);
 
     // Create mesh object
     PolyMesh *mesh = old_mesh_from_previous_example();
