+++ tutorials/Tutorial_601_Cloud_NormalEstimation/main.cpp
@@ -39,7 +39,7 @@ using namespace easy3d;
 
 int main(int argc, char **argv) {
     // Initialize logging.
-    logging::initialize();
+    logging::initialize(WARNING);
 
     const std::string file = resource::directory() + "/data/bunny.bin";
 
