+++ tutorials/Tutorial_105_SurfaceMesh_Connectivity/main.cpp
@@ -67,7 +67,7 @@ SurfaceMesh* old_mesh_from_previous_example() {
 
 int main(int argc, char** argv) {
     // Initialize logging.
-    logging::initialize();
+    logging::initialize(WARNING);
 
 	SurfaceMesh* mesh = old_mesh_from_previous_example();
 	
