+++ easy3d/renderer/soft_shadow.cpp
@@ -31,6 +31,7 @@
 #include <easy3d/renderer/frustum.h>
 #include <easy3d/renderer/drawable_triangles.h>
 #include <easy3d/renderer/setting.h>
+#include <easy3d/renderer/transform.h>
 #include <easy3d/renderer/clipping_plane.h>
 
 
@@ -127,8 +128,9 @@ namespace easy3d {
         const mat4& MVP = camera_->modelViewProjectionMatrix();
         // camera position is defined in world coordinate system.
         const vec3& wCamPos = camera_->position();
+
         program->bind();
-        program->set_uniform("MVP", MVP);                                  easy3d_debug_log_gl_error;
+        program->set_uniform("MVP", MVP);
         program->set_uniform("SHADOW", shadow_matrix_);                    easy3d_debug_log_gl_error;
         program->set_uniform("wLightPos", light_pos_);                     easy3d_debug_log_gl_error;
         program->set_uniform("wCamPos", wCamPos);                          easy3d_debug_log_gl_error;
@@ -148,13 +150,20 @@ namespace easy3d {
 
         for (auto d : surfaces) {
             if (d->is_visible()) {
+                // transformation introduced by manipulation
+                const mat4 MANIP = d->manipulated_matrix();
+                // needs be padded when using uniform blocks
+                const mat3 NORMAL = transform::normal_matrix(MANIP);
+                program->set_uniform("MANIP", MANIP)
+                        ->set_uniform( "NORMAL", NORMAL);
                 program->set_uniform("smooth_shading", d->smooth_shading());
                 program->set_block_uniform("Material", "ambient", d->material().ambient);
                 program->set_block_uniform("Material", "specular", d->material().specular);
                 program->set_block_uniform("Material", "shininess", &d->material().shininess);
                 program->set_uniform("default_color", d->color());				easy3d_debug_log_gl_error;
                 program->set_uniform("per_vertex_color", d->coloring_method() != State::UNIFORM_COLOR && d->color_buffer());		easy3d_debug_log_gl_error;
-                program->set_uniform("is_background", false);
+                program->set_uniform("is_background", false)
+                        ->set_uniform("selected", d->is_selected());
                 if (setting::clipping_plane)
                     setting::clipping_plane->set_program(program, d->plane_clip_discard_primitive());
                 d->gl_draw(false);
