+++ easy3d/renderer/shadow.cpp
@@ -220,8 +220,9 @@ namespace easy3d {
         const mat4& MVP = camera_->modelViewProjectionMatrix();
         // camera position is defined in world coordinate system.
         const vec3& wCamPos = camera_->position();
+
         program->bind();
-        program->set_uniform("MVP", MVP);                                  easy3d_debug_log_gl_error;
+        program->set_uniform("MVP", MVP);
         program->set_uniform("SHADOW", shadow_matrix_);                    easy3d_debug_log_gl_error;
         program->set_uniform("wLightPos", light_pos_);                     easy3d_debug_log_gl_error;
         program->set_uniform("wCamPos", wCamPos);                          easy3d_debug_log_gl_error;
@@ -229,13 +230,20 @@ namespace easy3d {
         program->bind_texture("shadowMap", fbo_->depth_texture(), 0);      easy3d_debug_log_gl_error;
         for (auto d : surfaces) {
             if (d->is_visible()) {
+                // transformation introduced by manipulation
+                const mat4 MANIP = d->manipulated_matrix();
+                // needs be padded when using uniform blocks
+                const mat3 NORMAL = transform::normal_matrix(MANIP);
+                program->set_uniform("MANIP", MANIP)
+                        ->set_uniform( "NORMAL", NORMAL);
                 program->set_uniform("smooth_shading", d->smooth_shading());
                 program->set_block_uniform("Material", "ambient", d->material().ambient);
                 program->set_block_uniform("Material", "specular", d->material().specular);
                 program->set_block_uniform("Material", "shininess", &d->material().shininess);
                 program->set_uniform("default_color", d->color());				easy3d_debug_log_gl_error;
                 program->set_uniform("per_vertex_color", d->coloring_method() != State::UNIFORM_COLOR && d->color_buffer());
-                program->set_uniform("is_background", false);
+                program->set_uniform("is_background", false)
+                        ->set_uniform("selected", d->is_selected());
                 if (setting::clipping_plane)
                     setting::clipping_plane->set_program(program, d->plane_clip_discard_primitive());
                 d->gl_draw(false);
