### 2021-01-10 957e8d1
instantiate StackTracer in logger initialization
