### 2021-01-10 127160a
improve logging
