+++ easy3d/util/logging.cpp
@@ -26,6 +26,7 @@
 #include <thread>
 #include <iomanip>
 #include <cstring>
+#include <mutex>
 
 
 #if defined(WIN32) || defined(_WIN32) || defined(__WIN32__)
@@ -35,45 +36,34 @@
 #endif
 
 
+
 namespace google {
 
-    // This is the set of log sinks. This must be in a separate library to ensure
-    // that there is only one instance of this across the entire program.
-    static std::set<google::LogSink *> log_sinks_global;
-
-    static google::LogSeverity log_severity_threshold_global;
-
-    void setLogSeverityThreshold(LogSeverity min_severity) {
-        log_severity_threshold_global = min_severity;
-    }
-
-
-    // Note: the Log sink functions are not thread safe.
-    void AddLogSink(LogSink *sink) {
-        // TODO(settinger): Add locks for thread safety.
-        log_sinks_global.insert(sink);
-    }
-
-    void RemoveLogSink(LogSink *sink) {
-        log_sinks_global.erase(sink);
-    }
-
-
-    const std::string MessageLogger::severity_labels[4] = {
+    static std::string global_severity_labels[5] = {
             "INFO    ",
             "WARNING ",
             "ERROR   ",
-            "FATAL   "
+            "FATAL   ",
+            "QUIET   "
     };
 
 
     // see https://gist.github.com/vratiu/9780109
-    const char *MessageLogger::severity_color_code[4] = {
+    static const char *global_severity_color_code[5] = {
             "",     // default -> INFO
             "3",    // yellow  -> WARNING
             "5",    // purple  -> ERROR
-            "1"     // red     -> FATAL
+            "1",    // red     -> FATAL
+            ""      // default -> QUIET
     };
+
+
+    // This is the set of log sinks. This must be in a separate library to ensure
+    // that there is only one instance of this across the entire program.
+    static std::set<google::LogSink *> global_log_sinks;
+
+    static google::LogSeverity global_log_severity_threshold;
+
 
     MessageLogger::MessageLogger(const char *file, int line, int severity)
             : full_path_(file), line_(line), severity_(severity) {
@@ -85,7 +75,10 @@ namespace google {
 
     // Output the contents of the stream to the proper channel on destruction.
     MessageLogger::~MessageLogger() {
-        if (severity_ < log_severity_threshold_global)
+        static std::mutex mutex;
+        const std::lock_guard<std::mutex> lock(mutex);
+
+        if (severity_ < global_log_severity_threshold)
             return;
 
         short_name_ = simple_name(std::string(full_path_));
@@ -109,7 +102,7 @@ namespace google {
                        << std::this_thread::get_id();
         pid_tid_str_ = pid_tid_stream.str();
 
-        // remove all the ending line breaks
+        // remove all trailing ending line breaks
         message_ = stream_.str();
         while (message_.back() == '\n' && !message_.empty()) {
             message_.erase(message_.end() - 1);
@@ -119,31 +112,22 @@ namespace google {
             // If fatal error occurred, stop the program.
             // After logging the error (before the program crashes), notify StackTracer not to log again.
             if (!backward::SignalHandling::failure_has_been_recored) {
-                std::stringstream stream;
-                stream << "\n================================================================================="
-                       << "\nEasy3D has encountered a fatal error and has to abort. ";
-
-                const std::string log_file = easy3d::logging::FileLogClient::log_file_name();
-                if (!log_file.empty()) { // a log file exists
-                    stream << "The error has been recorded \n"
-                           << "in the log file [" + easy3d::file_system::absolute_path(log_file) + "].";
-                }
-
-                stream << "\nPlease report this issue with the complete log, a description of how to reproduce";
-                stream << "\nthe issue, and possibly your data to Liangliang Nan (liangliang.nan@gmail.com).";
[…]
