### 2021-01-10 b7669ce
allow to print counter
