+++ easy3d/util/logging.h
@@ -105,8 +105,48 @@ namespace google {
         virtual void WaitTillSent() { /* client code should implement this */ }
     };
 
+}
 
     // ---------------------------- Logger Class --------------------------------
+
+    namespace easy3d {
+        namespace logging {
+            // We want the special COUNTER value available for LOG_EVERY_X()'ed messages
+            enum PRIVATE_Counter { COUNTER };
+        }
+    }
+
+namespace google {
+
+    class LogStream : public std::ostringstream {
+    public:
+        LogStream(int ctr) : ctr_(ctr) {}
+
+        int ctr() const { return ctr_; }
+
+    private:
+        LogStream(const LogStream &);
+
+        LogStream &operator=(const LogStream &);
+
+        int ctr_;  // Counter hack (for the LOG_EVERY_X() macro)
+    };
+
+}
+
+// Define global operator<< to declare using ::operator<<.
+// This allows to output the COUNTER value. This is only valid if ostream is a LogStream.
+// You must not use COUNTER with non-glog ostream
+inline std::ostream& operator<<(std::ostream &os, const easy3d::logging::PRIVATE_Counter&) {
+    auto log = dynamic_cast<google::LogStream*>(&os);
+    if (log)
+        os << log->ctr();
+    return os;
+}
+
+
+namespace google {
+
 
     // Class created for each use of the logging macros.
     // The logger acts as a stream and routes the final stream contents to the
@@ -117,14 +157,13 @@ namespace google {
     class MessageLogger {
     public:
         MessageLogger(const char *file, int line, google::LogSeverity severity);
-
         MessageLogger(const char *file, int line, google::LogSeverity severity, int ctr);
 
         // Output the contents of the stream to the proper channel on destruction.
         ~MessageLogger();
 
         // Return the stream associated with the logger object.
-        std::stringstream &stream() { return stream_; }
+        LogStream &stream() { return stream_; }
 
     private:
 
@@ -148,7 +187,7 @@ namespace google {
         std::string time_str_;
         std::string pid_tid_str_;
         std::string message_;
-        std::stringstream stream_;
+        LogStream stream_;
         int severity_;
     };
 
@@ -210,13 +249,14 @@ using google::LoggerVoidify;
 #endif
 
 
-#define SOME_KIND_OF_LOG_FIRST_N(severity, n) \
-  static int LOG_OCCURRENCES = 0; \
-  if (LOG_OCCURRENCES <= n) \
-    ++LOG_OCCURRENCES; \
-  if (LOG_OCCURRENCES <= n) \
-    MessageLogger( \
-        __FILE__, __LINE__, severity, LOG_OCCURRENCES).stream()
+
+// Use macro expansion to create, for each use of LOG_EVERY_N(), static
+// variables with the __LINE__ expansion as part of the variable name.
+#define LOG_EVERY_N_VARNAME(base, line) LOG_EVERY_N_VARNAME_CONCAT(base, line)
+#define LOG_EVERY_N_VARNAME_CONCAT(base, line) base ## line
+
+#define LOG_OCCURRENCES LOG_EVERY_N_VARNAME(occurrences_, __LINE__)
+#define LOG_OCCURRENCES_MOD_N LOG_EVERY_N_VARNAME(occurrences_mod_n_, __LINE__)
 
 #define SOME_KIND_OF_LOG_EVERY_N(severity, n) \
   static int LOG_OCCURRENCES = 0, LOG_OCCURRENCES_MOD_N = 0; \
@@ -233,6 +273,18 @@ using google::LoggerVoidify;
       ((LOG_OCCURRENCES_MOD_N=(LOG_OCCURRENCES_MOD_N + 1) % n) == (1 % n))) \
     MessageLogger( \
[…]
