### 2021-01-10 b36569b
pick only for left button and not in camera path creation mode
