### 2021-01-10 66b5ff0
allow to log line file name, line number, and address
