### 2021-01-10 741e628
change visibility also changes the selection
