### 2021-01-10 2eb4c8d
visualizing slice of a mesh
