### 2021-01-10 c7ad2de
add hint for normals
