### 2021-01-10 a116b1a
correct color property location
