### 2021-01-10 d5380c3
make new model current by default
