+++ applications/Mapple/paint_canvas.cpp
@@ -831,7 +831,7 @@ void PaintCanvas::addModel(Model *model) {
     model->set_manipulator(manipulator);
 
     models_.push_back(model);
-//    model_idx_ = static_cast<int>(models_.size()) - 1; // make the last one current
+    model_idx_ = static_cast<int>(models_.size()) - 1; // make the last one current
 }
 
 
