### 2021-01-10 107b6e2
allow select model from the model widget
