### 2021-01-10 cf59a0b
set reconstruction depth to 5 (faster)
