### 2021-01-11 0a53b29
Signal now allows fuctions with arguments
