+++ easy3d/core/signal.h
@@ -33,10 +33,10 @@
 namespace easy3d {
 
     /**
-     * \brief Implementation of a simple signal-slot mechanism. \par
-     * \details Multiple slots (classes and their member functions) can be connected to a signal object.
+     * \brief Implementation of a simple signal-slot mechanism.
+     * \details Multiple slots (functions, lambda functions, and member functions) can be connected to a signal object.
      *        You can connect functions to the signal which will be called when the trigger() method on the signal
-     *        object is invoked. Any argument passed to emit() will be passed to the given functions. A typical usage
+     *        object is invoked. Any argument passed to trigger() will be passed to the given functions. A typical usage
      *        of Signal in Easy3D is camera manipulation. When the camera has been manipulated, the viewer should be
      *        notified (e.g., a repaint event should be triggered). This is done by calling to the viewer's update()
      *        function. So in Easy3D, the viewer's update function is connected to the camera.
@@ -44,89 +44,152 @@ namespace easy3d {
      * \class Signal easy3d/core/signal.h
      * \todo Extend this class to accept multiple functions of a class.
      */
+
+
+// A signal object may call multiple slots with the
+// same signature. You can connect functions to the signal
+// which will be called when the emit() method on the
+// signal object is invoked. Any argument passed to emit()
+// will be passed to the given functions.
+
+    template <typename... Args>
     class Signal {
-	public:
-        /**
-         * \brief Connect an owned slot to this signal.
-         *  - If no overloaded function:  \par
-         *      \code
-         *      camera()->connect(this, &Viewer::update);
-         *      \endcode
-         *  - If the function has been overloaded (of course you can always treat it this way):   \par
-         *      \code
-         *      camera()->connect(this, static_cast<void(Viewer::*)(void) const>(&Viewer::update));
-         *      \endcode
-         */
-        template < class Class, class Function, class... Args >
-        void connect(Class&& owner, Function&& func, Args&&... args)  {
-            owned_slots_[owner] = std::bind(func, owner, std::forward<Args>(args)...);
-        }
 
-        /**
-         * \brief Disconnect all the slots of a owner from this signal.
-         */
-        template < class Class >
-        void disconnect(Class&& owner)  {
-            owned_slots_.erase(owner);
-        }
+    public:
+        Signal()  = default;
+        ~Signal() = default;
 
-        /**
-         * \brief Connect a free (i.e., not owned by any class) slot to this signal.
-         * A unique index of the function must be provided (to differentiate with other free functions).
-         */
-        template < class Function, class... Args >
-        void connect(int idx, Function&& func, Args&&... args) {
-            free_slots_[idx] = std::bind(func, std::forward<Args>(args)...);
-        }
+        // Copy constructor and assignment create a new signal.
+        Signal(Signal const& /*unused*/) {}
 
-        /**
-         * \brief Disconnect a free slot function from this signal.
-         * The slot is identified by its unique index.
-         */
-        void disconnect(int idx) { free_slots_.erase(idx); }
-
-        /**
-         * \brief Trigger all the execution of all connected slot functions.
-         * \todo A better function name can be 'emit', but occupied by other software like Qt.
-         *       Don't know how to still use 'emit'.
-         */
-        template < class ... Args >
-        void send(Args&&... args) {
-            for(auto& it : owned_slots_) {
-                it.second(std::forward<Args>(args)...);
+        Signal& operator=(Signal const& other) {
+            if (this != &other) {
+                disconnect_all();
             }
-            for(auto& it : free_slots_) {
-                it.second(std::forward<Args>(args)...);
+            return *this;
+        }
+
+        // Move constructor and assignment operator work as expected.
[…]
