+++ sandbox/Experiment/main.cpp
@@ -1,61 +1,52 @@
-#if 0
-
-#define ELPP_FEATURE_ALL
-#define ELPP_FEATURE_CRASH_LOG
-//#define ELPP_EXPERIMENTAL_ASYNC
-
-#include <easy3d/util/easylogging++.h>
-
-
-INITIALIZE_EASYLOGGINGPP
-
-void myCrashHandler(int sig) {
-    LOG(ERROR) << "Woops! Crashed!";
-    // FOLLOWING LINE IS OPTIONAL
-    el::Helpers::logCrashReason(sig, true);
-    // FOLLOWING LINE IS ABSOLUTELY NEEDED AT THE END IN ORDER TO ABORT APPLICATION
-    el::Helpers::crashAbort(sig);
-}
-int main(void) {
-    el::Helpers::setCrashHandler(myCrashHandler);
-
-    LOG(INFO) << "My crash handler!";
-
-    std::cout << std::endl << "    ======= Backtrace: =========" << std::endl << el::base::debug::StackTrace();
-
-    int* i;
-    *i = 0; // Crash!
-
-    return 0;
-}
-
-#else
-
-#include <easy3d/util/stack_tracer.h>
-#include <easy3d/util/logging.h>
-
+#include <easy3d/core/signal.h>
+#include <string>
 #include <iostream>
+
 
 using namespace easy3d;
 
 
+class Car {
+public:
+    Car() {
+        std::function<void(void)> stop1 = [this]() -> void { std::cout << "stopped\n" ; };
+        auto stop2 = [this]() -> void { std::cout << "stopped\n" ; };
 
-int main(void) {
+        stop_signal.connect(stop1);
+        stop_signal.connect(stop2);
+        easy3d::connect(&stop_signal, stop1);
+        easy3d::connect(&stop_signal, stop2);
+    }
 
-    logging::initialize(INFO);
-
-    LOG(INFO) << "crash test";
-    LOG(ERROR) << "error!";
-
-    StackTracer st;
-
-    int* i;
-    *i = 0; // Crash!
+    void stop() { stop_signal.send(); }
 
 
-    LOG(FATAL) << "bad";
+private:
+    Signal<>  stop_signal;
+};
+
+
+
+int main() {
+    Car car;
+    car.stop();
+    return 0;
+
+
+
+
+
+
+
+    // create new signal
+    Signal<std::string, int> signal;
+
+    // attach a slot
+    signal.connect([](std::string arg1, int arg2) {
+        std::cout << arg1 << " " << arg2 << std::endl;
+    });
+
+    signal.send("The answer:", 42);
 
     return 0;
 }
-
-#endif
