+++ sandbox/Experiment/main.cpp
@@ -1,5 +1,4 @@
 #include <easy3d/core/signal.h>
-#include <string>
 #include <iostream>
 
 
@@ -8,45 +7,145 @@ using namespace easy3d;
 
 class Car {
 public:
-    Car() {
-        std::function<void(void)> stop1 = [this]() -> void { std::cout << "stopped\n" ; };
-        auto stop2 = [this]() -> void { std::cout << "stopped\n" ; };
+    Car(int speed) : speed_(speed) {}
+public:
 
-        stop_signal.connect(stop1);
-        stop_signal.connect(stop2);
-        easy3d::connect(&stop_signal, stop1);
-        easy3d::connect(&stop_signal, stop2);
+    int speed() const { return speed_; }
+
+    void start() {
+        std::cout << "Car::" << __func__ << ". Class member: started\n";
     }
 
-    void stop() { stop_signal.send(); }
+    void report_speed(int max_allow_speed) const {
+        std::cout << "Car::" << __func__ << ". max allowed is " << max_allow_speed << ". I am at " << speed_ << "\n";
+    }
 
+    void stop(int hours, const std::string& msg) const {
+        std::cout << "Car::" << __func__ << ". " << msg << ". After driving for " << hours << " hours\n";
+    }
 
 private:
-    Signal<>  stop_signal;
+    int speed_;
 };
 
 
+void func_start() {
+    std::cout << __func__ << "\n";
+}
+
+void func_start_1arg(Car* car) {
+    std::cout << __func__ << ". speed is " << car->speed() << "\n";
+}
+
+void func_report_speed(int max_allow_speed, const Car* car) {
+    std::cout << __func__ << ". max allowed is " << max_allow_speed << ". I am at " << car->speed() << "\n";
+}
+
+void func_stop(const Car* car, int hours, const std::string& msg) {
+    std::cout << __func__ << ". " << msg << ". After driving for " << hours << " hours, I have to stopped. My speed was " << car->speed() << "\n";
+}
+
 
 int main() {
-    Car car;
-    car.stop();
-    return 0;
+
+    Car car(100);
+
+    // connect to a class member ---------------------------------------------------------------------------------------
+
+    Signal<>  start_signal;
+    Signal<int>  report_signal;
+    Signal<int, const std::string&>  stop_signal;
+
+    // ---- connect to a class member, no argument
+    easy3d::connect(&start_signal, &car, &Car::start);  // use the global function connect()
+    start_signal.connect(&car, &Car::start);  // use signal's connect()
+
+    // ---- connect to a const class member, one argument
+    easy3d::connect(&report_signal, &car, &Car::report_speed);  // use the global function connect()
+    report_signal.connect(&car, &Car::report_speed);  // use signal's connect()
+
+    // ---- connect to a const class member, two arguments
+    easy3d::connect(&stop_signal, &car, &Car::stop);  // use the global function connect()
+    stop_signal.connect(&car, &Car::stop);  // use signal's connect()
+
+    start_signal.send();
+    report_signal.send(80);
+    stop_signal.send(6, "I have to stop");
+
+    // connect to a function -------------------------------------------------------------------------------------------
+
+    Signal<> func_start_signal;
+    Signal<Car*> func_start_signal_1arg;
+    Signal<int, const Car*> func_report_signal;
+    Signal<const Car*, int, const std::string&> another_stop_signal;
+
+    // ---- connect to a function, no argument
+    easy3d::connect(&func_start_signal, func_start);  // use the global function connect()
+    func_start_signal.connect(func_start);  // use signal's connect()
+
+    // ---- connect to a function, one argument
+
+    easy3d::connect(&func_start_signal_1arg, func_start_1arg);  // use the global function connect()
[…]
