### 2021-01-11 447943f
add tests for Signal
