### 2021-01-11 a8534b8
remove redudant (non-const) functions
