### 2021-01-11 203701f
Timer is not templated
