+++ applications/Mapple/paint_canvas.cpp
@@ -285,8 +285,7 @@ void PaintCanvas::mousePressEvent(QMouseEvent *e) {
                     camera_->setPivotPoint(p);
                     // show, but hide the visual hint of pivot point after \p delay milliseconds.
                     show_pivot_point_ = true;
-                    const int delay = 10000;
-                    Timer::single_shot(delay, [&]() {
+                    Timer<>::single_shot(10000, [&]() {
                         show_pivot_point_ = false;
                         update();
                     });
