+++ easy3d/viewer/viewer.cpp
@@ -535,8 +535,7 @@ namespace easy3d {
                     camera_->setPivotPoint(p);
                     // show, but hide the visual hint of pivot point after \p delay milliseconds.
                     show_pivot_point_ = true;
-                    const int delay = 10000;
-                    Timer::single_shot(delay, [&]() {
+                    Timer<>::single_shot(10000, [&]() {
                         show_pivot_point_ = false;
                         update();
                     });
