### 2021-01-11 cdbe6f1
add tests for Timer
