+++ sandbox/Experiment/CMakeLists.txt
@@ -4,8 +4,8 @@ project(${PROJECT_NAME})
 add_executable(${PROJECT_NAME}
 #        ThreadPool/example.cpp
 #        test_logging.cpp
-        test_signal.cpp
-#        test_timer.cpp
+#        test_signal.cpp
+        test_timer.cpp
         )
 
 set_target_properties(${PROJECT_NAME} PROPERTIES FOLDER "SandBox")
