+++ sandbox/SomeTest/some_test.cpp
@@ -1,244 +1,0 @@
-/**
- * Copyright (C) 2015 by Liangliang Nan (liangliang.nan@gmail.com)
- * https://3d.bk.tudelft.nl/liangliang/
- *
- * This file is part of Easy3D. If it is useful in your research/work,
- * I would be grateful if you show your appreciation by citing it:
- * ------------------------------------------------------------------
- *      Liangliang Nan.
- *      Easy3D: a lightweight, easy-to-use, and efficient C++
- *      library for processing and rendering 3D data. 2018.
- * ------------------------------------------------------------------
- * Easy3D is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License Version 3
- * as published by the Free Software Foundation.
- *
- * Easy3D is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include "some_test.h"
-#include <easy3d/renderer/drawable.h>
-#include <easy3d/renderer/drawable_lines.h>
-#include <easy3d/renderer/drawable_triangles.h>
-#include <easy3d/renderer/renderer.h>
-#include <easy3d/core/surface_mesh.h>
-#include <easy3d/renderer/setting.h>
-#include <easy3d/core/random.h>
-#include <easy3d/fileio/ply_reader_writer.h>
-#include <easy3d/fileio/surface_mesh_io.h>
-#include <easy3d/util/dialogs.h>
-
-#include <3rd_party/glfw/include/GLFW/glfw3.h>    // for the KEYs
-
-
-namespace easy3d {
-
-
-    SomeTest::SomeTest(const std::string &title)
-            : Viewer(title) {
-    }
-
-
-    std::string SomeTest::usage() const {
-        return Viewer::usage() +
-               (" ----------------------- SomeTest Usage --------------------------- \n"
-                "\tCtrl+O: Open file\n"
-                "\tG: Ground truth\n"
-                "\tE: Edges\n"
-                "\tC: All candidate faces\n"
-                " ------------------------------------------------------------------- \n");
-    }
-
-
-    bool SomeTest::open() {
-        const std::string title("Please choose a file");
-        const std::string default_path("");
-        const std::vector<std::string> &filters = {
-                "Mesh Files (*.obj *.ply *.off *.stl)", "*.obj *.ply *.off *.stl",
-                "All Files (*.*)", "*"
-        };
-
-        const std::string &file_name = dialog::open(title, default_path, filters);
-        auto mesh = SurfaceMeshIO::load(file_name);
-        if (mesh) {
-            clear_scene();
-            add_model(mesh, false);
-            create_drawables(mesh);
-            fit_screen();
-            return true;
-        }
-        return false;
-    }
-
-
-    bool SomeTest::key_press_event(int key, int modifiers) {
-        if (key == GLFW_KEY_G) {
-            std::cout << "Now I am showing ground truth" << std::endl;
-            for (auto d : edges_ground_truth_)
-                d->set_visible(false);
-            for (auto d : candidate_faces_)
-                d->set_visible(false);
-            for (auto d : faces_ground_truth_)
-                d->set_visible(true);
-            update();
-            return true;
-        } else if (key == GLFW_KEY_E) {
-            std::cout << "Now I am showing edges" << std::endl;
-            for (auto d : faces_ground_truth_)
-                d->set_visible(false);
-            for (auto d : candidate_faces_)
-                d->set_visible(false);
[…]
