+++ easy3d/util/timer.h
@@ -59,6 +59,8 @@ namespace easy3d {
          * \param delay The time to be delayed, in milliseconds.
          * \param inst The pointer to \c Class instance, e.g., '&a' for 'Class a' or 'this' within Class.
          * \param func The pointer to the member function of \c inst, e.g., '&Class::foo'.
+         * \note When a member function has overloads or inheritance, use static_cast of the function to indicate
+         *      the template argument. For example, \code static_cast<void (Class::*)(void)>(&Class::func). \endcode
          */
         template < class Class >
         static void single_shot(int delay, Class* inst, void (Class::*func)(Args...), Args... args);
@@ -68,6 +70,8 @@ namespace easy3d {
          * \param delay The time to be delayed, in milliseconds.
          * \param inst The pointer to \c Class instance, e.g., '&a' for 'Class a' or 'this' within Class.
          * \param func The pointer to the member function of \c inst, e.g., '&Class::foo'.
+         * \note When a member function has overloads or inheritance, use static_cast of the function to indicate
+         *      the template argument. For example, \code static_cast<void (Class::*)(void)>(&Class::func). \endcode
          */
         template < class Class >
         static void single_shot(int delay, Class* inst, void (Class::*func)(Args...) const, Args... args);
@@ -86,6 +90,8 @@ namespace easy3d {
          * \param delay The time to be delayed, in milliseconds.
          * \param inst The pointer to \c Class instance, e.g., '&a' for 'Class a' or 'this' within Class.
          * \param func The pointer to the member function of \c inst, e.g., '&Class::foo'.    
+         * \note When a member function has overloads or inheritance, use static_cast of the function to indicate
+         *      the template argument. For example, \code static_cast<void (Class::*)(void)>(&Class::func). \endcode
          */
         template < class Class >
         void set_timeout(int delay, Class* inst, void (Class::*func)(Args...), Args... args) const;
@@ -96,6 +102,8 @@ namespace easy3d {
          * \param delay The time to be delayed, in milliseconds.
          * \param inst The pointer to \c Class instance, e.g., '&a' for 'Class a' or 'this' within Class.
          * \param func The pointer to the member function of \c inst, e.g., '&Class::foo'.    
+         * \note When a member function has overloads or inheritance, use static_cast of the function to indicate
+         *      the template argument. For example, \code static_cast<void (Class::*)(void)>(&Class::func). \endcode
          */
         template < class Class >
         void set_timeout(int delay, Class* inst, void (Class::*func)(Args...) const, Args... args) const;
@@ -112,6 +120,8 @@ namespace easy3d {
          * \param interval The interval, in milliseconds.
          * \param inst The pointer to \c Class instance, e.g., '&a' for 'Class a' or 'this' within Class.
          * \param func The pointer to the member function of \c inst, e.g., '&Class::foo'.
+         * \note When a member function has overloads or inheritance, use static_cast of the function to indicate
+         *      the template argument. For example, \code static_cast<void (Class::*)(void)>(&Class::func). \endcode
          */
         template < class Class >
         void set_interval(int interval, Class* inst, void (Class::*func)(Args...), Args... args);
@@ -121,6 +131,8 @@ namespace easy3d {
          * \param interval The interval, in milliseconds.
          * \param inst The pointer to \c Class instance, e.g., '&a' for 'Class a' or 'this' within Class.
          * \param func The pointer to the member function of \c inst, e.g., '&Class::foo'.
+         * \note When a member function has overloads or inheritance, use static_cast of the function to indicate
+         *      the template argument. For example, \code static_cast<void (Class::*)(void)>(&Class::func). \endcode
          */
         template < class Class >
         void set_interval(int interval, Class* inst, void (Class::*func)(Args...) const, Args... args);
