### 2021-01-11 1163066
allow to connect one signal to another signal
