### 2021-01-11 3c408c7
remove unnecessary calls to update()
