+++ applications/Mapple/dialogs/dialog_walk_through.cpp
@@ -27,7 +27,7 @@
 #include "main_window.h"
 
 #include <easy3d/core/model.h>
-#include <easy3d/renderer/walk_throuth.h>
+#include <easy3d/renderer/walk_through.h>
 #include <easy3d/renderer/key_frame_interpolator.h>
 #include <easy3d/renderer/manipulated_camera_frame.h>
 #include <easy3d/renderer/camera.h>
