### 2021-01-12 1c831b0
send modified signal after rotation
