+++ test/test_signal.cpp
@@ -30,22 +30,25 @@
 using namespace easy3d;
 
 
-// a simple class
+// A trivial class (that does not make sense, but just to show to use the Signal class).
 
 class Car {
 public:
     Car(int speed) : speed_(speed) {}
 
 public:
-
     int speed() const { return speed_; }
 
     void start() {
         std::cout << "started\n";
     }
 
+    void start(const std::string& msg) {
+        std::cout << "car started: " << msg << "\n";
+    }
+
     void report_speed(int max_allow_speed) const {
-        std::cout << "max allowed is " << max_allow_speed << ". I am at " << speed_ << "\n";
+        std::cout << "max allowed speed is " << max_allow_speed << ". I am at " << speed_ << "\n";
     }
 
     void stop(int hours, const std::string &msg) const {
@@ -59,20 +62,39 @@ private:
 
 void test_for_members(Car *car) {
     Signal<> start_signal;
+    Signal<const std::string&> start_signal_1arg;
     Signal<int> report_signal;
     Signal<int, const std::string &> stop_signal;
 
     // ---- connect to a class member, no argument
-    easy3d::connect(&start_signal, car, &Car::start);  // use the global function connect()
-    start_signal.connect(car, &Car::start);  // use signal's connect()
+
+    start_signal.connect(car, &Car::start);  // this works
+    start_signal.connect(car, static_cast<void (Car::*)(void)> (&Car::start));  // this works
+    start_signal.connect(car, overload<Car>(&Car::start));  // this works
+
+//    easy3d::connect(&start_signal, car, &Car::start);  // this won't work
+    easy3d::connect(&start_signal, car, static_cast<void (Car::*)(void)> (&Car::start));  // this works
+    easy3d::connect(&start_signal, car, overload<Car>(&Car::start));  // this works
+
+    start_signal_1arg.connect(car, &Car::start);  // this works
+    start_signal_1arg.connect(car, static_cast<void (Car::*)(const std::string&)> (&Car::start));  // this works
+    start_signal_1arg.connect(car, overload<Car, const std::string&>(&Car::start));  // this works
+
+//    easy3d::connect(&start_signal_1arg, car, &Car::start);  // this won't work
+    easy3d::connect(&start_signal_1arg, car, static_cast<void (Car::*)(const std::string&)> (&Car::start));  // this works
+    easy3d::connect(&start_signal_1arg, car, overload<Car, const std::string&>(&Car::start));  // this works
 
     // ---- connect to a const class member, one argument
+
     easy3d::connect(&report_signal, car, &Car::report_speed);  // use the global function connect()
     report_signal.connect(car, &Car::report_speed);  // use signal's connect()
 
     // ---- connect to a const class member, two arguments
+
     easy3d::connect(&stop_signal, car, &Car::stop);  // use the global function connect()
     stop_signal.connect(car, &Car::stop);  // use signal's connect()
+
+    // ---- emit all the signals
 
     start_signal.send();
     report_signal.send(80);
@@ -84,8 +106,8 @@ void func_start() {
     std::cout << "started\n";
 }
 
-void func_start(Car *car) {
-    std::cout << "speed is " << car->speed() << "\n";
+void func_start(const std::string& msg) {
+    std::cout << "car started: " << msg << "\n";
 }
 
 void func_report_speed(int max_allow_speed, const Car *car) {
@@ -99,30 +121,44 @@ void func_stop(const Car *car, int hours, const std::string &msg) {
 
 void test_for_functions(Car *car) {
     Signal<> func_start_signal;
-    Signal<Car *> func_start_signal_1arg;
+    Signal<const std::string &> func_start_signal_1arg;
     Signal<int, const Car *> func_report_signal;
     Signal<const Car *, int, const std::string &> another_stop_signal;
 
[…]
