+++ easy3d/core/signal.h
@@ -83,17 +83,25 @@ namespace easy3d {
         }
 
 
-        /// Connects a function to this signal.
-        /// The returned value can be used to disconnect the function again.
+        /**
+         * Connects a function to this signal.
+         * The returned value can be used to disconnect the function from this signal.
+         * \note When a function has overloads, explicitly cast the function to the right function type using
+         *     e.g., <tt> static_cast<void (*)(const std::string&, int)>(&print). </tt> Or use the helper function
+         *     \fn overload for a lighter syntax.
+         */
         int connect(std::function<void(Args...)> const &slot) const {
             slots_.insert(std::make_pair(++current_id_, slot));
             return current_id_;
         }
 
-        /// Connects a member function of an object to this Signal.
-        /// The returned value can be used to disconnect the function again.
-        /// \note When a member function has overloads or inheritance, you may need to explicit cast the function to indicate
-        ///         the template argument. For example, \code static_cast<void (Car::*)(void)>(&Vehicle::start). \endcode
+        /**
+         * Connects a member function of an object to this signal.
+         * The returned value can be used to disconnect the function from this signal.
+         * \note When a member function has overloads, explicitly cast the function to the right function type using
+         *     e.g., <tt> static_cast<void (Viewer::*)(const std::string&, int)>(Viewer::print). </tt>
+         *     Or use the helper function \fn overload for a lighter syntax.
+         */
         template<typename Class>
         int connect(Class *inst, void (Class::*func)(Args...)) {
             return connect([=](Args... args) {
@@ -101,10 +109,13 @@ namespace easy3d {
             });
         }
 
-        /// Connects a const member function of an object to this Signal.
-        /// The returned value can be used to disconnect the function again.
-        /// \note When a member function has overloads or inheritance, you may need to explicit cast the function to indicate
-        ///         the template argument. For example, \code static_cast<void (Car::*)(void)>(&Vehicle::start). \endcode
+        /**
+         * Connects a const member function of an object to this Signal.
+         * The returned value can be used to disconnect the function from this signal.
+         * \note When a member function has overloads, explicitly cast the function to the right function type using
+         *     e.g., <tt> static_cast<void (Viewer::*)(const std::string&, int)>(Viewer::print). </tt>
+         *     Or use the helper function \fn overload for a lighter syntax.
+         */
         template<typename Class>
         int connect(Class *inst, void (Class::*func)(Args...) const) {
             return connect([=](Args... args) {
@@ -112,9 +123,11 @@ namespace easy3d {
             });
         }
 
-        /// Connects this signal to another signal \p receiver.
-        /// Upon return, the emission of this signal will trigger \p receiver to emit.
-        /// The returned value can be used to disconnect the connected signal.
+        /**
+         * Connects this signal to another signal \p receiver.
+         * Upon return, the emission of this signal will trigger \p receiver to emit.
+         * The returned value can be used to disconnect the connected signal.
+         */
         int connect(Signal<Args...> *receiver) {
             return connect(receiver, &Signal<Args...>::send);
         }
@@ -162,25 +175,35 @@ namespace easy3d {
     /// \name  Global methods for connection and disconnection.
     //\{
 
-    /// Connects a function to the signal.
-    /// The returned value can be used to disconnect the function again.
+    /**
+     * Connects a function to the signal.
+     * The returned value can be used to disconnect the function from this signal.
+     * \note When a function has overloads, explicitly cast the function to the right function type using
[…]

[…]
