+++ applications/Mapple/paint_canvas_snapshot.cpp
@@ -361,8 +361,8 @@ void PaintCanvas::recordAnimation(const QString &file_name, int fps, int bit_rat
 //    easy3d::connect(&camera_->frame_modified, this, overload<PaintCanvas>(&PaintCanvas::update));  // this also works
 //    camera_->frame_modified.connect(this, static_cast<void (PaintCanvas::*)(void)> (&PaintCanvas::update));  // this also works
 //    camera_->frame_modified.connect(this, overload<PaintCanvas>(&PaintCanvas::update));  // this also works
-////    camera_->frame_modified.connect(this, &PaintCanvas::update);  // this won't work
-////    easy3d::connect(&camera_->frame_modified, this, &PaintCanvas::update);  // this won't work
+//    camera_->frame_modified.connect(this, &PaintCanvas::update);  // this won't work due to overloading
+//    easy3d::connect(&camera_->frame_modified, this, &PaintCanvas::update);  // this won't work due to overloading
 
     // restore the viewer size
     if (w != original_width || h != original_height) {
