### 2021-01-12 c062431
undefine ERROR (also defined in windows.h)
