+++ easy3d/fileio/surface_mesh_io_ply.cpp
@@ -348,7 +348,7 @@ namespace easy3d {
 			//-----------------------------------------------------
 
             binary = binary && (file_name.find("ascii") == std::string::npos);
-            LOG_IF(WARNING, !binary) << "you're writing an ASCII ply file. Use binary format for better performance";
+            LOG_IF(!binary, WARNING) << "you're writing an ASCII ply file. Use binary format for better performance";
 
             PlyWriter writer;
             return writer.write(file_name, elements, "", binary);
