+++ tutorials/Tutorial_503_SoftShadow/main.cpp
@@ -39,7 +39,7 @@ using namespace easy3d;
 
 int main(int argc, char **argv) {
     // Initialize logging.
-    logging::initialize(WARNING);
+    logging::initialize();
 
     const std::string file = resource::directory() + "/data/room.obj";
 
