+++ easy3d/util/logging.h
@@ -1,7 +1,33 @@
+/**
+ * Copyright (C) 2015 by Liangliang Nan (liangliang.nan@gmail.com)
+ * https://3d.bk.tudelft.nl/liangliang/
+ *
+ * This file is part of Easy3D. If it is useful in your research/work,
+ * I would be grateful if you show your appreciation by citing it:
+ * ------------------------------------------------------------------
+ *      Liangliang Nan.
+ *      Easy3D: a lightweight, easy-to-use, and efficient C++
+ *      library for processing and rendering 3D data. 2018.
+ * ------------------------------------------------------------------
+ * Easy3D is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License Version 3
+ * as published by the Free Software Foundation.
+ *
+ * Easy3D is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
 #ifndef EASY3D_UTIL_LOGGING_H
 #define EASY3D_UTIL_LOGGING_H
 
+
 #include <easy3d/util/easylogging++.h>
+
 
 namespace easy3d {
 
@@ -35,7 +61,12 @@ namespace easy3d {
 
         protected:
             void handle(const el::LogDispatchData* data) noexcept override {
-                send(data->logMessage()->level(), data->logMessage()->message());
+                // remove all trailing ending line breaks
+                std::string msg = data->logMessage()->message();
+                while (msg.back() == '\n' && !msg.empty()) {
+                    msg.erase(msg.end() - 1);
+                }
+                send(data->logMessage()->level(), msg);
             }
         };
 
