+++ easy3d/gui/picker_point_cloud.cpp
@@ -31,6 +31,7 @@
 #include <easy3d/renderer/framebuffer_object.h>
 #include <easy3d/renderer/opengl_error.h>
 #include <easy3d/renderer/drawable_points.h>
+#include <easy3d/renderer/manipulator.h>
 #include <easy3d/renderer/opengl_info.h>
 #include <easy3d/util/logging.h>
 
@@ -123,7 +124,7 @@ namespace easy3d {
 
         const auto &points = model->get_vertex_property<vec3>("v:point").vector();
         int num = static_cast<int>(points.size());
-        const mat4 &m = camera()->modelViewProjectionMatrix();
+        const mat4 &m = camera()->modelViewProjectionMatrix() * model->manipulator()->matrix();
 
         auto &select = model->vertex_property<bool>("v:select").vector();
 
@@ -164,7 +165,7 @@ namespace easy3d {
         glGetIntegerv(GL_VIEWPORT, viewport);
 
         vec4 rectangle(rect.x_min(), rect.x_max(), rect.y_min(), rect.y_max());
-        const mat4 &MVP = camera()->modelViewProjectionMatrix();
+        const mat4 &MVP = camera()->modelViewProjectionMatrix() * model->manipulator()->matrix();
 
         program->bind();
         program->set_uniform("viewport", viewport);
@@ -217,7 +218,7 @@ namespace easy3d {
 
         const auto &points = model->get_vertex_property<vec3>("v:point").vector();
         int num = static_cast<int>(points.size());
-        const mat4 &m = camera()->modelViewProjectionMatrix();
+        const mat4 &m = camera()->modelViewProjectionMatrix() * model->manipulator()->matrix();
 
         auto &select = model->vertex_property<bool>("v:select").vector();
 
@@ -272,7 +273,8 @@ namespace easy3d {
 
         program->bind();
         program->set_uniform("viewport", viewport);
-        program->set_uniform("MVP", camera()->modelViewProjectionMatrix());
+        program->set_uniform("MVP", camera()->modelViewProjectionMatrix())
+                ->set_uniform("MANIP", model->manipulator()->matrix());
         program->set_uniform("deselect", deselect);
         drawable->gl_draw(true);
         program->release();
