### 2021-01-13 d8342d2
compute vertex normals for rendering
