+++ applications/Mapple/paint_canvas.cpp
@@ -867,7 +867,7 @@ void PaintCanvas::fitScreen(const easy3d::Model *model) {
     }
 
     if (box.is_valid()) {
-        camera_->setSceneBoundingBox(box.min(), box.max());
+        camera_->setSceneBoundingBox(box.min_point(), box.max_point());
         camera_->showEntireScene();
         update();
     }
