+++ tutorials/Tutorial_204_CameraInterpolation/viewer.cpp
@@ -84,7 +84,7 @@ bool CameraIntrepolation::key_press_event(int key, int modifiers)
         Box3 box;
         for (auto m : models_)
             box.add_box(m->bounding_box());
-        camera_->setSceneBoundingBox(box.min(), box.max());
+        camera_->setSceneBoundingBox(box.min_point(), box.max_point());
         std::cout << "path deleted"<< std::endl;
         return true;
     }
