### 2021-01-13 8bcbce8
rename min() and max() in Box and Rect
