### 2021-01-13 a77f95b
add NOMINMAX (solve the min max conflicts on windows)
