### 2021-01-13 a5fb1a6
fix: fbo destroyed without an opengl context
