+++ applications/Mapple/widgets/widget_drawable_triangles.cpp
@@ -7,10 +7,10 @@
 #include <easy3d/renderer/drawable_lines.h>
 #include <easy3d/renderer/drawable_triangles.h>
 #include <easy3d/renderer/texture_manager.h>
-#include <easy3d/renderer/setting.h>
 #include <easy3d/renderer/buffers.h>
 #include <easy3d/renderer/renderer.h>
 #include <easy3d/core/surface_mesh.h>
+#include <easy3d/core/poly_mesh.h>
 #include <easy3d/util/file_system.h>
 #include <easy3d/util/logging.h>
 #include <easy3d/fileio/resources.h>
@@ -265,60 +265,87 @@ void WidgetTrianglesDrawable::updatePanel() {
 
 
 
+
+namespace details {
+
+    template <typename MODEL>
+    void collect_color_schemes(MODEL* model, const QString& scalar_prefix, std::vector<QString>& schemes) {
+        // color schemes from color properties and texture
+        for (const auto &name : model->face_properties()) {
+            if (name.find("f:color") != std::string::npos)
+                schemes.push_back(QString::fromStdString(name));
+        }
+        for (const auto &name : model->vertex_properties()) {
+            if (name.find("v:color") != std::string::npos || name.find("v:texcoord") != std::string::npos)
+                schemes.push_back(QString::fromStdString(name));
+        }
+
+        // scalar fields defined on faces
+        for (const auto &name : model->face_properties()) {
+            if (model->template get_face_property<float>(name))
+                schemes.push_back(scalar_prefix + QString::fromStdString(name));
+            else if (model->template get_face_property<double>(name))
+                schemes.push_back(scalar_prefix + QString::fromStdString(name));
+            else if (model->template get_face_property<unsigned int>(name))
+                schemes.push_back(scalar_prefix + QString::fromStdString(name));
+            else if (model->template get_face_property<int>(name))
+                schemes.push_back(scalar_prefix + QString::fromStdString(name));
+            else if (model->template get_face_property<char>(name))
+                schemes.push_back(scalar_prefix + QString::fromStdString(name));
+            else if (model->template get_face_property<unsigned char>(name))
+                schemes.push_back(scalar_prefix + QString::fromStdString(name));
+        }
+
+        // scalar fields defined on vertices
+        for (const auto &name : model->vertex_properties()) {
+            if (model->template get_vertex_property<float>(name))
+                schemes.push_back(scalar_prefix + QString::fromStdString(name));
+            else if (model->template get_vertex_property<double>(name))
+                schemes.push_back(scalar_prefix + QString::fromStdString(name));
+            else if (model->template get_vertex_property<unsigned int>(name))
+                schemes.push_back(scalar_prefix + QString::fromStdString(name));
+            else if (model->template get_vertex_property<int>(name))
+                schemes.push_back(scalar_prefix + QString::fromStdString(name));
+            else if (model->template get_vertex_property<char>(name))
+                schemes.push_back(scalar_prefix + QString::fromStdString(name));
+            else if (model->template get_vertex_property<unsigned char>(name))
+                schemes.push_back(scalar_prefix + QString::fromStdString(name));
+        }
+    }
+
+    // vector fields defined on faces
+    template <typename MODEL>
+    void vector_fields_on_faces(MODEL* model, std::vector<QString>& fields) {
+        fields.push_back("f:normal");
+        for (const auto &name : model->face_properties()) {
+            if (model->template get_face_property<vec3>(name)) {
+                if (name != "f:normal" && name != "f:color")
+                    fields.push_back(QString::fromStdString(name));
+            }
+        }
+    }
[…]
