### 2021-01-14 11d818c
reset manipulator after model has changed
