### 2021-01-14 0d33c88
add option to allow translation w.r.t. first point
