+++ easy3d/fileio/point_cloud_io_las.cpp
@@ -36,218 +36,210 @@
 namespace easy3d {
 
 
-	namespace io {
+    namespace io {
 
 
-		bool load_las(const std::string& file_name, PointCloud* cloud)
-		{
-			LASreadOpener lasreadopener;
-			lasreadopener.set_file_name(file_name.c_str(), true);
+        bool load_las(const std::string &file_name, PointCloud *cloud, bool transform, const std::string &prop_name) {
+            LASreadOpener lasreadopener;
+            lasreadopener.set_file_name(file_name.c_str(), true);
 
-			LASreader* lasreader = lasreadopener.open();
-			if (!lasreader || lasreader->npoints <= 0) {
+            LASreader *lasreader = lasreadopener.open();
+            if (!lasreader || lasreader->npoints <= 0) {
                 LOG(ERROR) << "could not open file: " << file_name;
-				lasreader->close();
-				delete lasreader;
-				return false;
-			}
+                lasreader->close();
+                delete lasreader;
+                return false;
+            }
 
-			std::size_t num = lasreader->npoints;
-			LOG(INFO) << "reading " << num << " points...";
+            std::size_t num = lasreader->npoints;
+            LOG(INFO) << "reading " << num << " points...";
 
-			// read the first point
-			if (!lasreader->read_point()) {
-				LOG(ERROR) << "failed reading point";
-				lasreader->close();
-				delete lasreader;
-				return false;
-			}
+            // read the first point
+            if (!lasreader->read_point()) {
+                LOG(ERROR) << "failed reading point";
+                lasreader->close();
+                delete lasreader;
+                return false;
+            }
 
-			PointCloud::VertexProperty<vec3> colors = cloud->add_vertex_property<vec3>("v:color");
-			PointCloud::VertexProperty<int>  classification = cloud->add_vertex_property<int>("v:classification");
+            LASpoint &p0 = lasreader->point;
+            // compute the actual coordinates as double floating point values
+            p0.compute_coordinates();
+            double x0 = p0.coordinates[0];
+            double y0 = p0.coordinates[1];
+            double z0 = p0.coordinates[2];
+            LOG_IF((!transform) && (x0 > 1e4 || y0 > 1e4 || z0 > 1e4), WARNING)
+                << "model has large coordinates (first point: "
+                << x0 << " " << y0 << " " << z0
+                << ") and some decimals may be lost. Hint: transform the model w.r.t. its first point";
 
-			// Liangliang: las format usually represent very large area of urban scenes and some coordinates
-			//			   may have very large values. In order to render the points properly in OpenGL, I
-			//			   record the relative positions to the first point stored in the file.
-			LASpoint& p0 = lasreader->point;
-			float r, g, b;
-			if (p0.have_rgb) { // some file may have rgb
-				r = float(p0.get_R()) / USHRT_MAX;
-				g = float(p0.get_G()) / USHRT_MAX;
-				b = float(p0.get_B()) / USHRT_MAX;
-			}
-			else // in case color doesn't exist, use intensity
-				r = g = b = p0.intensity % 255 / 255.0f;
+            float r = p0.have_rgb ? float(p0.get_R()) / USHRT_MAX : p0.intensity % 255 / 255.0f;
+            float g = p0.have_rgb ? float(p0.get_G()) / USHRT_MAX : p0.intensity % 255 / 255.0f;
+            float b = p0.have_rgb ? float(p0.get_B()) / USHRT_MAX : p0.intensity % 255 / 255.0f;
 
-			PointCloud::Vertex v = cloud->add_vertex(vec3(0, 0, 0));
-			colors[v] = vec3(r, g, b);
-			classification[v] = p0.classification;
+            auto colors = cloud->add_vertex_property<vec3>("v:color");
[…]
