+++ easy3d/renderer/buffers.cpp
@@ -76,9 +76,39 @@ namespace easy3d {
                               << "]";
             }
 
-            template<typename FT>
+//            template<typename FT>
+//            inline void
+//            update(PointCloud *model, PointsDrawable *drawable, PointCloud::VertexProperty<FT> prop) {
+//                assert(model);
+//                assert(drawable);
+//                assert(prop);
+//
+//                if (model->empty()) {
+//                    LOG(WARNING) << "model has no valid geometry";
+//                    return;
+//                }
+//
+//                const float dummy_lower = (drawable->clamp_range() ? drawable->clamp_lower() : 0.0f);
+//                const float dummy_upper = (drawable->clamp_range() ? drawable->clamp_upper() : 0.0f);
+//                float min_value = std::numeric_limits<float>::max();
+//                float max_value = -std::numeric_limits<float>::max();
+//                details::clamp_scalar_field(prop.vector(), min_value, max_value, dummy_lower, dummy_upper);
+//
+//                auto points = model->get_vertex_property<vec3>("v:point");
+//                std::vector<vec2> d_texcoords;
+//                d_texcoords.reserve(model->n_vertices());
+//                for (auto v : model->vertices()) {
+//                    float coord = (prop[v] - min_value) / (max_value - min_value);
+//                    d_texcoords.emplace_back(vec2(coord, 0.5f));
+//                }
+//                drawable->update_vertex_buffer(points.vector());
+//                drawable->update_texcoord_buffer(d_texcoords);
+//            }
+
+
+            template<typename Mesh, typename FT>
             inline void
-            update(PointCloud *model, PointsDrawable *drawable, PointCloud::VertexProperty<FT> prop) {
+            update_scalar(Mesh *model, PointsDrawable *drawable, typename Mesh::template VertexProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -94,7 +124,8 @@ namespace easy3d {
                 float max_value = -std::numeric_limits<float>::max();
                 details::clamp_scalar_field(prop.vector(), min_value, max_value, dummy_lower, dummy_upper);
 
-                auto points = model->get_vertex_property<vec3>("v:point");
+                auto points = model->template get_vertex_property<vec3>("v:point");
+
                 std::vector<vec2> d_texcoords;
                 d_texcoords.reserve(model->n_vertices());
                 for (auto v : model->vertices()) {
@@ -106,9 +137,9 @@ namespace easy3d {
             }
 
 
-            template<typename FT>
+            template<typename Mesh, typename FT>
             inline void
-            update(SurfaceMesh *model, PointsDrawable *drawable, SurfaceMesh::VertexProperty<FT> prop) {
+            update(Mesh *model, LinesDrawable *drawable, typename Mesh::template EdgeProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -124,45 +155,14 @@ namespace easy3d {
                 float max_value = -std::numeric_limits<float>::max();
                 details::clamp_scalar_field(prop.vector(), min_value, max_value, dummy_lower, dummy_upper);
 
-                auto points = model->get_vertex_property<vec3>("v:point");
-
-                std::vector<vec2> d_texcoords;
-                d_texcoords.reserve(model->n_vertices());
-                for (auto v : model->vertices()) {
-                    float coord = (prop[v] - min_value) / (max_value - min_value);
-                    d_texcoords.emplace_back(vec2(coord, 0.5f));
-                }
-                drawable->update_vertex_buffer(points.vector());
-                drawable->update_texcoord_buffer(d_texcoords);
-            }
-
-
-            template<typename FT>
-            inline void
[…]
