### 2021-01-14 ba7445f
pick only visible parts
