+++ easy3d/renderer/buffers.cpp
@@ -512,43 +512,43 @@ namespace easy3d {
             }
 
 
-            template<typename FT>
-            void update(Graph *model, LinesDrawable *drawable, Graph::EdgeProperty<FT> prop) {
-                assert(model);
-                assert(drawable);
-                assert(prop);
-
-                if (model->empty()) {
-                    LOG(WARNING) << "model has no valid geometry";
-                    return;
-                }
-
-                const float dummy_lower = (drawable->clamp_range() ? drawable->clamp_lower() : 0.0f);
-                const float dummy_upper = (drawable->clamp_range() ? drawable->clamp_upper() : 0.0f);
-                float min_value = std::numeric_limits<float>::max();
-                float max_value = -std::numeric_limits<float>::max();
-                details::clamp_scalar_field(prop.vector(), min_value, max_value, dummy_lower, dummy_upper);
-
-                auto points = model->get_vertex_property<vec3>("v:point");
-                std::vector<vec3> d_points;
-                d_points.reserve(model->n_edges() * 2);
-                std::vector<vec2> d_texcoords;
-                d_texcoords.reserve(model->n_edges() * 2);
-                for (auto e : model->edges()) {
-                    Graph::Vertex s = model->vertex(e, 0);
-                    Graph::Vertex t = model->vertex(e, 1);
-                    d_points.push_back(points[s]);
-                    d_points.push_back(points[t]);
-                    float coord_s = (prop[e] - min_value) / (max_value - min_value);
-                    d_texcoords.emplace_back(vec2(coord_s, 0.5f));
-                    float coord_t = (prop[e] - min_value) / (max_value - min_value);
-                    d_texcoords.emplace_back(vec2(coord_t, 0.5f));
-                }
-                drawable->update_vertex_buffer(d_points);
-                drawable->update_texcoord_buffer(d_texcoords);
-                drawable->release_element_buffer();
-                drawable->set_impostor_type(LinesDrawable::CYLINDER);
-            }
+//            template<typename FT>
+//            void update(Graph *model, LinesDrawable *drawable, Graph::EdgeProperty<FT> prop) {
+//                assert(model);
+//                assert(drawable);
+//                assert(prop);
+//
+//                if (model->empty()) {
+//                    LOG(WARNING) << "model has no valid geometry";
+//                    return;
+//                }
+//
+//                const float dummy_lower = (drawable->clamp_range() ? drawable->clamp_lower() : 0.0f);
+//                const float dummy_upper = (drawable->clamp_range() ? drawable->clamp_upper() : 0.0f);
+//                float min_value = std::numeric_limits<float>::max();
+//                float max_value = -std::numeric_limits<float>::max();
+//                details::clamp_scalar_field(prop.vector(), min_value, max_value, dummy_lower, dummy_upper);
+//
+//                auto points = model->get_vertex_property<vec3>("v:point");
+//                std::vector<vec3> d_points;
+//                d_points.reserve(model->n_edges() * 2);
+//                std::vector<vec2> d_texcoords;
+//                d_texcoords.reserve(model->n_edges() * 2);
+//                for (auto e : model->edges()) {
+//                    Graph::Vertex s = model->vertex(e, 0);
+//                    Graph::Vertex t = model->vertex(e, 1);
+//                    d_points.push_back(points[s]);
+//                    d_points.push_back(points[t]);
+//                    float coord_s = (prop[e] - min_value) / (max_value - min_value);
+//                    d_texcoords.emplace_back(vec2(coord_s, 0.5f));
+//                    float coord_t = (prop[e] - min_value) / (max_value - min_value);
+//                    d_texcoords.emplace_back(vec2(coord_t, 0.5f));
+//                }
[…]
