+++ easy3d/renderer/buffers.cpp
@@ -106,9 +106,9 @@ namespace easy3d {
 //            }
 
 
-            template<typename Mesh, typename FT>
+            template<typename Model, typename FT>
             inline void
-            update_scalar(Mesh *model, PointsDrawable *drawable, typename Mesh::template VertexProperty<FT> prop) {
+            update_scalar(Model *model, PointsDrawable *drawable, typename Model::template VertexProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -137,9 +137,9 @@ namespace easy3d {
             }
 
 
-            template<typename Mesh, typename FT>
+            template<typename Model, typename FT>
             inline void
-            update(Mesh *model, LinesDrawable *drawable, typename Mesh::template EdgeProperty<FT> prop) {
+            update_scalar(Model *model, LinesDrawable *drawable, typename Model::template EdgeProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -175,9 +175,9 @@ namespace easy3d {
             }
 
 
-            template<typename Mesh, typename FT>
+            template<typename Model, typename FT>
             inline void
-            update(Mesh *model, LinesDrawable *drawable, typename Mesh::template VertexProperty<FT> prop) {
+            update_scalar(Model *model, LinesDrawable *drawable, typename Model::template VertexProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -380,9 +380,9 @@ namespace easy3d {
             }
 
 
-            template<typename Mesh, typename FT>
+            template<typename FT>
             inline void
-            update(Mesh *model, TrianglesDrawable *drawable, typename Mesh::template VertexProperty<FT> prop) {
+            update(SurfaceMesh *model, TrianglesDrawable *drawable, SurfaceMesh::VertexProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -393,9 +393,9 @@ namespace easy3d {
                 }
 
                 if (model->is_triangle_mesh()) {
-                    auto points = model->template get_vertex_property<vec3>("v:point");
+                    auto points = model->get_vertex_property<vec3>("v:point");
                     model->update_vertex_normals();
-                    auto normals = model->template get_vertex_property<vec3>("v:normal");
+                    auto normals = model->get_vertex_property<vec3>("v:normal");
 
                     const float dummy_lower = (drawable->clamp_range() ? drawable->clamp_lower() : 0.0f);
                     const float dummy_upper = (drawable->clamp_range() ? drawable->clamp_upper() : 0.0f);
@@ -422,7 +422,7 @@ namespace easy3d {
                     drawable->update_normal_buffer(normals.vector());
                     drawable->update_texcoord_buffer(d_texcoords);
 
-                    auto triangle_range = model->template face_property<std::pair<int, int> >("f:triangle_range");
+                    auto triangle_range = model->face_property<std::pair<int, int> >("f:triangle_range");
                     int idx = 0;
                     for (auto face : model->faces()) {
                         triangle_range[face] = std::make_pair(idx, idx);
@@ -441,7 +441,7 @@ namespace easy3d {
                      * implemented by selecting triangle primitives using shaders. This allows data uploaded to the GPU
                      * for the rendering purpose be shared for selection. Yeah, performance gain!
                      */
-                    auto triangle_range = model->template face_property<std::pair<int, int> >("f:triangle_range");
+                    auto triangle_range = model->face_property<std::pair<int, int> >("f:triangle_range");
                     int count_triangles = 0;
 
                     /**
@@ -453,9 +453,9 @@ namespace easy3d {
                      * between flat and smooth shading without transferring different data to the GPU.
[…]
