+++ easy3d/renderer/buffers.cpp
@@ -76,39 +76,10 @@ namespace easy3d {
                               << "]";
             }
 
-//            template<typename FT>
-//            inline void
-//            update(PointCloud *model, PointsDrawable *drawable, PointCloud::VertexProperty<FT> prop) {
-//                assert(model);
-//                assert(drawable);
-//                assert(prop);
-//
-//                if (model->empty()) {
-//                    LOG(WARNING) << "model has no valid geometry";
-//                    return;
-//                }
-//
-//                const float dummy_lower = (drawable->clamp_range() ? drawable->clamp_lower() : 0.0f);
-//                const float dummy_upper = (drawable->clamp_range() ? drawable->clamp_upper() : 0.0f);
-//                float min_value = std::numeric_limits<float>::max();
-//                float max_value = -std::numeric_limits<float>::max();
-//                details::clamp_scalar_field(prop.vector(), min_value, max_value, dummy_lower, dummy_upper);
-//
-//                auto points = model->get_vertex_property<vec3>("v:point");
-//                std::vector<vec2> d_texcoords;
-//                d_texcoords.reserve(model->n_vertices());
-//                for (auto v : model->vertices()) {
-//                    float coord = (prop[v] - min_value) / (max_value - min_value);
-//                    d_texcoords.emplace_back(vec2(coord, 0.5f));
-//                }
-//                drawable->update_vertex_buffer(points.vector());
-//                drawable->update_texcoord_buffer(d_texcoords);
-//            }
-
 
             template<typename Model, typename FT>
             inline void
-            update_scalar(Model *model, PointsDrawable *drawable, typename Model::template VertexProperty<FT> prop) {
+            update_scalar_on_vertices(Model *model, PointsDrawable *drawable, typename Model::template VertexProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -139,7 +110,7 @@ namespace easy3d {
 
             template<typename Model, typename FT>
             inline void
-            update_scalar(Model *model, LinesDrawable *drawable, typename Model::template EdgeProperty<FT> prop) {
+            update_scalar_on_edges(Model *model, LinesDrawable *drawable, typename Model::template EdgeProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -177,7 +148,7 @@ namespace easy3d {
 
             template<typename Model, typename FT>
             inline void
-            update_scalar(Model *model, LinesDrawable *drawable, typename Model::template VertexProperty<FT> prop) {
+            update_scalar_on_vertices(Model *model, LinesDrawable *drawable, typename Model::template VertexProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -216,35 +187,98 @@ namespace easy3d {
             }
 
 
-//            template<typename FT>
-//            inline void
-//            update(Graph *model, PointsDrawable *drawable, Graph::VertexProperty<FT> prop) {
-//                assert(model);
-//                assert(drawable);
-//                assert(prop);
-//
-//                if (model->empty()) {
-//                    LOG(WARNING) << "model has no valid geometry";
-//                    return;
-//                }
-//
-//                const float dummy_lower = (drawable->clamp_range() ? drawable->clamp_lower() : 0.0f);
-//                const float dummy_upper = (drawable->clamp_range() ? drawable->clamp_upper() : 0.0f);
-//                float min_value = std::numeric_limits<float>::max();
-//                float max_value = -std::numeric_limits<float>::max();
-//                details::clamp_scalar_field(prop.vector(), min_value, max_value, dummy_lower, dummy_upper);
-//
[…]
