+++ easy3d/renderer/buffers.cpp
@@ -187,99 +187,6 @@ namespace easy3d {
             }
 
 
-            template<typename Model>
-            inline void
-            update_scalar_on_vertices(Model *model, PointsDrawable *drawable, const std::string &name) {
-                if (model->template get_vertex_property<float>(name)) {
-                    auto prop = model->template get_vertex_property<float>(name);
-                    details::update_scalar_on_vertices<Model>(model, drawable, prop);
-                } else if (model->template get_vertex_property<double>(name)) {
-                    auto prop = model->template get_vertex_property<double>(name);
-                    details::update_scalar_on_vertices<Model>(model, drawable, prop);
-                } else if (model->template get_vertex_property<int>(name)) {
-                    auto prop = model->template get_vertex_property<int>(name);
-                    details::update_scalar_on_vertices<Model>(model, drawable, prop);
-                } else if (model->template get_vertex_property<unsigned int>(name)) {
-                    auto prop = model->template get_vertex_property<unsigned int>(name);
-                    details::update_scalar_on_vertices<Model>(model, drawable, prop);
-                } else if (model->template get_vertex_property<char>(name)) {
-                    auto prop = model->template get_vertex_property<char>(name);
-                    details::update_scalar_on_vertices<Model>(model, drawable, prop);
-                } else if (model->template get_vertex_property<unsigned char>(name)) {
-                    auto prop = model->template get_vertex_property<unsigned char>(name);
-                    details::update_scalar_on_vertices<Model>(model, drawable, prop);
-                } else {
-                    LOG(WARNING) << "scalar field \'" << name
-                                 << "\' not found from vertex properties (use uniform coloring)";
-                    drawable->set_coloring_method(State::UNIFORM_COLOR);
-                    update(model, drawable);
-                    return;
-                }
-            }
-            
-
-            template<typename Model>
-            inline void
-            update_scalar_on_vertices(Model *model, LinesDrawable *drawable, const std::string &name) {
-                if (model->template get_vertex_property<float>(name)) {
-                    auto prop = model->template get_vertex_property<float>(name);
-                    details::update_scalar_on_vertices<Model>(model, drawable, prop);
-                } else if (model->template get_vertex_property<double>(name)) {
-                    auto prop = model->template get_vertex_property<double>(name);
-                    details::update_scalar_on_vertices<Model>(model, drawable, prop);
-                } else if (model->template get_vertex_property<int>(name)) {
-                    auto prop = model->template get_vertex_property<int>(name);
-                    details::update_scalar_on_vertices<Model>(model, drawable, prop);
-                } else if (model->template get_vertex_property<unsigned int>(name)) {
-                    auto prop = model->template get_vertex_property<unsigned int>(name);
-                    details::update_scalar_on_vertices<Model>(model, drawable, prop);
-                } else if (model->template get_vertex_property<char>(name)) {
-                    auto prop = model->template get_vertex_property<char>(name);
-                    details::update_scalar_on_vertices<Model>(model, drawable, prop);
-                } else if (model->template get_vertex_property<unsigned char>(name)) {
-                    auto prop = model->template get_vertex_property<unsigned char>(name);
-                    details::update_scalar_on_vertices<Model>(model, drawable, prop);
-                } else {
-                    LOG(WARNING) << "scalar field \'" << name
-                                 << "\' not found from vertex properties (use uniform coloring)";
-                    drawable->set_coloring_method(State::UNIFORM_COLOR);
-                    update(model, drawable);
-                    return;
-                }
-            }
-
-
-            template<typename Model>
-            inline void
[…]
