+++ easy3d/renderer/buffers.cpp
@@ -453,7 +453,114 @@ namespace easy3d {
             }
 
 
-            template<typename Mesh>
+            void update_uniform_coloring(PolyMesh* model, TrianglesDrawable* drawable, bool border) {
+                assert(model);
+                assert(drawable);
+
+                if (model->empty()) {
+                    LOG(WARNING) << "model has no valid geometry";
+                    return;
+                }
+
+                model->update_vertex_normals();
+                auto vnormal = model->get_vertex_property<vec3>("v:normal");
+
+                if (model->is_tetraheral_mesh()) {
+                    std::vector<unsigned int> d_indices;
+                    std::vector<vec3> d_normals;
+                    for (auto f : model->faces()) {
+                        if (model->is_border(f) == border) {
+                            for (auto v : model->vertices(f)) {
+                                d_indices.push_back(v.idx());
+                                d_normals.push_back(vnormal[v]);
+                            }
+                        }
+                    }
+
+                    drawable->update_vertex_buffer(model->points());
+                    drawable->update_normal_buffer(d_normals);
+                    drawable->update_element_buffer(d_indices);
+                } else {
+                    /**
+                     * We use the Tessellator to eliminate duplicate vertices. This allows us to take advantage of element
+                     * buffer to minimize the number of vertices sent to the GPU.
+                     */
+                    Tessellator tessellator;
+
+                    for (auto f : model->faces()) {
+                        if (model->is_border(f) == border) {
+                            tessellator.begin_polygon(model->compute_face_normal(model->halfface(f, 0)));
+                            // tessellator.set_winding_rule(Tessellator::WINDING_NONZERO);  // or POSITIVE
+                            tessellator.begin_contour();
+                            for (auto v : model->vertices(f)) {
+                                Tessellator::Vertex vertex(model->position(v), v.idx());
+                                vertex.append(vnormal[v]);
+                                tessellator.add_vertex(vertex);
+                            }
+                            tessellator.end_contour();
+                            tessellator.end_polygon();
+                        }
+                    }
+
+                    const std::vector<Tessellator::Vertex *> &vts = tessellator.vertices();
+                    std::vector<vec3> d_points;
+                    std::vector<vec3> d_normals;
+                    d_points.reserve(vts.size());
+                    d_normals.reserve(vts.size());
+
+                    for (auto v :vts) {
+                        d_points.emplace_back(v->data());
+                        d_normals.emplace_back(v->data() + 3);
+                    }
+
+                    const auto &d_indices = tessellator.elements();
+
+                    drawable->update_vertex_buffer(d_points);
+                    drawable->update_normal_buffer(d_normals);
+                    drawable->update_element_buffer(d_indices);
+
+                    DLOG(INFO) << "num of vertices in model/sent to GPU: " << model->n_vertices() << "/"
+                               << d_points.size();
+                }
+            }
+
+
+            template<typename FT>
+            inline void
+            update(PolyMesh *model, TrianglesDrawable *drawable, PolyMesh::VertexProperty<vec3> prop, bool border) {
+                LOG(WARNING) << "not implemented yet";
+            }
+
+
+            template<typename FT>
+            inline void
+            update(PolyMesh *model, TrianglesDrawable *drawable, PolyMesh::FaceProperty<vec3> prop, bool border) {
+                LOG(WARNING) << "not implemented yet";
+            }
+
+
+            template<typename FT>
+            inline void
+            update(PolyMesh *model, TrianglesDrawable *drawable, PolyMesh::VertexProperty<vec2> prop, bool border) {
+                LOG(WARNING) << "not implemented yet";
+            }
+
+
+            template<typename FT>
[…]
