### 2021-01-14 10b9146
allow to show per vertex and per face color
