+++ easy3d/fileio/CMakeLists.txt
@@ -32,7 +32,7 @@ set(${PROJECT_NAME}_SOURCES
         surface_mesh_io_obj.cpp
         surface_mesh_io_off.cpp
         surface_mesh_io_ply.cpp
-        surface_mesh_io_smesh.cpp
+        surface_mesh_io_sm.cpp
         surface_mesh_io_stl.cpp
         poly_mesh_io.cpp
         poly_mesh_io_mesh.cpp
