### 2021-01-14 9014361
allow to draw PolyMesh with texcoords on vertices
