+++ easy3d/renderer/buffers.cpp
@@ -189,7 +189,7 @@ namespace easy3d {
 
             template<typename FT>
             inline void
-            update(SurfaceMesh *model, TrianglesDrawable *drawable, SurfaceMesh::FaceProperty<FT> prop) {
+            update_scalar_on_faces(SurfaceMesh *model, TrianglesDrawable *drawable, SurfaceMesh::FaceProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -323,7 +323,7 @@ namespace easy3d {
 
             template<typename FT>
             inline void
-            update(SurfaceMesh *model, TrianglesDrawable *drawable, SurfaceMesh::VertexProperty<FT> prop) {
+            update_scalar_on_vertices(SurfaceMesh *model, TrianglesDrawable *drawable, SurfaceMesh::VertexProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -509,7 +509,7 @@ namespace easy3d {
 
 
             inline void
-            update_color_on_vertices(PolyMesh *model, TrianglesDrawable *drawable, PolyMesh::VertexProperty<vec3> colors, bool border) {
+            update_colors_on_vertices(PolyMesh *model, TrianglesDrawable *drawable, PolyMesh::VertexProperty<vec3> colors, bool border) {
                 assert(model);
                 assert(drawable);
                 assert(colors);
@@ -636,8 +636,65 @@ namespace easy3d {
 
 
             inline void
-            update_texcoords_on_vertices(PolyMesh *model, TrianglesDrawable *drawable, PolyMesh::VertexProperty<vec2> prop, bool border) {
-                LOG(WARNING) << "not implemented yet";
+            update_texcoords_on_vertices(PolyMesh *model, TrianglesDrawable *drawable, PolyMesh::VertexProperty<vec2> vtexcoords, bool border) {
+                assert(model);
+                assert(drawable);
+                assert(vtexcoords);
+
+                if (model->empty()) {
+                    LOG(WARNING) << "model has no valid geometry";
+                    return;
+                }
+
+                model->update_vertex_normals();
+                auto normals = model->get_vertex_property<vec3>("v:normal");
+                auto points = model->get_vertex_property<vec3>("v:point");
+
+                /**
+                 * We use the Tessellator to eliminate duplicate vertices. This allows us to take advantage of element
+                 * buffer to minimize the number of vertices sent to the GPU.
+                 */
+                Tessellator tessellator;
+
+                for (auto f : model->faces()) {
+                    if (model->is_border(f) == border) {
+                        tessellator.begin_polygon(model->compute_face_normal(model->halfface(f, 0)));
+                        // tessellator.set_winding_rule(Tessellator::WINDING_NONZERO);  // or POSITIVE
+                        tessellator.begin_contour();
+                        for (auto v : model->vertices(f)) {
+                            Tessellator::Vertex vertex(model->position(v), v.idx());
+                            vertex.append(normals[v]);
+                            vertex.append(vtexcoords[v]);
+                            tessellator.add_vertex(vertex);
+                        }
+                        tessellator.end_contour();
+                        tessellator.end_polygon();
+                    }
+                }
+
+                const std::vector<Tessellator::Vertex *> &vts = tessellator.vertices();
+                std::vector<vec3> d_points;
+                std::vector<vec3> d_normals;
+                std::vector<vec2> d_texcoords;
+                d_points.reserve(vts.size());
+                d_normals.reserve(vts.size());
+                d_texcoords.reserve(vts.size());
+
+                for (auto v :vts) {
+                    d_points.emplace_back(v->data());
+                    d_normals.emplace_back(v->data() + 3);
[…]
