### 2021-01-14 d423e95
allow to visualize scalar fields on vertices
