### 2021-01-14 530ef9a
allow to visualize scalar fields on faces
