### 2021-01-14 4bf3dc0
allow to visualize vector fields on faces and vertices of PolyMesh
