+++ easy3d/core/poly_mesh.h
@@ -1202,6 +1202,11 @@ namespace easy3d {
             return HalfFace((f.idx() << 1) + i);
         }
 
+        /// returns the face of HalfFace \c h.
+        Face face(HalfFace h) const {
+            return Face((h.idx() >> 1));
+        }
+
         /// returns the twin halfface of halfface \c h.
         HalfFace opposite(HalfFace h) const
         {
@@ -1330,8 +1335,8 @@ namespace easy3d {
         /// find the halfface defined by a sequence of \c vertices (orientation sensitive)
         HalfFace find_half_face(const std::vector<Vertex>& vertices) const;
 
-        /// returns whether halfface \c h is degenerate
-        bool is_degenerate(HalfFace h) const;
+        /// returns whether face \c f is degenerate
+        bool is_degenerate(Face f) const;
 
         /// extracts the boundary surface and return its faces
         void extract_boundary(std::vector< std::vector<Vertex> >& faces) const;
@@ -1354,8 +1359,8 @@ namespace easy3d {
         /// compute face normals by calling compute_face_normal(HalfFace) for each face.
         void update_face_normals();
 
-        /// compute normal vector of face \c h.
-        vec3 compute_face_normal(HalfFace h) const;
+        /// compute normal vector of face \c f.
+        vec3 compute_face_normal(Face f) const;
 
         /**
          * Computes vertex normals for each vertex.
@@ -1432,7 +1437,6 @@ namespace easy3d {
         CellProperty<CellConnectivity>          cconn_;
 
         VertexProperty<vec3>    vpoint_;
-        HalfFaceProperty<vec3>  fnormal_;
     };
 
 
