### 2021-01-14 c5f81db
allow to visualize vector fields on vertices of PolyMesh
