+++ easy3d/fileio/point_cloud_io_las.cpp
@@ -36,6 +36,7 @@
 namespace easy3d {
 
 
+    // \cond
     namespace io {
 
 
@@ -241,5 +242,5 @@ namespace easy3d {
         }
 
     }
-
+    // \endcond
 }
