+++ easy3d/fileio/surface_mesh_io_stl.cpp
@@ -37,6 +37,7 @@
 
 namespace easy3d {
 
+    // \cond
 	namespace io {
 
 
@@ -288,6 +289,7 @@ namespace easy3d {
 		}
 
 	} // namespace io
+    // \endcond
 
 } // namespace easy3d
 
