+++ easy3d/fileio/ply_reader_writer.cpp
@@ -37,6 +37,8 @@
 
 namespace easy3d {
 
+
+    // \cond
     namespace io {
 
 
@@ -672,5 +674,6 @@ namespace easy3d {
         }
 
     } // namespace io
+    // \endcond
 
 } // namespace easy3d
