### 2021-01-15 0884c3d
allow to disable select models
