### 2021-01-15 fddeb30
set default verbose level
