+++ tutorials/Tutorial_202_Viewer_Qt/viewer.cpp
@@ -148,11 +148,11 @@ void ViewerQt::initializeGL() {
 
     int major_requested = QSurfaceFormat::defaultFormat().majorVersion();
     int minor_requested = QSurfaceFormat::defaultFormat().minorVersion();
-    LOG(INFO) << "OpenGL vendor: " << func_->glGetString(GL_VENDOR);
-    LOG(INFO) << "OpenGL renderer: " << func_->glGetString(GL_RENDERER);
-    LOG(INFO) << "OpenGL version requested: " << major_requested << "." << minor_requested;
-    LOG(INFO) << "OpenGL version received: " << func_->glGetString(GL_VERSION);
-    LOG(INFO) << "GLSL version received: " << func_->glGetString(GL_SHADING_LANGUAGE_VERSION);
+    VLOG(1) << "OpenGL vendor: " << func_->glGetString(GL_VENDOR);
+    VLOG(1) << "OpenGL renderer: " << func_->glGetString(GL_RENDERER);
+    VLOG(1) << "OpenGL version requested: " << major_requested << "." << minor_requested;
+    VLOG(1) << "OpenGL version received: " << func_->glGetString(GL_VERSION);
+    VLOG(1) << "GLSL version received: " << func_->glGetString(GL_SHADING_LANGUAGE_VERSION);
 
     int major = 0;
     func_->glGetIntegerv(GL_MAJOR_VERSION, &major);
@@ -809,7 +809,7 @@ void ViewerQt::paintGL() {
                 LOG(WARNING) << "MSAA is available with " << samples_ << " samples (" << samples
                              << " requested but max support is " << max_num << ")";
         } else
-            LOG(INFO) << "Samples received: " << samples_ << " (" << samples << " requested, max support is "
+            VLOG(1) << "Samples received: " << samples_ << " (" << samples << " requested, max support is "
                       << max_num << ")";
 
         queried = true;
