+++ easy3d/util/logging.h
@@ -48,16 +48,17 @@ namespace easy3d {
 
         /**
          * @brief Initializes the logging module.
-         * @param verbose \c ture to log messages at a the INFO level to stderr.
-         *      \c false only log at WARNING, ERROR, FATAL levels.
+         * @param info_to_stderr \c ture to log messages at a the INFO level to stderr.
+         *      \c false only log at WARNING, ERROR, FATAL levels to stderr.
+         * @param verbose_level A value in the range [0, 9] for VLOG that will be written into the log file.
          * @param log_file A string specifying the full path to the log file.
-         *      If valid, the log messages will be written to this file in addition to logging to stderr.
-         *      No log file will be created if \p log_file is empty.
-         *      Passing "default" allows to creat a log file with a title "ApplicationName.log" in a
-         *      directory "logs" next to the executable file.
+         *      If \p log_file is a valid path: log messages will be written to this file in addition to stderr.
+         *      If \p log_file is empty: no log file will be created.
+         *      If \p log_file is "default": creat a log file with a title in the form "ApplicationName.log" in a
+         *          directory "logs" next to the executable file.
          * @note This initialization is optional. If not called, log messages will be written to stderr only.
          */
-        void initialize(bool verbose = false, const std::string &log_file = "default");
+        void initialize(bool info_to_stderr = false, int verbose_level = 0, const std::string &log_file = "default");
 
 
         /// Base class for a logger (that can log messages to whatever)
