### 2021-01-15 1fe447c
easylogging no default log file
