### 2021-01-15 6d20278
format logging
