### 2021-01-15 407936a
format logging
