+++ easy3d/renderer/clipping_plane.cpp
@@ -153,7 +153,6 @@ namespace easy3d {
         easy3d_debug_log_gl_error;
 
         // draw the face of the clipping plane
-        //glDisable(GL_DEPTH_TEST);
         glEnable(GL_BLEND);
         glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
         TrianglesDrawable face;
