+++ easy3d/core/surface_mesh_builder.cpp
@@ -22,7 +22,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <easy3d/core/manifold_builder.h>
+#include <easy3d/core/surface_mesh_builder.h>
 
 #include <set>
 
@@ -33,23 +33,23 @@
 namespace easy3d {
 
 
-    static const std::string name_known_nm_vertex("v:ManifoldBuilder:known_nm_vertex");
-    static const std::string name_visited_vertex("v:ManifoldBuilder:visited_vertex");
-    static const std::string name_visited_halfedge("h:ManifoldBuilder:visited_halfedge");
-    static const std::string name_original_vertex("h:ManifoldBuilder:original_vertex");
+    static const std::string name_known_nm_vertex("v:SurfaceMeshBuilder:known_nm_vertex");
+    static const std::string name_visited_vertex("v:SurfaceMeshBuilder:visited_vertex");
+    static const std::string name_visited_halfedge("h:SurfaceMeshBuilder:visited_halfedge");
+    static const std::string name_original_vertex("h:SurfaceMeshBuilder:original_vertex");
 
 
-    ManifoldBuilder::ManifoldBuilder(SurfaceMesh *mesh)
+    SurfaceMeshBuilder::SurfaceMeshBuilder(SurfaceMesh *mesh)
             : mesh_(mesh) {
     }
 
 
-    ManifoldBuilder::~ManifoldBuilder() {
+    SurfaceMeshBuilder::~SurfaceMeshBuilder() {
         LOG_IF(original_vertex_, ERROR) << "missing call to end_surface(), which must be in pair with begin_surface()";
     }
 
 
-    void ManifoldBuilder::begin_surface() {
+    void SurfaceMeshBuilder::begin_surface() {
         num_faces_less_three_vertices_ = 0;
         num_faces_duplicate_vertices = 0;
         num_faces_out_of_range_vertices_ = 0;
@@ -60,11 +60,11 @@ namespace easy3d {
         copied_vertices_for_linking_.clear();
         outgoing_halfedges_.clear();
 
-        original_vertex_ = mesh_->add_vertex_property<SurfaceMesh::Vertex>(name_original_vertex);
+        original_vertex_ = mesh_->add_vertex_property<Vertex>(name_original_vertex);
     }
 
 
-    void ManifoldBuilder::end_surface(bool log_issues) {
+    void SurfaceMeshBuilder::end_surface(bool log_issues) {
         // ----------------------------------------------------------------------------------
 
         // Step 1: Resolve non-manifold vertices (also collect data for the report).
@@ -223,15 +223,15 @@ namespace easy3d {
     }
 
 
-    SurfaceMesh::Vertex ManifoldBuilder::add_vertex(const vec3 &p) {
+    SurfaceMesh::Vertex SurfaceMeshBuilder::add_vertex(const vec3 &p) {
         DLOG_IF(!original_vertex_, ERROR) << "you must call begin_surface() before the constructing a surface mesh";
-        SurfaceMesh::Vertex v = mesh_->add_vertex(p);
+        Vertex v = mesh_->add_vertex(p);
         original_vertex_[v] = v;
         return v;
     }
 
 
-    bool ManifoldBuilder::vertices_valid(const std::vector<SurfaceMesh::Vertex> &vertices) {
+    bool SurfaceMeshBuilder::vertices_valid(const std::vector<Vertex> &vertices) {
         const std::size_t n = vertices.size();
 
         // Check #1: a face has less than 3 vertices
@@ -268,9 +268,9 @@ namespace easy3d {
     }
 
 
-    SurfaceMesh::Face ManifoldBuilder::add_face(const std::vector<SurfaceMesh::Vertex> &vertices) {
[…]
