+++ easy3d/fileio/surface_mesh_io_ply.cpp
@@ -25,7 +25,7 @@
 #include <easy3d/fileio/surface_mesh_io.h>
 #include <easy3d/fileio/ply_reader_writer.h>
 #include <easy3d/core/surface_mesh.h>
-#include <easy3d/core/manifold_builder.h>
+#include <easy3d/core/surface_mesh_builder.h>
 #include <easy3d/util/logging.h>
 
 
@@ -151,7 +151,7 @@ namespace easy3d {
 
 			mesh->clear();
 
-            ManifoldBuilder builder(mesh);
+            SurfaceMeshBuilder builder(mesh);
             builder.begin_surface();
 
             // add vertices
