+++ tutorials/Tutorial_308_TexturedMesh/viewer.cpp
@@ -27,7 +27,7 @@
 #include <unordered_map>
 
 #include <easy3d/core/surface_mesh.h>
-#include <easy3d/core/manifold_builder.h>
+#include <easy3d/core/surface_mesh_builder.h>
 #include <easy3d/renderer/texture_manager.h>
 #include <easy3d/renderer/camera.h>
 #include <easy3d/renderer/drawable_triangles.h>
@@ -85,7 +85,7 @@ namespace easy3d {
         SurfaceMesh *mesh = new SurfaceMesh;
         mesh->set_name(file_name);
 
-        ManifoldBuilder builder(mesh);
+        SurfaceMeshBuilder builder(mesh);
         builder.begin_surface();
 
         // add vertices

[…]
