+++ ReadMe.md
@@ -94,7 +94,7 @@ step</a>.
 
 
 ### Documentation
-The documentation for Easy3D-v2.4.1 is available [here](https://3d.bk.tudelft.nl/liangliang/software/easy3d_doc/html/index.html).
+The documentation for Easy3D-v2.4.2 is available [here](https://3d.bk.tudelft.nl/liangliang/software/easy3d_doc/html/index.html).
 
 The Easy3D Documentation is an on-going effort with more and more details being added. You can build the latest Easy3D 
 documentation from the source code.
