### 2021-01-15 2d3cda9
have both move() and drag() for tools
