+++ applications/Mapple/tools/tool_selection_surface_mesh.h
@@ -70,7 +70,7 @@ namespace easy3d {
             ToolSurfaceMeshFaceSelectionRect(ToolManager *mgr, SurfaceMeshPicker *picker, SelectMode mode = SM_SELECT);
 
             virtual void press(int x, int y) override;
-            virtual void move(int x, int y) override;
+            virtual void drag(int x, int y) override;
             virtual void release(int x, int y) override;
 
         protected:
@@ -86,7 +86,7 @@ namespace easy3d {
             ToolSurfaceMeshFaceSelectionLasso(ToolManager *mgr, SurfaceMeshPicker *picker, SelectMode mode = SM_SELECT);
 
             virtual void press(int x, int y) override;
-            virtual void move(int x, int y) override;
+            virtual void drag(int x, int y) override;
             virtual void release(int x, int y) override;
 
         protected:
