### 2021-01-16 0e8f7da
clean code & allow select primitives
