+++ easy3d/renderer/dual_depth_peeling.cpp
@@ -195,7 +195,7 @@ namespace easy3d {
                     program->set_uniform("MANIP", d->manipulated_matrix());
                     if (setting::clipping_plane)
                         setting::clipping_plane->set_program(program, d->plane_clip_discard_primitive());
-                    d->gl_draw(false);
+                    d->gl_draw();
                 }
             }
             program->release();
@@ -250,7 +250,7 @@ namespace easy3d {
                     program->set_uniform("selected", d->is_selected());
                     if (setting::clipping_plane)
                         setting::clipping_plane->set_program(program, d->plane_clip_discard_primitive());
-                    d->gl_draw(false);
+                    d->gl_draw();
                 }
             }
             program->release_texture();
