### 2021-01-16 9e1c82c
disable selection when switching non-uniform color rendering
