+++ easy3d/renderer/buffers.cpp
@@ -77,9 +77,9 @@ namespace easy3d {
             }
 
 
-            template<typename Model, typename FT>
+            template<typename MODEL, typename FT>
             inline void
-            update_scalar_on_vertices(Model *model, PointsDrawable *drawable, typename Model::template VertexProperty<FT> prop) {
+            update_scalar_on_vertices(MODEL *model, PointsDrawable *drawable, typename MODEL::template VertexProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -108,9 +108,9 @@ namespace easy3d {
             }
 
 
-            template<typename Model, typename FT>
+            template<typename MODEL, typename FT>
             inline void
-            update_scalar_on_edges(Model *model, LinesDrawable *drawable, typename Model::template EdgeProperty<FT> prop) {
+            update_scalar_on_edges(MODEL *model, LinesDrawable *drawable, typename MODEL::template EdgeProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -146,9 +146,9 @@ namespace easy3d {
             }
 
 
-            template<typename Model, typename FT>
+            template<typename MODEL, typename FT>
             inline void
-            update_scalar_on_vertices(Model *model, LinesDrawable *drawable, typename Model::template VertexProperty<FT> prop) {
+            update_scalar_on_vertices(MODEL *model, LinesDrawable *drawable, typename MODEL::template VertexProperty<FT> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -465,6 +465,8 @@ namespace easy3d {
                 model->update_vertex_normals();
                 auto normals = model->get_vertex_property<vec3>("v:normal");
 
+                // since we have two parts, no need to transfer all vertices and normals
+                // I just use the tessellator
                 /*
                 if (model->is_tetraheral_mesh()) {
                     std::vector<unsigned int> d_indices;
@@ -874,8 +876,8 @@ namespace easy3d {
             }
 
 
-            template<typename Model>
-            void update_texcoords_on_vertices(Model *model, PointsDrawable *drawable, typename Model::template VertexProperty<vec2> prop) {
+            template<typename MODEL>
+            void update_texcoords_on_vertices(MODEL *model, PointsDrawable *drawable, typename MODEL::template VertexProperty<vec2> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -1445,8 +1447,8 @@ namespace easy3d {
             }
 
 
-            template <typename Model>
-            void update_colors_on_edges(Model *model, LinesDrawable *drawable, typename Model::template EdgeProperty<vec3> prop) {
+            template <typename MODEL>
+            void update_colors_on_edges(MODEL *model, LinesDrawable *drawable, typename MODEL::template EdgeProperty<vec3> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -1474,9 +1476,9 @@ namespace easy3d {
             }
 
 
-            template <typename Model>
+            template <typename MODEL>
             void
-            update_colors_on_vertices(Model *model, LinesDrawable *drawable, typename Model::template VertexProperty<vec3> prop) {
+            update_colors_on_vertices(MODEL *model, LinesDrawable *drawable, typename MODEL::template VertexProperty<vec3> prop) {
                 assert(model);
                 assert(drawable);
                 assert(prop);
@@ -1504,9 +1506,9 @@ namespace easy3d {
             }
 
 
-            template <typename Model>
+            template <typename MODEL>
             void
-            update_texcoords_on_vertices(Model *model, LinesDrawable *drawable, typename Model::template VertexProperty<vec2> prop) {
[…]
