### 2021-01-17 b5f8f27
slection visualized as boolean type scalar fields
