### 2021-01-17 b8c2d46
sm format for surface meshes
