+++ easy3d/viewer/viewer.cpp
@@ -1084,7 +1084,7 @@ namespace easy3d {
             is_ply_mesh = (io::PlyReader::num_instances(file_name, "face") > 0);
 
         Model *model = nullptr;
-        if ((ext == "ply" && is_ply_mesh) || ext == "obj" || ext == "off" || ext == "stl" || ext == "smesh" ||
+        if ((ext == "ply" && is_ply_mesh) || ext == "obj" || ext == "off" || ext == "stl" || ext == "sm" ||
             ext == "plg") { // mesh
             model = SurfaceMeshIO::load(file_name);
         } else if (ext == "ply" && io::PlyReader::num_instances(file_name, "edge") > 0) {
@@ -1240,7 +1240,7 @@ namespace easy3d {
         const std::string title("Please choose a file");
         const std::string &default_path = resource::directory() + "/data/";
         const std::vector<std::string> &filters = {
-                "Surface Mesh (*.obj *.ply *.off *.stl *.smesh)", "*.obj *.ply *.off *.stl *.smesh",
+                "Surface Mesh (*.obj *.ply *.off *.stl *.sm)", "*.obj *.ply *.off *.stl *.sm",
                 "Point Cloud (*.bin *.ply *.xyz *.bxyz *.las *.laz *.vg *.bvg *.ptx)",
                 "*.bin *.ply *.xyz *.bxyz *.las *.laz *.vg *.bvg *.ptx",
                 "Polytope Mesh (*.plm *.pm *.mesh)", "*.plm *.pm *.mesh",
@@ -1271,7 +1271,7 @@ namespace easy3d {
 
         const std::string &title = "Please choose a file name";
         const std::vector<std::string> &filters = {
-                "Surface Mesh (*.obj *.ply *.off *.stl *.smesh)", "*.obj *.ply *.off *.stl *.smesh",
+                "Surface Mesh (*.obj *.ply *.off *.stl *.sm)", "*.obj *.ply *.off *.stl *.sm",
                 "Point Cloud (*.bin *.ply *.xyz *.bxyz *.las *.laz *.vg *.bvg)",
                 "*.bin *.ply *.xyz *.bxyz *.las *.laz *.vg *.bvg",
                 "Polytope Mesh (*.plm *.pm *.mesh)", "*.plm *.pm *.mesh",
