### 2021-01-17 5f4c1a6
update only the texcoord buffer for primitive selection
