+++ applications/Mapple/tools/tool_selection_point_cloud.h
@@ -43,6 +43,8 @@ namespace easy3d {
             ToolPointCloudSelection(ToolManager *mgr);
             virtual ~ToolPointCloudSelection() {}
             std::string name() const { return "PointCloudSelect"; }
+
+            void update_render_buffer(PointCloud* cloud) const;
         };
 
         // -------------------- Rect Select ----------------------
