### 2021-01-17 aa86811
visualize selection when invert selection triggered
