### 2021-01-17 2650b42
consistent color for selection
