### 2021-01-17 818e077
log clamped scalar field range
