### 2021-01-17 b2b2c8d
avoid updating the texcoord buffer twice
