### 2021-01-17 8dfaf4b
show drawables of each model on the model list panel
