### 2021-01-17 18e62ea
change drawable visibility from the model list widget
