+++ applications/Mapple/widgets/widget_model_list.h
@@ -5,6 +5,7 @@
 
 namespace easy3d {
     class Model;
+    class Drawable;
 }
 
 class MainWindow;
@@ -32,6 +33,8 @@ public:
 
     void prepareContextMenu(QMenu *menu);
 
+    void updateDrawableVisibility(easy3d::Drawable* d);
+
 private Q_SLOTS :
 
     void setAutoFocus(bool b);
@@ -54,6 +57,9 @@ private Q_SLOTS :
     void modelItemSelectionChanged();
 
     void showContextMenu(const QPoint &p);
+
+signals:
+    void currentDrawableChanged(easy3d::Drawable*);
 
 protected:
     void mousePressEvent(QMouseEvent* e);
