+++ easy3d/viewer/viewer.cpp
@@ -1153,6 +1153,8 @@ namespace easy3d {
             int pre_idx = model_idx_;
             const std::string name = model->name();
             models_.erase(pos);
+            delete model->renderer();
+            delete model->manipulator();
             delete model;
             model_idx_ = static_cast<int>(models_.size()) - 1; // make the last one current
             LOG(INFO) << "model deleted: " << name;
