### 2021-01-18 c09ab14
mark selection (instead of deleting selection)
