### 2021-03-22 #2885 @RealPolitiX
Dear [ https://github.com/Kevin-Mattheus-Moerman | @Kevin-Mattheus-Moerman
],

Thanks for the invitation, I'm unfortunately not working explicitly in this
area, so I decline the invitation.

Cheers

R. Patrick Xian

On Mon, Mar 22, 2021 at 3:28 AM Monique Teillaud ***@***.***>
wrote:

> Dear [ https://github.com/Kevin-Mattheus-Moerman |
> @Kevin-Mattheus-Moerman ]
>
> Just a short email to decline your invitation, as the submission does not
> fall in my precise area of research.
>
> It could help to ask some other CGAL developers
> [ https://github.com/CGAL/ | https://www.cgal.org/people.html ] [
> https://github.com/CGAL/ ]
> [ https://github.com/CGAL/ | https://github.com/CGAL/ ]
> [ https://www.cgal.org/people.html ]
> Best regards,
> --
> Monique Teillaud
> https://members.loria.fr/Monique.Teillaud/
> INRIA Nancy - Grand Est, LORIA
> Institut National de Recherche en Informatique et Automatique
>
> ----- Le 21 Mar 21, à 11:55, Kevin Mattheus Moerman ***@***.***> a écrit :
>
> > [ https://github.com/gpeyre | @gpeyre ] , [ https://github.com/jizecn |
> @jizecn
> > ] , [ https://github.com/RealPolitiX | @RealPolitiX ] , [
> > https://github.com/jcarpent | @jcarpent ] , [ https://github.com/jni |
> @jni ] ,
> > [ https://github.com/SimonDanisch | @SimonDanisch ] , [
> > https://github.com/bryevdv | @bryevdv ] , [ https://github.com/jsignell
> |
> > @jsignell ] , [ https://github.com/afabri | @afabri ] , [
> > https://github.com/janetournois | @janetournois ] , [
> > https://github.com/MoniqueTeillaud | @MoniqueTeillaud ] , are you able
> to help
> > review this submission for [ https://joss.theoj.org/ | JOSS ] ? The [
> > https://joss.readthedocs.io/en/latest/submitting.html#the-review-process
> |
> > review process ] focuses on [
> > https://joss.readthedocs.io/en/latest/submitting.html#the-review-process
> | the
> > software ] and this [
> >
> https://github.com/openjournals/joss-papers/blob/joss.02885/joss.02885/10.21105.joss.02885.pdf
> > | short paper ] . If you are interested you can comment here. Thanks!
>
> > Note, given the global COVID19 pandemic we can be very flexible in terms
> of
> > review times.
>
> > Note, this is a so-called "Pre-review issue" where reviewers/editors are
> > assigned. Once I've attracted sufficient reviewers I'll trigger a "Review
> > issue" where the review will take place.
>
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, [
> >
> https://github.com/openjournals/joss-reviews/issues/2885#issuecomment-803554049
> > | view it on GitHub ] , or [
> >
> https://github.com/notifications/unsubscribe-auth/ACHKOTPDWGXQLPVZOFH7SBLTEXGBZANCNFSM4UODJIHA
> > | unsubscribe ] .
>
