### 2021-04-19 #2885 @Kevin-Mattheus-Moerman
@keenancrane @nmwsharp @ranahanocka @dcoeurjo @MarkGillespie @ctlee @rohan-sawhney are you able to help review this submission  for [JOSS](https://joss.theoj.org/)? The [review process](https://joss.readthedocs.io/en/latest/submitting.html#the-review-process) focuses on [the software](https://github.com/LiangliangNan/Easy3D) and this [short paper](https://github.com/openjournals/joss-papers/blob/joss.02885/joss.02885/10.21105.joss.02885.pdf). If you are interested you can comment here. Thanks!

Note, given the global COVID19 pandemic **we can be very flexible in terms of review times**.

Note, this is a so-called "Pre-review issue" where reviewers/editors are assigned. Once I've attracted sufficient reviewers I'll trigger a "Review issue" where the review will take place.
