### 2021-05-05 #3255 @whedon
[REVIEW]: Easy3D: a lightweight, easy-to-use, and efficient C++ library for processing and rendering 3D data
**Submitting author:** @LiangliangNan (<a href="http://orcid.org/0000-0002-5629-9975">Liangliang Nan</a>)
**Repository:** <a href="https://github.com/LiangliangNan/Easy3D" target ="_blank">https://github.com/LiangliangNan/Easy3D</a>
**Version:** v2.4.5

@adi3 & @vissarion & @ @bhajay, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:

- [x] #3472

- [x] #3473
- [x] #3480
- [x] #3481
- [x] #3479

- [x] #3474
- [x] #3475
- [x] #3476
- [x] #3477
- [x] #3478
