### 2021-05-28 LiangliangNan/Easy3D#79 @vissarion
Build dependencies missing from documentation
In the README it is stated that
> For the core functionalities and the basic viewer, all dependencies have already been included in the distribution.

however there are some dependencies not included in the distribution such as the CGAL library that the user should install manually and a few libraries/packages that are usually are available via OS package managers.

I manage to build on  `5.4.0-73-generic #82-Ubuntu SMP ` with gcc version `9.3.0` and cmake version `3.18.3` after install cgal 5.1 manually and
```
sudo apt-get install libxinerama-dev libxrandr-dev libxcursor-dev libxi-dev
```
There are probably other dependencies that my system has met.

I think it would be useful if all needed dependencies are listed in the documentation.
