### 2021-05-28 LiangliangNan/Easy3D#80 @vissarion
DOIs are missing from JOSS manuscript
DOIs are missing from JOSS manuscript

e.g. MeshLab: an Open-Source Mesh Processing Tool
http://diglib.eg.org/handle/10.2312/LocalChapterEvents.ItalChap.ItalianChapConf2008.129-136
