### 2021-05-28 LiangliangNan/Easy3D#81 @vissarion
Missing test-suite
A test-suite is either missing or not easy to find. I couldn't find instructions on how to build and run tests manually.

There is only a directory `tests` that you can build some tests and tutorials.
