### 2021-05-28 LiangliangNan/Easy3D#80 @vissarion
Also it seems that a reference to CGAL library which Easy3D depends on is missing.
