### 2021-05-28 LiangliangNan/Easy3D#82 @vissarion
Some community guidelines are missing 
I cannot find clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

The following guide could be helpful

https://github.com/LiangliangNan/Easy3D/community
