### 2021-05-28 LiangliangNan/Easy3D#82 @vissarion
This is part of this JOSS review openjournals/joss-reviews#3255
