### 2021-05-28 #3255 @vissarion
Hi @LiangliangNan first of all thanks for this very useful feature-rich package!

As part of my review I created a few issues in the repository
- https://github.com/LiangliangNan/Easy3D/issues/79
- https://github.com/LiangliangNan/Easy3D/issues/80
- https://github.com/LiangliangNan/Easy3D/issues/81
- https://github.com/LiangliangNan/Easy3D/issues/82

I also have two more comments:
- Performance claims: it seems that in the intro of your paper there are some performance claims about easy3D's efficiency by taking advantage of the powerful GPUs. Can this be clarified? Do you have for example benchmarks for this efficiency or a reference to some scientific publication or some evidence that convince the reader about that claim?
- [OPTIONAL] Automated tests/availability on various platforms: I have only tested the software on debian/ubuntu but there is a claim in the README that the software builds on mac and windows platforms. Unfortunately, I do not have availability on those platforms. My suggestion here is to create automated tests (e.g. using github actions or travis or some other available CI tool) that build and test easy3D on all the claimed platforms. Of course this is not an obligation for the acceptance to JOSS, just my recommendation for the package.
