### 2021-05-28 LiangliangNan/Easy3D#79 @LiangliangNan
The use of CGAL (and also Qt) is optional.
I have clarified this in the ReadMe file:

> Like most software, Easy3D depends on some third-party libraries. Easy3D has made this easier for the users by including the source code of most third-party libraries (for the core functionalities and the basic viewer), and it leaves very few optional (for a few additional features that are typically not needed by most users).
>
> The optional third-party libraries are:
>
> - CGAL (optional): Easy3D has implemented a few algorithms for advanced surface mesh processing, such as surface (re)orientation, detecting and resolving duplicate vertices/faces and self-intersection, and clipping/splitting/slicing surface meshes. To enable these features, CGAL must be installed and is visible to CMake. Without CGAL, these features are disabled by default and you can still use all other features.
>
> - Qt (optional): Easy3D supports Qt for UI creation, which can help develop sophisticated applications for 3D data processing and visualization. You can switch on the CMake option EASY3D_ENABLE_QT to include the examples and applications that depend on Qt (e.g., Tutorial_202_Viewer_Qt and Mapple)
