### 2021-05-28 #3255 @LiangliangNan
@vissarion Thanks for the very constructive comments and suggestions.

I've revised the ReadMe file for the "missing dependencies". All other issues will be addressed shortly in the individual issues.

> * Performance claims: it seems that in the intro of your paper there are some performance claims about easy3D's efficiency by taking advantage of the powerful GPUs. Can this be clarified? Do you have for example benchmarks for this efficiency or a reference to some scientific publication or some evidence that convince the reader about that claim?

This was observed and reported by many users: Easy3D can render larger models at a good framerate.
But indeed I haven't done any formal evaluation.
