### 2021-05-28 LiangliangNan/Easy3D#79 @vissarion
I see. However, even if it is optional, if the user have it installed then cmake will try to include it by default and then if the version is of your installed cgal is <5.1 the cmake will complain. So the question here is: if I have already installed CGAL (version<5.1) in my system then how could I install easy3D ? As it is now I should either uninstall CGAL or upgrade to >=5.1. I think it would be useful to have a cmake flag that disables the inclusion of CGAL defaulted to disabled since CGAL is optional.

The same issue could have been possible with `Qt` but I didn't experience it.

Also minimum supported versions of Qt and CGAL should be included in the README, as well as links to the repository where the user can download and build them.
