### 2021-05-28 LiangliangNan/Easy3D#82 @LiangliangNan
Thanks!

This has been added here (and is also referred to in the ReadMe):
https://github.com/LiangliangNan/Easy3D/blob/main/CONTRIBUTING.md
