### 2021-05-28 LiangliangNan/Easy3D#79 @LiangliangNan
Oh... Actually, an earlier version of Easy3D had that CMake option for CGAL :sob:

Anyway. I've added back the CMake option `EASY3D_ENABLE_CGAL`, allowing users to explicitly request CGAL.

The main changes in the CMakeLists file:
```
# Build advanced features that require CGAL (>= v5.1)
option(EASY3D_ENABLE_CGAL    "Build advanced features that require CGAL (>= v5.1)"    OFF)
```

and

```
if (EASY3D_ENABLE_CGAL)
      find_package(CGAL REQUIRED)
      if (CGAL_FOUND)
          if (CGAL_VERSION VERSION_GREATER_EQUAL "5.1")
              set(EASY3D_HAS_CGAL TRUE)
              add_definitions(-DHAS_CGAL)
              message(STATUS "Found CGAL v${CGAL_VERSION}")
          else ()
              set(EASY3D_HAS_CGAL FALSE)
              message(ERROR "CGAL v${CGAL_VERSION} is found but Easy3D requires at least version 5.1. Please install CGAL v5.1 or later.")
          endif ()
      else ()
          set(EASY3D_HAS_CGAL FALSE)
          message(ERROR "Some advanced features require CGAL (v5.1 or later) but CGAL was not found. Easy3D will be built without these features. "
                  "You can set 'CGAL_DIR' to the directory containing 'CGALConfig.cmake' or 'cgal-config.cmake' to help find CGAL.")
      endif ()
  endif ()
```

And in the README file, the minimum supported versions of Qt and CGAL are also made clear.
