### 2021-05-29 LiangliangNan/Easy3D#80 @LiangliangNan
Thanks for pointing out these.

DOIs have been added.
The reference to CGAL has also been added.
