### 2021-06-06 LiangliangNan/Easy3D#81 @LiangliangNan
A test suite has been added.

Tests of data structures, file-io, algorithms, and visualization techniques are automatic. Model/element picking functionalities require minimum user input (e.g., clicking on the model).
