### 2021-06-09 LiangliangNan/Easy3D#81 @vissarion
Thanks, please add some explanation for how the user can run those tests, e.g. in the README after the "build" section.
