### 2021-06-09 #3255 @vissarion
@Kevin-Mattheus-Moerman almost all of my comments/issues above are addressed by the author (thanks @LiangliangNan !). There is only one left https://github.com/LiangliangNan/Easy3D/issues/81 where I asked for some more clarifications.
