### 2021-06-09 LiangliangNan/Easy3D#81 @LiangliangNan
Thanks for the comments. I've added a 'Test' section after 'Build':
> ### Test
> A test-suite is provided in the `tests` subfolder, which contains a collection of automated test cases (for data
> structures, IO, algorithms, visualization, etc.) and some semi-automated test cases (for GUI-related functionalities
> that require interactive user input). You can build and run the `tests` executable file to quickly test Easy3D on your
> machine.

Besides, I've also added a 'Repository layout' before 'Build'. I think this will be helpful for users.
> ### Repository layout
> The repository contains a `CMakeLists.txt` file that serves as anchor for configuring and building programs,
> and a set of subfolders:
> * `3rd_party` - source code of third party libraries
> * `applications` - applications built on top of Easy3D
> * `cmake` - CMake-related configuration files
> * `docs` - documentation configuration file (Doxygen)
> * `easy3d` - source code of Easy3D
> * `resources` - test data, images, shaders, and textures, etc.
> * `tests` - a collection of test cases
