### 2021-06-09 LiangliangNan/Easy3D#81 @vissarion
I am getting the following error when I try to build tests

```
ThinkPad-T490:~/workspace/Easy3D/tests$ make
[  1%] Building CXX object CMakeFiles/tests.dir/graph.cpp.o
/home/workspace/Easy3D/tests/graph.cpp:25:10: fatal error: easy3d/core/graph.h: No such file or directory
   25 | #include <easy3d/core/graph.h>
      |          ^~~~~~~~~~~~~~~~~~~~~
compilation terminated.
make[2]: *** [CMakeFiles/tests.dir/build.make:82: CMakeFiles/tests.dir/graph.cpp.o] Error 1
make[1]: *** [CMakeFiles/Makefile2:95: CMakeFiles/tests.dir/all] Error 2
make: *** [Makefile:103: all] Error 2
```
gcc version 9.3.0 (Ubuntu 9.3.0-17ubuntu1~20.04)
