### 2021-06-09 LiangliangNan/Easy3D#81 @LiangliangNan
I guess you were using the CMakeList file at the `tests` subdirectory?

Could you please download the entire code and use the main `CMakeList file` in the root directory?  I will clarify this in ReadMe.
