### 2021-06-09 LiangliangNan/Easy3D#81 @vissarion
Building works with the root `CMakeList` file but there is no explanation of how to run the tests. Please add this also in `README`
