### 2021-06-09 LiangliangNan/Easy3D#81 @LiangliangNan
I've added the following:

> To build and run the test suite, download the entire source tree, use `CMakeLists.txt` in the root directory of the repository, switch on the CMake option `EASY3D_BUILD_TESTS`, and run CMake. After CMake, you can build all or only the`tests` executable. Finally, you can run the `tests` executable for the test.
