### 2021-06-09 LiangliangNan/Easy3D#81 @vissarion
Cloning from scratch I tried

```
mkdir Release
cd Release
cmake -DCMAKE_BUILD_TYPE=Release -DEASY3D_BUILD_TESTS=On ..
make
```
to get
```
[ 87%] Building CXX object tests/CMakeFiles/tests.dir/linear_solvers.cpp.o
In file included from /home/vissarion/workspace/Easy3D/easy3d/core/types.h:31,
                 from /home/vissarion/workspace/Easy3D/tests/linear_solvers.cpp:26:
/home/vissarion/workspace/Easy3D/easy3d/core/vec.h: In instantiation of ‘easy3d::Vec<N, T>::thisclass& easy3d::Vec<N, T>::operator=(const thisclass&) [with long unsigned int N = 6; T = double; easy3d::Vec<N, T>::thisclass = easy3d::Vec<6, double>]’:
/home/vissarion/workspace/Easy3D/easy3d/core/mat.h:1274:16:   required from ‘void easy3d::lu_back_substitution(const easy3d::Mat<N, N, T>&, const easy3d::Vec<N, T>&, const easy3d::Vec<N, T>&, easy3d::Vec<N, T>*) [with long unsigned int N = 6; T = double]’
/home/vissarion/workspace/Easy3D/tests/linear_solvers.cpp:72:42:   required from here
/home/vissarion/workspace/Easy3D/easy3d/core/vec.h:85:19: error: ‘memcpy’ was not declared in this scope, and no declarations were found by argument-dependent lookup at the point of instantiation [-fpermissive]
   85 |             memcpy(data_, rhs.data(), N*sizeof(T));
      |             ~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
In file included from /usr/include/features.h:461,
                 from /usr/include/x86_64-linux-gnu/c++/9/bits/os_defines.h:39,
                 from /usr/include/x86_64-linux-gnu/c++/9/bits/c++config.h:524,
                 from /usr/include/c++/9/cstdint:38,
                 from /home/vissarion/workspace/Easy3D/easy3d/core/types.h:28,
                 from /home/vissarion/workspace/Easy3D/tests/linear_solvers.cpp:26:
/usr/include/x86_64-linux-gnu/bits/string_fortified.h:31:1: note: ‘void* memcpy(void*, const void*, size_t)’ declared here, later in the translation unit
   31 | __NTH (memcpy (void *__restrict __dest, const void *__restrict __src,
      | ^~~~~
make[2]: *** [tests/CMakeFiles/tests.dir/build.make:95: tests/CMakeFiles/tests.dir/linear_solvers.cpp.o] Error 1
make[1]: *** [CMakeFiles/Makefile2:3868: tests/CMakeFiles/tests.dir/all] Error 2
make: *** [Makefile:149: all] Error 2
```
