### 2021-06-09 LiangliangNan/Easy3D#81 @LiangliangNan
Solved by adding a missing `#include <cstring>`. On Mac it was not required: the C++ standard library is messy on Mac.
It should work now.
