### 2021-06-10 LiangliangNan/Easy3D#81 @vissarion
Ok thanks, I created a PR to add CI builds to make sure that your package builds after every commit.
It is minimal in the sense that only supports `gcc` on `ubuntu` and only builds without testing since tests involve user interaction which is difficult to automate. This could be extended to build on different platforms and compilers as well as include a minimal test suite with not interactions. But this could be done in the future or left as a `good-first-issue` for new contributors.
