### 2021-07-11 #3255 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman : There is no paper.md anymore in the repository

Is on the 'joss' branch. I've also just updated it. Does that help?
