### 2021-07-11 #3255 @danielskatz
@bhajay - please do not use the automated "open an issue" option that GitHub provides. The issues are opened in the wrong repo and have no content.  Instead, please open issue manually in https://github.com/LiangliangNan/Easy3D as needed.  I'm going to close the issues that were incorrectly opened in the wrong repo.  (We would like to disable this GitHub functionality, but cannot)
