### 2021-08-16 #3255 @Kevin-Mattheus-Moerman
@LiangliangNan congratulations, given the positive reviews above we will now process this work for acceptance in JOSS.

- [ ] Since this work is about to be accepted we recommend you proofread the paper yourself one more time and also that you verify that the author names and affiliations are accurate.
- [ ] The paper currently lacks an `Acknowledgements` section, this section is not required but if the work was funded please do include sources of funding in an `Acknowledgements` section.
- [ ] Please archive a copy of the software on ZENODO and report back here with the archive DOI. You may process the upload manually or use these automated steps which many find useful: https://guides.github.com/activities/citable-code/.
- [ ] Please amend (if needed) the ZENODO archive (you can edit information manually) such that meta-data like the title and the author list match that of the JOSS paper. The license also needs to match your software license. FYI you can also ensure, where possible, at this point that ORCID ids are also added to the ZENODO authors.
- [ ] If during the review, or at the point of archiving, the version tag has changed (currently `v2.4.1`) please let us know.
