### 2021-08-19 #3255 @Kevin-Mattheus-Moerman
@LiangliangNan thanks for working on those steps. Once point remains, the ZENODO archive title still does not match that of the paper. Please change it to match: `Easy3D: a lightweight, easy-to-use, and efficient C++ library for processing and rendering 3D data`
