### 2021-08-23 #3255 @LiangliangNan
> 🚨🚨🚨 **THIS IS NOT A DRILL, YOU HAVE JUST ACCEPTED A PAPER INTO JOSS!** 🚨🚨🚨
>
> Here's what you must now do:
>
> 1. Check final PDF and Crossref metadata that was deposited 👉 [Creating pull request for 10.21105.joss.03255 joss-papers#2522](https://github.com/openjournals/joss-papers/pull/2522)
> 2. Wait a couple of minutes, then verify that the paper DOI resolves https://doi.org/10.21105/joss.03255
> 3. If everything looks good, then close this review issue.
> 4. Party like you just published a paper! 🎉🌈🦄💃👻🤘
>
> Any issues? Notify your editorial technical team...

Yes. Everything looks good.
Thank you again for all the help and efforts!
